% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.package_betapart.R
\name{beta.pair}
\alias{beta.pair}
\alias{betapart.core}
\title{From betapart package 1.5.4 : beta.pair and betapart.core functions}
\usage{
betapart.core(x)
}
\description{
From betapart package 1.5.4 : beta.pair and betapart.core functions
}
\seealso{
\code{\link[betapart]{beta.pair}}
}
\keyword{betapart}
