% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.package_fpc.R
\name{cluster.stats}
\alias{cluster.stats}
\title{From fpc package 2.2-9 : cluster.stats function}
\usage{
cluster.stats(
  d = NULL,
  clustering,
  alt.clustering = NULL,
  noisecluster = FALSE,
  silhouette = TRUE,
  G2 = FALSE,
  G3 = FALSE,
  wgap = TRUE,
  sepindex = TRUE,
  sepprob = 0.1,
  sepwithnoise = TRUE,
  compareonly = FALSE,
  aggregateonly = FALSE
)
}
\description{
From fpc package 2.2-9 : cluster.stats function
}
\seealso{
\code{\link[fpc]{cluster.stats}}
}
\keyword{fpc}
