% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.abgFunctions.R
\name{divLeinster}
\alias{divLeinster}
\title{Find cutoff to transform abundance values into binary values}
\usage{
divLeinster(spxp, Z = NULL, q = 2, check = TRUE)
}
\arguments{
\item{spxp}{a site (row) by species (cols) \code{matrix} with or without 
rownames and colnames}

\item{Z}{default \code{NULL}. \cr A species by species similarity 
\code{matrix}}

\item{q}{default \code{2}. \cr An \code{integer} corresponding to the 
importance attributed to relative abundances}

\item{check}{(\emph{optional}) default \code{TRUE}. \cr If \code{TRUE}, the 
given arguments will be checked}
}
\description{
This function calculates the diversity of each site of a site 
by species matrix according to the q parameter according to 
\href{http://www.jstor.org/stable/23143936}{Leinster & Cobbold 2012 Ecology}.
}
