% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.do_habitat_validation.R
\name{do.habitat.validation}
\alias{do.habitat.validation}
\title{Compare observed and simulated habitat of a \code{FATE} simulation
at the last simulation year.}
\usage{
do.habitat.validation(
  output.path,
  RF.model,
  habitat.FATE.map,
  validation.mask,
  simulation.map,
  predict.all.map,
  sim.version,
  name.simulation,
  perStrata,
  hab.obs,
  year,
  list.strata.releves,
  list.strata.simulations,
  opt.no_CPU = 1,
  studied.habitat = NULL
)
}
\arguments{
\item{output.path}{access path to the for the folder where output files
will be created.}

\item{RF.model}{random forest model trained on CBNA data (train.RF.habitat
function)}

\item{habitat.FATE.map}{a raster map of the observed habitat in the
studied area with same projection & resolution than validation mask and simulation mask.}

\item{validation.mask}{a raster mask that specified 
which pixels need validation, with same projection & resolution than simulation mask.}

\item{simulation.map}{a raster map of the whole studied area (provides by FATE parameters functions).}

\item{predict.all.map}{\code{Logical}. If TRUE, the script will predict 
habitat for the whole map.}

\item{sim.version}{name of the simulation to validate.}

\item{name.simulation}{simulation folder name.}

\item{perStrata}{\code{Logical}. If TRUE, the PFG abundance is defined
by strata in each pixel. If FALSE, PFG abundance is defined for all strata.}

\item{hab.obs}{a raster map of the extended studied map in the simulation, with same projection 
& resolution than simulation mask.}

\item{year}{simulation year selected for validation.}

\item{list.strata.releves}{a character vector which contain the observed strata 
definition, extracted from observed PFG releves.}

\item{list.strata.simulations}{a character vector which contain \code{FATE} 
strata definition and correspondence with observed strata definition.}

\item{opt.no_CPU}{default \code{1}. \cr The number of 
resources that can be used to parallelize the computation of performance of
habitat prediction.}

\item{studied.habitat}{default \code{NULL}. If \code{NULL}, the function will
take into account of habitats define in the \code{hab.obs} map. Otherwise, please specify 
in a 2 columns data frame the habitats (2nd column) and the ID (1st column) for each of them which will be taken 
into account for the validation.}
}
\value{
Habitat performance file. \cr
If option selected, the function also returns an habitat prediction file with 
observed and simulated habitat for each pixel of the whole map.
}
\description{
To compare observations and simulations, this function compute 
confusion matrix between observation and prediction and then compute the TSS 
for each habitat.
}
\details{
After several preliminary checks, the function is going to prepare the observations
database by extracting the observed habitat from a raster map. Then, for the
simulation \code{sim.version}, the script take the evolution abundance for each PFG
and all strata (or for each PFG & each strata if option selected) file and predict 
the habitat for the whole map (if option selected) thanks to the RF model. 
Finally, the function computes habitat performance based on TSS for each habitat.
}
\author{
Matthieu Combaud & Maxime Delprat
}
