% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.do_PFG_composition_validation.R
\name{do.PFG.composition.validation}
\alias{do.PFG.composition.validation}
\title{Compute distance between observed and simulated distribution}
\usage{
do.PFG.composition.validation(
  name.simulation,
  sim.version,
  hab.obs,
  PFG.considered_PFG.compo,
  strata.considered_PFG.compo,
  habitat.considered_PFG.compo,
  observed.distribution,
  perStrata,
  validation.mask,
  year,
  list.strata.simulations,
  list.strata.releves,
  habitat.FATE.map,
  studied.habitat
)
}
\arguments{
\item{name.simulation}{simulation folder name.}

\item{sim.version}{name of the simulation we want to validate (it works with
only one \code{sim.version}).}

\item{hab.obs}{a raster map of the extended studied map in the simulation, with same projection 
& resolution than simulation mask.}

\item{PFG.considered_PFG.compo}{a character vector of the list of PFG considered
in the validation.}

\item{strata.considered_PFG.compo}{a character vector of the list of precise 
strata considered in the validation.}

\item{habitat.considered_PFG.compo}{a character vector of the list of habitat(s)
considered in the validation.}

\item{observed.distribution}{PFG observed distribution table provides by \code{get.observed.distribution} function.}

\item{validation.mask}{a raster mask that specified 
which pixels need validation, with same projection & resolution than simulation mask.}

\item{year}{year of simulation to validate.}

\item{list.strata.simulations}{a character vector which contain \code{FATE} 
strata definition and correspondence with observed strata definition.}

\item{list.strata.releves}{a character vector which contain the observed strata 
definition, extracted from observed PFG releves.}

\item{habitat.FATE.map}{a raster map of the observed habitat in the
studied area with same projection & resolution than validation mask and simulation mask.}

\item{studied.habitat}{default \code{NULL}. If \code{NULL}, the function will
take into account of habitats define in the \code{hab.obs} map. Otherwise, please specify 
in a 2 columns data frame the habitats (2nd column) and the ID (1st column) for each of them which will be taken 
into account for the validation.}

\item{perStrata.compo}{\code{Logical}. All strata together (FALSE) or per strata (TRUE).}
}
\value{

}
\description{
This script is designed to compare the difference between the
PFG distribution in observed and simulated data. For a set of PFG, strata and
habitats chosen, the function compute distance between observed and simulated
distribution for a precise \code{FATE} simulation.
}
\details{
After preliminary checks, this code extract observed habitat from the \code{hab.obs}
map and, then, merge it with the simulated PFG abundance file (with or without strata definition) 
from results of the \code{FATE} simulation selected with \code{sim.version}. 
After filtration of the required PFG, strata and habitats, the function transforms 
the data into relative metrics and, then, compute distribution per PFG, strata
and habitat (if necessary). Finally, the code computes proximity between observed 
and simulated data, per PFG, strata and habitat.
}
\author{
Matthieu Combaud, Maxime Delprat
}
