% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.do_PFG_composition_validation.R
\name{do_PFG_composition_validation}
\alias{do_PFG_composition_validation}
\title{Compute distance between observed and simulated distribution.}
\usage{
do_PFG_composition_validation(
  sim,
  PFG.considered_PFG.compo,
  strata.considered_PFG.compo,
  habitat.considered_PFG.compo,
  observed.distribution,
  simu_PFG,
  habitat.whole.area.df
)
}
\arguments{
\item{sim}{Name of the single simulation to validate.}

\item{PFG.considered_PFG.compo}{A character vector of the list of PFG considered
in the validation.}

\item{strata.considered_PFG.compo}{A character vector of the list of precise 
strata considered in the validation.}

\item{habitat.considered_PFG.compo}{A character vector of the list of habitat(s)
considered in the validation.}

\item{observed.distribution}{A \code{data.frame} provides by \code{get_observed_distribution} with 5 columns :
\cr \code{PFG}, \code{habitat}, \code{strata}, \code{rank}, \code{observed.quantile}
\cr (see \code{\link{get_observed_distribution}})}

\item{simu_PFG}{A \code{data.frame} provides by \code{POST_FATE.temporalEvolution} with simulated abundance for each PFG and strata 
(if option selected) and pixel ID (see \code{\link{POST_FATE.temporalEvolution}}).}

\item{habitat.whole.area.df}{A \code{data.frame} provides by \code{POST_FATE.validation} with 3 columns : 
\cr \code{pixel} which contains the ID of each pixel in the study area.
\cr \code{code.habitat} which contains the ID of the habitat in each pixel.
\cr \code{for.validation} for each pixel, \code{0} if does not need validation, \code{1} if needs validation.}
}
\value{

}
\description{
This script is designed to compare the difference between the
PFG distribution in observed and simulated data. For a set of PFG, strata and
habitats chosen, the function computes distance between observed and simulated
distribution for a precise \code{FATE} simulation.
}
\details{
Firstly, this code merges \code{habitat.whole.area.df} data frame with the simulated PFG abundance 
\code{simu_PFG} data frame (with or without strata definition). 
After filtration of the required PFG, strata and habitat(s), the function transforms 
the data into relative metrics and, then, compute distribution per PFG, strata
and habitat (if necessary). Finally, the code computes proximity between observed 
and simulated data, per PFG, strata and habitat.
}
\author{
Matthieu Combaud, Maxime Delprat
}
