% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.createParams.R
\name{.createParams}
\alias{.createParams}
\title{Create a parameters file}
\usage{
.createParams(params.file, params.list, separator = " ")
}
\arguments{
\item{params.file}{a \code{string} corresponding to the name of the file 
that will be created}

\item{params.list}{a \code{list} containing all the parameters that will be 
included into \code{params.file}, and whose names correspond to the 
parameter names}

\item{separator}{a \code{string} to separate each parameter values within 
the parameter file}
}
\description{
Create a text file containing value(s) for one or several 
parameters and separated by a same character value.
}
\author{
Maya Guéguen
}
