% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.getCutoff.R
\name{.getCutoff}
\alias{.getCutoff}
\title{Find cutoff to transform abundance values into binary values}
\usage{
.getCutoff(Obs, Fit)
}
\arguments{
\item{Obs}{a \code{vector} containing binary observed values (\code{0} or 
\code{1})}

\item{Fit}{a \code{vector} containing relative abundance values (between 
\code{0} and \code{1})}
}
\description{
This function finds the best cutoff to transform abundance 
values into binary values while optimising sensitivity and specificity 
values based on observations
}
\examples{

vec.obs = c(rep(0, 60), rep(1, 40))
vec.pred = c(sample(x = seq(0, 0.2, 0.01), size = 50, replace = TRUE)
             , sample(x = seq(0.15, 1, 0.01), size = 50, replace = TRUE))
## plot(vec.obs, vec.pred)

cutoff = .getCutoff(Obs = vec.obs, Fit = vec.pred)
str(cutoff)

}
