% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.getELLIPSE.R
\name{.getELLIPSE}
\alias{.getELLIPSE}
\alias{util.ellipse}
\title{Obtain ellipse coordinates from (PCO) X,Y and a factor value}
\usage{
util.ellipse(mx, my, vx, cxy, vy, coeff)
}
\arguments{
\item{xy}{a \code{data.frame} or \code{matrix} with 2 columns corresponding 
to individuals coordinates, extracted from example from 
\code{\link[ade4]{dudi.pco}} analysis}

\item{fac}{a \code{vector} containing group labels for individuals (with 
\code{length(fac) = nrow(xy)})}
}
\description{
Obtain ellipse coordinates from (PCO) X,Y and a factor value
}
