% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.getOS.R
\name{.getOS}
\alias{.getOS}
\title{Find operating system of your computer}
\usage{
.getOS()
}
\description{
This functions finds on which operating sytem the user is 
currently using \code{RFate}. It notably allows to set the use of 
parallelisation.
}
