% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.getParam.R
\name{.getParam}
\alias{.getParam}
\title{Extract parameter value(s) from a parameter file}
\usage{
.getParam(params.lines, flag, flag.split, is.num = TRUE)
}
\arguments{
\item{params.lines}{a \code{string} corresponding to the name of the file 
from which to extract the parameter value}

\item{flag}{a \code{string} corresponding to the parameter name to be 
extracted and that must be present into the \code{param.lines} file}

\item{flag.split}{a \code{string} to choose the concerned type of parameter 
(either "\code{ }" or "\code{^--.*--$}"), depending on the type of parameter 
file (containing values or filenames)}

\item{is.num}{default \code{TRUE}. \cr If \code{TRUE}, the extracted 
parameter is considered to be \code{numeric} and will be processed as such}
}
\value{
A \code{vector} containing one or more values of type \code{string} 
(if \code{is.num = FALSE}) or \code{numeric} (if \code{is.num = TRUE}).
}
\description{
This function extracts from a text file the value(s) of a 
given parameter.
}
\examples{

## Create a skeleton folder with the default name ('FATE_simulation')
if (dir.exists("FATE_simulation")) unlink("FATE_simulation", recursive = TRUE)
PRE_FATE.skeletonDirectory()

## Create a Global_parameters file
PRE_FATE.params_globalParameters(name.simulation = "FATE_simulation"
                                 , required.no_PFG = 6
                                 , required.no_strata = 5
                                 , required.simul_duration = 100
                                 , required.seeding_duration = c(10,50)
                                 , required.seeding_timestep = 1
                                 , required.seeding_input = 100
                                 , required.max_abund_low = 30000
                                 , required.max_abund_medium = 50000
                                 , required.max_abund_high = 90000)
                                 
## Extract number of PFG
.getParam(params.lines = "FATE_simulation/DATA/GLOBAL_PARAMETERS/Global_parameters_V1.txt"
          , flag = "NO_PFG"
          , flag.split = " "
          , is.num = TRUE)


## ----------------------------------------------------------------------------------------- ##
                                
## Load example data
          

}
\author{
Maya Guéguen
}
