% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.setPattern.R
\name{.setPattern}
\alias{.setPattern}
\title{Replace a pattern with a new within all parameter files of a 
\code{FATE-HD} simulation folder}
\usage{
.setPattern(name.simulation, pattern.tofind, pattern.toreplace)
}
\arguments{
\item{name.simulation}{a \code{string} that corresponds to the main
directory or simulation name of the \code{FATE-HD} simulation}

\item{pattern.tofind}{a \code{string} that corresponds to the pattern to 
find}

\item{pattern.toreplace}{a \code{string} that corresponds to the pattern to 
replace}
}
\description{
This function scans all the files within a \code{FATE-HD} 
simulation folder to find a specific pattern and replace it with a new one
}
\author{
Maya Guéguen
}
