% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.zipUnzip.R
\name{.unzip_ALL}
\alias{.unzip_ALL}
\alias{.unzip}
\alias{.zip_ALL}
\alias{.zip}
\title{Compress (\code{.tif}, \code{.img}) or decompress (\code{.gz}) files 
contained in results folder}
\usage{
.unzip_ALL(folder_name, no_cores)
.unzip(folder_name, list_files, no_cores)
.zip_ALL(folder_name, no_cores)
.zip(folder_name, list_files, no_cores)
}
\arguments{
\item{folder_name}{a \code{string} corresponding to the directory to be 
scanned}

\item{list_files}{a \code{vector} containing filenames to be compress or 
decompress, in order not to scan the whole given directory}

\item{no_cores}{default \code{1}. \cr an \code{integer} corresponding to the 
number of computing resources that can be used to parallelize the 
(de)compression}
}
\description{
These functions compress (\code{.tif}, \code{.img}) or 
decompress (\code{.gz}) files contained in a given folder.
}
