% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.package_clValid.R
\name{dunn}
\alias{dunn}
\title{From clValid package O.7 : dunn function}
\usage{
dunn(distance = NULL, clusters)
}
\description{
From clValid package O.7 : dunn function
}
\seealso{
\code{\link[clValid]{dunn}}
}
\keyword{clValid}
