% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.package_ecospat.R
\name{ecospat.niche.overlap}
\alias{ecospat.niche.overlap}
\alias{ascgen}
\alias{ecospat.grid.clim.dyn}
\alias{ecospat.kd}
\title{From ecospat package 3.2 : ecospat.kd, ecospat.grid.clim.dyn and 
ecospat.niche.overlap functions (and sp 1.4-5, adehabitatMA 0.3.14, 
adehabitatHR 0.4.19 packages}
\usage{
ascgen(xy, cellsize = NULL, nrcol = NULL, count = TRUE)
}
\description{
From ecospat package 3.2 : ecospat.kd, ecospat.grid.clim.dyn and 
ecospat.niche.overlap functions (and sp 1.4-5, adehabitatMA 0.3.14, 
adehabitatHR 0.4.19 packages
}
\seealso{
\code{\link[ecospat]{ecospat.grid.clim.dyn}},
\code{\link[ecospat]{ecospat.niche.overlap}}
}
\keyword{ecospat,}
\keyword{niche}
\keyword{overlap}
