% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FATE}
\alias{FATE}
\title{FATE Wrapper}
\usage{
FATE(simulParam, no_CPU = 1L, verboseLevel = 0L)
}
\arguments{
\item{simulParam}{a \code{string} corresponding to the name of a 
parameter file that will be contained into the \code{PARAM_SIMUL} folder 
of the \code{FATE} simulation}

\item{no_CPU}{(\emph{optional}) default \code{1}. \cr The number of 
resources that can be used to parallelize the simulation}

\item{verboseLevel}{(\emph{optional}) default \code{0}. \cr The logger 
verbose level : a \code{FATE} simulation can render different levels of 
information (from \code{0} to \code{4}, see 
\href{FATE#details}{\code{Details}}).}
}
\value{
None
}
\description{
This function runs a \code{FATE} dynamical landscape vegetation simulation.
}
\details{
This function allows to run a vegetation simulation with the 
\code{FATE} model, based on a simulation folder and a species simulation 
parameter file.

A \code{FATE} simulation can be parallelized, using the \code{no_CPU} 
parameter, given that the user machine is multi-core !

Quantity of informations are rendered by the software into the \code{R} 
console, and the \code{verboseLevel} parameter allows to filter which 
level of information is printed :

\describe{
  \item{0. }{shows any message}
  \item{1. }{shows any message, except debug}
  \item{2. }{shows only warning and error messages}
  \item{3. }{shows only error messages}
  \item{4. }{mute}
}
}
\examples{
\dontrun{FATE()}

}
\author{
Damien Georges

Isabelle Boulangeat, \email{isabelle.boulangeat@inrae.fr}

Maya Guéguen, \email{maya.gueguen@univ-grenoble-alpes.fr}
}
