% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FATE.run.R
\name{FATE.run}
\alias{FATE.run}
\title{Run a \code{FATE} simulation}
\description{
This function allows to run a \code{FATE} simulation from a 
folder of parameter files.
}
\author{
Maya Guéguen
}
\keyword{FATE}
\keyword{internal}
