% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FATE_Bauges}
\alias{FATE_Bauges}
\title{Bauges simulation folder}
\format{A \code{list} object with 2 elements :

\itemize{
  \item 
  \item 
}}
\usage{
FATE_Bauges
}
\description{
Bauges simulation folder
}
\author{
Maya Guéguen
}
\keyword{datasets}
