% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FATE_RS.R
\name{FATE_RS}
\alias{FATE_RS}
\title{S}
\usage{
FATE_RS(
  name.simulation,
  file.simulParam,
  opt.no_CPU = 1,
  verbose.level = 2,
  PPM.names_PFG,
  PPM.ras_env,
  PPM.xy,
  PPM.mod,
  PATCH.threshold = 80,
  PATCH.buffer = 500,
  PATCH.min_m2 = 3e+05,
  COST.ras_elev,
  COST.ras_barr,
  COST.wei = c(0.43, 0.53, 0.04),
  COST.range = c(1, 10),
  RS.name_simul,
  RS.dens_dep,
  RS.sim,
  RS.demo,
  RS.disp,
  RS.init,
  POP.carrying_capacity = 0.115,
  year.start,
  year.end,
  year.step
)
}
\description{
...
}
\author{
Maya Guéguen
}
\keyword{RangeShifter}
