% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.get_observed_distribution.R
\name{get.observed.distribution}
\alias{get.observed.distribution}
\title{Compute distribution of relative abundance over observed relevés}
\usage{
get.observed.distribution(
  name.simulation,
  obs.path,
  releves.PFG,
  releves.sites,
  hab.obs,
  PFG.considered_PFG.compo,
  strata.considered_PFG.compo,
  habitat.considered_PFG.compo,
  perStrata.compo,
  sim.version
)
}
\arguments{
\item{name.simulation}{simulation folder name.}

\item{obs.path}{the function needs observed data, please create a folder for them in your 
simulation folder and then indicate in this parameter the access path to this new folder.}

\item{releves.PFG}{file which contain the observed Braund-Blanquet abundance at each site
and each PFG and strata.}

\item{releves.sites}{file which contain coordinates and a description of
the habitat associated with the dominant species of each site in the studied map.}

\item{hab.obs}{raster map of the extended studied area in the simulation.}

\item{PFG.considered_PFG.compo}{a character vector of the list of PFG considered
in the validation.}

\item{strata.considered_PFG.compo}{a character vector of the list of precise 
strata considered in the validation.}

\item{habitat.considered_PFG.compo}{a character vector of the list of habitat(s)
considered in the validation.}

\item{perStrata.compo}{Logical. All strata together (FALSE) or per strata (TRUE).}

\item{sim.version}{name of the simulation we want to validate (it works with
only one \code{sim.version}).}
}
\value{

}
\description{
This script is designed to compute distribution, per PFG/strata/habitat,
of relative abundance, from observed data.
}
\details{
The function takes the \code{releves.PFG} and \code{releves.sites} files and 
aggregate coverage per PFG. Then, the code get habitat information from also
the \code{hab.obs} map, keep only interesting habitat, strata and PFG, and
transform the data into relative metrics. Finally, the script computes distribution
per PFG, and if require per strata/habitat (else all strata/habitat will be considered together).
}
\author{
Matthieu Combaud, Maxime Delprat
}
