% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.get_observed_distribution.R
\name{get_observed_distribution}
\alias{get_observed_distribution}
\title{Compute distribution of relative abundance over observed releves.}
\usage{
get_observed_distribution(
  releves.PFG,
  hab.obs,
  studied.habitat = NULL,
  PFG.considered_PFG.compo,
  strata.considered_PFG.compo,
  habitat.considered_PFG.compo,
  perStrata,
  output.path
)
}
\arguments{
\item{releves.PFG}{A \code{data.frame} with at least 5 columns : \cr
\code{site}, \code{x}, \code{y}, which contain respectively ID, x coordinate & y coordinate of each site of the study area. \cr
\code{abund} & \code{PFG} which contain respectively abundance (can be absolute abundance, Braun-Blanquet abundance or presence-absence) 
& name of PFG.
\cr (\emph{and optionally, \code{strata}}) which contains the number of strata at each the abundance is noted. 
(habitat & PFG composition validation).}

\item{hab.obs}{A raster map of the extended studied map in the simulation, with same projection 
& resolution than simulation mask.}

\item{studied.habitat}{A \code{data.frame} with 2 columns : 
\cr \code{ID} which contains the habitat ID, & \code{habitat} which contains the habitat names which will be taken into account 
for the validation (habitat & PFG composition validation).}

\item{PFG.considered_PFG.compo}{A character vector which contains the list of PFG considered
in the validation.}

\item{strata.considered_PFG.compo}{A character vector which contains the list of precise 
strata considered in the validation.}

\item{habitat.considered_PFG.compo}{A character vector which contains the list of habitat(s)
considered in the validation.}

\item{perStrata}{\code{Logical}. 
\cr All strata together (FALSE) or per strata (TRUE).}

\item{name.simulation}{Simulation folder name.}
}
\value{

}
\description{
This script is designed to compute distribution, per PFG/strata/habitat,
of relative abundance, from observed data.
}
\details{
The function takes the \code{releves.PFG} file and aggregate coverage per PFG. 
Then, the code gets habitat information from the \code{hab.obs} map & the \code{studied.habitat}
data frame, keep only interesting habitat(s), strata and PFG, and transforms 
the data into relative metrics. Finally, the script computes distribution per PFG 
and if required per strata/habitat (else all strata/habitat will be considered together).
}
\author{
Matthieu Combaud, Maxime Delprat
}
