% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MontBlanc}
\alias{MontBlanc}
\title{Mont Blanc dataset}
\format{A \code{list} object with 3 elements :

\itemize{
  \item 
  \item 
  \item 
}}
\usage{
MontBlanc
}
\description{
MontBlanc dataset
}
\author{
Maya Guéguen
}
\keyword{datasets}
