% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.plot_predicted_habitat.R
\name{plot_predicted_habitat}
\alias{plot_predicted_habitat}
\title{Create a raster map of habitat prediction for a set of \code{FATE} 
simulations at a specific simulation year.}
\usage{
plot_predicted_habitat(
  predicted.habitat,
  col.df,
  simulation.map,
  output.path,
  sim.version
)
}
\arguments{
\item{predicted.habitat}{a csv file created by the \code{\link{do.habitat.validation}} function
which contains, for each pixel of the studied map, the simulated and observed habitat.}

\item{col.df}{a data frame with all the colors associated with the failure or 
success of each studied habitat prediction.}

\item{simulation.map}{a raster map of the whole studied area.}

\item{output.path}{access path to the for the folder where output files
will be created.}

\item{sim.version}{a character vector with the name(s) of the simulation(s) to validate.}
}
\value{
a synthetic.prediction.png file which contain the final prediction map.
}
\description{
This script is designed to create a raster map of habitat prediction
based on an habitat prediction file. For each pixel, the habitat failure or success value
is associated to a color and then, the map is built.
}
\details{
The function determines true/false prediction ('failure' if false, 'success' if true)
and prepare a data frame containing colors and habitats codes. Then, the script merge
the prediction data frame with the color and code habitat data frame. Finally,
the function draw a raster map and a plot of prediction habitat over it, thanks
to the data prepared before.
}
\author{
Matthieu Combaud, Maxime Delprat
}
