% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PNE_PARAM}
\alias{PNE_PARAM}
\title{Ecrins National Park (PNE) data : building of parameter files to run 
a \code{FATE-HD} simulation}
\format{A \code{list} object with 13 elements to help 
  building the simulation files and folders to run a \code{FATE-HD} 
  simulation :
  
  \describe{
    \item{masks}{a \code{\link[raster]{stack}} object of dimension 
    \code{782 x 619} with a resolution of \code{100m} and Lambers (lcc) 
    projection, containing 7 mask layers with binary values (0 or 1) to be 
    used in a \code{FATE-HD} simulation :
    \itemize{
      \item \strong{maskEcrins} : simulation map, where occurs succession
      \item \strong{noDisturb} : perturbation map, when there is none
      \item \strong{mowing} : perturbation map, where occurs mowing
      \item \strong{grazing1} : perturbation map, where occurs light grazing
      \item \strong{grazing2} : perturbation map, where occurs extensive 
      grazing
      \item \strong{grazing3} : perturbation map, where occurs intensive 
      grazing
      \item \strong{grazingAll} :  perturbation map, where occurs grazing, 
      all types combined
    }
    }
    \item{HS_0}{a \code{\link[raster]{stack}} object of dimension 
    \code{782 x 619} with a resolution of \code{100m} and Lambers (lcc) 
    projection, containing 24 layers with probability values (between 0 and 
    1) representing Habitat Suitability for initialization phase and 
    \emph{equilibrium} or \emph{current} time (0) for each PFG 
    and to be used in a \code{FATE-HD} simulation. \cr
    These maps are coming from Species Distribution Models and methods to 
    obtain them are described in Supplementary Materials of 
    \emph{\href{https://mayagueguen.github.io/FATE-WEBSITE/papers/Boulangeat_2014_GCB.pdf}{Boulangeat et al. 2014 GCB}}.}
    \item{HS_15}{same as HS_0 but 15 years after equilibrium}
    \item{HS_30}{same as HS_0 but 30 years after equilibrium}
    \item{HS_45}{same as HS_0 but 45 years after equilibrium}
    \item{HS_60}{same as HS_0 but 60 years after equilibrium}
    \item{HS_75}{same as HS_0 but 75 years after equilibrium}
    \item{HS_90}{same as HS_0 but 90 years after equilibrium}
    \item{strata_limits}{a \code{vector} of length \code{5} \cr
    containing height of stratum limits in centimeters \cr
    to be used with the \link{PRE_FATE.params_PFGsuccession} and 
    \link{PRE_FATE.params_PFGlight} functions \cr}
    \item{succ_light}{a \code{data.frame} of dimension \code{24 x 6} \cr
    containing traits for plant functional groups \cr
    obtained from the \link{PRE_FATE.speciesClustering_step3} function \cr
    to be used with the \link{PRE_FATE.params_PFGsuccession} and 
    \link{PRE_FATE.params_PFGlight} functions \cr}
    \item{disp}{a \code{data.frame} of dimension \code{24 x 4} \cr
    containing dispersal values (in meters) for plant functional groups \cr
    to be used with the \link{PRE_FATE.params_PFGdispersal} function \cr}
    \item{dist}{a \code{data.frame} of dimension \code{384 x 5} \cr
    containing response of plant functional groups to disturbances \cr
    to be used with the \link{PRE_FATE.params_PFGdisturbance} function \cr}
    \item{global}{a \code{vector} of length \code{18} \cr
    containing global parameter values for a \code{FATE-HD} simulation in 
    the PNE \cr
    to be used with the \link{PRE_FATE.params_globalParameters} function \cr
    }
  }}
\source{

}
\usage{
PNE_PARAM
}
\description{
This dataset contains all necessary files to build the 
\code{FATE-HD} simulation folder as well as the parameter files 
(\emph{\href{https://mayagueguen.github.io/FATE-WEBSITE/papers/Boulangeat_2014_GCB.pdf}{Boulangeat et al. 2014 GCB}}).
}
\author{
Maya Guéguen
}
\keyword{datasets}
