% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{PNE_PFG}
\alias{PNE_PFG}
\title{Ecrins National Park (PNE) data : building of Plant Functional Groups}
\format{A \code{list} object with 6 elements to help building 
  the Plant Functional Group :
  
  \describe{
    \item{sp.observations}{a \code{data.frame} of dimension \code{168313 x 7} \cr
    containing releves data about plant species in the PNE \cr
    to be used with the \link{PRE_FATE.selectDominant} function \cr
    \itemize{
      \item \strong{sites} : sites ID
      \item \strong{X} : x-axis coordinates in Lambers (lcc)
      \item \strong{Y} : y-axis coordinates in Lambers (lcc)
      \item \strong{habitat} : habitat ID from 
      \href{http://www.ecrins-parcnational.fr/sites/ecrins-parcnational.com/files/fiche_doc/12083/2006-atlas-delphine.pdf}{DELPHINE} :
      \itemize{
        \item 0 : glaciers and eternal snows
        \item 31 : uncolonized rocks
        \item 40 : lawns and meadows
        \item 50 : low moors
        \item 60 : open areas, brush
        \item 70 : semi-closed areas
        \item 81 : closed areas
        \item 83 : forests
      }
      \item \strong{species} : species ID
      \item \strong{abund_BB} : Braun-Blanquet abundance \cr 
      (see \code{\link{PRE_FATE.abundBraunBlanquet}} function for details)
      \item \strong{abund} : relative abundance obtained from the 
      \code{abund_BB} column with the 
      \code{\link{PRE_FATE.abundBraunBlanquet}} function
    }
    }
    \item{dom.traits}{\code{data.frame} of dimension \code{359 x 6} \cr
    containing traits for dominant species \cr
    to be used with the \link{PRE_FATE.speciesDistance} function \cr
    \itemize{
      \item \strong{species} : species ID
      \item \strong{GROUP} : rough generalization of Raunkier life-forms :
      \itemize{
        \item H : herbaceous
        \item C : chamaephytes
        \item P : phanerophytes
      }
      \item \strong{height} : mean plant height (cm)
      \item \strong{dispersal} : classes (from 1 to 7) based on 
      dispersal distances and types (Vittoz & Engler)
      \item \strong{palatability} : classes (from 0 to 3) (CBNA)
      \item \strong{light} : Ellenberg indicator value for light (from 1 
      to 9)
    }
    }
    \item{dom.dist_overlap}{\code{matrix} of dimension \code{358 x 358} \cr
    containing niche overlap distance for dominant species \cr
    to be used with the \link{PRE_FATE.speciesDistance} function \cr}
    \item{dom.determ}{\code{data.frame} of dimension \code{359 x 5} \cr
    containing dominant species information relative to PFG \cr
    obtained from the \link{PRE_FATE.speciesClustering_step2} function \cr
    \itemize{
      \item \strong{species} : species ID
      \item \strong{name} : species name (taxonomic)
      \item \strong{GROUP} : rough generalization of Raunkier life-forms :
      \itemize{
        \item H : herbaceous
        \item C : chamaephytes
        \item P : phanerophytes
      }
      \item \strong{PFG} : name of assigned Plant Functional Group
      \item \strong{determinant} : is the species kept as determinant 
      species within the PFG ? (see \link{PRE_FATE.speciesClustering_step2} 
      function for details)
    }
    }
    \item{nb.clusters}{\code{vector} of length \code{3} \cr
    number of groups kept for each life-form class, obtained by cutting the 
    hierarchical tree obtained from species distances}
    \item{PFG.traits}{\code{data.frame} of dimension \code{24 x 10} \cr
    containing traits for plant functional groups \cr
    obtained from the \link{PRE_FATE.speciesClustering_step3} function \cr
    \itemize{
      \item \strong{PFG_name} : full descriptive Plant Functional Group name
      \item \strong{PFG} : Plant Functional Group short name
      \item \strong{type} : rough generalization of Raunkier life-forms :
      \itemize{
        \item H : herbaceous
        \item C : chamaephytes
        \item P : phanerophytes
      }
      \item \strong{strata} : maximum height stratum that can be reached by 
      the PFG (from 1 to 5)
      \item \strong{disp} : MEDIAN classes (from 1 to 7) based on 
      dispersal distances and types (Vittoz & Engler)
      \item \strong{light} : MEDIAN Ellenberg indicator value for light 
      (from 1 to 9)
      \item \strong{height} : MEAN PFG height (cm)
      \item \strong{palatability} : MEDIAN classes (from 0 to 3) (CBNA)
      \item \strong{longevity} : MEAN age of lifespan
      \item \strong{maturity} : MEAN age of maturity
    }
    }
  }}
\source{

}
\description{
This dataset contains elements to create the Plant Functional 
Groups (PFG) over the Ecrins National Park (PNE) 
(\emph{\href{https://mayagueguen.github.io/FATE-WEBSITE/papers/Boulangeat_2012_GCB.pdf}{Boulangeat et al. 2012 GCB}}).
}
\author{
Maya Guéguen
}
