% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{PNE_RESULTS}
\alias{PNE_RESULTS}
\title{Ecrins National Park (PNE) data : building of Plant Functional Groups}
\format{A \code{list} object with 2 elements :}
\source{

}
\description{
This dataset contains all data for the Ecrins National Park 
(PNE) area. It contains both the data to create the Plant Functional 
Groups (PFG) over this area 
(\emph{\href{https://mayagueguen.github.io/FATE-WEBSITE/papers/Boulangeat_2012_GCB.pdf}{Boulangeat et al. 2012 GCB}}), 
and the necessary files to build the \code{FATE-HD} simulation folder as well as 
the parameter files 
(\emph{\href{https://mayagueguen.github.io/FATE-WEBSITE/papers/Boulangeat_2014_GCB.pdf}{Boulangeat et al. 2014 GCB}}).
}
\author{
Maya Guéguen
}
