% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POST_FATE.binaryMaps.R
\name{POST_FATE.binaryMaps}
\alias{POST_FATE.binaryMaps}
\title{Create binary maps for each Plant Functional Group for one (or 
several) specific year of a \code{FATE} simulation}
\usage{
POST_FATE.binaryMaps(
  name.simulation,
  file.simulParam = NULL,
  years,
  method,
  method1.threshold = 0.05,
  method2.cutoff = NULL,
  opt.no_CPU = 1
)
}
\arguments{
\item{name.simulation}{a \code{string} corresponding to the main directory 
or simulation name of the \code{FATE} simulation}

\item{file.simulParam}{default \code{NULL}. \cr A \code{string} 
corresponding to the name of a parameter file that will be contained into 
the \code{PARAM_SIMUL} folder of the \code{FATE} simulation}

\item{years}{an \code{integer}, or a \code{vector} of \code{integer}, 
corresponding to the simulation year(s) that will be used to extract PFG 
abundance maps}

\item{method}{an \code{integer} to choose the transformation method : \cr 
\code{1} (relative abundance) or \code{2} (optimizing TSS) (see 
\href{POST_FATE.binaryMaps#details}{\code{Details}})}

\item{method1.threshold}{default \code{0.05}. \cr If \code{method = 1}, 
minimum relative abundance required for each PFG to be considered as present 
in the concerned pixel}

\item{method2.cutoff}{default \code{NULL}. \cr If \code{method = 2}, a 
\code{data.frame} with 3 columns : \code{year}, \code{PFG}, \code{cutoff} \cr
(see \href{POST_FATE.binaryMaps#details}{\code{Details}})}

\item{opt.no_CPU}{(\emph{optional}) default \code{1}. \cr The number of 
resources that can be used to parallelize the \code{unzip/zip} of raster 
files}
}
\value{
Two folders are created :
\describe{
  \item{\file{BIN_perPFG \cr_allStrata}}{containing presence / absence 
  raster maps for each PFG across all strata}
  \item{\file{BIN_perPFG \cr_perStrata}}{containing presence / absence 
  raster maps for each PFG for each stratum}
}
}
\description{
This script is designed to produce raster maps of PFG presence 
/ absence for one (or several) specific \code{FATE} simulation year.
}
\details{
This function allows to obtain, for a specific \code{FATE} simulation and 
a specific parameter file within this simulation, \strong{raster maps of PFG 
presence / absence}. \cr \cr


For each PFG and each selected simulation year, raster maps are retrieved 
from the results folder \code{ABUND_REL_perPFG_allStrata} and lead to the 
production of as many maps as those found :

\describe{
  \item{\code{1} fixed threshold}{relative abundance maps are transformed 
  into binary maps according to the threshold given by 
  \code{method1.threshold} : 
  \deqn{abund\_rel_{\text{ PFG}_i} > \text{method1.threshold} \;\; 
  \Leftrightarrow \;\; 1}}
  \item{\code{2} optimizing TSS}{relative abundance maps are transformed 
  into binary maps according to the \href{.getCutoff}{\code{cutoff}} found 
  with the \code{\link{POST_FATE.graphic_validationStatistics}} function : 
  \deqn{abund\_rel_{\text{ PFG}_i} > \text{method2.cutoff}_{\text{ PFG}_i} \;\; 
  \Leftrightarrow \;\; 1}}
}

Binary maps per stratum are obtained by multiplying raster maps from 
\code{ABUND_perPFG_perStrata} folder by corresponding raster maps from 
\code{BIN_perPFG_allStrata} folder.

\strong{It requires} that the \code{\link{POST_FATE.relativeAbund}} 
function has been run and that the folder \code{ABUND_REL_perPFG_allStrata} 
exists. \cr If \code{method = 2}, it requires that the 
\code{\link{POST_FATE.graphic_validationStatistics}} function has been run. 
\cr \cr

\strong{These binary \code{raster} files can then be used by other 
functions} :

\itemize{
  \item to produce graphics of \emph{PFG modelled presence} \code{vs} 
  \emph{PFG Habitat Suitability} maps \cr (see 
  \code{\link{POST_FATE.graphic_mapPFGvsHS}})
}
}
\examples{

                                                        
## ----------------------------------------------------------------------------------------- ##
## Load example data
                                                                         

}
\seealso{
\code{\link{POST_FATE.relativeAbund}},
\code{\link{POST_FATE.graphic_validationStatistics}}
\code{\link{.getCutoff}},
\code{\link{POST_FATE.graphic_mapPFGvsHS}}
}
\author{
Maya Guéguen
}
\keyword{FATE,}
\keyword{binary}
\keyword{outputs,}
