% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POST_FATE.graphic_evolutionAbund_pixels.R
\name{POST_FATE.graphic_evolutionAbund_pixels}
\alias{POST_FATE.graphic_evolutionAbund_pixels}
\title{Create a graphical representation of the evolution of PFG abundance
through time for 5 (or more) pixels of a \code{FATE-HD} simulation}
\usage{
POST_FATE.graphic_evolutionAbund_pixels(name.simulation,
  file.simulParam = NULL, no.years = 10, opt.abund_fixedScale = TRUE,
  opt.no_CPU = 1, opt.cells_ID = NULL)
}
\arguments{
\item{name.simulation}{a \code{string} that corresponds to the main directory
or simulation name of the \code{FATE-HD} simulation}

\item{file.simulParam}{a \code{string} that corresponds to the name of a
parameter file that will be contained into the \code{PARAM_SIMUL} folder
of the \code{FATE-HD} simulation}

\item{no.years}{an \code{integer} corresponding to the number of simulation
years that will be used to extract PFG abundance maps}

\item{opt.abund_fixedScale}{default \code{TRUE}. If \code{FALSE}, the ordinate
scale will be adapted for each PFG for the graphical representation of the 
evolution of abundances through time}

\item{opt.no_CPU}{default 1 (\emph{optional}). The number of resources that 
can be used to parallelize the \code{unzip/zip} of raster files}

\item{opt.cells_ID}{default NULL (\emph{optional}). The cells ID of the 
studied area for which PFG abundances will be extracted.
number of resources that 
can be used to parallelize the \code{unzip/zip} of raster files}
}
\value{
One \code{POST_FATE_[...].pdf} file is created : 
\describe{
  \item{\file{GRAPHIC_B \cr abundance}}{to visualize for each PFG the
  evolution of its abundance within each selected pixel through
  simulation time}
}
}
\description{
This script is designed to produce one graphical 
representations for a \code{FATE-HD} simulation : the evolution through time
of the abundance of each PFG for 5 (or more) randomly selected cells of the
studied area.
}
\details{
This function allows one to obtain, for a specific \code{FATE-HD} simulation
and a specific parameter file within this simulation, one preanalytical
graphics. \cr

For each PFG and each selected simulation year, raster maps are retrieved
from the results folder \code{ABUND_perPFG_allStrata} and unzipped.
Informations extracted lead to the production of one graphic before the
maps are compressed again :

\itemize{
  \item{the evolution of \strong{abundance} of each Plant Functional
  Group through simulation time, within 5 (or more) randomly selected pixels
  of the studied area (\code{FATE-HD} \emph{arbitrary unit})
  }
}
}
\examples{

\dontrun{                      
POST_FATE.graphic_evolutionAbund_pixels(name.simulation = "FATE_simulation"
                                    , file.simulParam = "Simul_parameters_V1.txt"
                                    , opt.no_CPU = 1)
                                    
POST_FATE.graphic_evolutionAbund_pixels(name.simulation = "FATE_simulation"
                                    , file.simulParam = "Simul_parameters_V1.txt"
                                    , no.years = 50
                                    , opt.no_CPU = 4)
                                    
POST_FATE.graphic_evolutionAbund_pixels(name.simulation = "FATE_simulation"
                                    , file.simulParam = "Simul_parameters_V1.txt"
                                    , no.years = 50
                                    , opt.abund_fixedScale = FALSE
                                    , opt.no_CPU = 1)
}
                                    
                                    

}
\seealso{
\code{\link{POST_FATE.relativeAbund_presenceAbsence}}
}
\author{
Maya Guéguen
}
\keyword{FATE,}
\keyword{abundance}
\keyword{outputs,}
\keyword{through}
\keyword{time}
