% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POST_FATE.graphic_evolutionCoverage.R
\name{POST_FATE.graphic_evolutionCoverage}
\alias{POST_FATE.graphic_evolutionCoverage}
\title{Create a graphical representation of the evolution of PFG coverage
for a \code{FATE-HD} simulation}
\usage{
POST_FATE.graphic_evolutionCoverage(name.simulation, file.simulParam = NULL,
  opt.no_CPU = 1)
}
\arguments{
\item{name.simulation}{a \code{string} that corresponds to the main directory
or simulation name of the \code{FATE-HD} simulation}

\item{file.simulParam}{\emph{to be filled}}

\item{opt.no_CPU}{default 1 (\emph{optional}). The number of resources that 
can be used to parallelize the \code{unzip/zip} of raster files}
}
\value{
T
}
\description{
This script is designed to create a parameter file containing
simulation years at which the \code{FATE-HD} software must save rasters of
PFG abundances or simulation objects.
}
\details{
\emph{to be filled}
}
\examples{

## Create a skeleton folder with the default name ('FATE_simulation')
if (dir.exists("FATE_simulation")) system("rm -r FATE_simulation/")
PRE_FATE.skeletonDirectory()

## Create a Namespace_constants parameter file
PRE_FATE.params_namespaceConstants(name.simulation = "FATE_simulation"
                                   , global.abund.low = 1000000
                                   , global.abund.med = 5000000
                                   , global.abund.high = 8000000
                                   , global.max.by.cohort = 5000000
                                   , global.resource.thresh.med = 13000000
                                   , global.resource.thresh.low = 19000000)

## Create a Global_parameters file
PRE_FATE.params_globalParameters(name.simulation = "FATE_simulation"
                                 , required.succ_option = "fateh"
                                 , required.no_PFG = 6
                                 , required.no_STRATA = 5
                                 , required.hs_option = 1
                                 , required.seeding_timestep = 1
                                 , required.seeding_duration = c(10,50)
                                 , required.simul_duration = 100)

## Create a SAVE_year_maps or/and SAVE_year_objects parameter file
PRE_FATE.params_saveYears(name.simulation = "FATE_simulation"
                          , years.maps = c(100, 150, 200)
                          , years.objects = 200)

## Create a Changing_times parameter file
PRE_FATE.params_changingYears(name.simulation = "FATE_simulation"
                              , type.changing = "DIST"
                              , mat.changing = data.frame(year = c(50,50,80,80)
                                                          , order = c(1,2,1,2)
                                                          , file.name = c("MASK_DIST1_50.tif"
                                                                          , "MASK_DIST2_50.tif"
                                                                          , "MASK_DIST1_80.tif"
                                                                          , "MASK_DIST2_80.tif")))

## Create PFG succession parameter files
PRE_FATE.params_PFGsuccession(name.simulation = "FATE_simulation"
                              , mat.PFG.succ = data.frame(PFG = paste0("PFG",1:6)
                                                          , type = c("C", "C", "H", "H", "P", "P")
                                                          , height = c(10, 250, 36, 68, 1250, 550)
                                                          , maturity = c(5, 5, 3, 3, 8, 9)
                                                          , longevity = c(12, 200, 25, 4, 110, 70)
                                                          , dispersal = 1
                                                          , light = c(4, 6, 3, 6, 5, 5)))

## Create PFG dispersal parameter files
PRE_FATE.params_PFGdispersal(name.simulation = "FATE_simulation"
                             , mat.PFG.disp = data.frame(PFG = paste0("PFG", 1:6)
                                                         , d50 = rep(c(500, 500, 100),2)
                                                         , d99 = rep(c(10000, 15000, 20000),2)
                                                         , ldd = rep(c(100000, 50000, 100000),2)))

## Create PFG disturbance parameter files
PRE_FATE.params_PFGdisturbance(name.simulation = "FATE_simulation"
                               , mat.PFG.dist = data.frame(name = rep(c("DIST1","DIST2"), each = 4)
                                                           , responseStage = rep(1:4, 2)
                                                           , KilledIndiv_H = c(0,0,0,0,1,1,0,0)
                                                           , KilledIndiv_C = c(0,10,10,10,1,1,0,0)
                                                           , KilledIndiv_P = c(10,10,10,10,10,0,0,0)
                                                           , ResproutIndiv_H = c(0,0,9,10,0,0,5,1)
                                                           , ResproutIndiv_C = c(0,0,0,0,0,0,5,1)
                                                           , ResproutIndiv_P = c(0,0,0,0,0,0,0,0)))

## Create a Simulation parameter file
file.create("FATE_simulation/DATA/MASK/mask.tif")
PRE_FATE.params_simulParameters(name.simulation = "FATE_simulation"
                                , name.mask = "mask.tif")

}
\author{
Maya Guéguen
}
