% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POST_FATE.graphic_evolutionLight_pixels.R
\name{POST_FATE.graphic_evolutionLight_pixels}
\alias{POST_FATE.graphic_evolutionLight_pixels}
\title{Create a graphical representation of the evolution of light
resources through time for 5 (or more) pixels of a \code{FATE-HD} simulation}
\usage{
POST_FATE.graphic_evolutionLight_pixels(name.simulation,
  file.simulParam = NULL, no.years = 10, opt.abund_fixedScale = TRUE,
  opt.cells_ID = NULL, opt.no_CPU = 1, opt.doPlot = TRUE)
}
\arguments{
\item{name.simulation}{a \code{string} that corresponds to the main
directory or simulation name of the \code{FATE-HD} simulation}

\item{file.simulParam}{a \code{string} that corresponds to the name of a
parameter file that will be contained into the \code{PARAM_SIMUL} folder
of the \code{FATE-HD} simulation}

\item{no.years}{an \code{integer} corresponding to the number of simulation
years that will be used to extract PFG abundance maps}

\item{opt.abund_fixedScale}{default \code{TRUE}. If \code{FALSE}, the
ordinate scale will be adapted for each stratum for the graphical
representation of the evolution of light resources through time}

\item{opt.cells_ID}{default NULL (\emph{optional}). The cells ID of the 
studied area for which light resources will be extracted.}

\item{opt.no_CPU}{default 1 (\emph{optional}). The number of resources that 
can be used to parallelize the \code{unzip/zip} of raster files}

\item{opt.doPlot}{default TRUE (\emph{optional}). If TRUE, plot(s) will be
processed, otherwise only the calculation and reorganization of outputs
will occur, be saved and returned.}
}
\value{
One \code{POST_FATE_[...].pdf} file is created : 
\describe{
  \item{\file{GRAPHIC_A \cr light resources}}{to visualize for each stratum
  the evolution of its light resource within each selected pixel through
  simulation time}
}
}
\description{
This script is designed to produce one graphical representation
for a \code{FATE-HD} simulation : the evolution through time of the light
resources for 5 (or more) randomly selected cells of the studied area.
}
\details{
This function allows one to obtain, for a specific \code{FATE-HD} simulation
and a specific parameter file within this simulation, one preanalytical
graphics. \cr

For each stratum and each selected simulation year, raster maps are
retrieved from the results folder \code{LIGHT} and unzipped.
Informations extracted lead to the production of one graphic before the
maps are compressed again :

\itemize{
  \item{the evolution of \strong{light resources} of each stratum through
  simulation time, within 5 (or more) randomly selected pixels
  of the studied area (\emph{1: Low, 2: Medium, 3: High})
  }
}
}
\examples{

\dontrun{                      
POST_FATE.graphic_evolutionLight_pixels(name.simulation = "FATE_simulation"
                                    , file.simulParam = "Simul_parameters_V1.txt"
                                    , opt.no_CPU = 1)
                                    
POST_FATE.graphic_evolutionLight_pixels(name.simulation = "FATE_simulation"
                                    , file.simulParam = "Simul_parameters_V1.txt"
                                    , no.years = 50
                                    , opt.no_CPU = 4)
                                    
POST_FATE.graphic_evolutionLight_pixels(name.simulation = "FATE_simulation"
                                    , file.simulParam = "Simul_parameters_V1.txt"
                                    , no.years = 50
                                    , opt.abund_fixedScale = FALSE
                                    , opt.no_CPU = 1)
}
                                    
                                    

}
\seealso{
\code{\link{POST_FATE.relativeAbund}}, 
\code{\link{POST_FATE.graphic_validationStatistics}}
}
\author{
Maya Guéguen
}
\keyword{FATE,}
\keyword{light}
\keyword{outputs,}
\keyword{resources}
\keyword{through}
\keyword{time}
