% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POST_FATE.graphic_evolutionPixels.R
\name{POST_FATE.graphic_evolutionPixels}
\alias{POST_FATE.graphic_evolutionPixels}
\title{Create a graphical representation of the evolution of PFG abundance 
through time for 5 (or more) pixels of a \code{FATE-HD} simulation}
\usage{
POST_FATE.graphic_evolutionPixels(name.simulation,
  file.simulParam = NULL, opt.abund_fixedScale = TRUE,
  opt.cells_ID = NULL, opt.doPlot = TRUE)
}
\arguments{
\item{name.simulation}{a \code{string} that corresponds to the main 
directory or simulation name of the \code{FATE-HD} simulation}

\item{file.simulParam}{a \code{string} that corresponds to the name of a 
parameter file that will be contained into the \code{PARAM_SIMUL} folder 
of the \code{FATE-HD} simulation}

\item{opt.abund_fixedScale}{default \code{TRUE}. If \code{FALSE}, the 
ordinate scale will be adapted for each PFG for the graphical 
representation of the evolution of abundances through time}

\item{opt.cells_ID}{default \code{NULL} (\emph{optional}). The cells ID of 
the studied area for which PFG abundances will be extracted.}

\item{opt.doPlot}{default \code{TRUE} (\emph{optional}). If \code{TRUE}, 
plot(s) will be processed, otherwise only the calculation and 
reorganization of outputs will occur, be saved and returned.}
}
\value{
A \code{list} containing one \code{data.frame} object with the 
following columns, and one \code{ggplot2} object :

\describe{
  \item{tab}{
    \describe{
      \item{\code{TYPE}}{the concerned information (either '\code{light}', 
      '\code{abundance}' or '\code{soil}')}
      \item{\code{GROUP}}{the concerned entity (either 
      '\code{STRATUM_[...]}', PFG name or '\code{soil}')}
      \item{\code{ID}}{the id of the concerned pixel}
      \item{\code{HAB}}{the habitat of the concerned pixel}
      \item{\emph{YEAR}}{the concerned simulation year}
      \item{\code{Occupancy}}{the number of occupied pixels divided by the 
      total number of pixels within the studied area}
    }
  }
  \item{plot}{\code{ggplot2} object, representing the evolution 
  of each PFG abundance, \emph{and light and soil resources if those 
  modules are activated} \cr \cr}
}


One \code{POST_FATE_TABLE_PIXEL_evolution_pixels_[...].csv} file is created : 
\describe{
  \item{\file{pixels ids}}{always, containing the \code{data.frame} detailed above}
}


One \code{POST_FATE_[...].pdf} file is created : 
\describe{
  \item{\file{GRAPHIC_A \cr pixels}}{to visualize for each PFG the 
  evolution of its abundance within each selected pixel through 
  simulation time}
}
}
\description{
This script is designed to produce one graphical 
representation for a \code{FATE-HD} simulation : the evolution through time 
of the abundance of each PFG for 5 (or more) randomly selected cells of the 
studied area.
}
\details{
This function allows one to obtain, for a specific \code{FATE-HD} 
simulation and a specific parameter file within this simulation, one 
preanalytical graphics :

\itemize{
  \item{the evolution of \strong{abundance} of each Plant Functional Group 
  through simulation time, within 5 (or more) randomly selected pixels of 
  the studied area (\code{FATE-HD} \emph{arbitrary unit})
  }
  \item{\strong{if LIGHT is activated} (see 
  \code{\link{PRE_FATE.params_globalParameters}}), evolution of 
  \strong{light resources} within the selected pixels is also represented 
  (\emph{1: Low, 2: Medium, 3: High})
  }
  \item{\strong{if SOIL is activated} (see 
  \code{\link{PRE_FATE.params_globalParameters}}), evolution of 
  \strong{soil resources} within the selected pixels is also represented 
  (user-defined scale) \cr \cr
  }
}


It requires that the \code{\link{POST_FATE.temporalEvolution}} function has 
been run and that the file 
\code{POST_FATE.evolution_abundance_PIXEL_[...].csv} exists (as well as the 
files \code{POST_FATE.evolution_light_PIXEL_[...].csv} and 
\code{POST_FATE.evolution_soil_PIXEL_[...].csv} if those modules are 
activated).
}
\examples{

\dontrun{                      
POST_FATE.graphic_evolutionPixels(name.simulation = "FATE_simulation"
                                    , file.simulParam = "Simul_parameters_V1.txt")
                                    
POST_FATE.graphic_evolutionPixels(name.simulation = "FATE_simulation"
                                    , file.simulParam = "Simul_parameters_V1.txt"
                                    , opt.abund_fixedScale = FALSE)
}
                                    
                                    

}
\seealso{
\code{\link{POST_FATE.temporalEvolution}}
}
\author{
Maya Guéguen
}
\keyword{FATE,}
\keyword{abundance}
\keyword{outputs,}
\keyword{through}
\keyword{time}
