% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POST_FATE.graphic_evolutionStability.R
\name{POST_FATE.graphic_evolutionStability}
\alias{POST_FATE.graphic_evolutionStability}
\title{Create a graphical representation of the evolution of habitat 
composition through time for a \code{FATE} simulation}
\usage{
POST_FATE.graphic_evolutionStability(
  name.simulation,
  file.simulParam = NULL,
  movingWindow_size = 3,
  movingWindow_step = 1,
  opt.doPlot = TRUE
)
}
\arguments{
\item{name.simulation}{a \code{string} corresponding to the main directory 
or simulation name of the \code{FATE} simulation}

\item{file.simulParam}{default \code{NULL}. \cr A \code{string} 
corresponding to the name of a parameter file that will be contained into 
the \code{PARAM_SIMUL} folder of the \code{FATE} simulation}

\item{movingWindow_size}{default \code{3}. \cr An \code{integer} 
corresponding to the size (\emph{in years}) of the moving window that will 
be used to calculate metrics of habitat stability}

\item{movingWindow_step}{default \code{1}. \cr An \code{integer} 
corresponding to the step (\emph{in years}) between the years of the moving 
window}

\item{opt.doPlot}{(\emph{optional}) default \code{TRUE}. \cr If \code{TRUE}, 
plot(s) will be processed, otherwise only the calculation and reorganization 
of outputs will occur, be saved and returned}
}
\value{
A \code{list} containing two \code{data.frame} objects with the 
following columns, and one \code{ggplot2} object :

\describe{
  \item{tab.hab}{
    \describe{
      \item{\code{HAB}}{concerned habitat}
      \item{\code{year}}{concerned simulation year}
      \item{\code{totalAbundance}}{total abundance over all the pixels 
      within the concerned habitat}
      \item{\code{no.PFG}}{number of PFG over all the pixels within the 
      concerned habitat}
      \item{\code{evenness}}{evenness over all the pixels within the 
      concerned habitat}
    }
  }
  \item{tab.stab}{
    \describe{
      \item{\code{HAB}}{concerned habitat}
      \item{\code{no.years}}{number of simulation years used (\emph{moving 
      window size})}
      \item{\code{yearStep}}{step between each simulation year of the moving 
      window}
      \item{\code{yearStart}}{first simulation year of the moving window}
      \item{\code{yearEnd}}{last simulation year of the moving window}
      \item{\code{metric}}{concerned metric (either \code{totalAbundance} or 
      \code{evenness})}
      \item{\code{mean}}{mean value of the concerned metric over the years 
      of the concerned moving window}
      \item{\code{sd}}{value of standard deviation of the concerned metric 
      over the years of the concerned moving window}
      \item{\code{cv}}{value of coefficient of variation of the concerned 
      metric over the years of the concerned moving window}
    }
  }
  \item{plot.stab}{\code{ggplot2} object, representing the evolution of 
  total abundance and evenness of each habitat \cr \cr}
}

Two \code{POST_FATE_TABLE_HAB_evolution_[...].csv} files are created : 
\describe{
  \item{\file{stability1}}{always, containing \code{tab.hab}}
  \item{\file{stability2}}{\emph{if successive years available}, containing 
  \code{tab.stab}}
}


One \code{POST_FATE_[...].pdf} files is created : 
\describe{
  \item{\file{GRAPHIC_A \cr stability}}{to visualize for each habitat the 
  evolution of its total abundance and its evenness through simulation time}
}
}
\description{
This script is designed to produce one graphical representation 
for a \code{FATE} simulation : the evolution through time of the total 
abundance and evenness of each habitat.
}
\details{
This function allows to obtain, for a specific \code{FATE} simulation and 
a specific parameter file within this simulation, one preanalytical graphic :

\itemize{
  \item{the evolution of \strong{total abundance} (\code{FATE} 
  \emph{arbitrary unit}) and \strong{evenness} (\emph{between \code{0} and 
  \code{1}}) of each habitat through simulation time, with \emph{evenness} 
  representing the uniformity of the species composition of the habitat 
  (similar to \strong{Shannon entropy}) :
  \deqn{
  \text{evenness} = - \frac{\Sigma(\text{proportion}_{\text{ PFG}_i} *
  log(\text{proportion}_{\text{ PFG}_i}))}{log(\text{no.PFG})}
  }
  with \deqn{
  \text{proportion}_{\text{ PFG}_i} = \frac{abund_{\text{ PFG}_i
  \text{, }\text{Habitat}_j}}{abund_{\text{ PFG}_{all}\text{, }
  \text{Habitat}_j}}
  }
  }
}

If the information has been provided (see 
\code{\link{POST_FATE.temporalEvolution}}), the graphics will be also done 
per habitat. \cr \cr

\strong{It requires} that the \code{\link{POST_FATE.temporalEvolution}} 
function has been run and that the file 
\code{POST_FATE_TABLE_PIXEL_evolution_abundance.csv} exists.
}
\examples{

\dontrun{                      
POST_FATE.graphic_evolutionStability(name.simulation = "FATE_simulation"
                                    , file.simulParam = "Simul_parameters_V1.txt")
                                    
POST_FATE.graphic_evolutionStability(name.simulation = "FATE_simulation"
                                    , file.simulParam = "Simul_parameters_V1.txt"
                                    , opt.fixedScale = FALSE)
}



## ----------------------------------------------------------------------------------------- ##
## Load example data


}
\seealso{
\code{\link{POST_FATE.temporalEvolution}}
}
\author{
Maya Guéguen
}
\keyword{FATE,}
\keyword{abundance}
\keyword{evenness}
\keyword{outputs,}
\keyword{through}
\keyword{time,}
