% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POST_FATE.graphic_mapPFGcover.R
\name{POST_FATE.graphic_mapPFGcover}
\alias{POST_FATE.graphic_mapPFGcover}
\title{Create a map of the Plant Functional Group cover 
 \cr for one (or several) specific year of a \code{FATE-HD} simulation}
\usage{
POST_FATE.graphic_mapPFGcover(name.simulation, file.simulParam = NULL,
  year, strata_min, opt.mat.cover.obs = NULL, opt.ras.cover.obs = NULL,
  opt.no_CPU = 1, opt.doPlot = TRUE)
}
\arguments{
\item{name.simulation}{a \code{string} that corresponds to the main 
directory or simulation name of the \code{FATE-HD} simulation}

\item{file.simulParam}{a \code{string} that corresponds to the name of a 
parameter file that will be contained into the \code{PARAM_SIMUL} folder 
of the \code{FATE-HD} simulation}

\item{year}{an \code{integer} corresponding to the simulation year(s) that 
will be used to extract PFG abundance and binary maps}

\item{strata_min}{an \code{integer} corresponding to the lowest stratum 
from which PFG abundances are summed up to the highest stratum}

\item{opt.mat.cover.obs}{default \code{NULL} (\emph{optional}). A 
\code{data.frame} with 3 columns : \code{X}, \code{Y}, \code{obs}}

\item{opt.ras.cover.obs}{default \code{NULL} (\emph{optional}). A 
\code{string} that corresponds to the file name of a raster containing 
observed values for vegetation cover}

\item{opt.no_CPU}{default \code{1} (\emph{optional}). The number of 
resources that can be used to parallelize the \code{unzip/zip} of raster 
files}

\item{opt.doPlot}{default \code{TRUE} (\emph{optional}). If \code{TRUE}, 
plot(s) will be processed, otherwise only the calculation and 
reorganization of outputs will occur, be saved and returned.}
}
\value{
A \code{list} containing one or several (one for each simulation 
year) \code{list} of \code{raster} and \code{data.frame} with the following 
columns, and one or two \code{ggplot2} objects :

\describe{
  \item{tab}{ 
    \describe{ 
      \item{raster}{}
      \item{ccr}{ 
        \describe{
          \item{\code{PFG}}{the concerned Plant Functional Group}
          \item{\code{AUC.sd}}{standard deviation of the AUC values}
        }
      }
    }
  }
  \item{plot}{
    \describe{
      \item{plot.cover}{\code{ggplot2} object, representing the }
      \item{plot.ccr}{\code{ggplot2} object, representing the \cr \cr}
    }
  }
}


One \code{POST_FATE_[...].pdf} file is created : 
\describe{
  \item{\file{GRAPHIC_C \cr map_PFGcover}}{to visualize the PFG cover
  within the studied area}
}

A raster file is created into simulation results folder.
}
\description{
This script is designed to produce a raster map of PFG cover 
for one (or several) specific \code{FATE-HD} simulation year.
}
\details{
This function allows one to obtain, for a specific \code{FATE-HD} simulation
and a specific parameter file within this simulation, one raster 
maps of PFG relative cover and one preanalytical graphic. \cr \cr

For each PFG and each selected simulation year, raster maps are retrieved 
from the results folders \code{ABUND_perPFG_perStrata} and 
\code{BIN_perPFG_perStrata} and unzipped. 
Informations extracted lead to the production of one graphic before the
maps are compressed again :

\itemize{
  \item{the map of \strong{Plant Functional Group cover} for each selected
  simulation year(s), representing the relative cumulated abundance of PFG 
  present in each pixel above a height threshold
  }
}
}
\examples{

\dontrun{                      
POST_FATE.graphic_mapPFGcover(name.simulation = "FATE_simulation"
                              , file.simulParam = "Simul_parameters_V1.txt"
                              , year = 850
                              , strata_min = 3
                              , opt.no_CPU = 1)
                                    
POST_FATE.graphic_mapPFGcover(name.simulation = "FATE_simulation"
                              , file.simulParam = "Simul_parameters_V1.txt"
                              , year = c(850, 950)
                              , strata_min = 3
                              , opt.no_CPU = 1)
}
                                    
                                    

}
\seealso{
\code{\link{POST_FATE.relativeAbund}}, 
\code{\link{POST_FATE.graphic_validationStatistics}}
}
\author{
Maya Guéguen
}
\keyword{FATE,}
\keyword{abundance,}
\keyword{area}
\keyword{classification}
\keyword{correct}
\keyword{cover,}
\keyword{curve,}
\keyword{forest}
\keyword{outputs,}
\keyword{rate}
\keyword{relative}
\keyword{sensitivity,}
\keyword{skill}
\keyword{specificity,}
\keyword{statistic,}
\keyword{true}
\keyword{under}
