% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POST_FATE.graphic_mapPFGsoil.R
\name{POST_FATE.graphic_mapPFGsoil}
\alias{POST_FATE.graphic_mapPFGsoil}
\title{Create a map of the community weighted mean (CWM) of soil
 \cr for one (or several) specific year of a \code{FATE-HD} simulation}
\usage{
POST_FATE.graphic_mapPFGsoil(name.simulation, file.simulParam = NULL,
  year, strata_min = 1, mat.PFG.succ, opt.mat.soil.obs = NULL,
  opt.ras.soil.obs = NULL, opt.no_CPU = 1, opt.doPlot = TRUE)
}
\arguments{
\item{name.simulation}{a \code{string} that corresponds to the main directory
or simulation name of the \code{FATE-HD} simulation}

\item{file.simulParam}{a \code{string} that corresponds to the name of a
parameter file that will be contained into the \code{PARAM_SIMUL} folder
of the \code{FATE-HD} simulation}

\item{year}{an \code{integer} corresponding to the simulation year(s) that 
will be used to extract PFG abundance and binary maps}

\item{strata_min}{an \code{integer} corresponding to the lowest stratum from
which PFG abundances are summed up to the highest stratum}

\item{mat.PFG.succ}{a \code{data.frame} with 2 columns : PFG, soil_contrib}

\item{opt.mat.soil.obs}{default NULL (\emph{optional}). A \code{data.frame}
with 3 columns : X, Y, obs}

\item{opt.ras.soil.obs}{default NULL (\emph{optional}). A \code{string} that
corresponds to the file name of a raster containing observed values for 
community weighted mean of soil}

\item{opt.no_CPU}{default 1 (\emph{optional}). The number of resources that 
can be used to parallelize the \code{unzip/zip} of raster files}

\item{opt.doPlot}{default TRUE (\emph{optional}). If TRUE, plot(s) will be
processed, otherwise only the calculation and reorganization of outputs
will occur, be saved and returned.}
}
\value{
One \code{POST_FATE_[...].pdf} file is created : 
\describe{
  \item{\file{GRAPHIC_B \cr PFGlsoil}}{to visualize the soil CWM
  within the studied area}
}
}
\description{
This script is designed to produce a raster map of community
weighted mean (CWM) of soil for one (or several) specific \code{FATE-HD}
simulation year.
}
\details{
This function allows one to obtain, for a specific \code{FATE-HD} simulation
and a specific parameter file within this simulation, one preanalytical
graphic. \cr

For each PFG and each selected simulation year, raster maps are retrieved
from the results folders \code{ABUND_perPFG_perStrata} and 
\code{BIN_perPFG_perStrata} and unzipped.
Informations extracted lead to the production of one graphic before the
maps are compressed again :

\itemize{
  \item{the map of \strong{soil Community Weighted Mean} for each selected
  simulation year(s), representing the simulated value of soil (Landolt)
  within each pixel
  }
}
}
\examples{

\dontrun{                      
POST_FATE.graphic_mapPFGsoil(name.simulation = "FATE_simulation"
                              , file.simulParam = "Simul_parameters_V1.txt"
                              , year = 850
                              , strata_min = 1
                              , mat.PFG.succ = data.frame(PFG = paste0("PFG", 1:3)
                                                          , soil_contrib = c(1.8, 2, 3.1))
                              , opt.no_CPU = 1)
                                    
POST_FATE.graphic_mapPFGsoil(name.simulation = "FATE_simulation"
                              , file.simulParam = "Simul_parameters_V1.txt"
                              , year = c(850, 950)
                              , strata_min = 1
                              , mat.PFG.succ = data.frame(PFG = paste0("PFG", 1:3)
                                                          , soil_contrib = c(1.8, 2, 3.1))
                              , opt.no_CPU = 1)
}
                                    
                                    

}
\seealso{
\code{\link{POST_FATE.relativeAbund}}, 
\code{\link{POST_FATE.graphic_validationStatistics}}
}
\author{
Maya Guéguen
}
\keyword{FATE,}
\keyword{abundance,}
\keyword{community}
\keyword{mean,}
\keyword{outputs,}
\keyword{relative}
\keyword{soil,}
\keyword{weighted}
