% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POST_FATE.graphic_mapPFGvsHS.R
\name{POST_FATE.graphic_mapPFGvsHS}
\alias{POST_FATE.graphic_mapPFGvsHS}
\title{Create maps of both habitat suitability and simulated occurrences of 
each Plant Functional Group for one (or several) specific year of a 
\code{FATE} simulation}
\usage{
POST_FATE.graphic_mapPFGvsHS(
  name.simulation,
  file.simulParam = NULL,
  years,
  opt.stratum = "all"
)
}
\arguments{
\item{name.simulation}{a \code{string} corresponding to the main directory 
or simulation name of the \code{FATE} simulation}

\item{file.simulParam}{default \code{NULL}. \cr A \code{string} 
corresponding to the name of a parameter file that will be contained into 
the \code{PARAM_SIMUL} folder of the \code{FATE} simulation}

\item{years}{an \code{integer}, or a \code{vector} of \code{integer}, 
corresponding to the simulation year(s) that will be used to extract PFG 
binary maps}

\item{opt.stratum}{(\emph{optional}) default \code{all}. \cr The stratum 
number from which to extract PFG binary maps}
}
\value{
A \code{list} containing one or several (one for each simulation 
year) \code{list} of \code{ggplot2} objects, representing for each plant 
functional group its map of modelled presence / absence \code{vs} its 
habitat suitability map. \cr \cr


One \code{POST_FATE_[...].pdf} file is created : 
\describe{
  \item{\file{GRAPHIC_B \cr map_PFGvsHS}}{to visualize the PFG presence 
  within the studied area (probability and simulated occurrence)}
}
}
\description{
This script is designed to produce raster maps of PFG habitat 
suitability and simulated occurrences for one (or several) specific 
\code{FATE} simulation year.
}
\details{
This function allows to obtain, for a specific \code{FATE} simulation and 
a specific parameter file within this simulation, one preanalytical graphic. 
\cr \cr

For each PFG and each selected simulation year, raster maps are retrieved 
from the results folder \code{BIN_perPFG_allStrata} (unless the 
\code{opt.stratum} is used, then it will be from the folder 
\code{BIN_perPFG_perStrata}) and unzipped. 
Informations extracted lead to the production of one graphic before the 
maps are compressed again :

\itemize{
  \item{the comparison between each \strong{PFG habitat suitability map and 
  its simulated map of presence} \cr \cr
  }
}


\strong{It requires} that the \code{\link{POST_FATE.relativeAbund}} and the 
\code{\link{POST_FATE.binaryMaps}} function have been run 
and that the folders \code{BIN_perPFG_allStrata} and 
\code{BIN_perPFG_perStrata} exist.
}
\examples{

\dontrun{                      
POST_FATE.graphic_mapPFGvsHS(name.simulation = "FATE_simulation"
                             , file.simulParam = "Simul_parameters_V1.txt"
                             , years = 850)
                                    
POST_FATE.graphic_mapPFGvsHS(name.simulation = "FATE_simulation"
                             , file.simulParam = "Simul_parameters_V1.txt"
                             , years = c(850, 950))
                                    
POST_FATE.graphic_mapPFGvsHS(name.simulation = "FATE_simulation"
                             , file.simulParam = "Simul_parameters_V1.txt"
                             , years = 850
                             , opt.stratum = 2)
}
                                    
                                    

}
\seealso{
\code{\link{POST_FATE.relativeAbund}}, 
\code{\link{POST_FATE.binaryMaps}}
}
\author{
Maya Guéguen
}
\keyword{FATE,}
\keyword{binary}
\keyword{habitat}
\keyword{outputs,}
\keyword{suitability,}
