% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POST_FATE.graphic_validationStatistics.R
\name{POST_FATE.graphic_validationStatistics}
\alias{POST_FATE.graphic_validationStatistics}
\title{Create a graphical representation of several statistics for each PFG 
to asses the quality of the model for one (or several) specific year of a 
\code{FATE} simulation}
\usage{
POST_FATE.graphic_validationStatistics(
  name.simulation,
  file.simulParam = NULL,
  years,
  mat.PFG.obs,
  opt.ras_habitat = NULL,
  opt.doPlot = TRUE
)
}
\arguments{
\item{name.simulation}{a \code{string} corresponding to the main directory 
or simulation name of the \code{FATE} simulation}

\item{file.simulParam}{default \code{NULL}. \cr A \code{string} 
corresponding to the name of a parameter file that will be contained into 
the \code{PARAM_SIMUL} folder of the \code{FATE} simulation}

\item{years}{an \code{integer}, or a \code{vector} of \code{integer}, 
corresponding to the simulation year(s) that will be used to extract PFG 
relative abundance maps}

\item{mat.PFG.obs}{a \code{data.frame} with 4 columns : \code{PFG}, 
\code{X}, \code{Y}, \code{obs}}

\item{opt.ras_habitat}{(\emph{optional}) default \code{NULL}. \cr 
A \code{string} corresponding to the file name of a raster mask, with an 
\code{integer} value within each pixel, corresponding to a specific habitat}

\item{opt.doPlot}{(\emph{optional}) default \code{TRUE}. \cr If \code{TRUE}, 
plot(s) will be processed, otherwise only the calculation and reorganization 
of outputs will occur, be saved and returned}
}
\value{
A \code{list} containing one \code{data.frame} object with the 
following columns, and one \code{ggplot2} object :

\describe{
  \item{tab}{ 
    \describe{
      \item{\code{PFG}}{concerned plant functional group}
      \item{\code{AUC.sd}}{standard deviation of the AUC values}
      \item{\code{sensitivity.sd}}{standard deviation of the sensitivity 
      values}
      \item{\code{specificity.sd}}{standard deviation of the specificity 
      values}
      \item{\code{variable}}{name of the calculated statistic among 
      \code{sensitivity}, \code{specificity}, \code{TSS} and \code{AUC}}
      \item{\code{value}}{value of the corresponding statistic}
    }
  }
  \item{plot}{\code{ggplot2} object, representing the values for each PFG 
  of these four validation statistics (sensitivity, specificity, TSS, AUC) 
  \cr \cr}
}

One \file{POST_FATE_TABLE_YEAR_[...].csv} file is created : 
\describe{
  \item{\file{validationStatistics}}{containing the \code{data.frame} 
  detailed above}
}

One \file{POST_FATE_[...].pdf} file is created : 
\describe{
  \item{\file{GRAPHIC_B \cr validationStatistics}}{to assess the modeling 
  quality of each PFG based on given observations within the studied area}
}
}
\description{
This script is designed to produce a graphical representation 
of several statistics (sensitivity, specificity, TSS, AUC) for quality 
assessment for one (or several) specific \code{FATE} simulation year.
}
\details{
This function allows to obtain, for a specific \code{FATE} simulation and 
a specific parameter file within this simulation, \strong{PFG validation 
statistic values} and one preanalytical graphic. \cr \cr


Observation records (presences and absences) are required for each PFG 
within the \code{mat.PFG.obs} object :

\describe{
  \item{\code{PFG}}{the concerned plant functional group}
  \item{\code{X, Y}}{the coordinates of each observation, matching with the 
  projection of the mask of \code{name.simulation}}
  \item{\code{obs}}{either \code{0} or \code{1} to indicate presence or 
  absence \cr \cr}
}


For each PFG and each selected simulation year, raster maps are retrieved 
from the results folder \code{ABUND_REL_perPFG_allStrata} and lead to the 
production of one table :

\itemize{
  \item{the value of \strong{several statistics to evaluate the predictive 
  quality of the model} for each plant functional group \cr 
  (\code{\link[PresenceAbsence]{sensitivity}}, 
  \code{\link[PresenceAbsence]{specificity}}, 
  \code{\link[PresenceAbsence]{auc}}, 
  \code{TSS = sensitivity + specificity - 1})}
}

\strong{If a raster mask for habitat has been provided}, the values and 
graphics will be also calculated per habitat. \cr \cr

\strong{It requires} that the \code{\link{POST_FATE.relativeAbund}} 
function has been run and that the folder \code{ABUND_REL_perPFG_allStrata} 
exists. \cr \cr

\strong{This \code{.csv} file can then be used by other functions} :

\itemize{
  \item to produce maps of PFG \emph{presence / absence} from modelled 
  abundances \cr (see \code{\link{POST_FATE.binaryMaps}})
}
}
\examples{

                                                        
                                                        
## ----------------------------------------------------------------------------------------- ##
## Load example data

}
\seealso{
\code{\link{POST_FATE.relativeAbund}},
\code{\link[PresenceAbsence]{cmx}},
\code{\link[PresenceAbsence]{sensitivity}},
\code{\link[PresenceAbsence]{specificity}},
\code{\link[PresenceAbsence]{auc}},
\code{\link{.getCutoff}},
\code{\link{POST_FATE.binaryMaps}}
}
\author{
Maya Guéguen
}
\keyword{FATE,}
\keyword{area}
\keyword{curve,}
\keyword{outputs,}
\keyword{sensitivity,}
\keyword{skill}
\keyword{specificity,}
\keyword{statistic}
\keyword{true}
\keyword{under}
