% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POST_FATE.graphics.R
\name{POST_FATE.graphics}
\alias{POST_FATE.graphics}
\title{Create all possible graphical representations for a \code{FATE} 
simulation}
\usage{
POST_FATE.graphics(
  name.simulation,
  file.simulParam,
  years,
  no_years,
  opt.ras_habitat = NULL,
  opt.doFunc.evolutionCoverage = TRUE,
  opt.doFunc.evolutionPixels = TRUE,
  opt.cells_ID = NULL,
  opt.abund_fixedScale = TRUE,
  opt.doFunc.validation = TRUE,
  opt.mat.PFG.obs,
  opt.stratum = "all",
  opt.doFunc.mapPFGvsHS = TRUE,
  opt.doFunc.mapPFG = TRUE,
  opt.method,
  opt.method1.threshold = 0.05,
  opt.method2.cutoff = NULL,
  opt.stratum_min = 1,
  opt.stratum_max = 10,
  opt.doBinary = TRUE,
  opt.doPlot = TRUE,
  opt.no_CPU = 1
)
}
\arguments{
\item{name.simulation}{a \code{string} corresponding to the main directory 
or simulation name of the \code{FATE} simulation}

\item{file.simulParam}{default \code{NULL}. \cr A \code{string} 
corresponding to the name of a parameter file that will be contained into 
the \code{PARAM_SIMUL} folder of the \code{FATE} simulation}

\item{years}{an \code{integer}, or a \code{vector} of \code{integer}, 
corresponding to the simulation year(s) that will be used to extract PFG 
abundance maps}

\item{no_years}{an \code{integer} corresponding to the number of simulation 
years that will be used to extract PFG abundance / light / soil maps}

\item{opt.ras_habitat}{(\emph{optional}) default \code{NULL}. \cr 
A \code{string} corresponding to the file name of a raster mask, with an 
\code{integer} value within each pixel, corresponding to a specific habitat}

\item{opt.doFunc.evolutionCoverage}{default \code{TRUE} (\emph{optional}). 
If \code{TRUE}, corresponding \code{POST_FATE.graphic_...} function will 
be run.}

\item{opt.doFunc.evolutionPixels}{default \code{TRUE} (\emph{optional}). 
If \code{TRUE}, corresponding \code{POST_FATE.graphic_...} function will 
be run.}

\item{opt.cells_ID}{(\emph{optional}) default \code{NULL}. \cr The cells ID 
of the studied area for which PFG abundances will be extracted}

\item{opt.abund_fixedScale}{default \code{TRUE}. If \code{FALSE}, the 
ordinate scale will be adapted for each PFG for the graphical representation 
of the  evolution of abundances through time}

\item{opt.doFunc.validation}{default \code{TRUE} (\emph{optional}). 
If \code{TRUE}, corresponding \code{POST_FATE.graphic_...} function will 
be run.}

\item{opt.mat.PFG.obs}{a \code{data.frame} with 4 columns : \code{PFG}, 
\code{X}, \code{Y}, \code{obs}}

\item{opt.stratum}{(\emph{optional}) default \code{all}. \cr The stratum 
number from which to extract PFG binary maps}

\item{opt.doFunc.mapPFGvsHS}{default \code{TRUE} (\emph{optional}). 
If \code{TRUE}, corresponding \code{POST_FATE.graphic_...} function will 
be run.}

\item{opt.doFunc.mapPFG}{default \code{TRUE} (\emph{optional}). 
If \code{TRUE}, corresponding \code{POST_FATE.graphic_...} function will 
be run.}

\item{opt.method}{an \code{integer} to choose the transformation method : \cr 
\code{1} (relative abundance) or \code{2} (optimizing TSS) (see 
\code{\href{POST_FATE.binaryMaps#details}{Details}})}

\item{opt.method1.threshold}{default \code{0.05}. \cr If \code{method = 1}, 
minimum relative abundance required for each PFG to be considered as present 
in the concerned pixel}

\item{opt.method2.cutoff}{default \code{NULL}. \cr If \code{method = 2}, a 
\code{data.frame} with 3 columns : \code{year}, \code{PFG}, \code{cutoff} \cr
(see \code{\href{POST_FATE.binaryMaps#details}{Details}})}

\item{opt.stratum_min}{(\emph{optional}) default \code{1}. \cr An 
\code{integer} corresponding to the lowest stratum from which PFG 
abundances will be summed up}

\item{opt.stratum_max}{(\emph{optional}) default \code{10}. \cr An 
\code{integer} corresponding to the highest stratum from which PFG 
abundances will be summed up}

\item{opt.doBinary}{(\emph{optional}) default \code{TRUE}. \cr If 
\code{TRUE}, abundance maps (absolute or relative) are systematically 
multiplied by binary maps (see 
\code{\href{POST_FATE.graphic_mapPFG#details}{Details}})}

\item{opt.doPlot}{(\emph{optional}) default \code{TRUE}. \cr If \code{TRUE}, 
plot(s) will be processed, otherwise only the calculation and reorganization 
of outputs will occur, be saved and returned}

\item{opt.no_CPU}{(\emph{optional}) default \code{1}. \cr The number of 
resources that can be used to parallelize the \code{unzip/zip} of raster 
files, as well as the extraction of values from raster files}
}
\value{
The following \code{POST_FATE_[...].pdf} files are created : 
\describe{
  \item{\file{GRAPHIC_A \cr spaceOccupancy}}{to visualize for each PFG the 
  evolution of its occupation of the studied area through simulation time}
  \item{\file{GRAPHIC_A \cr abundance}}{to visualize for each PFG the 
  evolution of its abundance within the whole studied area through 
  simulation time}
  \item{\file{GRAPHIC_A \cr pixels}}{to visualize for each PFG the 
  evolution of its abundance within each selected pixel through 
  simulation time, as well as the evolution of light and soil resources}
  \item{\file{GRAPHIC_B \cr validationStatistics}}{to assess the modeling 
  quality of each PFG based on given observations within the studied area}
  \item{\file{GRAPHIC_B \cr PFGvsHS}}{to visualize the PFG presence 
  within the studied area (probability and simulated occurrence)}
  \item{\file{GRAPHIC_C \cr PFGrichness}}{to visualize the PFG richness 
  within the studied area}
  \item{\file{GRAPHIC_C \cr PFGcover}}{to visualize the PFG cover 
  within the studied area}
  \item{\file{GRAPHIC_C \cr PFGlight}}{to visualize the light CWM 
  within the studied area}
  \item{\file{GRAPHIC_C \cr PFGlsoil}}{to visualize the soil CWM 
  within the studied area}
}

Three folders are created :
\describe{
  \item{\file{ABUND_REL_perPFG \cr_allStrata}}{containing relative 
  abundance raster maps for each PFG across all strata (see 
  \code{\link{POST_FATE.relativeAbund}})}
  \item{\file{BIN_perPFG \cr_allStrata}}{containing presence / absence 
  raster maps for each PFG across all strata (see 
  \code{\link{POST_FATE.graphic_validationStatistics}})}
  \item{\file{BIN_perPFG \cr_perStrata}}{containing presence / absence 
  raster maps for each PFG for each stratum (see 
  \code{\link{POST_FATE.graphic_validationStatistics}})}
}
}
\description{
This script is designed to produce a set of graphical 
representations for a \code{FATE} simulation. Graphics can be of three 
types : 1) representing an evolution through time (of abundance, light, 
soil) ; 2) vizualising the goodness of the modelisation (presence/absence, 
validation statistics) : 3) or representing a spatial distribution for a 
specific year (richness, abundance, light, soil).
}
\details{
This function allows one to obtain, for a specific \code{FATE} 
simulation and a specific parameter file within this simulation, up to ten 
preanalytical graphics. \cr \cr

For each PFG and each selected simulation year, raster maps are retrieved 
from the results folders (\code{ABUND_perPFG_allStrata} or 
\code{ABUND_perPFG_allStrata} or \code{ABUND_REL_perPFG_allStrata} or 
\code{BIN_perPFG_perStrata} or \code{BIN_perPFG_allStrata} or \code{LIGHT} 
or \code{SOIL}) and unzipped.
Informations extracted lead to the production of the following graphics 
before the maps are compressed again :

\itemize{
  \item{the evolution of \strong{space occupancy} of each Plant Functional 
  Group through simulation time, with \emph{space occupancy} representing 
  the percentage of pixels within the mask of studied area where the PFG 
  is present (see \code{\link{POST_FATE.graphic_evolutionCoverage}})
  }
  \item{the evolution of \strong{abundance} of each Plant Functional Group 
  through simulation time, with \emph{abundance} being the sum over the 
  whole studied area of the PFG abundances (\code{FATE} \emph{arbitrary 
  unit}) (see \code{\link{POST_FATE.graphic_evolutionCoverage}})
  }
  \item{the evolution of \strong{abundance} of each Plant Functional Group 
  through simulation time, within 5 (or more) randomly selected pixels of 
  the studied area (\code{FATE} \emph{arbitrary unit}), as well as 
  \strong{light resources} of each stratum (\emph{1: Low, 2: Medium, 3: 
  High}) and \strong{soil resources} if these modules are selected (see 
  \code{\link{POST_FATE.graphic_evolutionPixels}})
  }
  \item{the value of \strong{several statistics for the predictive quality 
  of the model for each Plant Functional Group} and for each selected 
  simulation year(s) (see 
  \code{\link{POST_FATE.graphic_validationStatistics}})
  }
  \item{the maps of \strong{Plant Functional Group habitat suitability and 
  occurrences} for each selected simulation year(s), representing the 
  probability of presence of each PFG in each pixel compared to its 
  simulated presence (see \code{\link{POST_FATE.graphic_mapPFGvsHS}})
  }
  \item{the map of \strong{Plant Functional Group richness} for each 
  selected simulation year(s), representing the number of PFG present in 
  each pixel (see \code{\link{POST_FATE.graphic_mapPFG}})
  }
  \item{the map of \strong{Plant Functional Group cover} for each selected 
  simulation year(s), representing the relative cumulated abundance of PFG 
  present in each pixel above a height threshold (see 
  \code{\link{POST_FATE.graphic_mapPFG}})
  }
  \item{the map of \strong{light Community Weighted Mean} for each selected 
  simulation year(s), representing the simulated value of light within each 
  pixel above a height threshold (see 
  \code{\link{POST_FATE.graphic_mapPFG}})
  }
  \item{the map of \strong{soil Community Weighted Mean} for each selected 
  simulation year(s), representing the simulated value of soil within each 
  pixel (see \code{\link{POST_FATE.graphic_mapPFG}})
  }
}
}
\examples{

\dontrun{                      
POST_FATE.graphics(name.simulation = "FATE_simulation"
                                    , file.simulParam = "Simul_parameters_V1.txt"
                                    , year = 100
                                    , no.years = 10)
                                    
POST_FATE.graphics(name.simulation = "FATE_simulation"
                                    , file.simulParam = "Simul_parameters_V1.txt"
                                    , year = 100
                                    , no.years = 10
                                    , opt.abund_fixedScale = FALSE
                                    , opt.no_CPU = 4)
}
                                    
                                    

}
\seealso{
\code{\link{POST_FATE.temporalEvolution}}, 
\code{\link{POST_FATE.graphic_evolutionCoverage}}, 
\code{\link{POST_FATE.graphic_evolutionPixels}},
\code{\link{POST_FATE.relativeAbund}}, 
\code{\link{POST_FATE.binaryMaps}}, 
\code{\link{POST_FATE.graphic_validationStatistics}}, 
\code{\link{POST_FATE.graphic_mapPFGvsHS}}, 
\code{\link{POST_FATE.graphic_mapPFG}}
}
\author{
Maya Guéguen
}
\keyword{FATE,}
\keyword{abundance}
\keyword{outputs,}
\keyword{through}
\keyword{time}
