% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POST_FATE.relativeAbund.R
\name{POST_FATE.relativeAbund}
\alias{POST_FATE.relativeAbund}
\title{Create relative abundance maps for each Plant Functional Group for 
one (or several) specific year of a \code{FATE} simulation}
\usage{
POST_FATE.relativeAbund(
  name.simulation,
  file.simulParam = NULL,
  years,
  opt.no_CPU = 1
)
}
\arguments{
\item{name.simulation}{a \code{string} corresponding to the main directory 
or simulation name of the \code{FATE} simulation}

\item{file.simulParam}{default \code{NULL}. \cr A \code{string} 
corresponding to the name of a parameter file that will be contained into 
the \code{PARAM_SIMUL} folder of the \code{FATE} simulation}

\item{years}{an \code{integer}, or a \code{vector} of \code{integer}, 
corresponding to the simulation year(s) that will be used to extract PFG 
abundance maps}

\item{opt.no_CPU}{(\emph{optional}) default \code{1}. \cr The number of 
resources that can be used to parallelize the \code{unzip/zip} of raster 
files}
}
\value{
One result folder is created :
\describe{
  \item{\file{ABUND_REL_perPFG \cr_allStrata}}{containing relative 
  abundance raster maps for each PFG across all strata}
}
}
\description{
This script is designed to produce raster maps of PFG 
simulated relative abundances for one (or several) specific \code{FATE} 
simulation year.
}
\details{
This function allows to obtain, for a specific \code{FATE} simulation and 
a specific parameter file within this simulation, \strong{raster maps of PFG 
relative abundance}. \cr \cr

For each PFG and each selected simulation year, raster maps are retrieved 
from the results folder \code{ABUND_perPFG_allStrata} and unzipped. 
Informations extracted lead to the production of the same number of raster 
before the maps are compressed again :

\itemize{
  \item{for each selected simulation year(s), \strong{relative abundances} 
  for all strata combined are calculated :
  \deqn{\frac{abund_{\text{ PFG}_i\text{, }\text{Stratum}_{all}}}
  {abund_{\text{ PFG}_{all}\text{, }\text{Stratum}_{all}}}} \cr \cr
  }
}

\strong{These \code{raster} files can then be used by other functions} :

\itemize{
  \item to produce \emph{presence/absence maps} and \emph{validation 
  statistics}, and associated graphics \cr (see 
  \code{\link{POST_FATE.graphic_validationStatistics}})
}
}
\examples{

\dontrun{                      
POST_FATE.relativeAbund(name.simulation = "FATE_simulation"
                        , file.simulParam = "Simul_parameters_V1.txt"
                        , years = 850
                        , opt.no_CPU = 1)
                                    
POST_FATE.relativeAbund(name.simulation = "FATE_simulation"
                        , file.simulParam = "Simul_parameters_V1.txt"
                        , years = c(850, 950)
                        , opt.no_CPU = 1)
}
                                                        
                                                        
                                                        
## ----------------------------------------------------------------------------------------- ##
## Load example data
                                    

}
\seealso{
\code{\link{POST_FATE.graphic_validationStatistics}}
}
\author{
Maya Guéguen
}
\keyword{FATE,}
\keyword{abundance}
\keyword{outputs,}
\keyword{relative}
