% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POST_FATE.validation.R
\name{POST_FATE.validation}
\alias{POST_FATE.validation}
\title{Compute validation data for habitat, PFG richness and composition for a 
\code{FATE} simulation.}
\usage{
POST_FATE.validation(
  name.simulation,
  file.simulParam,
  year,
  mat.obs,
  mat.hab,
  ras.habitat,
  doRichness = TRUE,
  doComposition = TRUE,
  doHabitat = TRUE,
  RF.seed = 123,
  RF.training = 0.7,
  doHabitat.allMap = FALSE,
  opt.ras_validation = NULL,
  opt.keep_PFG = NULL,
  opt.keep_strata = NULL
)
}
\arguments{
\item{name.simulation}{a \code{string} corresponding to the main directory 
or simulation name of the \code{FATE} simulation}

\item{file.simulParam}{default \code{NULL}. \cr A \code{string} 
corresponding to the name of a parameter file that will be contained into 
the \code{PARAM_SIMUL} folder of the \code{FATE} simulation}

\item{year}{an \code{integer} corresponding to the simulation year that will 
be used to extract PFG abundance table}

\item{mat.obs}{a \code{data.frame} with at least 5 columns : \cr 
\code{site}, \code{x}, \code{y}, \code{PFG}, \code{abund} 
\cr (\emph{and optionally, \code{strata}, \code{code.habitat}}) 
(see \href{POST_FATE.validation#details}{\code{Details}})}

\item{mat.hab}{a \code{data.frame} with 2 columns : \code{code.habitat}, 
\code{habitat}}

\item{ras.habitat}{a \code{string} corresponding to the file name of a 
raster mask, with an \code{integer} value within each pixel, corresponding 
to a specific habitat}

\item{doRichness}{(\code{logical}) default \code{TRUE}. \cr If \code{TRUE}, 
difference in observed and simulated PFG richness will be run.}

\item{doComposition}{(\code{logical}) default \code{TRUE}. \cr If \code{TRUE}, 
difference in observed and simulated PFG distribution over habitat will be run.}

\item{doHabitat}{(\code{logical}) default \code{TRUE}. \cr If \code{TRUE}, 
a random forest model to predict habitat from PFG abundances will be run.}

\item{RF.seed}{default \code{123}. \cr An \code{integer} to be given to 
\code{\link[base]{set.seed}} function, in order to fix the produced results 
if needed, as dataset will be divided randomly into training and testing 
datasets for the random forest model}

\item{RF.training}{default \code{0.7}. \cr A \code{numeric} between \code{0} 
and \code{1} corresponding to the percentage of data that will be used as 
training dataset in the random forest model}

\item{doHabitat.allMap}{(\code{logical}) default \code{TRUE}. \cr If 
\code{TRUE}, habitat prediction over the whole simulation map will be run.}

\item{opt.ras_validation}{(\emph{optional}) default \code{NULL}. \cr 
A \code{string} corresponding to the file name of a raster mask, with either 
\code{0} or \code{1} within each pixel, \code{1} corresponding to the cells 
of the studied area in which the validation will take place}

\item{opt.keep_PFG}{(\emph{optional}) default \code{NULL}. \cr 
A \code{vector} of \code{character} corresponding to the names of the PFG to 
keep for the validation}

\item{opt.keep_strata}{(\emph{optional}) default \code{NULL}. \cr 
A \code{list} with names corresponding to the strata to keep for the 
validation, and each element containing correspondence with \code{FATE} 
strata definition}
}
\value{
A \code{list} containing two to ten elements depending on the 
options selected :
\describe{
  \item{PFG richness}{three \code{vector} objects :
  \describe{
    \item{\code{rich.obs}}{\code{vector} containing names of PFG in observed 
    data}
    \item{\code{rich.sim}}{\code{vector} containing names of PFG in 
    simulated data}
    \item{\code{rich.diff}}{\code{vector} containing names of PFG in 
    observed but not in simulated data}
  }
  }
  \item{PFG composition}{
  \describe{
    \item{\code{compo.distrib}}{\code{data.frame} object with the following 
    columns :
    \describe{
      \item{\code{PFG}}{concerned plant functional group}
      \item{\code{code.habitat}}{concerned habitat code}
      \item{\code{strata}}{concerned height stratum}
      \item{\code{quantile.perc}}{concerned quantile}
      \item{\code{quantile.obs}}{corresponding observed abundance}
      \item{\code{quantile.sim}}{corresponding simulated abundance}
    }
    }
    \item{\code{compo.proximity}}{\code{data.frame} object with the 
    following columns :
    \describe{
      \item{\code{PFG}}{concerned plant functional group}
      \item{\code{code.habitat}}{concerned habitat code}
      \item{\code{strata}}{concerned height stratum}
      \item{\code{proximity}}{composition similarity between each 
      habitat/strata combination}
    }
    }
  }
  }
  \item{Habitat}{
  \describe{
    \item{\code{hab.RF.model}}{\code{randomForest} object obtained from 
    \code{\link[randomForest]{randomForest}} function}
    \item{\code{hab.RF.perf}}{\code{data.frame} object with the following 
    columns :
    \describe{
      \item{\code{dataset}}{\code{train}, \code{test} or \code{valid}}
      \item{\code{habitat}}{concerned habitat code}
      \item{\code{sensitivity}}{true positive rate}
      \item{\code{specificity}}{true negative rate}
      \item{\code{weight}}{share of each habitat in the observed habitat 
      distribution}
      \item{\code{TSS}}{mean True Skill Statistic}
      \item{\code{TSSw}}{mean weighted True Skill Statistic}
    }
    }
    \item{\code{hab.tab.pred}}{\code{data.frame} object with the following columns :
    \describe{
      \item{\code{pixel}}{concerned pixel}
      \item{\code{x}}{corresponding x-coordinate}
      \item{\code{y}}{corresponding y-coordinate}
      \item{\code{code.habitat}}{concerned habitat code}
      \item{\code{habitat.obs}}{observed habitat}
      \item{\code{habitat.sim}}{simulated habitat through random forest model}
      \item{\code{habitat.final}}{simulated habitat through random forest 
      model, with wrong predictions indicated as \code{failure}}
      \item{\code{fail_succ}}{wether or not the simulated habitat is the 
      same as the observed}
      \item{\code{color}}{corresponding color code}
    }
    }
    \item{\code{hab.ras.pred}}{\code{raster} object containing habitat 
    predictions}
    \item{\code{hab.plot}}{\code{ggplot2} object, representing 
    \code{hab.ras.pred} raster}
  }
  }
}


One to four \file{POST_FATE_TABLE_[...].csv} files are created : 
\describe{
  \item{\file{HAB_validation_compo_distribution_}}{\emph{if composition 
  module was activated}, containing \code{compo.distrib}}
  \item{\file{HAB_validation_compo_proximity_}}{\emph{if composition module 
  was activated}, containing \code{compo.proximity}}
  \item{\file{HAB_validation_RF_performance_}}{\emph{if habitat module was 
  activated}, containing \code{hab.RF.perf}}
  \item{\file{PIXEL_validation_RF_prediction_}}{\emph{if habitat module was 
  activated}, containing \code{hab.tab.pred}}
}

\file{HabitatPrediction_YEAR_[...].tif} and 
\file{POST_FATE_GRAPHIC_D_map_habitat_[...].pdf} files are created 
containing random forest predicted habitat raster and plot respectively.
}
\description{
This script is designed to compare observed and simulated PFG 
distribution for one specific \code{FATE} simulation year by computing 
a) difference in PFG richness ; 
b) similarity between distribution quantiles ; 
c) a random forest model to predict habitat classes.
}
\details{
This function allows to obtain, for a specific \code{FATE} simulation and 
a specific parameter file within this simulation, \strong{PFG validation 
habitat model and predictions} and one preanalytical graphic. \cr \cr

\describe{
  \item{PFG richness}{
  check if all simulated PFG abundances are superior to 0 within the 
  simulated area (\emph{only a subset of PFG can be examined with the 
  \code{opt.keep_PFG} parameter}).
  }
  
  \item{PFG composition}{
  for each PFG and each habitat / stratum combination, abundances for each 
  quartiles (\code{25}, \code{50}, \code{75} and \code{100} \%) is 
  calculated for both observed and simulated distributions. \cr
  (\emph{Only a subset of PFG and strata can be examined with the 
  \code{opt.keep_PFG} and \code{opt.keep_strata} parameters.}) \cr
  
  Then, a composition similarity between each PFG / habitat / strata 
  combination is calculated as a pseudo-distance between observed and 
  simulated quartiles \code{Q} :
  \deqn{S_{\text{ Habitat}_j \text{, }\text{Stratum}_k} = 
  \sum S_{\text{ PFG}_i \text{, }\text{Habitat}_j \text{, }\text{Stratum}_k}}
  with
  \deqn{S_{\text{ PFG}_i \text{, }\text{Habitat}_j \text{, }\text{Stratum}_k} 
  = 1 - \frac{\text{1}}{4} * \sum abs(Q_{\text{ q} \text{, }sim} - Q_{\text{ q} \text{, }obs})}
  with \code{q} varying from 1 to 4.
  }

  \item{Habitat validation}{
  a \code{\link[randomForest]{randomForest}} model is built to predict 
  habitat (provided through \code{ras.habitat}) in function of observed PFG 
  abundances. Only habitat code in \code{mat.hab} are effectively used and 
  predicted. TSS (True Skill Statistic) is computed between observed and 
  simulated habitat, and averaged over each habitat. TSS can also be 
  weighted by the share of each habitat in the observed habitat distribution.
  }
}
}
\examples{

library(raster)

## Create a simulation folder
PRE_FATE.skeletonDirectory(name.simulation = "FATE_Champsaur")

## Load example data
Champsaur_params = .loadData("Champsaur_params", "RData")
.loadData("Champsaur_results_V1", "7z")

## Please extract results files in the 'FATE_Champsaur/RESULTS' folder

## Define a vector to choose habitats taken into account
mat.hab = data.frame(ID = c(6, 5, 7, 8)
                     , habitat = c("coniferous.forest"
                                   , "deciduous.forest"
                                   , "natural.grassland"
                                   , "woody.heatland"))

## Habitat & validation maps
ras_simulation = Champsaur_params$stk.mask$Champsaur
ras.habitat = Champsaur_params$stk.mask$habitat
ras.habitat = projectRaster(from = ras.habitat, to = ras_simulation, method = "ngb")
# writeRaster(ras_simulation, filename = "FATE_Champsaur/DATA/MASK/MASK_Champsaur.tif")

## Observed data
mat.obs = Champsaur_params$tab.releves

## Transform observed PFG abundances into relative abundances
mat.obs$abund = PRE_FATE.abundBraunBlanquet(mat.obs$abund) / 100
mat.obs = aggregate(abund ~ site + PFG + strata + x + y
                    , data = mat.obs, FUN = "sum")
                        
## Create Global and Simulation parameters
PRE_FATE.params_globalParameters(name.simulation = "FATE_Champsaur"
                                 , opt.saving_abund_PFG_stratum = TRUE
                                 , opt.saving_abund_PFG = TRUE
                                 , opt.saving_abund_stratum = FALSE
                                 , required.no_PFG = 15
                                 , required.no_strata = 7
                                 , required.simul_duration = 2000
                                 , required.seeding_duration = 1000
                                 , required.seeding_timestep = 1
                                 , required.seeding_input = 100
                                 , required.potential_fecundity = 1
                                 , required.max_abund_low = 1000
                                 , required.max_abund_medium = 2000
                                 , required.max_abund_high = 3000
                                 , doDispersal = TRUE
                                 , DISPERSAL.mode = 1
                                 , DISPERSAL.saving = FALSE
                                 , doHabSuitability = TRUE
                                 , HABSUIT.mode = 1)
                                
PRE_FATE.params_simulParameters(name.simulation = "FATE_Champsaur"
                                , name.MASK = "MASK_Champsaur.tif")

simul.param = "Simul_parameters_V1.txt"
# simul.param = paste0("FATE_Champsaur/PARAM_SIMUL/", simul.param)

POST_FATE.validation(name.simulation = "FATE_Champsaur"
                     , file.simulParam = simul.param
                     , year = 2000
                     , mat.obs = mat.obs
                     , mat.hab = mat.hab
                     , ras.habitat = ras.habitat
                     , doHabitat = TRUE
                     , doHabitat.allMap = TRUE
                     , doComposition = TRUE
                     , doRichness = TRUE)


}
\author{
Matthieu Combaud, Maxime Delprat, Maya Guéguen
}
