% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRE_FATE.abundBraunBlanquet.R
\name{PRE_FATE.abundBraunBlanquet}
\alias{PRE_FATE.abundBraunBlanquet}
\title{Transform Braun-Blanquet abundances into relative abundances}
\usage{
PRE_FATE.abundBraunBlanquet(abund)
}
\arguments{
\item{abund}{a vector with abundance values from Braun-Blanquet
(+, r, 1, 2, 3, 4, 5) with NA when no information}
}
\value{
a vector with numerical values between 0 and 100 corresponding to the median of each recovery class.
}
\description{
This script is designed to transform Braun-Blanquet abundance
information into relative abundances or average recovery values
(between 0 and 100).
}
\details{
Braun-Blanquet values allow to estimate the abundance-dominance of a species
based on ans estimation of the number of individuals and the covering surface.
A correspondence has been defined between this index and average recovery values :

\tabular{crr}{
  Braun-Blanquet \tab Recovery class (\%) \tab Average recovery (\%) \cr
  + \tab <1 \tab 0.5 \cr
  1 \tab 1-5 \tab 2.5 \cr
  2 \tab 5-25 \tab 15 \cr
  3 \tab 25-50 \tab 37.5 \cr
  4 \tab 50-75 \tab 62.5 \cr
  5 \tab 75-100 \tab 87.5
}

Braun-Blanquet J., Roussine N. & Nègre R., 1952. Les groupements végétaux de la France méditerranéenne.
Dir. Carte Group. Vég. Afr. Nord , CNRS, 292 p.

Baudière A. & Serve L., 1975. Les groupements végétaux du Plade Gorra-Blanc (massif du Puigmal, Pyrénées Orientales).
Essai d'interprétation phytosociologique et phytogéographique. Nat. Monsp., sér. Bot., 25, 5-21.

Foucault B. (de), 1980. Les prairies du bocage virois (Basse-Normandie, France). Typologie phytosociologique et 
essai de reconstitution des séries évolutives herbagères. Doc. Phytosoc., N.S., 5, 1-109.
}
\examples{

## Load example data
data(MontBlanc)

}
\author{
Maya Guéguen
}
\keyword{Braun-Blanquet}
\keyword{abundance,}
