% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRE_FATE.params_globalParameters.R
\name{PRE_FATE.params_globalParameters}
\alias{PRE_FATE.params_globalParameters}
\title{Create \emph{Global_parameters} parameter file for a \code{FATE-HD}
simulation}
\usage{
PRE_FATE.params_globalParameters(name.simulation, opt.nb_CPU = 1,
  required.succ_option, required.nb_PFG, required.nb_STRATA, required.hs_option,
  required.seeding_timestep, required.seeding_duration, required.simul_duration,
  doDisturbances = FALSE, DIST.nb, DIST.nb_sub, DIST.freq, doSoil = FALSE,
  doFire = FALSE, doDrought = FALSE, doHabStability = FALSE,
  doAliens = FALSE)
}
\arguments{
\item{name.simulation}{a \code{string} that corresponds to the main directory
or simulation name of the \code{FATE-HD} simulation}

\item{opt.nb_CPU}{default 1 (\emph{optional}). The number of resources that can be used 
to parallelize the \code{FATE-HD} simulation}

\item{required.succ_option}{a \code{string} to choose if the simulation will
include habitat suitability (\emph{fateh}) or not (\emph{fate})}

\item{required.nb_PFG}{an \code{integer} corresponding to the number of PFG}

\item{required.nb_STRATA}{an \code{integer} corresponding to the number of 
height strata}

\item{required.hs_option}{an \code{integer} corresponding to the way of 
simulating the habitat suitability variation between years for each PFG, 
either random (1) or PFG specific (2)}

\item{required.seeding_timestep}{an \code{integer} corresponding to the 
time interval at which occurs the seeding, and until the seeding duration
 is not over (\emph{in years})}

\item{required.seeding_duration}{an \code{integer} corresponding to the 
duration of seeding (\emph{in years})}

\item{required.simul_duration}{an \code{integer} corresponding to the 
duration of simulation (\emph{in years})}

\item{doDisturbances}{default \code{FALSE}. If \code{TRUE}, disturbances 
are applied in the \code{FATE-HD} simulation, and associated parameters 
are required}

\item{DIST.nb}{the number of disturbances}

\item{DIST.nb_sub}{the number of way a PFG should react to a disturbance}

\item{DIST.freq}{the frequency of each disturbance (\emph{in years})}
}
\value{
A \code{.txt} file into the \code{name.simulation/DATA/GLOBAL_PARAMETERS}
directory with the following parameters :

\itemize{
  \item NB_CPUS
  \item SUCC_MOD
  \item NB_FG
  \item NB_STRATUM
  \item ENVSUIT_OPTION
  \item SEEDING_TIMESTEP
  \item SEEDING_DURATION
  \item SIMULATION_TIME
}

If the simulation includes disturbances :

\itemize{
  \item DO_DISTURBANCES
  \item NB_DISTURBANCES
  \item NB_SUBDISTURBANCES
  \item FREQ_DISTURBANCES
}

If the simulation includes soil competition :

\itemize{
  \item DO_SOIL_COMPETITION
  \item SOIL_DEFAULT_VALUE
  \item NB_SOIL_CATEGORIES
  \item SOIL_CATEGORIES_TRESHOLDS
}

If the simulation includes drought disturbances :

\itemize{
  \item DO_DROUGHT_DISTURBANCES
  \item NB_SUBDROUGHT
  \item CHRONO_POST_DROUGHT
  \item CHRONO_CURR_DROUGHT
}

If the simulation includes habitat stability check :

\itemize{
  \item DO_HAB_STABILITY
  \item NB_HABITATS
}

If the simulation includes aliens introduction :

\itemize{
  \item DO_ALIENS_DISTURBANCE
  \item FREQ_ALIENS
}
}
\description{
This script is designed to create one parameter file
containing \code{GLOBAL PARAMETERS} used in \code{FATE-HD} model.
}
\details{
\describe{
  \item{To get abundances \cr per PFG per pixel}{
  \itemize{
    \item per strata
    \item for all strata
  }
  \deqn{abund = 10 000 * totalAbund / \text{GLOBAL HIGH ABUND}}
  }
  \item{To transform PFG \cr abundances into \cr light resources}{
    \deqn{abund < \text{GLOBAL MEDIUM RESOURCES THRESH} \Leftrightarrow Light = High}
    \deqn{abund > \text{GLOBAL MEDIUM RESOURCES THRESH } \& \\
    abund < \text{GLOBAL LOW RESOURCES THRESH} \Leftrightarrow Light = Medium}
    \deqn{abund > \text{GLOBAL LOW RESOURCES THRESH} \Leftrightarrow Light = Low}
  }
}
}
\examples{

## Create a skeleton folder with the default name ('FATE_simulation')
PRE_FATE.skeletonDirectory()

## Create a Global_parameters file
PRE_FATE.params_globalParameters(name.simulation = "FATE_simulation"
                                 , required.succ_option = "fateh"
                                 , required.nb_PFG = 3
                                 , required.nb_STRATA = 5
                                 , required.hs_option = 1
                                 , required.seeding_timestep = 1
                                 , required.seeding_duration = 10
                                 , required.simul_duration = 100)
                                   
## Create SEVERAL Global_parameters files
PRE_FATE.params_globalParameters(name.simulation = "FATE_simulation"
                                 , required.succ_option = "fateh"
                                 , required.nb_PFG = 3
                                 , required.nb_STRATA = 5
                                 , required.hs_option = c(1,2)
                                 , required.seeding_timestep = 1
                                 , required.seeding_duration = 10
                                 , required.simul_duration = 100)

}
\author{
Maya Guéguen
}
