% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRE_FATE.params_multipleSet.R
\name{PRE_FATE.params_multipleSet}
\alias{PRE_FATE.params_multipleSet}
\title{Create multiple set(s) of parameter files for a \code{FATE} 
simulation}
\usage{
PRE_FATE.params_multipleSet(
  name.simulation.1,
  name.simulation.2 = NULL,
  file.simulParam.1,
  file.simulParam.2 = NULL,
  no_simulations,
  opt.folder.name = "FATE_simulation_MULTIPLE_SET",
  opt.seed = NULL,
  opt.percent_seeding = 0.5,
  opt.percent_light = 0.5,
  opt.percent_soil = 0.5,
  do.seeding_duration = TRUE,
  do.seeding_timestep = TRUE,
  do.seeding_input = TRUE,
  do.potential_fecundity = TRUE,
  do.no_strata = TRUE,
  do.LIGHT.thresh_medium = TRUE,
  do.LIGHT.thresh_low = TRUE,
  do.SOIL.init = TRUE,
  do.SOIL.retention = TRUE,
  do.DISPERSAL.mode = TRUE,
  do.HABSUIT.mode = TRUE
)
}
\arguments{
\item{name.simulation.1}{a \code{string} corresponding to the main directory 
or simulation name of the \code{FATE} simulation from which to 
retrieve the first parameter simulation file (\code{file.simulParam.1}), 
and the second if given (\code{file.simulParam.2}) and no other directory 
provided (\code{name.simulation.2 = NULL})}

\item{name.simulation.2}{(\emph{optional}) default \code{NULL}. \cr 
A \code{string} corresponding to the main directory or simulation name of 
the \code{FATE} simulation from which to retrieve the second parameter 
simulation file (\code{file.simulParam.2})}

\item{file.simulParam.1}{a \code{string} corresponding to the name of the 
simulation parameter file from which to retrieve parameter values that will 
be used to build the multiple set of new parameters}

\item{file.simulParam.2}{(\emph{optional}) default \code{NULL}. \cr 
A \code{string} corresponding to the name of the second simulation parameter 
file from which to retrieve parameter values that will be used to build 
parameter ranges in comparison with values from \code{file.simulParam.1}}

\item{no_simulations}{an \code{integer} corresponding to the number of set 
of parameters that will be produced according to Latin Hypercube Sampling 
(LHS)}

\item{opt.folder.name}{(\emph{optional}) default 
\code{FATE_simulation_MULTIPLE_SET}. \cr A \code{string} corresponding 
to the name of the folder that will be created to store the results}

\item{opt.seed}{(\emph{optional}) default \code{NULL}. \cr
An \code{integer} to be given to \code{\link[base]{set.seed}} function, in 
order to fix the produced results if needed, as 
\href{https://www.rdocumentation.org/packages/SPOT/versions/2.0.3/topics/designLHD}{\code{designLHD}} 
is also a random value generator}

\item{opt.percent_seeding}{default \code{0.5}. Amount of variation (between 
\code{0} and \code{1}) around the original value of \code{SEEDING_DURATION}, 
\code{SEEDING_TIMESTEP}, \code{SEEDING_INPUT}, \code{POTENTIAL_FECUNDITY} 
if selected}

\item{opt.percent_light}{default \code{0.5}. Amount of variation (between 
\code{0} and \code{1}) around the original value of 
\code{LIGHT_THRESH_MEDIUM}, \code{LIGHT_THRESH_LOW} if selected}

\item{opt.percent_soil}{default \code{0.5}. Amount of variation (between 
\code{0} and \code{1}) around the original value of \code{SOIL_INIT}, 
\code{SOIL_RETENTION} if selected}

\item{do.seeding_duration}{default \code{TRUE}. If \code{TRUE}, 
\code{SEEDING_DURATION} parameter within \emph{Global_parameters} file will 
be declined into a range of values}

\item{do.seeding_timestep}{default \code{TRUE}. If \code{TRUE}, 
\code{SEEDING_TIMESTEP} parameter within \emph{Global_parameters} file will 
be declined into a range of values}

\item{do.seeding_input}{default \code{TRUE}. If \code{TRUE}, 
\code{SEEDING_INPUT} parameter within \emph{Global_parameters} file will be 
declined into a range of values}

\item{do.potential_fecundity}{default \code{TRUE}. If \code{TRUE}, 
\code{POTENTIAL_FECUNDITY} parameter within \emph{Global_parameters} file 
will be declined into a range of values}

\item{do.no_strata}{default \code{TRUE}. If \code{TRUE}, \code{NO_STRATA} 
parameter within \emph{Global_parameters} file will be declined into a range 
of values, with potential impact on some parameters within PFG succession 
(and light) files (parameters \code{STRATA}, \code{MAX_ABUNDANCE}, 
\code{IMM_SIZE}, \code{CHANG_STR_AGES} (and \code{LIGHT_TOL}), see 
\code{\link{PRE_FATE.params_PFGsuccession}} (and 
\code{\link{PRE_FATE.params_PFGlight}}))}

\item{do.LIGHT.thresh_medium}{default \code{TRUE}. If \code{TRUE}, 
\code{LIGHT_THRESH_MEDIUM} parameter within \emph{Global_parameters} file 
will be declined into a range of values}

\item{do.LIGHT.thresh_low}{default \code{TRUE}. If \code{TRUE}, 
\code{LIGHT_THRESH_LOW} parameter within \emph{Global_parameters} file 
will be declined into a range of values}

\item{do.SOIL.init}{default \code{TRUE}. If \code{TRUE}, \code{SOIL_INIT} 
parameter within \emph{Global_parameters} file will be declined into a range 
of values}

\item{do.SOIL.retention}{default \code{TRUE}. If \code{TRUE}, 
\code{SOIL_RETENTION} parameter within \emph{Global_parameters} file will be 
declined into a range of values}

\item{do.DISPERSAL.mode}{default \code{TRUE}. If \code{TRUE}, 
\code{DISPERSAL_MODE} parameter within \emph{Global_parameters} file will be 
declined into its three possible values (either either packets kernel 
(\code{1}), exponential kernel (\code{2}) or exponential kernel with 
probability (\code{3}), see \code{\link{PRE_FATE.params_globalParameters}})}

\item{do.HABSUIT.mode}{default \code{TRUE}. If \code{TRUE}, 
\code{HABSUIT_MODE} parameter within \emph{Global_parameters} file will be 
declined into its two possible values (either random (\code{1}) or PFG 
specific (\code{2}), see \code{\link{PRE_FATE.params_globalParameters}})}
}
\value{
A new folder containing the different sets of parameters asked. 

Depending on what elements have been asked to be varied, three types of 
files can have been modified :
\enumerate{
  \item the global parameter file
  \item the PFG succession files
  \item the PFG light succession files \cr \cr
}

Below are listed the parameters that can change (if selected) within each 
file :

\itemize{
  \item Into the \code{name.simulation/DATA/GLOBAL_PARAMETERS} folder :
  \itemize{
    \item NO_STRATA
    \item SEEDING_DURATION
    \item SEEDING_TIMESTEP
    \item SEEDING_INPUT
    \item POTENTIAL_FECUNDITY \cr \cr
  }
  If the simulation includes \emph{light interaction} :
  \itemize{
    \item LIGHT_THRESH_MEDIUM
    \item LIGHT_THRESH_LOW
  }
  If the simulation includes \emph{soil interaction} :
  \itemize{
    \item SOIL_INIT
    \item SOIL_RETENTION
  }
  If the simulation includes \emph{dispersal} :
  \itemize{
    \item DISPERSAL_MODE
  }
  If the simulation includes \emph{habitat suitability} :
  \itemize{
    \item HABSUIT_MODE \cr \cr
  }
  \item Into the \code{name.simulation/DATA/PFGS/SUCC} folder :
  \itemize{
    \item STRATA
    \item MAX_ABUNDANCE
    \item IMM_SIZE
    \item CHANG_STR_AGES \cr \cr
  }
  \item Into the \code{name.simulation/DATA/PFGS/LIGHT} folder :
  \itemize{
    \item LIGHT_TOL
  }
}
}
\description{
This script is designed to create multiple sets of parameters 
using Latin Hypercube Sampling to help find best combination of parameters 
(see \code{Details})
}
\details{
A \code{FATE} simulation requires several parameters to define general 
characteristics of the simulation : they are saved within a 
\emph{Global_parameters} file (see 
\code{\link{PRE_FATE.params_globalParameters}}). To fit the model to a 
particular area and set of Plant Functional Groups (PFG), these are the 
parameters that should be optimized, since they are not data-dependant, 
unlike, for example, parameters related to PFG (height, maturity, dispersal 
distances, soil tolerance, etc).

(\emph{Note : this is true, except when varying the number of strata, which 
will have an impact on some parameters within SUCC and LIGHT PFG parameter 
files.})

\strong{The main idea is to start from a complete simulation folder, to 
select the parameters that should vary, and to create new parameter files 
with new parameter values based on pre-existing values. \cr \cr}

Three possible scenarios are available :
\describe{
  \item{1 folder - \cr 1 simulation file}{
  \itemize{
    \item requested parameter values are extracted from the given 
    simulation file
    \item ranges are assigned to each parameter according to the specified 
    value \cr e.g. : if \code{opt.percent_seeding = 0.5}, and 
    \code{do.seeding_duration} is asked, values will be generated for this 
    parameter between : 
    \deqn{\text{SEEDING_DURATION} \pm \text{SEEDING_DURATION} * 
    \frac{50}{100}}
    \item according to the required number of parameter sets to be produced 
    (\code{no_simulations}), Latin Hypercube Sampling is applied to select 
    each new parameter values
    \item parameter files are created for these new parameter values \cr \cr
  }
  }
  \item{1 folder - \cr 2 simulation files}{
  \itemize{
    \item same as 1st scenario
    \item ranges assigned to each parameter correspond to the extracted 
    values (e.g. : if do.seeding_duration is asked, values will be 
    generated for this parameter between : \code{SEEDING_DURATION 
    (file_simulation.1)} and \code{SEEDING_DURATION (file_simulation.2)} 
    \cr \cr
  }
  }
  \item{2 folders - \cr 2 simulation files}{
  \itemize{
    \item same as 2nd scenario, except that the two given simulation files 
    come from two different simulation folders \cr \cr
  }
  }
}

\strong{Latin Hypercube Sampling} is a statistical method to generate a 
sampling of parameter values from a multidimensional space, while ensuring 
a good representation of the real variability. The range of each parameter 
is known, and depending on the number of set of parameters asked to be 
obtained at the end, each range is more or less finely cut and values are 
drawn in order to explore the whole space of combinations.
}
\seealso{
\href{https://www.rdocumentation.org/packages/SPOT/versions/2.0.3/topics/designLHD}{\code{designLHD}},
\code{\link{PRE_FATE.skeletonDirectory}},
\code{\link{PRE_FATE.params_globalParameters}},
\code{\link{PRE_FATE.params_PFGsuccession}},
\code{\link{PRE_FATE.params_PFGlight}},
}
\author{
Billur Bektas, Maya Guéguen
}
\keyword{FATE,}
\keyword{Hypercube}
\keyword{Latin}
\keyword{Sampling}
\keyword{simulation,}
