% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRE_FATE.params_PFGdisturbance.R
\name{PRE_FATE.params_PFGdisturbance}
\alias{PRE_FATE.params_PFGdisturbance}
\title{Create \emph{DISTURBANCE} parameter files for a \code{FATE-HD}
simulation}
\usage{
PRE_FATE.params_PFGdisturbance(name.simulation, mat.PFG.dist,
  mat.PFG.succ = paste0(name.simulation,
  "/DATA/PFGS/SUCC_COMPLETE_TABLE.csv"))
}
\arguments{
\item{name.simulation}{a \code{string} that corresponds to the main directory
or simulation name of the \code{FATE-HD} simulation}

\item{mat.PFG.dist}{a \code{data.frame} with at least 4 columns : \cr name, 
responseStage, KilledIndiv_[...], ResproutIndiv_[...] (see \code{Details})}

\item{mat.PFG.succ}{a \code{data.frame} with at least 6 columns : \cr
PFG, type, MATURITY, LONGEVITY, STRATA and CHANG_STR_AGES_to_str_3. \cr
Such an object can be obtained with the \code{PRE_FATE.params_PFGsuccession}
 function.}
}
\value{
A \code{.txt} file per PFG into the \code{name.simulation/DATA/PFGS/DIST/}
directory with the following parameters :

\itemize{
  \item PROP_KILLED : the proportion of propagules killed by each disturbance \cr
  \emph{(0: 0\% 1: 10\% 2: 20\% 3: 30\% 4: 40\% 5: 50\% 6: 60\% 7: 70\% 8: 80\% 9: 
  90\% 10: 100\%)}
  \item BREAK_AGE : the age when each PFG changes of response stage \emph{(in years)}
  \item RESPR_AGE : the PFG resprouting age table (in a single row) \cr
  This is a vector of \code{no.DIST * no.responseStages} numbers \emph{(in years)}
  corresponding to the age at which each living PFG can be rejuvenated (younger 
  than the actual one) :
  \itemize{
    \item at different response stages \emph{(!currently 4, not yet editable!)}
    \item for each disturbance.
  }
  \item FATES : the PFG disturbance response table (in a single row) \cr
  This is a vector of \code{no.DIST * no.responseStages * 2} numbers \cr
  \emph{(0: 0\% 1: 10\% 2: 20\% 3: 30\% 4: 40\% 5: 50\% 6: 60\% 7: 70\% 8: 80\% 9: 
  90\% 10: 100\%)} \cr
  corresponding to the proportion of individuals :
  \itemize{
    \item that will be killed or resprout
    \item at different response stages \emph{(!currently 4, not yet editable!)}
    \item for each disturbance.
  }
  \item ACTIVATED_SEED : the proportion of seeds activated by each disturbance \cr
  \emph{(0: 0\% 1: 10\% 2: 20\% 3: 30\% 4: 40\% 5: 50\% 6: 60\% 7: 70\% 8: 80\% 9: 
  90\% 10: 100\%)} \cr \cr
}

A \code{DIST_COMPLETE_TABLE.csv} file summarizing information for all groups into the
\code{name.simulation/DATA/PFGS/} directory.
}
\description{
This script is designed to create parameter files containing 
response to disturbance parameters for each PFG (one file for each of them) 
used in the disturbance module of \code{FATE-HD}.
}
\details{
The disturbance module of \code{FATE-HD} allows the user to simulate spatial
perturbation(s) that will impact each PFG in terms of resprouting and mortality
on the different response stages \emph{(!currently 4, not yet editable!)}. \cr
Several parameters are required for each PFG in order to set up these responses :

\describe{
  \item{type}{or life-form, based on Raunkier. It should be either \code{H} 
  (herbaceous), \code{C} (chamaephyte) or \code{P} (phanerophyte) for now}
  \item{MATURITY}{the age from which the PFG can reproduce}
  \item{LONGEVITY}{the maximum or average lifespan of the PFG}
  \item{STRATA}{the maximum height stratum that the PFG can reach}
  \item{CHANG_STR_AGES}{at what age each PFG goes into the upper stratum}
}


These values will allow to calculate or define a set of characteristics for each
PFG :

\describe{
  \item{PROP_KILLED}{ = the proportion of propagules killed by each 
  disturbance \cr \cr
  It is currently set to 0 for all PFG and disturbances.
  }
  \item{BREAK_AGE}{ = each PFG can respond to a disturbance in several 
  different ways \cr \emph{(!currently 4, not yet editable!)} that depend 
  on the PFG age \cr
   = ages at which each PFG changes of response stage \cr \cr
   
   These response classes are the same for all disturbances for each PFG. \cr
   These ages are defined according to the maturity and longevity of the PFG,
   its life form and its ability to reach the 3rd height stratum :
   \itemize{
     \item age from class 1 to 2 :
     \itemize{
       \item \code{maturity - 2} for herbaceous
       \item \code{1} for chamaephytes and phanerophytes
     }
     \item age from class 2 to 3 : \code{min(CHANG_STR_AGES_to_str_3, maturity)}
     \item age from class 3 to 4 : \code{min(CHANG_STR_AGES_to_str_3, longevity - 2)}
   }
   Some corrections are made for short-living plants (annuals, biennials...) :
   as they die after 1 or 2 years, they are not affected differently according
   to life stages. All ages are the same and correspond to their longevity.
  }
  \item{RESPR_AGE}{ = when subject to a perturbation, each PFG can either stay 
  undisturbed, be killed, or resprout \emph{(in years)} \cr
   = ages at which each PFG will be rejuvenated by a disturbance \cr \cr
  It does not impact dead inviduals, only living ones. \cr
  It is defined according to the maturity and longevity of the PFG, and its 
  ability to reach the 3rd height stratum :
  \itemize{
    \item individuals within 1st and 2nd response stage are too young to resprout
    \item individuals within 3rd response stage : \cr
    \code{min(CHANG_STR_AGES_to_str_3, maturity - 2)}
    \item individuals within 4th response stage : \code{longevity - 2}
    \item short-living plants (annuals, biennials...) always start back at 0
  }
  }
  \item{FATES}{ = proportion of killed or resprouting individuals \cr
   = for each disturbance and for each response stage   
  }
  \item{ACTIVATED_SEED}{ = the proportion of seeds activated by each 
  disturbance \cr \cr
  It is currently set to 0 for all PFG and disturbances.}
}
}
\examples{

## Create a skeleton folder with the default name ('FATE_simulation')
PRE_FATE.skeletonDirectory()

## Create PFG succession parameter files
PRE_FATE.params_PFGsuccession(name.simulation = "FATE_simulation"
                            , mat.PFG.succ = data.frame(PFG = paste0("PFG",1:6)
                                                        , type = c("C", "C", "H", "H", "P", "P")  
                                                        , height = c(10, 250, 36, 68, 1250, 550)
                                                        , maturity = c(5, 5, 3, 3, 8, 9)
                                                        , longevity = c(12, 200, 25, 4, 110, 70)
                                                        , dispersal = 1
                                                        , light = c(4, 6, 3, 6, 5, 5)))
                                                        
## Create PFG disturbance parameter files
PRE_FATE.params_PFGdisturbance(name.simulation = "FATE_simulation"
                               , mat.PFG.dist = data.frame(name = rep(c("DIST1","DIST2"), each = 4)
                                                           , responseStage = rep(1:4, 2)
                                                           #, KilledPropagule_H = 0
                                                           #, KilledPropagule_C = 0
                                                           #, KilledPropagule_P = 0
                                                           , KilledIndiv_H = c(0,0,0,0,1,1,0,0)
                                                           , KilledIndiv_C = c(0,10,10,10,1,1,0,0)
                                                           , KilledIndiv_P = c(10,10,10,10,10,0,0,0)
                                                           , ResproutIndiv_H = c(0,0,9,10,0,0,5,1)
                                                           , ResproutIndiv_C = c(0,0,0,0,0,0,5,1)
                                                           , ResproutIndiv_P = c(0,0,0,0,0,0,0,0)))
                                                           

}
\author{
Maya Guéguen
}
