% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRE_FATE.params_PFGdrought.R
\name{PRE_FATE.params_PFGdrought}
\alias{PRE_FATE.params_PFGdrought}
\title{Create \emph{DROUGHT} parameter files for a \code{FATE}
simulation}
\usage{
PRE_FATE.params_PFGdrought(
  name.simulation,
  mat.PFG.dist = NULL,
  mat.PFG.tol,
  mat.PFG.drought,
  opt.folder.name = NULL
)
}
\arguments{
\item{name.simulation}{a \code{string} corresponding to the main directory 
or simulation name of the \code{FATE} simulation}

\item{mat.PFG.dist}{(\emph{optional}) \cr 
a \code{data.frame} with 5 columns : \cr 
\code{PFG}, \code{type}, \code{maturity}, \code{longevity}, 
\code{age_above_150cm} (see 
\href{PRE_FATE.params_PFGdrought.html#details}{\code{Details}})}

\item{mat.PFG.tol}{a \code{data.frame} with 3 to 7 columns : \cr 
\itemize{
  \item \code{nameDist},
  \item \code{PFG},
  \item (\emph{\code{responseStage}, \code{breakAge}, \code{resproutAge}}), 
  \item \code{responseStage}, \code{killedIndiv}, \code{resproutIndiv}  
  (\emph{or \code{strategy_tol}})
}
(see \href{PRE_FATE.params_PFGdrought.html#details}{\code{Details}})}

\item{mat.PFG.drought}{a \code{data.frame} with 4 or 6 columns : \cr 
\itemize{
  \item \code{PFG},
  \item \code{threshold_moderate}, \code{threshold_severe},
  \item \code{counter_recovery}, \code{counter_sens}, \code{counter_cum}
  (\emph{or \code{strategy_drou}})
}}

\item{opt.folder.name}{(\emph{optional}) \cr a \code{string} corresponding 
to the name of the folder that will be created into the 
\code{name.simulation/DATA/PFGS/DROUGHT/} directory to store the results}
}
\value{
A \code{.txt} file per PFG into the 
\code{name.simulation/DATA/PFGS/DROUGHT/} directory with the following 
parameters :

\describe{
  \item{BREAK_AGE}{ages at which the PFG changes of response stage 
  \emph{(in years)}}
  \item{RESPR_AGE}{resprouting age table (in a single row) \cr
  This is a vector of \code{no.DIST (=2) * no.responseStages} numbers 
  corresponding \cr to the age at which the PFG can be rejuvenated 
  (younger than the actual one) :
  \itemize{
    \item at different response stages \emph{(\code{RS})}
    \item for each disturbance \emph{(\code{DI})}.
  }
  These parameters should be given in this order (e.g. with 3 response 
  stages) : \cr \code{DI1_RS1, DI1_RS2, DI1_RS3, DI2_RS1...} \emph{(in 
  years)}. 
  }
  \item{FATES}{disturbance response table (in a single row) \cr
  This is a vector of \code{no.DIST (=2) * no.responseStages * 2} numbers 
  corresponding \cr to the proportion of individuals :
  \itemize{
    \item that will be killed \emph{(\code{Ki})} or resprout 
    \emph{(\code{Re})}
    \item at different response stages \emph{(\code{RS})}
    \item for each disturbance \emph{(\code{DI})}.
  }
  These parameters should be given in this order (e.g. with 3 response 
  stages) : \cr \code{DI1_RS1_Ki, DI1_RS1_Re, DI1_RS2_Ki, DI1_RS2_Re, 
  DI1_RS3_Ki, DI1_RS3_Re, DI2_RS1_Ki...}
  \cr \emph{(from \code{0} to \code{10}, corresponding to 0 to 100\%)}. 
  }
  \item{PROP_KILLED}{proportion of propagules killed by each disturbance \cr
  \emph{(from \code{0} to \code{10}, corresponding to 0 to 100\%)}}
  \item{ACTIVATED_SEED}{proportion of seeds activated by each disturbance \cr
  \emph{(from \code{0} to \code{10}, corresponding to 0 to 100\%)} \cr \cr}
  \item{THRESHOLD_MOD}{threshold below which the PFG will experience 
  moderate drought \cr \emph{(same unit as that of the map given with the 
  \code{DROUGHT_MASK} flag in 
  \code{\link{PRE_FATE.params_globalParameters}})}}
  \item{THRESHOLD_SEV}{threshold below which the PFG will experience 
  severe drought \cr \emph{(same unit as that of the map given with the 
  \code{DROUGHT_MASK} flag in 
  \code{\link{PRE_FATE.params_globalParameters}})}}
  \item{COUNTER_RECOVERY}{number of years removed from the PFG counter of 
  cumulated consecutive years of drought events, during non-drought years}
  \item{COUNTER_SENS}{number of consecutive years of drought the PFG must 
  experience before suffering severe effects due to a severe drought \cr 
  (\emph{sensitivity to severe drought})}
  \item{COUNTER_CUM}{number of consecutive years of drought the PFG must 
  experience before any subsequent drought event start having severe effects 
  \cr (\emph{cumulative drought response}) \cr \cr}
}

A \code{DROUGHT_COMPLETE_TABLE.csv} file summarizing information for all 
groups into the \code{name.simulation/DATA/PFGS/} directory.  

If the \code{opt.folder.name} has been used, the files will be into the 
folder \code{name.simulation/DATA/PFGS/DROUGHT/opt.folder.name/}.
}
\description{
This script is designed to create parameter files containing 
response to drought disturbance parameters for each PFG (one file for each 
of them) used in the drought disturbance module of \code{FATE}.
}
\details{
The \strong{drought disturbance module} is a specific case of the 
\code{disturbance module}. It also allows the user to simulate spatial 
perturbation(s) that will impact each PFG in terms of \emph{resprouting} and 
\emph{mortality} at different response stages, but with specific rules to 
determine when the PFG is affected (see 
\code{\link{PRE_FATE.params_globalParameters}}). \cr \cr


Several parameters, given within \code{mat.PFG.dist} or \code{mat.PFG.tol}, 
are required for each PFG in order to set up these responses. The 
explanations are the same than those that can be found in 
\code{\link{PRE_FATE.params_PFGdisturbance}} function. Therefore, 
\strong{only parameters whose values or descriptions change are detailed 
below :}

\describe{
  \item{nameDist}{a \code{string} to choose the concerned drought 
  disturbance : \cr \code{immediate} or \code{delayed} \cr \cr}
  
  \item{(\emph{strategy_tol})}{a \code{string} to choose the response to 
  drought strategy : \cr \code{herbs_cham_1}, \code{herbs_cham_2}, 
  \code{herbs_cham_3}, \code{trees_1}, \code{trees_2}, \code{trees_3} 
  \cr \cr}
}


These values will allow to calculate or define a set of characteristics for 
each PFG :

\describe{
  \item{FATES}{ = proportion of killed and resprouting individuals \cr
   = for each disturbance and for each response stage \cr \cr
  Two methods to define these tolerances are available :
  \itemize{
    \item from \strong{predefined scenarios} (using 
    \code{strategy_tol}) : \cr
      \itemize{
        \item the values give the percentage of killed or resprouting 
        individuals
        \item with \code{1, 2, 3, 4}: response classes
        \item with \code{K}: killed individuals, \code{R}: resprouting 
        individuals \cr \cr
      }
      \describe{
        \item{}{\strong{\code{| ___1___ | ___2___ | ___3___ | ___4___ |}}}
        \item{}{\strong{\code{| _K_ _R_ | _K_ _R_ | _K_ _R_ | _K_ _R_ |}}}
        \item{}{\code{________________IMMEDIATE________________}}
        \item{herbs_cham_1}{\code{| 10\% _0_ | _0_ _0_ | _0_ _0_ | _0_ _0_ |}}
        \item{herbs_cham_2}{\code{| 20\% _0_ | _0_ _0_ | _0_ _0_ | 10\% _0_ |}}
        \item{herbs_cham_3}{\code{| 40\% _0_ | 10\% _0_ | 10\% _0_ | 20\% _0_ |}}
        \item{trees_1}{\code{| 10\% _0_ | _0_ _0_ | _0_ 40\% | _0_ 40\% |}}
        \item{trees_2}{\code{| 20\% _0_ | _0_ 10\% | _0_ 50\% | 10\% 50\% |}}
        \item{trees_3}{\code{| 40\% _0_ | 10\% 40\% | 10\% 80\% | 20\% 80\% |}}
        \item{}{\code{_________________DELAYED_________________}}
        \item{herbs_cham_1}{\code{| _0_ _0_ | _0_ 10\% | _0_ 10\% | _0_ 10\% |}}
        \item{herbs_cham_2}{\code{| _0_ _0_ | _0_ 10\% | _0_ 10\% | _0_ 10\% |}}
        \item{herbs_cham_3}{\code{| _0_ _0_ | _0_ 10\% | _0_ 10\% | _0_ 10\% |}}
        \item{trees_1}{\code{| _0_ _0_ | _0_ 10\% | _0_ 40\% | _0_ 40\% |}}
        \item{trees_2}{\code{| 10\% _0_ | _0_ 40\% | _0_ 40\% | _0_ 40\% |}}
        \item{trees_3}{\code{| 20\% _0_ | 10\% 40\% | 10\% 50\% | 10\% 50\% |}}
      }
    \item from \strong{user data} : \cr
      \emph{with the values contained within the \code{responseStage}, 
      \code{killedIndiv} and \code{resproutIndiv} columns, if provided \cr
      The \code{PFG} column can contain either the life form (\code{H}, 
      \code{C} or \code{P}) or the PFG name. Both methods can be combined 
      (but are applied in the order given by the \code{PFG} column). \cr \cr
      }
  }
  }
}


Supplementary parameters related to drought, given within 
\code{mat.PFG.drought}, are required for each PFG :

\describe{
  \item{threshold_moderate}{a value corresponding to the threshold below 
  which the PFG will experience moderate drought (on the same scale than 
  \code{threshold_severe} and the map given with the \code{DROUGHT_MASK} 
  flag in \code{\link{PRE_FATE.params_globalParameters}})}
  \item{threshold_severe}{a value corresponding to the threshold below 
  which the PFG will experience severe drought (on the same scale than 
  \code{threshold_moderate} and the map given with the \code{DROUGHT_MASK} 
  flag in \code{\link{PRE_FATE.params_globalParameters}}). It should be 
  inferior or equal to \code{threshold_moderate}. \cr \cr}
  \item{counter_recovery}{an \code{integer} corresponding to the number of 
  years removed from the PFG counter of cumulated consecutive years of 
  drought events, during non-drought years}
  \item{counter_sens}{an \code{integer} corresponding to the number of 
  consecutive years of drought the PFG must experience before suffering 
  severe effects due to a severe drought (\emph{sensitivity to severe 
  drought})}
  \item{counter_cum}{an \code{integer} corresponding to the number of 
  consecutive years of drought the PFG must experience before any subsequent 
  drought event start having severe effects (\emph{cumulative drought 
  response}). It should be superior or equal to \code{counter_sens}.}
  \item{(\emph{strategy_drou})}{a \code{string} to choose the "counter" 
  strategy : \cr \code{herbs}, \code{chamaephytes}, \code{trees_shrubs} 
  \cr \cr}
}

These values will allow to define a set of characteristics for each PFG :
\describe{
  \item{sensitivity to values}{with the \strong{THRESHOLD_MODERATE} and 
  \strong{THRESHOLD_SEVERE} parameters}
  \item{sensitivity through time}{with the \strong{COUNTER_RECOVERY}, 
  \strong{COUNTER_SENS} and \strong{COUNTER_CUM} parameters}
}
}
\examples{


## ----------------------------------------------------------------------------------------- ##
                                    
## Load example data
                                                           

}
\seealso{
\code{\link{PRE_FATE.skeletonDirectory}}, 
\code{\link{PRE_FATE.params_globalParameters}},
\code{\link{PRE_FATE.params_PFGdisturbance}}
}
\author{
Maya Guéguen
}
\keyword{FATE,}
\keyword{disturbance,}
\keyword{killing,}
\keyword{resprouting}
\keyword{simulation,}
