% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRE_FATE.params_PFGlight.R
\name{PRE_FATE.params_PFGlight}
\alias{PRE_FATE.params_PFGlight}
\title{Create \emph{LIGHT} parameter files for a \code{FATE} simulation}
\usage{
PRE_FATE.params_PFGlight(
  name.simulation,
  mat.PFG.light,
  mat.PFG.tol = NULL,
  opt.folder.name = NULL
)
}
\arguments{
\item{name.simulation}{a \code{string} corresponding to the main directory 
or simulation name of the \code{FATE} simulation}

\item{mat.PFG.light}{a \code{data.frame} with 2 to 6 columns : \cr 
\itemize{
  \item \code{PFG},
  \item \code{type} (\emph{or \code{shade_factor}})
  \item \code{type}, (\emph{or \code{active_germ_low}, 
  \code{active_germ_medium}, \code{active_germ_high}}) (\emph{or
  \code{strategy_ag}})
  \item \emph{\code{type}, \code{light_need}}
}
(see \href{PRE_FATE.params_PFGlight.html#details}{\code{Details}})}

\item{mat.PFG.tol}{(\emph{optional}) \cr 
a \code{data.frame} with 2 to 4 columns : \cr 
\itemize{
  \item \code{PFG},
  \item \code{lifeStage}, \code{resources}, \code{tolerance} 
  (\emph{or \code{strategy_tol}})
}
(see \href{PRE_FATE.params_PFGlight.html#details}{\code{Details}})}

\item{opt.folder.name}{(\emph{optional}) \cr a \code{string} corresponding 
to the name of the folder that will be created into the 
\code{name.simulation/DATA/PFGS/LIGHT/} directory to store the results}
}
\value{
A \code{.txt} file per PFG into the 
\code{name.simulation/DATA/PFGS/LIGHT/} directory with the following 
parameters :

\describe{
  \item{NAME}{name of the PFG}
  \item{LIGHT}{light value or strategy of the PFG}
  \item{SHADE_FACTOR}{index of shade quantity to weight PFG abundance and 
  transform it into light resources}
  \item{ACTIVE_GERM}{germination rates depending on light conditions 
  \cr \emph{(integer between \code{0} and \code{100}\%)}}
  \item{LIGHT_TOL}{light tolerance table (in a single row). \cr 
  This is a vector of 9 numbers corresponding to the ability of the PFG to 
  survive or not :
  \itemize{
    \item at different life stages \emph{(Germinant (\code{Ge}), Immature 
    (\code{Im}), Mature (\code{Ma}))}
    \item under different light conditions \emph{(Low (\code{L}), Medium 
    (\code{M}) or High (\code{H}))}.
  }
  These parameters should be given in this order : \code{GeL, GeM, GeH, ImL, 
  ImM, ImH, MaL, MaM, MaH}
  \cr \emph{(integer between \code{0} and \code{100}\%)}. 
  \cr \cr}
}

A \code{LIGHT_COMPLETE_TABLE.csv} file summarizing information for all 
groups into the \code{name.simulation/DATA/PFGS/} directory.  

If the \code{opt.folder.name} has been used, the files will be into the 
folder \code{name.simulation/DATA/PFGS/LIGHT/opt.folder.name/}.
}
\description{
This script is designed to create parameter files containing 
light-related parameters for each PFG (one file for each of them) used in 
the light module of \code{FATE}.
}
\details{
The \strong{light module} allows the user to add the effect of 
\strong{light interaction} within a primary vegetation succession. \cr \cr

Several parameters, given within \code{mat.PFG.light} or \code{mat.PFG.tol}, 
are required for each PFG in order to set up this light interaction :

\describe{
  \item{PFG}{the concerned plant functional group \cr \cr}
  
  \item{type}{or life-form, based on Raunkier. \cr It should be either 
  \code{H} (herbaceous), \code{C} (chamaephyte) or \code{P} (phanerophyte) 
  for now \cr \cr}
  
  \item{(\emph{shade_factor})}{an \code{integer} between \code{1} and 
  \code{Inf} corresponding to an index of shade quantity to weight PFG 
  abundance and transform it into light resources (\emph{e.g. if two PFG 
  have shade factors of \code{1} and \code{5} respectively, for the same 
  abundances, the second PFG will produce 5 times more shade than the first 
  one}) \cr \cr}
  
  \item{(\emph{active_germ_low})}{an \code{integer} between \code{0} and 
  \code{100} corresponding to the proportion of seeds that will germinate for 
  \code{Low} light condition}
  \item{(\emph{active_germ_medium})}{an \code{integer} between \code{0} and 
  \code{100} corresponding to the proportion of seeds that will germinate for 
  \code{Medium} light condition}
  \item{(\emph{active_germ_high})}{an \code{integer} between \code{0} and 
  \code{100} corresponding to the proportion of seeds that will germinate for 
  \code{High} light condition}
  \item{(\emph{strategy_ag})}{a \code{string} to choose the germination 
  strategy : \cr \code{light_lover}, \code{indifferent}, \code{shade_lover} 
  \cr \cr}
  
  \item{(\emph{light_need})}{an \code{integer} between \code{0} and \code{5} 
  corresponding to the light preference of the PFG (e.g. from Flora 
  Indicativa)\cr \cr}
  
  \item{lifeStage}{the concerned life stage (\code{Germinant}, 
  \code{Immature}, \code{Mature})}
  \item{resources}{the concerned light condition (\code{Low}, 
  \code{Medium}, \code{High})}
  \item{tolerance}{an \code{integer} between \code{0} and \code{100} 
  corresponding to the proportion of surviving individuals}
  \item{(\emph{strategy_tol})}{a \code{string} to choose the tolerance 
  strategy : \cr \code{full_light}, \code{pioneer}, \code{ubiquist}, 
  \code{semi_shade}, \code{undergrowth} \cr \cr}
}


These values will allow to calculate or define a set of characteristics for 
each PFG :

\describe{
  \item{SHADE_FACTOR}{index of shade quantity to weight PFG abundance and 
  transform it into light resources \cr \cr
  Two methods to define these proportions are available :
  \itemize{
    \item from \strong{predefined rules} (using \code{type}) :
    \itemize{
      \item for \code{H} (herbaceous) : \code{1}
      \item for \code{C} (chamaephyte) : \code{5}
      \item for \code{P} (phanerophyte) : \code{100}
    }
    \item from \strong{user data} : \cr
    \emph{with the values contained within the \code{shade_factor} column, 
    if provided \cr \cr}
  }
  }

  \item{ACTIVE_GERM}{proportion of seeds that will germinate for each light 
  condition (\code{Low}, \code{Medium}, \code{High}) \cr \cr
  Three methods to define these proportions are available :
  \itemize{
    \item from \strong{predefined scenarios} (using \code{strategy_ag}) : \cr \cr
    \strong{\code{| _L_ _M_ _H_ |}} \cr
    \code{_______________} \cr
    \code{| 50\% 80\% 90\% |} \strong{light_lover} \cr
    \code{| 90\% 90\% 90\% |} \strong{indifferent} \cr
    \code{| 90\% 80\% 50\% |} \strong{shade_lover} \cr \cr
    
    \item from \strong{predefined rules} (using \code{type}) :
    \itemize{
      \item for \code{H} (herbaceous) : \code{50\%, 80\%, 90\%}
      \item for \code{C} (chamaephyte) or \code{P} (phanerophyte): 
      \code{90\%, 90\%, 90\%}
    }
    \item from \strong{user data} : \cr
    \emph{with the values contained within the \code{active_germ_low}, 
    \code{active_germ_medium} and \code{active_germ_high} columns, if 
    provided \cr \cr}
  }
  }

  \item{LIGHT_TOL}{ defined for each life stage (\code{Germinant}, 
  \code{Immature}, \code{Mature}) \cr and each soil condition (\code{Low}, 
  \code{Medium}, \code{High}) \cr \cr
  Three methods to define these tolerances are available :
  \itemize{
    \item from \strong{predefined scenarios} (using 
    \code{strategy_tol}) : \cr
      \itemize{
        \item \code{.} means \emph{Not tolerant}, \code{1} means 
        \emph{Tolerant} (\code{100\%})
        \item with \code{g}: Germinant, \code{i}: Immature, \code{m}: Mature
        \item with \code{L}: low light, \code{M}: medium light, \code{H}: 
        high light \cr \cr
      }
      \strong{\code{| _ g _ | _ i _ | _ m _ |}} \cr
      \strong{\code{| L M H | L M H | L M H |}} \cr
      \code{_________________________} \cr
      \code{| 1 1 1 | . . 1 | . . 1 |} \strong{full_light} \cr
      \code{| 1 1 1 | . 1 1 | . 1 1 |} \strong{pioneer} \cr
      \code{| 1 1 1 | 1 1 1 | 1 1 1 |} \strong{ubiquist} \cr
      \code{| 1 1 . | 1 1 . | 1 1 1 |} \strong{semi_shade} \cr
      \code{| 1 1 . | 1 1 . | 1 1 . |} \strong{undergrowth} \cr \cr
      
    \item from \strong{predefined rules} (using \code{type} and 
    \code{light_need}):
      \describe{
        \item{(A)}{PFG are tolerant to \code{Low} light if \code{light <= 2}}
        \item{(A)}{PFG are tolerant to \code{Medium} light if 
        \code{2 <= light <= 4}}
        \item{(A)}{PFG are tolerant to \code{High} light if 
        \code{light >= 3}}
        \item{(B)}{all germinants are assumed to be tolerant to \code{Low} 
        light}
        \item{(C)}{all mature trees or chamaephytes are assumed to be 
        tolerant to \code{Medium} and \code{High} light conditions}
        \item{\emph{(D)}}{\emph{all immature trees that grow in the 
        penultimate stratum are assumed to be tolerant to \code{High} light} 
        \strong{!! desactivated !!}}
      }
      \itemize{
        \item \code{.} means \emph{Not tolerant}
        \item \code{A, B, C, D} mean \emph{Tolerant} (\code{100\%}) according 
        to one of the rule defined above
        \item with \code{g}: Germinant, \code{i}: Immature, \code{m}: Mature
        \item with \code{L}: low light, \code{M}: medium light, \code{H}: 
        high light \cr \cr
      }
      \strong{\code{| _ g _ | _ i _ | _ m _ |}} \cr
      \strong{\code{| L M H | L M H | L M H |}} \cr
      \code{_________________________} \cr
      \code{| A . . | A . D | A C C |} \strong{1} \cr
      \code{| A A . | A A D | A A C |} \strong{2} \cr
      \code{| B A . | . A D | . A C |} \strong{3} \cr
      \code{| B A A | . A A | . A A |} \strong{4} \cr
      \code{| B . A | . . A | . C A |} \strong{5} \cr \cr
      
    \item from \strong{user data} : \cr
      \emph{with the values contained within the \code{lifeStage}, 
      \code{resources} and \code{tolerance} columns, if provided}
  }
  }
}
}
\examples{

## Create a skeleton folder with the default name ('FATE_simulation')
PRE_FATE.skeletonDirectory()


## Create PFG light parameter files (with strategies) ----------------------------------------
mat.ag = data.frame(PFG = paste0('PFG', 1:6)
                    , type = c('C', 'C', 'H', 'H', 'P', 'P')
                    , strategy_ag = c('shade_lover', 'indifferent'
                                      , 'indifferent', 'shade_lover'
                                      , 'light_lover', 'light_lover'))

mat.tol = data.frame(PFG = paste0('PFG', 1:6)
                     , strategy_tol = c('undergrowth', 'ubiquist'
                                        , 'ubiquist', 'semi_shade'
                                        , 'pioneer', 'full_light'))

PRE_FATE.params_PFGlight(name.simulation = 'FATE_simulation'
                         , mat.PFG.light = mat.ag
                         , mat.PFG.tol = mat.tol)
                                                        

## Create PFG light parameter files (with all values) ----------------------------------------
mat.ag = data.frame(PFG = paste0('PFG', 1:6)
                    , shade_factor = c(5, 3, 1, 1, 12, 18)
                    , active_germ_low = c(90, 80, 80, 80, 50, 50)
                    , active_germ_medium = rep(80, 6)
                    , active_germ_high = c(40, 80, 80, 50, 90, 90))

mat.tol = expand.grid(resources = c('Low', 'Medium', 'High')
                      , lifeStage = c('Germinant', 'Immature', 'Mature')
                      , PFG = paste0('PFG', 1:6))
mat.tol$tolerance = c(100, 100, 0, 100, 0, 0, 100, 0, 0
                      , rep(100, 9)
                      , rep(100, 9)
                      , 100, 100, 100, 100, 100, 100, 100, 0, 0
                      , 100, 100, 100, 0, 100, 100, 0, 0, 100
                      , 100, 100, 100, 0, 0, 100, 0, 0, 100)

PRE_FATE.params_PFGlight(name.simulation = 'FATE_simulation'
                         , mat.PFG.light = mat.ag
                         , mat.PFG.tol = mat.tol)
                                                        
                                                        
## -------------------------------------------------------------------------------------------

## Load example data
Champsaur_params = .loadData('Champsaur_params', 'RData')

## Create a skeleton folder
PRE_FATE.skeletonDirectory(name.simulation = 'FATE_Champsaur')


## PFG traits for light
tab.light = Champsaur_params$tab.LIGHT
str(tab.light)

## Create PFG light parameter files ----------------------------------------------------------
PRE_FATE.params_PFGlight(name.simulation = 'FATE_Champsaur'
                         , mat.PFG.light = tab.light[, c('PFG', 'type')]
                         , mat.PFG.tol = tab.light[, c('PFG', 'strategy_tol')])


}
\seealso{
\code{\link{PRE_FATE.skeletonDirectory}},
\code{\link{PRE_FATE.params_globalParameters}}, 
\code{\link{PRE_FATE.params_PFGsuccession}}
}
\author{
Maya Guéguen
}
\keyword{FATE,}
\keyword{light}
\keyword{simulation,}
\keyword{tolerance}
