% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRE_FATE.params_PFGsuccession.R
\name{PRE_FATE.params_PFGsuccession}
\alias{PRE_FATE.params_PFGsuccession}
\title{Create \emph{SUCCESSION} parameter files for a \code{FATE} simulation}
\usage{
PRE_FATE.params_PFGsuccession(
  name.simulation,
  mat.PFG.succ,
  strata.limits = c(0, 20, 50, 150, 400, 1000, 2000, 5000, 10000),
  strata.limits_reduce = TRUE,
  opt.folder.name = NULL
)
}
\arguments{
\item{name.simulation}{a \code{string} corresponding to the main directory 
or simulation name of the \code{FATE} simulation}

\item{mat.PFG.succ}{a \code{data.frame} with at least 5 columns : \cr 
\code{PFG}, \code{type}, \code{height}, \code{maturity}, \code{longevity} 
\cr (\emph{and optionally, \code{max_abundance}, \code{potential_fecundity}, 
\code{immature_size}, \code{is_alien}, \code{flammability}}) 
\cr (see \href{PRE_FATE.params_PFGsuccession.html#details}{\code{Details}})}

\item{strata.limits}{a \code{vector} of \code{integer} containing values 
among which height strata limits will be chosen}

\item{strata.limits_reduce}{default \code{TRUE}. \cr If \code{TRUE}, stratum 
height limits are checked to try and bring several PFGs together in a same 
stratum}

\item{opt.folder.name}{(\emph{optional}) \cr a \code{string} corresponding 
to the name of the folder that will be created into the 
\code{name.simulation/DATA/PFGS/SUCC/} directory to store the results}
}
\value{
A \code{.txt} file per PFG into the 
\file{name.simulation/DATA/PFGS/SUCC/} directory with the following 
parameters :

\describe{
  \item{NAME}{name of the PFG}
  \item{TYPE}{PFG life-form \emph{(\code{H}: herbaceous \code{C}: 
  chamaephyte \code{P}: phanerophyte)}}
  \item{HEIGHT}{PFG maximum height \emph{(in cm)}}
  \item{MATURITY}{PFG maturity age \emph{(in years)}}
  \item{LONGEVITY}{PFG life span \emph{(in years)}}
  \item{MAX_STRATUM}{maximum height stratum that the PFG can reach}
  \item{MAX_ABUNDANCE}{maximum abundance / space (qualitative) that the PFG 
  is able to produce / occupy \cr \emph{(\code{1}: Low \code{2}: Medium 
  \code{3}: High)}}
  \item{IMM_SIZE}{PFG immature relative size \emph{(from \code{0} to 
  \code{10}, corresponding to 0 to 100\%)}}
  \item{CHANG_STR_AGES}{ages at which the PFG goes in the upper stratum 
  \cr \emph{(in years, put a value higher than the PFG life span if it is 
  not supposed to rise a stratum)}}
  \item{SEED_POOL_LIFE}{maximum number of years seeds are able to survive 
  (for active and dormant pool)}
  \item{SEED_DORMANCY}{are the seeds dormant or not \emph{(\code{0}: No 
  \code{1}: Yes)}}
  \item{POTENTIAL_\cr FECUNDITY}{maximum number of seeds produced by the 
  PFG \cr \emph{(set by default to \code{100})}}
  \item{IS_ALIEN}{is the PFG an alien or not \emph{(\code{0}: No \code{1}: 
  Yes)}}
  \item{FLAMMABILITY}{how easily the PFG burns \emph{(\code{numeric})}
  \cr \cr}
}

A \file{SUCC_COMPLETE_TABLE.csv} file summarizing information for all 
groups into the \file{name.simulation/DATA/PFGS/} directory. \cr
This file can be used to parameterize the disturbance files (see 
\code{\link{PRE_FATE.params_PFGdisturbance}}).

If the \code{opt.folder.name} has been used, the files will be into the 
folder \file{name.simulation/DATA/PFGS/SUCC/opt.folder.name/}.
}
\description{
This script is designed to create parameter files containing 
succession parameters for each PFG (one file for each of them) used in the 
core module of \code{FATE}.
}
\details{
The \strong{core module} of \code{FATE} allows the user to simulate a 
primary vegetation succession based on a \strong{demography model}. \cr \cr

Several parameters, given within \code{mat.PFG.succ}, are required for 
each PFG in order to set up this life cycle :

\describe{
  \item{type}{or life-form, based on Raunkier. \cr It should be either 
  \code{H} (herbaceous), \code{C} (chamaephyte) or \code{P} (phanerophyte) 
  for now}
  \item{height}{the maximum or average height that reach the PFG}
  \item{maturity}{the age from which the PFG can reproduce}
  \item{longevity}{the maximum or average lifespan of the PFG \cr \cr}
  \item{(\emph{max_abundance})}{the maximum abundance of mature PFG in 
  favorable conditions}
  \item{(\emph{potential_fecundity})}{the maximum number of seeds produced 
  by the PFG}
  \item{(\emph{immature_size})}{the relative size of immature versus mature 
  plants}
  \item{(\emph{is_alien})}{if the PFG is to be considered as an alien 
  (\code{1}) or not (\code{0})}
  \item{(\emph{flammability})}{how easily the PFG burns \cr \cr}
}


These values will allow to calculate or define a set of characteristics for 
all PFG :

\describe{
  \item{STRATA_LIMITS}{= height values that define each stratum.\cr \cr
  Two methods to define these limits are available :
  \itemize{
    \item from \strong{predefined rules} (using \code{strata.limits_reduce 
    = TRUE}, \code{strata.limits}, \code{height}) : \cr \cr
    \itemize{
      \item limits should go exponentially and are selected among 
      \code{strata.limits}
      \item PFG are separated according to these \code{strata.limits} and 
      then grouped making sure to have 
      \deqn{\text{number of PFG per stratum} \geq \sqrt{\text{total number 
      of PFG}} - 2}
      to try to homogenize the number of PFG within each stratum. \cr \cr
    }
    \item from \strong{user data} : (using \code{strata.limits_reduce = 
    FALSE}) \cr
      \emph{with the values contained within the \code{strata.limits} 
      column, if provided \cr \cr}
  }
  }
}

and a set of characteristics for each PFG :

\describe{
  \item{MAX_STRATUM}{= maximum stratum that each PFG can reach}
  \item{MAX_ABUNDANCE}{= maximum abundance of mature PFG in favorable 
  conditions \cr
   = It can be seen as a proxy of maximum carrying capacity for mature 
  individuals \cr (\emph{and therefore as a proxy a quantity of shade 
  produced as well, if the light module is activated}). \cr \cr
  Two methods to define these abundances are available :
  \itemize{
    \item from \strong{predefined rules} (using \code{type}, 
    \code{MAX_STRATUM}) : \cr \cr
    \tabular{rcccc}{
      \strong{MAX_STRATUM} \tab \strong{1} \tab \strong{2} \tab 
      \strong{3} \tab \strong{+}\cr
      \strong{\code{H} (herbaceous)} \tab \code{1} \tab \code{1} \tab 
      \code{2} \tab \code{2} \cr
      \strong{\code{C} (chamaephyte)} \tab \code{1} \tab \code{2} \tab 
      \code{2} \tab \code{3} \cr
      \strong{\code{P} (phanerophyte)} \tab \code{1} \tab \code{2} \tab 
      \code{3} \tab \code{3}
    }
    \item from \strong{user data} : \cr
      \emph{with the values contained within the \code{max_abundance} 
      column, if provided \cr \cr}
  }
  }
  \item{IMM_SIZE}{= relative size of immature versus mature plants \cr
  = for example, immature herbaceous take as much space as mature 
  herbaceous, while immature phanerophytes take less space (\emph{and 
  contribute to shade half less}) than mature individuals \cr \cr
  Two methods to define these sizes are available :
  \itemize{
    \item from \strong{predefined rules} (using \code{type}, \code{MAX_STRATUM}) 
    : \cr \cr
    
    \tabular{rcccc}{
      \strong{MAX_STRATUM} \tab \strong{1} \tab \strong{2} \tab 
      \strong{3} \tab \strong{+}\cr
      \strong{\code{H} (herbaceous)} \tab \code{100\%} \tab \code{80\%} 
      \tab \code{50\%} \tab \code{50\%} \cr
      \strong{\code{C} (chamaephyte)} \tab \code{100\%} \tab \code{50\%} 
      \tab \code{50\%} \tab \code{50\%} \cr
      \strong{\code{P} (phanerophyte)} \tab \code{50\%} \tab \code{50\%} 
      \tab \code{50\%} \tab \code{10\%}
    }
    \item from \strong{user data} : \cr
      \emph{with the values contained within the \code{immature_size} 
      column, if provided \cr \cr}
  }
  }
  \item{CHANG_STR_AGES}{= at what age each PFG goes into the upper stratum 
  \cr \cr It is defined using a logistic growth curve with 2 points to 
  parameterize it :
  \enumerate{
    \item at \eqn{age = \text{maturity}/2}, \eqn{height = \text{IMM_SIZE} * \text{height}}
    \item at \eqn{age = \text{longevity}}, \eqn{height = \text{height}} \cr \cr
  }
  }
  \item{POTENTIAL_FECUNDITY}{= maximum number of seeds produced by the PFG 
  \cr \cr
  Two methods to define this number are available :
  \itemize{
    \item from \strong{predefined rules} : same value for all PFG 
    (\code{100\%})
    \item from \strong{user data} : \cr
      \emph{with the values contained within the \code{potential_fecundity} 
      column, if provided \cr \cr}
  }
  }
  \item{IS_ALIEN}{= if the PFG is to be considered as an alien (\code{1}) or 
  not (\code{0})}
  \item{FLAMMABILITY}{= how easily the PFG burns}
}
}
\examples{

## Create a skeleton folder with the default name ('FATE_simulation')
PRE_FATE.skeletonDirectory()

## Create PFG succession parameter files
PRE_FATE.params_PFGsuccession(name.simulation = "FATE_simulation"
                              , mat.PFG.succ = data.frame(PFG = paste0("PFG", 1:6)
                                                          , type = c("C", "C", "H", "H", "P", "P")
                                                          , height = c(10, 250, 36, 68, 1250, 550)
                                                          , maturity = c(5, 5, 3, 3, 8, 9)
                                                          , longevity = c(12, 200, 25, 4, 110, 70)))
                                                        

## Create PFG succession parameter files
PRE_FATE.params_PFGsuccession(name.simulation = "FATE_simulation"
                              , mat.PFG.succ = data.frame(PFG = paste0("PFG", 1:6)
                                                          , type = c("C", "C", "H", "H", "P", "P")
                                                          , height = c(10, 250, 36, 68, 1250, 550)
                                                          , maturity = c(5, 5, 3, 3, 8, 9)
                                                          , longevity = c(12, 200, 25, 4, 110, 70)
                                                          , immature_size = c(10, 8, 10, 10, 1, 5)))
                                                        
                                                        
## ----------------------------------------------------------------------------------------- ##

## Load example data
                                                        

}
\seealso{
\code{\link{PRE_FATE.skeletonDirectory}}, 
\code{\link{PRE_FATE.params_globalParameters}}, 
\code{\link{PRE_FATE.params_PFGdisturbance}}
}
\author{
Maya Guéguen
}
\keyword{FATE,}
\keyword{height,}
\keyword{longevity,}
\keyword{maturity}
\keyword{simulation,}
