% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRE_FATE.params_saveYears.R
\name{PRE_FATE.params_saveYears}
\alias{PRE_FATE.params_saveYears}
\title{Create \emph{SAVE} parameter files for a \code{FATE-HD}
simulation}
\usage{
PRE_FATE.params_saveYears(name.simulation, years.maps = NULL,
  years.objects = NULL)
}
\arguments{
\item{name.simulation}{a \code{string} that corresponds to the main directory
or simulation name of the \code{FATE-HD} simulation}

\item{years.maps}{a \code{vector} of simulation years at which PFG abundance
maps will be saved}

\item{years.objects}{a \code{vector} of simulation years at which \code{FATE-HD}
simulation state will be saved}
}
\value{
Two \code{.txt} files into the \code{name.simulation/DATA/SAVE/}
directory with one line for each simulation year for which the raster maps /
\code{FATE-HD} object are to be saved.
}
\description{
This script is designed to create a parameter file containing
simulation years at which the \code{FATE-HD} software must save rasters of
PFG abundances or simulation objects.
}
\details{
\code{FATE-HD} software allows the user to save two different types of outputs :

\describe{
  \item{Raster maps}{PFG maps can be saved for all specified simulation years.
  It includes maps per PFG per strata and summary maps per PFG for all height
  strata combined. \cr Raster format used is depending on input data format. It
  can be either \code{.img} or \code{.tif}.}
  \item{Model objects}{using \code{BOOST} library and its serialization functions,
  \code{FATE-HD} is able to save a simulation at a specific time. This object
  allows the user to restart a simulation from this precise state.}
}
}
\examples{

## Create a skeleton folder with the default name ('FATE_simulation')
PRE_FATE.skeletonDirectory()

## Create a Namespace_constants parameter file
PRE_FATE.params_saveYears(name.simulation = "FATE_simulation"
                          , years.maps = c(100, 150, 200)
                          , years.objects = 200)

}
\author{
Maya Guéguen
}
