% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRE_FATE.params_savingYears.R
\name{PRE_FATE.params_savingYears}
\alias{PRE_FATE.params_savingYears}
\title{Create \emph{SAVE} parameter files for a \code{FATE} simulation}
\usage{
PRE_FATE.params_savingYears(
  name.simulation,
  years.maps = NULL,
  years.objects = NULL,
  opt.folder.name = NULL
)
}
\arguments{
\item{name.simulation}{a \code{string} corresponding to the main directory
or simulation name of the \code{FATE} simulation}

\item{years.maps}{(\emph{optional}) \cr a \code{vector} of simulation years 
at which PFG abundance maps will be saved \cr (\emph{as well as maps of 
light and soil resources if these modules are activated})}

\item{years.objects}{(\emph{optional}) \cr a \code{vector} of simulation 
years at which \code{FATE} simulation state will be saved}

\item{opt.folder.name}{(\emph{optional}) \cr a \code{string} corresponding 
to the name of the folder that will be created into the 
\code{name.simulation/DATA/SAVE/} directory to store the results}
}
\value{
Two \code{.txt} files into the \code{name.simulation/DATA/SAVE/}
directory :

 \itemize{
   \item \file{SAVE_YEARS_maps.txt} : one line for each simulation year for 
   which the raster maps are to be saved
   \item \file{SAVE_YEARS_objects.txt} : one line for each simulation year 
   for which the \code{FATE} objects are to be saved \cr \cr
 }

If the \code{opt.folder.name} has been used, the files will be into the 
folder \code{name.simulation/DATA/SAVE/opt.folder.name/}.
}
\description{
This script is designed to create a parameter file containing 
simulation years at which the \code{FATE} software must save rasters of 
PFG abundances (as well as light and soil resources if these modules are 
activated) and/or simulation objects.
}
\details{
\code{FATE} software allows the user to save two different types of 
outputs :

\describe{
  \item{Raster maps}{PFG abundance maps can be saved for all specified 
  simulation years. \cr It includes maps per PFG per strata 
  (\code{ABUND_perPFG_perStrata} folder) and summary maps 
  per PFG for all height strata combined (\code{ABUND_perPFG_allStrata} 
  folder). \cr If the light and / or soil modules are activated (see 
  \code{\link{PRE_FATE.params_globalParameters}}), maps for light and / or 
  soil resources are also saved. \cr Raster format used is depending on 
  input data format. It can be either \code{.img} or \code{.tif}.}
  \item{Model objects}{using \code{BOOST} library and its serialization 
  functions, \code{FATE} is able to save a simulation at a specific 
  time. This object allows the user to restart a simulation from this 
  precise state by specifying its name within the \emph{Simul_parameters} 
  file with the \code{SAVED_STATE} flag (see 
  \code{\link{PRE_FATE.params_simulParameters}}).}
}
}
\examples{

## Create a skeleton folder with the default name ('FATE_simulation')
PRE_FATE.skeletonDirectory()

## Create a SAVE_year_maps or/and SAVE_year_objects parameter file
PRE_FATE.params_savingYears(name.simulation = "FATE_simulation"
                          , years.maps = c(100, 150, 200)
                          , years.objects = 200)

}
\seealso{
\code{\link{PRE_FATE.skeletonDirectory}}, 
\code{\link{PRE_FATE.params_globalParameters}}
}
\author{
Maya Guéguen
}
\keyword{FATE,}
\keyword{simulation}
