% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRE_FATE.speciesClustering_step1.R
\name{PRE_FATE.speciesClustering_step1}
\alias{PRE_FATE.speciesClustering_step1}
\title{Create clusters based on dissimilarity matrix}
\usage{
PRE_FATE.speciesClustering_step1(mat.species.DIST)
}
\arguments{
\item{mat.species.DIST}{a \code{dist} object, or a \code{list} of 
\code{dist} objects (one for each \code{GROUP} value), corresponding to the 
distance between each pair of species. \cr Such an object can be obtained 
with the \code{\link{PRE_FATE.speciesDistance}} function.}
}
\value{
A \code{list} containing one \code{list} and one \code{data.frame} 
with the following columns :

\describe{
  \item{clust.dendrograms}{a \code{list} with as many objects of 
  class \code{\link[stats]{hclust}} as data subsets}
  \item{clust.evaluation}{ \cr
  \describe{
    \item{\code{GROUP}}{name of data subset}
    \item{\code{no.clusters}}{number of clusters used for the clustering}
    \item{\code{variable}}{evaluation metrics' name}
    \item{\code{value}}{value of evaluation metric}
  }
  }
}

Two \file{PRE_FATE_CLUSTERING_[...].pdf} files are created : 
\describe{
  \item{\file{STEP_1A \cr clusteringMethod}}{to account for the chosen 
  clustering method}
  \item{\file{STEP_1B \cr numberOfClusters}}{for decision support, to help 
  the user to choose the adequate number of clusters to be given to the 
  \code{\link{PRE_FATE.speciesClustering_step2}} function}
}
}
\description{
This script is designed to create clusters of species based 
on a distance matrix between those species. Several metrics are computed 
to evaluate these clusters and a graphic is produced to help the user to 
choose the best number of clusters..
}
\details{
This function allows to \strong{obtain dendrograms based on a distance 
matrix between species}.

As for the \code{\link{PRE_FATE.speciesDistance}} method, clustering can be 
run for data subsets, conditioning that \code{mat.species.DIST} is given as 
a \code{list} of \code{dist} objects (instead of a \code{dist} object alone).
\cr \cr

The process is as follows :

\describe{
  \item{\strong{1. Choice of the \cr optimal \cr clustering method}}{
  hierarchical clustering on the dissimilarity matrix is realized with the 
  \code{\link[stats]{hclust}}.
  \itemize{
    \item Several methods are available for the agglomeration : 
    \emph{complete}, \emph{ward.D}, \emph{ward.D2}, \emph{single}, 
    \emph{average (UPGMA)}, \emph{mcquitty (WPGMA)}, \emph{median (WPGMC)} 
    and \emph{centroid (UPGMC)}.
    \item \emph{Mouchet et al. (2008)} proposed a similarity measure between 
    the input distance and the one obtained with the clustering which must 
    be minimized to help finding the best clustering method :
    \deqn{ 1 - cor( \text{mat.species.DIST}, \text{clustering.DIST} ) ^ 2}
  }
  \strong{For each agglomeration method, this measure is calculated. The 
  method that minimizes it is kept and used for further analyses (see 
  \file{STEP_1A} output file). \cr \cr}
  }
  
  \item{\strong{2. Evaluation of the \cr clustering}}{once the hierarchical 
  clustering is done, the number of clusters to keep should be chosen. \cr 
  To do that, several metrics are computed :
  \itemize{
    \item{\emph{Dunn index (\code{mdunn}) : }}{ratio of the smallest 
    distance between observations not in the same cluster to the largest 
    intra-cluster distance. Value between \code{0} and \eqn{\infty}, and 
    should be maximized.}
    \item{\emph{Meila's Variation of Information index (\code{mVI}) : }}
    {measures the amount of information lost and gained in changing 
    between two clusterings. Should be minimized.}
    \item{\emph{Coefficient of determination (\code{R2}) : }}{value 
    between \code{0} and \code{1}. Should be maximized.}
    \item{\emph{Calinski and Harabasz index (\code{ch}) : }}{the higher 
    the value, the "better" is the solution.}
    \item{\emph{Corrected rand index (\code{Rand}) : }}{measures the 
    similarity between two data clusterings. Value between \code{0} and 
    \code{1}, with \code{0} indicating that the two data clusters do not 
    agree on any pair of points and \code{1} indicating that the data 
    clusters are exactly the same.}
    \item{\emph{Average silhouette width (\code{av.sil}) : }}{Observations 
    with a large \code{s(i)} (almost \code{1}) are very well clustered, a 
    small \code{s(i)} (around \code{0}) means that the observation lies 
    between two clusters, and observations with a negative \code{s(i)} are 
    probably placed in the wrong cluster. Should be maximized.}
  }
  \strong{A graphic is produced, giving the values of these metrics in 
  function of the number of clusters used. Number of clusters with 
  evaluation metrics' values among the 3 best are highlighted to help the 
  user to make his/her optimal choice (see \file{STEP_1B} output file).}
  }
}

\emph{\cr \cr
Mouchet M., Guilhaumon f., Villeger S., Mason N.W.H., Tomasini J.A. & 
Mouillot D., 2008. Towards a consensus for calculating dendrogam-based 
functional diversity indices. Oikos, 117, 794-800.}
}
\note{
\strong{The function does not return ONE dendrogram} (or as many as 
given dissimilarity structures) \strong{but a LIST with all tested numbers 
of clusters.} One final dendrogram can then be obtained using this result 
as a parameter in the \code{\link{PRE_FATE.speciesClustering_step2}} 
function.
}
\examples{

## Load example data
PNE_PFG = .loadData("PNE_PFG")

## PNE_PFG$dom.traits : data.frame
## PNE_PFG$dom.dist_overlap : niolap object

## Calculate distance between species
sp.DIST = PRE_FATE.speciesDistance(mat.species.traits = PNE_PFG$dom.traits
                                   , mat.species.overlap = PNE_PFG$dom.dist_overlap
                                   , opt.max.percent.NA = 0.9
                                   , opt.max.percent.similarSpecies = 0.25
                                   , opt.min.sd = 0.3)
                                   
str(sp.DIST)
             
## Run hierarchical clustering and number of clusters' selection                      
sp.CLUST = PRE_FATE.speciesClustering_step1(mat.species.DIST = sp.DIST)

## 2 .pdf files have been produced
str(sp.CLUST)

}
\seealso{
\code{\link[stats]{hclust}},
\code{\link[stats]{cutree}},
\code{\link[fpc]{cluster.stats}},
\code{\link[clValid]{dunn}},
\code{\link{PRE_FATE.speciesDistance}},
\code{\link{PRE_FATE.speciesClustering_step2}}
}
\author{
Maya Guéguen
}
\keyword{Average}
\keyword{Calinski}
\keyword{Corrected}
\keyword{Dunn}
\keyword{Harabasz}
\keyword{Information}
\keyword{Meila's}
\keyword{R2,}
\keyword{Variation}
\keyword{and}
\keyword{clustering,}
\keyword{hierarchical}
\keyword{index,}
\keyword{of}
\keyword{rand}
\keyword{silhouette}
\keyword{width}
