% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRE_FATE.speciesClustering_step2.R
\name{PRE_FATE.speciesClustering_step2}
\alias{PRE_FATE.speciesClustering_step2}
\title{Choose clusters and select determinant species}
\usage{
PRE_FATE.speciesClustering_step2(
  clust.dendrograms,
  no.clusters,
  mat.species.DIST
)
}
\arguments{
\item{clust.dendrograms}{a dendrogram, or a \code{list} of dendrograms (one 
for each \code{GROUP} value). \cr Such an object can be obtained 
with the \code{\link{PRE_FATE.speciesClustering_step1}} function.}

\item{no.clusters}{an \code{integer}, or a \code{vector} of \code{integer} 
(one for each \code{GROUP} value), with the number of clusters to be kept}

\item{mat.species.DIST}{a \code{dist} object, or a \code{list} of 
\code{dist} objects (one for each \code{GROUP} value), corresponding to the 
distance between each pair of species. \cr Such an object can be obtained 
with the \code{\link{PRE_FATE.speciesDistance}} function.}
}
\value{
A \code{list} containing one \code{vector} and one 
\code{data.frame} with the following columns :

\describe{
  \item{determ.sp}{the names of all determinant species}
  \item{determ.all}{(\emph{determinant and non-determinant species})
  \describe{
    \item{\code{PFG}}{ID of the plant functional group 
    (\code{GROUP} + \code{ID.cluster})}
    \item{\code{GROUP}}{name of data subset}
    \item{\code{ID.cluster}}{cluster number}
    \item{\code{species}}{name of species}
    \item{\code{ID.species}}{species number in each PFG}
    \item{\code{sp.mean.dist}}{species mean distance to other species of 
    the same PFG}
    \item{\code{allSp.mean}}{\eqn{mean(\text{sp.mean.dist})} within the PFG}
    \item{\code{allSp.min}}{\eqn{mean(\text{sp.mean.dist}) - 1.64 * 
    sd(\text{sp.mean.dist})} within the PFG}
    \item{\code{allSp.max}}{\eqn{mean(\text{sp.mean.dist}) + 1.64 * 
    sd(\text{sp.mean.dist})} within the PFG}
    \item{\code{DETERMINANT}}{\code{TRUE} if determinant species, \code{FALSE} 
    otherwise}
  }
  }
}

Two \file{PRE_FATE_CLUSTERING_[...].pdf} files are created : 
\describe{
  \item{\file{STEP_2C \cr distantSpecies}}{to visualize in each PFG the 
  distribution of mean distance of each species to other species, and 
  non-determinant species which are outside the distribution}
  \item{\file{STEP_2C_PCO}}{to visualize in each PFG the distribution 
  of species, with and without non-determinant species}
}
}
\description{
This script is designed to obtain functional groups by : 1) 
selecting the number of clusters to be kept from an object obtained with 
the \code{\link{PRE_FATE.speciesClustering_step1}} function ; 2) refining 
these groups by identifying determinant species in each of them.
}
\details{
This function allows to obtain a classification of \strong{dominant} 
species into Plant Functional Groups (PFG), and the \strong{determinant} 
species based on these PFG. \cr \cr

\strong{What is the difference between \code{dominant} and
\code{determinant} species ?}

\itemize{
  \item{\strong{Dominant} species are species representative of an 
  environment or a studied area, in terms of number of releves or 
  abundance values. They can be found with the 
  \code{\link{PRE_FATE.selectDominant}} function of this package. These 
  dominant species are used to build PFG with the 
  \code{\link{PRE_FATE.speciesClustering_step1}} function.
  }
  \item{Once PFG are built, \strong{determinant} species are defined as 
  refined subsets of dominant species within each PFG. \cr The process is 
  detailed below :
  \itemize{
    \item each dominant species is assigned to a PFG
    \item within each PFG :
    \itemize{
      \item for each species, compute its mean distance to the other 
      species within the PFG (\code{sp.mean.dist})
      \item calculate the mean value of all these mean distances
      (\code{allSp.mean})
      \item calculate the deviation values around this mean value
      (\code{allSp.min} and \code{allSp.max})
      \item determinant species are the ones that are included between 
      those deviation values
    }
  }
  }
}
}
\examples{

## Load example data
PNE_PFG = .loadData("PNE_PFG")

## PNE_PFG$dom.traits : data.frame
## PNE_PFG$dom.dist_overlap : niolap object

## Calculate distance between species
sp.DIST = PRE_FATE.speciesDistance(mat.species.traits = PNE_PFG$dom.traits
                                   , mat.species.overlap = PNE_PFG$dom.dist_overlap
                                   , opt.max.percent.NA = 0.9
                                   , opt.max.percent.similarSpecies = 0.25
                                   , opt.min.sd = 0.3)
                                   
str(sp.DIST)
             
## Run hierarchical clustering and number of clusters' selection                      
sp.CLUST = PRE_FATE.speciesClustering_step1(mat.species.DIST = sp.DIST)

## 2 .pdf files have been produced
str(sp.CLUST)

## Select number of clusters and find determinant species
sp.DETERM = PRE_FATE.speciesClustering_step2(clust.dendrograms = sp.CLUST$clust.dendrograms
                                             , no.clusters = PNE_PFG$nb.clusters
                                             , mat.species.DIST = sp.DIST)

## 2 .pdf files have been produced
str(sp.DETERM)

}
\seealso{
\code{\link[stats]{cutree}},
\code{\link[ade4]{quasieuclid}},
\code{\link[ade4]{dudi.pco}},
\code{\link{PRE_FATE.speciesDistance}},
\code{\link{PRE_FATE.speciesClustering_step1}},
\code{\link{PRE_FATE.speciesClustering_step3}}
}
\author{
Maya Guéguen
}
\keyword{Component}
\keyword{Ordination}
\keyword{Principal}
\keyword{clustering,}
\keyword{hierarchical}
