% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRE_FATE.speciesClustering_step3.R
\name{PRE_FATE.speciesClustering_step3}
\alias{PRE_FATE.speciesClustering_step3}
\title{Calculate PFG traits values based on determinant species traits 
values}
\usage{
PRE_FATE.speciesClustering_step3(mat.traits, opt.mat.PA = NULL)
}
\arguments{
\item{mat.traits}{a \code{data.frame} with at least 3 columns :
\describe{
  \item{\code{species}}{the ID of each determinant species (see 
  \code{\link{PRE_FATE.speciesClustering_step2}})}
  \item{\code{PFG}}{the corresponding Plant Functional Group (see 
  \code{\link{PRE_FATE.speciesClustering_step2}})}
  \item{\code{...}}{one column for each functional trait (see 
  \href{PRE_FATE.speciesClustering_step3#details}{\code{Details}})}
}}

\item{opt.mat.PA}{(\emph{optional}) default \code{NULL}. \cr 
a \code{data.frame} with sites in rows and species in columns, 
containing either \code{NA}, \code{0} or \code{1} (see 
\code{\link{PRE_FATE.selectDominant}})}
}
\value{
A \code{list} containing one or two \code{data.frame} with the 
following columns, and one \code{list} with as many \code{ggplot2} objects 
as functional traits given in \code{mat.traits} :

\describe{
  \item{tab.PFG.traits}{ \cr
  \describe{
    \item{\code{PFG}}{the concerned plant functional group}
    \item{\code{no.species}}{the number of species contained in this PFG}
    \item{\code{...}}{one column for each functional trait, computed as the 
    \code{mean} (for numeric traits) or the \code{median} (for categorical 
    traits) of the values of the determinant species of this PFG}
  }
  }
  \item{tab.PFG.PA}{table containing counts of presences for all PFG 
  (sites in rows, PFG in columns)}
  \item{plot}{\cr
  \describe{
    \item{\code{...}}{one for each functional trait, 'specific' cases 
    excepted (see 
    \href{PRE_FATE.speciesClustering_step3#details}{\code{Details}})}
  }
  }
}
    
The information is written in \file{PRE_FATE_PFG_TABLE_traits.csv}, 
\file{PRE_FATE_PFG_TABLE_sitesXPFG_PA.csv} and 
\file{PRE_FATE_CLUSTERING_STEP_3_PFGtraitsValues.pdf} files. \cr
This \code{.csv} file can be used to build parameter files to run a 
\code{FATE} simulation (e.g. \code{\link{PRE_FATE.params_PFGsuccession}}).
}
\description{
This script is designed to calculate PFG traits values based 
on determinant species traits values. Either the mean or the 
median is used depending on the trait class (i.e. numeric or 
categorical).
}
\details{
This function allows to \strong{obtain '\emph{average}' functional trait 
values for each Plant Functional Group}, based on values at the determinant 
species level. \cr \cr


\emph{A graphic is automatically produced for each functional 
trait given, with boxplot representing the values of determinant species, 
and colored points the values calculated for each PFG. \cr
However, some traits can have 'specific' representation, as long as their 
names within \code{mat.traits} match one of the configuration detailed 
below :}
\describe{
  \item{\code{maturity}, \code{longevity}}{to visualize the difference 
  between these two values, for the maturity time has an impact on the 
  fecundity of the PFG within \code{FATE} (see 
  \href{../articles/fate_tutorial_3_MODULES.html#core-module-succession}{CORE module})
  \cr \emph{If there is NO values for longevity within one PFG, and some 
  maturity values are available, some values might be inferred as 
  \eqn{\text{maturity} * 2}. If there is NO values for maturity within 
  one PFG, and some longevity values are available, some values might be 
  inferred as \eqn{\text{longevity} / 2}.}
  }
  \item{\code{height}, \code{light}}{to visualize the PFG light 
  preference, and help decide and understand the choice of the height 
  limits of strata in \code{FATE} (see 
  \href{../articles/fate_tutorial_3_MODULES.html#light-module-interaction}{LIGHT 
  interaction module})}
  \item{\code{soil_contrib}, \cr \code{soil_tol_min}, \cr 
  \code{soil_tol_max}}{to visualize the PFG soil preference, and help 
  parameterize the global parameters of the soil interaction module 
  within \code{FATE} (see 
  \href{../articles/fate_tutorial_3_MODULES.html#soil-module-interaction}{SOIL 
  interaction module})}
  \item{\cr \code{soil_contrib}, \cr \code{soil_tolerance}}{same as the 
  previous one, but \code{soil_tol_min} and \code{soil_tol_max} values 
  are obtained by adding or removing \code{soil_tolerance} to 
  \code{soil_contrib} \cr \cr}
}

If a sites x species table is provided (\code{opt.mat.PA}), a sites x PFG 
table (containing \code{NA}, \code{0} or \code{1}) is also returned.
}
\examples{

## Load example data
.loadData("Champsaur_PFG")

## Species traits
tab.traits = Champsaur_PFG$sp.traits
str(tab.traits)

## Determinant species
tab.PFG = Champsaur_PFG$PFG.species
str(tab.PFG)

## Merge traits and PFG informations
mat.traits = merge(tab.PFG[which(tab.PFG$DETERMINANT==TRUE), c("species","PFG")]
                   , tab.traits
                   , by = "species", all.x = TRUE)
str(mat.traits)

## Keep only traits of interest
mat.traits = mat.traits[, c("PFG", "species", "MATURITY", "LONGEVITY"
                           , "HEIGHT", "LIGHT", "DISPERSAL"
                           , "NITROGEN", "NITROGEN_TOLERANCE", "LDMC", "LNC")]
colnames(mat.traits) = c("PFG", "species", "maturity", "longevity"
                         , "height", "light", "dispersal"
                         , "soil_contrib", "soil_tolerance", "LDMC", "LNC")
mat.traits$soil_contrib = as.numeric(mat.traits$soil_contrib)
mat.traits$soil_tolerance = ifelse(mat.traits$soil_tolerance == 1, 0.5, 1)

## Compute traits per PFG : with one specific graphic ------------------------
PFG.traits = PRE_FATE.speciesClustering_step3(mat.traits = mat.traits)

names(PFG.traits)
str(PFG.traits$tab.PFG.traits)
str(PFG.traits$tab.PFG.PA)
names(PFG.traits$plot)
plot(PFG.traits$plot$maturity_longevity)
plot(PFG.traits$plot$height_light)
plot(PFG.traits$plot$soil)





}
\seealso{
\code{\link{PRE_FATE.speciesClustering_step1}},
\code{\link{PRE_FATE.speciesClustering_step2}}
}
\author{
Maya Guéguen
}
\keyword{functional}
\keyword{group,}
\keyword{traits}
