% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRE_FATE.speciesDistance.R
\name{PRE_FATE.speciesDistance}
\alias{PRE_FATE.speciesDistance}
\title{Computation of distances between species based on traits and niche 
overlap}
\usage{
PRE_FATE.speciesDistance(
  mat.traits,
  mat.overlap.option,
  mat.overlap.object,
  opt.weights = NULL,
  opt.maxPercent.NA = 0,
  opt.maxPercent.similarSpecies = 0.25,
  opt.min.sd = 0.3
)
}
\arguments{
\item{mat.traits}{a \code{data.frame} with at least 3 columns :
\describe{
\item{\code{species}}{the ID of each studied species}
\item{\code{GROUP}}{a factor variable containing grouping information to 
divide the species into data subsets (see 
\href{PRE_FATE.speciesDistance#details}{\code{Details}})}
\item{\code{...}}{one column for each functional trait}
}}

\item{mat.overlap.option}{a \code{string} corresponding to the way to 
calculate the distance between species based on niche overlap (either 
\code{PCA}, \code{raster} or \code{dist}, see 
\href{PRE_FATE.speciesDistance#details}{\code{Details}})}

\item{mat.overlap.object}{three options, depending on the value of 
\code{mat.overlap.option} :
\itemize{
  \item (\code{PCA} option) a \code{list} with 2 elements :
  \describe{
    \item{\code{tab.dom.PA}}{a \code{matrix} or \code{data.frame} with 
    sites in rows and species in columns, containing either \code{NA}, 
    \code{0} or \code{1} (see \code{\link{PRE_FATE.selectDominant}})}
    \item{\code{tab.env}}{a \code{matrix} or \code{data.frame} with 
    sites in rows and environmental variables in columns}
  }
  \item (\code{raster} option) a \code{data.frame} with 2 columns :
  \describe{
    \item{\code{species}}{the ID of each studied species}
    \item{\code{raster}}{path to raster file with species distribution}
  }
  \item (\code{dist} option) a similarity structure representing the 
  niche overlap between each pair of species. It can be a \code{dist} 
  object, a \code{niolap} object, or simply a \code{matrix}.
}}

\item{opt.weights}{(\emph{optional}) default \code{NULL}. \cr 
A \code{vector} of two \code{double} (between \code{0} and \code{1}) 
corresponding to the weights for traits and overlap distances 
respectively. They must sum up to \code{1}.}

\item{opt.maxPercent.NA}{(\emph{optional}) default \code{0}. \cr Maximum 
percentage of missing values (\code{NA}) allowed for each trait (between 
\code{0} and \code{1})}

\item{opt.maxPercent.similarSpecies}{(\emph{optional}) default \code{0.25}. 
\cr Maximum percentage of similar species (same value) 
allowed for each trait (between \code{0} and \code{1})}

\item{opt.min.sd}{(\emph{optional}) default \code{0.5}. \cr Minimum 
standard deviation allowed for each trait (trait unit)}
}
\value{
A \code{list} of 3 \code{dist} objects (functional distances, 
overlap distances, and combination of both according to the weights given 
(or not) by the \code{opt.weights} parameter), each of them corresponding 
to : the distance between each pair of species, or a \code{list} of 
\code{dist} objects, one for each \code{GROUP} value. \cr \cr

The information for the combination of both distances is written in 
\file{PRE_FATE_DOMINANT_speciesDistance.csv} file (or if necessary, one 
file is created for each group).
}
\description{
This script is designed to create a distance matrix between 
species, combining functional distances (based on functional trait values) 
and niche overlap (based on co-occurrence of species).
}
\details{
This function allows to obtain a \strong{distance matrix between species}, 
based on two types of distance information :

\enumerate{
  \item \strong{Functional traits : }
  \itemize{
    \item The \code{GROUP} column is required if species must be separated 
    to have one final distance matrix per \code{GROUP} value. \cr If the 
    column is missing, all species will be considered as part of a unique 
    dataset.
    \item The traits can be qualitative or quantitative, but previously 
    identified as such \cr (i.e. with the use of functions such as 
    \code{as.numeric}, \code{as.factor} and \code{ordered}).
    \item Functional distance matrix is calculated with Gower dissimilarity, 
    using the \code{\link[FD]{gowdis}} function.
    \item This function allows \code{NA} values. \cr However, too many 
    missing values lead to misleading results. Hence, 3 parameters allow the 
    user to play with the place given to missing values, and therefore the 
    selection of traits that will be used for the distance computation :
    \describe{
      \item{opt.maxPercent.NA}{traits with too many missing values are 
      removed}
      \item{opt.maxPercent \cr .similarSpecies}{traits with too many 
      similar values are removed}
      \item{opt.min.sd}{traits with too little variability are removed}
    }
  }
  \item \strong{Niche overlap : }
  \itemize{
    \item If \code{PCA} option is selected, the degree of niche overlap will 
    be computed using the \code{\link[ecospat]{ecospat.niche.overlap}}. 
    \item If \code{raster} option is selected, the degree of niche overlap will 
    be computed using the 
    \code{\href{niche.overlap}{https://www.rdocumentation.org/packages/phyloclim/versions/0.9-4/topics/niche.overlap}}
    . \cr \cr \cr
  }
}

Functional distances and niche overlap informations are then 
\strong{combined} according to the following formula :

\deqn{\text{mat.DIST}_{sub-group} = \frac{[\text{wei.FUNC} * 
\text{mat.FUNCTIONAL}_{sub-group} + \text{wei.OVER} * 
\text{mat.OVERLAP}_{sub-group}]}{[ \text{wei.FUNC} + \text{wei.OVER} ]}}

with :

\deqn{\text{wei.FUNC} = \text{opt.weights}[1]}
\deqn{\text{wei.OVER} = \text{opt.weights}[2]}

if \code{opt.weights} is given, otherwise :

\deqn{\text{wei.FUNC} = n_{traits}}
\deqn{\text{wei.OVER} = 1}

meaning that \strong{distance matrix obtained from functional information 
is weighted by the number of traits used}.
}
\examples{

## Load example data
Champsaur_PFG = .loadData('Champsaur_PFG', 'RData')

## Species traits
tab.traits = Champsaur_PFG$sp.traits
tab.traits = tab.traits[, c('species', 'GROUP', 'MATURITY', 'LONGEVITY'
                            , 'HEIGHT', 'DISPERSAL', 'LIGHT', 'NITROGEN')]
str(tab.traits)

## Species niche overlap (dissimilarity distances)
tab.overlap = 1 - Champsaur_PFG$mat.overlap ## transform into similarity
tab.overlap[1:5, 1:5]

## Give warnings -------------------------------------------------------------
sp.DIST = PRE_FATE.speciesDistance(mat.traits = tab.traits
                                   , mat.overlap.option = 'dist'
                                   , mat.overlap.object = tab.overlap)
str(sp.DIST)

## Change parameters to allow more NAs (and change traits used) --------------
sp.DIST = PRE_FATE.speciesDistance(mat.traits = tab.traits
                                   , mat.overlap.option = 'dist'
                                   , mat.overlap.object = tab.overlap
                                   , opt.maxPercent.NA = 0.05
                                   , opt.maxPercent.similarSpecies = 0.3
                                   , opt.min.sd = 0.3)
str(sp.DIST)

\dontrun{
require(foreach); require(ggplot2); require(ggdendro)
pp = foreach(x = names(sp.DIST$mat.ALL)) \%do\%
  {
    hc = hclust(sp.DIST$mat.ALL[[x]])
    pp = ggdendrogram(hc, rotate = TRUE) +
      labs(title = paste0('Hierarchical clustering based on species distance '
                          , ifelse(length(names(sp.DIST$mat.ALL)) > 1
                                   , paste0('(group ', x, ')')
                                   , '')))
    return(pp)
  }
plot(pp[[1]])
plot(pp[[2]])
plot(pp[[3]])
}


}
\seealso{
\code{\link[FD]{gowdis}},
\code{\link[ecospat]{ecospat.niche.overlap}}, 
\code{\href{niche.overlap}{https://www.rdocumentation.org/packages/phyloclim/versions/0.9-4/topics/niche.overlap}}
}
\author{
Maya Guéguen
}
\keyword{Gower}
\keyword{distance,}
\keyword{functional}
\keyword{niche}
\keyword{overlap}
\keyword{traits,}
