% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRE_FATE.speciesDistanceCombine.R
\name{PRE_FATE.speciesDistanceCombine}
\alias{PRE_FATE.speciesDistanceCombine}
\title{Combine several dissimilarity distance matrices}
\usage{
PRE_FATE.speciesDistanceCombine(list.mat.dist, opt.weights = NULL)
}
\arguments{
\item{list.mat.dist}{a \code{list} of matrices containing dissimilarity 
distance values between each pair of species.}

\item{opt.weights}{(\emph{optional}) default \code{NULL}. \cr 
A \code{vector} of \code{double} (between \code{0} and \code{1}) 
corresponding to the weights for each distance matrix provided in 
\code{list.mat.dist}. They must sum up to \code{1}.}
}
\value{
A \code{matrix} containing the weighted (or not) combination of the 
different distance matrices given. \cr \cr

The information for the combination of all distances is written in 
\file{PRE_FATE_DOMINANT_speciesDistance.csv} file.
}
\description{
This script is designed to create a distance matrix between 
species, combining functional distances (based on functional trait values) 
and niche overlap (based on co-occurrence of species).
}
\details{
This function allows to obtain a \strong{distance matrix between species}, 
based on several dissimilarity distance matrices \strong{combined} 
according to the following formula :

\deqn{\text{mat.DIST} = \Sigma (\text{wei.i} * \text{mat.DIST}_{i})}
}
\examples{

## Load example data
Champsaur_PFG = .loadData('Champsaur_PFG', 'RData')

## Species traits
tab.traits = Champsaur_PFG$sp.traits
tab.traits = tab.traits[, c('species', 'GROUP', 'MATURITY', 'LONGEVITY'
                            , 'HEIGHT', 'DISPERSAL', 'LIGHT', 'NITROGEN')]
str(tab.traits)

## Species niche overlap (dissimilarity distances)
DIST.overlap = Champsaur_PFG$mat.overlap
DIST.overlap[1:5, 1:5]

## Species functional distances (dissimilarity)
DIST.traits = PRE_FATE.speciesDistanceTraits(mat.traits = tab.traits
                                             , opt.maxPercent.NA = 0.05
                                             , opt.maxPercent.similarSpecies = 0.3
                                             , opt.min.sd = 0.3)
DIST.traits$Chamaephyte[1:5, 1:5]

## Combine distances ---------------------------------------------------------
list.DIST = list(DIST.overlap, DIST.traits$Chamaephyte)
sp.DIST = PRE_FATE.speciesDistanceCombine(list.mat.dist = list.DIST
                                          , opt.weights = c(0.2, 0.8))
str(sp.DIST)

\dontrun{
require(foreach); require(ggplot2); require(ggdendro)
hc = hclust(as.dist(sp.DIST))
pp = ggdendrogram(hc, rotate = TRUE) +
  labs(title = 'Hierarchical clustering based on species distances')
plot(pp)
}


}
\author{
Maya Guéguen
}
\keyword{distance}
\keyword{functional}
