% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRE_FATE.speciesDistanceTraits.R
\name{PRE_FATE.speciesDistanceTraits}
\alias{PRE_FATE.speciesDistanceTraits}
\title{Computation of traits distances between species}
\usage{
PRE_FATE.speciesDistanceTraits(
  mat.traits,
  opt.maxPercent.NA = 0,
  opt.maxPercent.similarSpecies = 0.25,
  opt.min.sd = 0.3
)
}
\arguments{
\item{mat.traits}{a \code{data.frame} with at least 3 columns :
\describe{
\item{\code{species}}{the ID of each studied species}
\item{\code{GROUP}}{a factor variable containing grouping information to 
divide the species into data subsets (see 
\href{PRE_FATE.speciesDistance#details}{\code{Details}})}
\item{\code{...}}{one column for each functional trait}
}}

\item{opt.maxPercent.NA}{(\emph{optional}) default \code{0}. \cr Maximum 
percentage of missing values (\code{NA}) allowed for each trait (between 
\code{0} and \code{1})}

\item{opt.maxPercent.similarSpecies}{(\emph{optional}) default \code{0.25}. 
\cr Maximum percentage of similar species (same value) 
allowed for each trait (between \code{0} and \code{1})}

\item{opt.min.sd}{(\emph{optional}) default \code{0.5}. \cr Minimum 
standard deviation allowed for each trait (trait unit)}
}
\value{
A \code{matrix} containing functional distances between each pair 
of species, calculated as \code{1 - Schoeners D}.
}
\description{
This script is designed to create a distance matrix between 
species, based on functional trait values.
}
\details{
This function allows to obtain a \strong{distance matrix between species} 
(1 - Schoeners D), based on functional traits information :

\itemize{
  \item The \code{GROUP} column is required if species must be separated 
  to have one final distance matrix per \code{GROUP} value. \cr If the 
  column is missing, all species will be considered as part of a unique 
  dataset.
  \item The traits can be qualitative or quantitative, but previously 
  identified as such \cr (i.e. with the use of functions such as 
  \code{as.numeric}, \code{as.factor} and \code{ordered}).
  \item Functional distance matrix is calculated with Gower dissimilarity, 
  using the \code{\link[FD]{gowdis}} function.
  \item This function allows \code{NA} values. \cr However, too many 
  missing values lead to misleading results. Hence, 3 parameters allow the 
  user to play with the place given to missing values, and therefore the 
  selection of traits that will be used for the distance computation :
  \describe{
    \item{opt.maxPercent.NA}{traits with too many missing values are 
    removed}
    \item{opt.maxPercent \cr .similarSpecies}{traits with too many 
    similar values are removed}
    \item{opt.min.sd}{traits with too little variability are removed}
  }
}
}
\examples{

## Load example data
Champsaur_PFG = .loadData('Champsaur_PFG', 'RData')

## Species traits
tab.traits = Champsaur_PFG$sp.traits
tab.traits = tab.traits[, c('species', 'GROUP', 'MATURITY', 'LONGEVITY'
                            , 'HEIGHT', 'DISPERSAL', 'LIGHT', 'NITROGEN')]
str(tab.traits)

## Give warnings -------------------------------------------------------------
DIST.traits = PRE_FATE.speciesDistanceTraits(mat.traits = tab.traits)
str(DIST.traits)

## Change parameters to allow more NAs (and change traits used) --------------
DIST.traits = PRE_FATE.speciesDistanceTraits(mat.traits = tab.traits
                                             , opt.maxPercent.NA = 0.05
                                             , opt.maxPercent.similarSpecies = 0.3
                                             , opt.min.sd = 0.3)
str(DIST.traits)

\dontrun{
require(foreach); require(ggplot2); require(ggdendro)
pp = foreach(x = names(DIST.traits)) \%do\%
  {
    hc = hclust(as.dist(DIST.traits[[x]]))
    pp = ggdendrogram(hc, rotate = TRUE) +
      labs(title = paste0('Hierarchical clustering based on species distance '
                          , ifelse(length(names(DIST.traits)) > 1
                                   , paste0('(group ', x, ')')
                                   , '')))
    return(pp)
  }
plot(pp[[1]])
plot(pp[[2]])
plot(pp[[3]])
}


}
\seealso{
\code{\link[FD]{gowdis}}
}
\author{
Maya Guéguen
}
\keyword{Gower}
\keyword{distance}
\keyword{functional}
\keyword{traits,}
