% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFATE.R
\name{RFATE}
\alias{RFATE}
\title{Shiny application to apply \code{RFate} functions and run 
\code{FATE} simulation}
\usage{
RFATE()
}
\description{
This \code{shiny} application allows to use all 
the \code{RFate} functions (\code{PRE_FATE}, \code{FATE} and 
\code{POST_FATE}), from the building of PFG to the treatment of \code{FATE} 
output files.
}
\author{
Maya Guéguen
}
\keyword{GUI}
\keyword{application,}
\keyword{interface,}
\keyword{shiny}
