% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFATE.run.R
\name{RFATE.run}
\alias{RFATE.run}
\title{Shiny application to apply \code{RFate} functions and run 
\code{FATE} simulation}
\usage{
RFATE.run()
}
\description{
This \code{\link[shiny]{shiny}} application allows to use all 
the \code{RFate} functions (\code{PRE_FATE}, \code{FATE} and 
\code{POST_FATE}), from the building of PFG to the treatment of \code{FATE} 
output files.
}
\author{
Maya Guéguen
}
\keyword{GUI}
\keyword{application,}
\keyword{interface,}
\keyword{shiny}
