% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFATE_SHINY.run.R
\name{RFATE_SHINY.run}
\alias{RFATE_SHINY.run}
\title{Shiny application to apply \code{RFate} functions}
\usage{
RFATE_SHINY.run()
}
\description{
This SHINY application allows to use all the \code{RFate} 
functions (PRE and POST FATE), from the building of PFG to the treatment of
FATE output files.
}
\author{
Maya Guéguen
}
\keyword{GUI}
\keyword{application,}
\keyword{interface,}
\keyword{shiny}
