% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{runFate}
\alias{runFate}
\title{FATE-HD Wrapper}
\usage{
runFate(paramFile, nbCpus = 1L, verboseLevel = 0L)
}
\arguments{
\item{paramFile}{path to the parameter file of the simulation.}

\item{nbCpus}{number of CPUs affected to the simulation (default 1).}

\item{verboseLevel}{logger verbose level.}
}
\value{
None
}
\description{
This function runs a FATE-HD dynamical landscape vegetation simulation.
}
\examples{
\dontrun{runFate()}

}
\references{
\url{http://www.will.chez-alice.fr/pdf/BoulangeatGCB2014.pdf}
}
\author{
Isabelle Boulangeat, \email{isabelle.boulangeat@inrae.fr}

Maya Gueguen, \email{maya.gueguen@univ-grenoble-alpes.fr}
}
