% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAVE_FATE.step1_PFG.R
\name{SAVE_FATE.step1_PFG}
\alias{SAVE_FATE.step1_PFG}
\title{Save data to reproduce building of Plant Functional Groups}
\usage{
SAVE_FATE.step1_PFG(
  name.dataset,
  mat.observations,
  rules.selectDominant = c(doRuleA = NA, rule.A1 = NA, rule.A2_quantile = NA, doRuleB =
    NA, rule.B1_percentage = NA, rule.B1_number = NA, rule.B2 = NA, doRuleC = NA),
  mat.traits,
  mat.overlap = NA,
  rules.speciesDistance = c(opt.maxPercent.NA = NA, opt.maxPercent.similarSpecies = NA,
    opt.min.sd = NA),
  mat.species.DIST,
  clust.evaluation = NA,
  no.clusters,
  determ.all,
  mat.traits.PFG
)
}
\arguments{
\item{name.dataset}{a \code{string} corresponding to the name to give to 
archive folder}

\item{mat.observations}{a \code{data.frame} with at least 3 columns : \cr
\code{sites}, \code{species}, \code{abund} (\emph{and optionally, 
\code{habitat}}) (see \code{\link{PRE_FATE.selectDominant}})}

\item{rules.selectDominant}{(\emph{optional}) default \code{NA}. \cr A 
\code{vector} containing all the parameter values given to the 
\code{\link{PRE_FATE.selectDominant}} function, if used 
(\code{doRuleA}, \code{rule.A1}, \code{rule.A2_quantile}, \code{doRuleB}, 
\code{rule.B1_percentage}, \code{rule.B1_number}, \code{rule.B2}, 
\code{doRuleC}).}

\item{mat.traits}{a \code{data.frame} with at least 3 columns :
\code{species}, \code{GROUP}, \code{...} (one column for each functional 
trait) \cr (see \code{\link{PRE_FATE.speciesDistance}})}

\item{mat.overlap}{(\emph{optional}) default \code{NA}. \cr 
Otherwise, two options :
\itemize{
  \item a \code{data.frame} with 2 columns : \code{species}, \code{raster}
  \item a dissimilarity structure representing the niche overlap between 
  each pair of species. \cr It can be a \code{dist} object, a 
  \code{niolap} object, or simply a \code{matrix}.
}
(see \code{\link{PRE_FATE.speciesDistance}})}

\item{rules.speciesDistance}{(\emph{optional}) default \code{NA}. \cr A 
\code{vector} containing all the parameter values given to the 
\code{\link{PRE_FATE.speciesDistance}} function, if used \cr (
\code{opt.maxPercent.NA}, \code{opt.maxPercent.similarSpecies}, 
\code{opt.min.sd}).}

\item{mat.species.DIST}{a \code{dist} object, or a \code{list} of 
\code{dist} objects (one for each \code{GROUP} value), corresponding to the 
distance between each pair of species. \cr Such an object can be obtained 
with the \code{\link{PRE_FATE.speciesDistance}} function.}

\item{clust.evaluation}{(\emph{optional}) default \code{NA}. \cr A 
\code{data.frame} with 4 columns : \cr
\code{GROUP}, \code{no.clusters}, \code{variable}, \code{value}. \cr Such an 
object can be obtained with the 
\code{\link{PRE_FATE.speciesClustering_step1}} function.}

\item{no.clusters}{an \code{integer}, or a \code{vector} of \code{integer} 
(one for each \code{GROUP} value), with the number of clusters to be kept 
(see \code{\link{PRE_FATE.speciesClustering_step2}})}

\item{determ.all}{a \code{data.frame} with 6 or 10 columns : \cr
\code{PFG}, \code{GROUP}, \code{ID.cluster}, \code{species}, 
\code{ID.species}, \code{DETERMINANT} \cr (\emph{and optionally, 
\code{sp.mean.dist}, \code{allSp.mean}, \code{allSp.min}, 
\code{allSp.max}}). \cr Such an object can be obtained 
with the \code{\link{PRE_FATE.speciesClustering_step2}} function.}

\item{mat.traits.PFG}{a \code{data.frame} with at least 3 columns :
\code{PFG}, \code{no.species}, \code{...} (one column for each functional 
trait, computed as the \code{mean} (for numeric traits) or the \code{median} 
(for categorical traits) of the values of the determinant species of this 
PFG). \cr Such an object can be obtained with the 
\code{\link{PRE_FATE.speciesClustering_step3}} function.}
}
\value{
A \code{list} containing all the elements given to the function and 
checked :

\describe{
  \item{name.dataset}{name of the dataset}
  \item{mat.observations}{(see \code{\link{PRE_FATE.selectDominant}}) \cr
  \describe{
    \item{\code{sites}}{name of sampling site}
    \item{\code{(x, y)}}{coordinates of sampling site}
    \item{\code{species}}{name of the concerned species}
    \item{\code{abund}}{abundance of the concerned species}
    \item{\code{(habitat)}}{habitat of sampling site \cr \cr}
  }
  }
  \item{rules.selectDominant}{a \code{vector} containing values for the 
  parameters \code{doRuleA}, \code{rule.A1}, \code{rule.A2_quantile}, 
  \code{doRuleB}, \code{rule.B1_percentage}, \code{rule.B1_number}, 
  \code{rule.B2}, \code{doRuleC} (see \code{\link{PRE_FATE.selectDominant}})}
  \item{mat.traits}{(see \code{\link{PRE_FATE.speciesDistance}}) \cr
  \describe{
    \item{\code{species}}{name of the concerned species}
    \item{\code{GROUP}}{name of the concerned data subset}
    \item{\code{...}}{one column for each functional trait \cr \cr}
  }
  }
  \item{mat.overlap}{a \code{dist} object corresponding to the distance 
  between each pair of species in terms of niche overlap (see 
  \code{\link{PRE_FATE.speciesDistance}})}
  \item{rules.speciesDistance}{a \code{vector} containing values for the 
  parameters \code{opt.maxPercent.NA}, \code{opt.maxPercent.similarSpecies}, 
  \code{opt.min.sd} (see \code{\link{PRE_FATE.speciesDistance}})}
  \item{mat.species.DIST}{a \code{dist} object corresponding to the distance 
  between each pair of species, or a \code{list} of \code{dist} objects, one 
  for each \code{GROUP} value (see \code{\link{PRE_FATE.speciesDistance}})}
  \item{clust.evaluation}{(see \code{\link{PRE_FATE.speciesClustering_step1}}) \cr
  \describe{
    \item{\code{GROUP}}{name of data subset}
    \item{\code{no.clusters}}{number of clusters used for the clustering}
    \item{\code{variable}}{evaluation metrics' name}
    \item{\code{value}}{value of evaluation metric \cr \cr}
  }
  }
  \item{no.clusters}{number of clusters to be kept for each data subset}
  \item{determ.all}{(see \code{\link{PRE_FATE.speciesClustering_step2}}) \cr
  \describe{
    \item{\code{PFG}}{ID of the plant functional group 
    (\code{GROUP} + \code{ID.cluster})}
    \item{\code{GROUP}}{name of data subset}
    \item{\code{ID.cluster}}{cluster number}
    \item{\code{species}}{name of species}
    \item{\code{ID.species}}{species number in each PFG}
    \item{\code{DETERMINANT}}{\code{TRUE} if determinant species, \code{FALSE} 
    otherwise}
    \item{\code{(sp.mean.dist)}}{species mean distance to other species of 
    the same PFG}
    \item{\code{(allSp.mean)}}{\eqn{mean(\text{sp.mean.dist})} within the PFG}
    \item{\code{(allSp.min)}}{\eqn{mean(\text{sp.mean.dist}) - 1.64 * 
    sd(\text{sp.mean.dist})} within the PFG}
    \item{\code{(allSp.max)}}{\eqn{mean(\text{sp.mean.dist}) + 1.64 * 
    sd(\text{sp.mean.dist})} within the PFG}
  }
  }
  \item{mat.traits.PFG}{(see \code{\link{PRE_FATE.speciesClustering_step3}}) \cr
  \describe{
    \item{\code{PFG}}{name of the concerned functional group}
    \item{\code{no.species}}{number of species in the concerned PFG}
    \item{\code{...}}{one column for each functional trait \cr \cr}
  }
  }
}

The information is written in \file{FATE_dataset_[name.dataset]_step1_PFG.RData} file.
}
\description{
This script is designed to gather all data and parameters 
used to build a set of Plant Functional Groups.
}
\examples{

## Load example data

}
\seealso{
\code{\link{PRE_FATE.abundBraunBlanquet}}, 
\code{\link{PRE_FATE.selectDominant}}, 
\code{\link{PRE_FATE.speciesDistance}}, 
\code{\link{PRE_FATE.speciesClustering_step1}}, 
\code{\link{PRE_FATE.speciesClustering_step2}}, 
\code{\link{PRE_FATE.speciesClustering_step3}}
}
\author{
Maya Guéguen
}
