% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAVE_FATE.step2_parameters.R
\name{SAVE_FATE.step2_parameters}
\alias{SAVE_FATE.step2_parameters}
\title{Save data to reproduce building of parameter files}
\usage{
SAVE_FATE.step2_parameters(
  name.dataset,
  name.simulation = NA,
  strata.limits,
  mat.PFG.succ,
  mat.PFG.light = NA,
  mat.PFG.light.tol = NA,
  mat.PFG.soil = NA,
  mat.PFG.soil.tol = NA,
  mat.PFG.disp = NA,
  mat.PFG.dist = NA,
  mat.PFG.dist.tol = NA,
  mat.PFG.drought = NA,
  mat.PFG.drought.tol = NA,
  rasters = list(name.MASK = NA, name.DIST = NA, name.DROUGHT = NA, name.FIRE = NA,
    name.ELEVATION = NA, name.SLOPE = NA),
  multipleSet = list(name.simulation.1 = NA, name.simulation.2 = NA, file.simulParam.1
    = NA, file.simulParam.2 = NA, no_simulations = NA, opt.percent_maxAbund = NA,
    opt.percent_seeding = NA, opt.percent_light = NA, opt.percent_soil = NA,
    do.max_abund_low = NA, do.max_abund_medium = NA, do.max_abund_high = NA,
    do.seeding_duration = NA, do.seeding_timestep = NA, do.seeding_input = NA,
    do.no_strata = NA, do.LIGHT.thresh_medium = NA, do.LIGHT.thresh_low = NA,
    do.SOIL.init = NA, do.SOIL.retention = NA, do.DISPERSAL.mode = NA,     
    do.HABSUIT.mode = NA)
)
}
\arguments{
\item{name.dataset}{a \code{string} corresponding to the name to give to 
archive folder}

\item{name.simulation}{(\emph{optional}) default \code{NA}. \cr 
A \code{string} corresponding to the name of the simulation folder}

\item{strata.limits}{a \code{vector} of \code{integer} containing height 
strata limits}

\item{mat.PFG.succ}{a \code{data.frame} with at least 5 columns : \cr 
\code{PFG}, \code{type}, \code{height}, \code{maturity}, \code{longevity} 
\cr (\emph{and optionally, \code{max_abundance}, \code{potential_fecundity}, 
\code{immature_size}, \code{is_alien}, \code{flammability}}) 
\cr (see \code{\link{PRE_FATE.params_PFGsuccession}})}

\item{mat.PFG.light}{(\emph{optional}) default \code{NA}. \cr 
A \code{data.frame} with 2 to 6 columns : \cr 
\itemize{
  \item \code{PFG},
  \item \code{type}, (\emph{or \code{active_germ_low}, 
  \code{active_germ_medium}, \code{active_germ_high}}) (\emph{or
  \code{strategy_ag}})
  \item \emph{\code{type}, \code{light_need}}
}
(see \code{\link{PRE_FATE.params_PFGlight}})}

\item{mat.PFG.light.tol}{(\emph{optional}) default \code{NA}. \cr 
A \code{data.frame} with 2 to 4 columns : \cr 
\itemize{
  \item \code{PFG},
  \item \code{lifeStage}, \code{resources}, \code{tolerance} 
  (\emph{or \code{strategy_tol}})
}
(see \code{\link{PRE_FATE.params_PFGlight}})}

\item{mat.PFG.soil}{(\emph{optional}) default \code{NA}. \cr 
A \code{data.frame} with 3 to 7 columns : \cr 
\itemize{
  \item \code{PFG},
  \item \code{type}, (\emph{or \code{active_germ_low}, 
  \code{active_germ_medium}, \code{active_germ_high}}) (\emph{or
  \code{strategy_ag}})
  \item \code{soil_contrib}, \code{soil_tol_min}, \code{soil_tol_max} 
  (\emph{or \code{strategy_contrib}})
}
(see \code{\link{PRE_FATE.params_PFGsoil}})}

\item{mat.PFG.soil.tol}{(\emph{optional}) default \code{NA}. \cr 
A \code{data.frame} with 2 to 4 columns : \cr 
\itemize{
  \item \code{PFG},
  \item \code{lifeStage}, \code{resources}, \code{tolerance} 
  (\emph{or \code{strategy_tol}})
}
(see \code{\link{PRE_FATE.params_PFGsoil}})}

\item{mat.PFG.disp}{(\emph{optional}) default \code{NA}. \cr 
A \code{data.frame} with 4 columns : \code{PFG}, \code{d50}, \code{d99}, 
\code{ldd} (see \code{\link{PRE_FATE.params_PFGdispersal}})}

\item{mat.PFG.dist}{(\emph{optional}) default \code{NA}. \cr 
A \code{data.frame} with 5 columns : \cr 
\code{PFG}, \code{type}, \code{maturity}, \code{longevity}, 
\code{age_above_150cm} (see \code{\link{PRE_FATE.params_PFGdisturbance}})}

\item{mat.PFG.dist.tol}{(\emph{optional}) default \code{NA}. \cr 
A \code{data.frame} with 3 to 7 columns : \cr 
\itemize{
  \item \code{nameDist},
  \item \code{PFG},
  \item (\emph{\code{responseStage}, \code{breakAge}, \code{resproutAge}}), 
  \item \code{responseStage}, \code{killedIndiv}, \code{resproutIndiv}  
  (\emph{or \code{strategy_tol}})
}
(see \code{\link{PRE_FATE.params_PFGdisturbance}})}

\item{mat.PFG.drought}{(\emph{optional}) default \code{NA}. \cr 
A \code{data.frame} with 4 or 6 columns : \cr 
\itemize{
  \item \code{PFG},
  \item \code{threshold_moderate}, \code{threshold_severe},
  \item \code{counter_recovery}, \code{counter_sens}, \code{counter_cum}
  (\emph{or \code{strategy_drou}})
}
(see \code{\link{PRE_FATE.params_PFGdrought}})}

\item{mat.PFG.drought.tol}{(\emph{optional}) default \code{NA}. \cr 
A \code{data.frame} with 3 to 7 columns : \cr 
\itemize{
  \item \code{nameDist},
  \item \code{PFG},
  \item (\emph{\code{responseStage}, \code{breakAge}, \code{resproutAge}}), 
  \item \code{responseStage}, \code{killedIndiv}, \code{resproutIndiv}  
  (\emph{or \code{strategy_tol}})
}
(see \code{\link{PRE_FATE.params_PFGdrought}})}

\item{rasters}{a \code{list} containing all the rasters given to the 
\code{\link{PRE_FATE.params_simulParameters}} function, if used 
(\code{name.MASK}, \code{name.DIST}, \code{name.DROUGHT}, \code{name.FIRE}, 
\code{name.ELEVATION}, \code{name.SLOPE})}

\item{multipleSet}{a \code{list} containing all the parameter values given 
to the \code{\link{PRE_FATE.params_multipleSet}} function, if used 
(\code{name.simulation.1}, \code{name.simulation.2}, 
\code{file.simulParam.1}, \code{file.simulParam.2}, 
\code{no_simulations}, \code{opt.percent_maxAbund}, 
\code{opt.percent_seeding}, \code{opt.percent_light}, 
\code{opt.percent_soil}, \code{do.max_abund_low}, 
\code{do.max_abund_medium}, \code{do.max_abund_high}, 
\code{do.seeding_duration}, \code{do.seeding_timestep}, 
\code{do.seeding_input}, \code{do.no_strata}, 
\code{do.LIGHT.thresh_medium}, \code{do.LIGHT.thresh_low}, 
\code{do.SOIL.init}, \code{do.SOIL.retention}, 
\code{do.DISPERSAL.mode}, \code{do.HABSUIT.mode})}
}
\value{
A \code{list} containing all the elements given to the function and 
checked, and two archive files :

\describe{
  \item{name.dataset}{name of the dataset}
  \item{strata.limits}{height strata limits}
  \item{mat.PFG.succ}{}
  \item{(mat.PFG.light)}{}
  \item{(mat.PFG.light.tol)}{}
  \item{(mat.PFG.soil)}{}
  \item{(mat.PFG.soil.tol)}{}
  \item{(mat.PFG.disp)}{}
  \item{(mat.PFG.dist)}{}
  \item{(mat.PFG.dist.tol)}{}
  \item{(mat.PFG.drought)}{}
  \item{(mat.PFG.drought.tol)}{}
  \item{rasters}{raster files of all simulation masks}
  \item{(multipleSet)}{ \cr \cr}
  \item{(name.simulation)}{name of the simulation folder}
  \item{(\code{DATA} folder)}{contained in \code{name.simulation} folder 
  and archived}
  \item{(\code{PARAM_SIMUL} folder)}{contained in \code{name.simulation} 
  folder and archived \cr \cr}
}

The information is written in \file{FATE_dataset_[name.dataset]_step2_parameters.RData} file.
}
\description{
This script is designed to gather all data and parameters 
used to build a \code{FATE} simulation folder.
}
\examples{

## Load example data

}
\seealso{
\code{\link{PRE_FATE.skeletonDirectory}}, 
\code{\link{PRE_FATE.params_PFGsuccession}}, 
\code{\link{PRE_FATE.params_PFGlight}}, 
\code{\link{PRE_FATE.params_PFGsoil}}, 
\code{\link{PRE_FATE.params_PFGdispersal}}, 
\code{\link{PRE_FATE.params_PFGdisturbance}}, 
\code{\link{PRE_FATE.params_PFGdrought}}, 
\code{\link{PRE_FATE.params_changingYears}}, 
\code{\link{PRE_FATE.params_savingYears}}, 
\code{\link{PRE_FATE.params_globalParameters}}, 
\code{\link{PRE_FATE.params_simulParameters}}, 
\code{\link{PRE_FATE.params_multipleSet}}
}
\author{
Maya Guéguen
}
