% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHINY_FATE.run.R
\name{SHINY_FATE.run}
\alias{SHINY_FATE.run}
\title{Shiny application to apply \code{RFate} functions}
\usage{
SHINY_FATE.run()
}
\description{
This SHINY application allows to use all the \code{RFate} 
functions (PRE and POST FATE), from the building of PFG to the treatment of
FATE output files.
}
\author{
Maya Guéguen
}
\keyword{GUI}
\keyword{application,}
\keyword{interface,}
\keyword{shiny}
