% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.package_betapart.R
\name{beta.pair}
\alias{beta.pair}
\alias{betapart.core}
\title{From betapart package 1.5.4 : beta.pair and betapart.core functions}
\usage{
betapart.core(x)
beta.pair(x, index.family = "sorensen")
}
\arguments{
\item{x}{data frame, where rows are sites and columns are species. 
Alternatively x can be a betapart object derived from the betapart.core 
function}

\item{index.family}{family of dissimilarity indices, partial match of 
"sorensen" or "jaccard".}
}
\description{
From betapart package 1.5.4 : beta.pair and betapart.core functions
}
\seealso{
\code{\link[betapart]{beta.pair}}
}
\keyword{betapart}
