% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.package_SPOT.R
\name{designLHD}
\alias{designLHD}
\alias{designLHDNorm}
\title{From SPOT package 2.5.0 : designLHD and designLHDNorm functions}
\usage{
designLHDNorm(dim, size, calcMinDistance = FALSE, nested = NULL, 
inequalityConstraint = NULL)
designLHD(x = NULL, lower, upper, control = list())
}
\arguments{
\item{dim}{number, dimension of the problem (will be no. of columns of 
the result matrix)}

\item{size}{number of points with that dimension needed. (will be no. 
of rows of the result matrix).}

\item{calcMinDistance}{Boolean to indicate whether a minimal distance 
should be calculated.}

\item{nested}{nested design to be considered during distance calculation.}

\item{inequalityConstraint}{inequality constraint function, smaller zero 
for infeasible points. Used to replace infeasible points with random 
points. Has to evaluate points in interval [0;1].}

\item{x}{optional matrix x, rows for points, columns for dimensions. This 
can contain one or more points which are part of the design, but specified 
by the user. These points are added to the design, and are taken into 
account when calculating the pair-wise distances. They do not count for 
the design size. E.g., if x has two rows, control$replicates is one and 
control$size is ten, the returned design will have 12 points (12 rows). 
The first two rows will be identical to x. Only the remaining ten rows are 
guaranteed to be a valid LHD.}

\item{lower}{vector with lower boundary of the design variables (in case 
of categorical parameters, please map the respective factor to a set of 
contiguous integers, e.g., with lower = 1 and upper = number of levels)}

\item{upper}{vector with upper boundary of the design variables (in case 
of categorical parameters, please map the respective factor to a set of 
contiguous integers, e.g., with lower = 1 and upper = number of levels)}

\item{control}{list of controls: see 
\code{\href{designLHD}{https://www.rdocumentation.org/packages/SPOT/versions/2.0.3/topics/designLHD}}}
}
\description{
From SPOT package 2.5.0 : designLHD and designLHDNorm functions
}
\seealso{
\code{\href{designLHD}{https://www.rdocumentation.org/packages/SPOT/versions/2.0.3/topics/designLHD}}
}
\keyword{Hypercube}
\keyword{Latin}
\keyword{SPOT,}
\keyword{Sampling,}
\keyword{internal}
