% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.package_SPOT.R
\name{designLHD}
\alias{designLHD}
\alias{designLHDNorm}
\title{From SPOT package 2.5.0 : designLHD and designLHDNorm functions}
\usage{
designLHDNorm(
  dim,
  size,
  calcMinDistance = FALSE,
  nested = NULL,
  inequalityConstraint = NULL
)
}
\description{
From SPOT package 2.5.0 : designLHD and designLHDNorm functions
}
\seealso{
\code{\link[SPOT]{designLHD}}
}
\keyword{Hypercube}
\keyword{Latin}
\keyword{SPOT,}
\keyword{Sampling}
