% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.do_habitat_validation.R
\name{do.habitat.validation}
\alias{do.habitat.validation}
\title{Compare observed and simulated habitat of a \code{FATE} simulation
at the last simulation year.}
\usage{
do.habitat.validation(
  output.path,
  RF.model,
  habitat.FATE.map,
  validation.mask,
  simulation.map,
  predict.all.map,
  sim.version,
  name.simulation,
  perStrata,
  hab.obs,
  year,
  list.strata.releves,
  list.strata.simulations
)
}
\arguments{
\item{output.path}{access path to the for the folder where output files
will be created.}

\item{RF.model}{random forest model trained on CBNA data (train.RF.habitat
function)}

\item{habitat.FATE.map}{a raster map of the observed habitat in the
studied area.}

\item{validation.mask}{a raster mask that specified which pixels need validation.}

\item{simulation.map}{a raster map of the whole studied area use to check
the consistency between simulation map and the observed habitat map.}

\item{predict.all.map}{\code{Logical}. If TRUE, the script will predict 
habitat for the whole map.}

\item{sim.version}{name of the simulation to validate.}

\item{name.simulation}{simulation folder name.}

\item{perStrata}{\code{Logical}. If TRUE, the PFG abundance is defined
by strata in each pixel. If FALSE, PFG abundance is defined for all strata.}

\item{hab.obs}{a raster map of the observed habitat in the
extended studied area.}

\item{year}{simulation year selected for validation.}

\item{list.strata.releves}{a character vector which contain the observed strata 
definition, extracted from observed PFG releves.}

\item{list.strata.simulations}{a character vector which contain \code{FATE} 
strata definition and correspondence with observed strata definition.}
}
\value{
Habitat performance file. \cr
If option selected, the function also returns an habitat prediction file with 
observed and simulated habitat for each pixel of the whole map.
}
\description{
To compare observations and simulations, this function compute 
confusion matrix between observation and prediction and then compute the TSS 
for each habitat.
}
\details{
After several preliminary checks, the function is going to prepare the observations
database by extracting the observed habitat from a raster map. Then, for the
simulation \code{sim.version}, the script take the evolution abundance for each PFG
and all strata (or for each PFG & each strata if option selected) file and predict 
the habitat for the whole map (if option selected) thanks to the RF model. 
Finally, the function computes habitat performance based on TSS for each habitat.
}
\author{
Matthieu Combaud & Maxime Delprat
}
