% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.do_PFG_composition_validation.R
\name{do.PFG.composition.validation}
\alias{do.PFG.composition.validation}
\title{Compute distance between observed and simulated distribution}
\usage{
\method{do}{PFG.composition.validation}(
  name.simulation,
  obs.path,
  sim.version,
  hab.obs,
  PFG.considered_PFG.compo,
  strata.considered_PFG.compo,
  habitat.considered_PFG.compo,
  observed.distribution,
  perStrata.compo,
  validation.mask,
  year
)
}
\arguments{
\item{name.simulation}{simulation folder name.}

\item{obs.path}{the function needs observed data, please create a folder for them in your 
simulation folder and then indicate in this parameter the access path to this new folder.}

\item{sim.version}{name of the simulation we want to validate (it works with
only one \code{sim.version}).}

\item{hab.obs}{file which contain the extended studied map in the simulation.}

\item{PFG.considered_PFG.compo}{a character vector of the list of PFG considered
in the validation.}

\item{strata.considered_PFG.compo}{a character vector of the list of precise 
strata considered in the validation.}

\item{habitat.considered_PFG.compo}{a character vector of the list of habitat(s)
considered in the validation.}

\item{observed.distribution}{PFG observed distribution table.}

\item{perStrata.compo}{Logical. All strata together (FALSE) or per strata (TRUE).}

\item{validation.mask}{file which contain a raster mask that specified 
which pixels need validation.}

\item{year}{year of simulation to validate.}
}
\value{

}
\description{
This script is designed to compare the difference between the
PFG distribution in observed and simulated data. For a set of PFG, strata and
habitats chosen, the function compute distance between observed and simulated
distribution for a precise \code{FATE} simulation.
}
\details{
After preliminary checks, this code extract observed habitat from the \code{hab.obs}
map and, then, merge it with the simulated PFG abundance file from results of a \code{FATE}
simulation. After filtration of the required PFG, strata and habitats, the function
transform the data into relative metrics and, then, compute distribution per PFG, strata
and habitat (if necessary). Finally, the code computes proximity between observed 
and simulated data, per PFG, strata and habitat.
}
\author{
Matthieu Combaud, Maxime Delprat
}
