% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.do_habitat_validation.R
\name{do_habitat_validation}
\alias{do_habitat_validation}
\title{Compare observed and simulated habitat of a \code{FATE} simulation
to a chosen year.}
\usage{
do_habitat_validation(
  output.path,
  RF.model,
  predict.all.map,
  sim,
  simu_PFG,
  habitat.whole.area.df,
  list.strata,
  perStrata
)
}
\arguments{
\item{output.path}{Access path to the for the folder where output files
will be created.}

\item{RF.model}{Random forest model trained on observed abundance data (\code{\link{train_RF_habitat}}
function)}

\item{predict.all.map}{\code{Logical}. If TRUE, the script will predict 
habitat for the whole map.}

\item{sim}{Name of the single simulation to validate.}

\item{simu_PFG}{A \code{data.frame} provides by \code{POST_FATE.temporalEvolution} with simulated abundance for each PFG and strata 
(if option selected) and pixel ID (see \code{\link{POST_FATE.temporalEvolution}}).}

\item{habitat.whole.area.df}{A \code{data.frame} provides by \code{POST_FATE.validation} with 3 columns : 
\cr \code{pixel} which contains the ID of each pixel in the study area.
\cr \code{code.habitat} which contains the ID of the habitat in each pixel.
\cr \code{for.validation} for each pixel, \code{0} if does not need validation, \code{1} if needs validation.}

\item{list.strata}{If abundance file is defined by strata : a character vector which contains \code{FATE} 
strata definition and correspondence with observed strata definition. \cr
If abundance file is defined for all strata : a character vector with value "all".}

\item{perStrata}{\code{Logical}. Default \code{TRUE}. If \code{TRUE}, PFG abundance is defined by strata. 
If \code{FALSE}, PFG abundance defined for all strata.}
}
\value{
Habitat performance file. \cr
If option selected, the function also returns an habitat prediction file with 
observed and simulated habitat for each pixel of the whole map.
}
\description{
To compare observations and simulations, this function compute 
confusion matrix between observation and prediction and then compute the TSS 
for each habitats.
}
\details{
For a given simulation \code{sim}, this script takes the evolution abundance for each PFG
and all strata (or for each PFG & each strata if option selected) data frame and predicts 
the habitat for the whole map (if option selected) thanks to the RF model. 
Finally, the function computes habitat performance based on TSS for each habitat.
}
\author{
Matthieu Combaud & Maxime Delprat
}
