% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.do_PFG_composition_validation.R
\name{do_PFG_composition_validation}
\alias{do_PFG_composition_validation}
\title{Compute distance between observed and simulated distribution}
\usage{
do_PFG_composition_validation(
  sim,
  PFG.considered_PFG.compo,
  strata.considered_PFG.compo,
  habitat.considered_PFG.compo,
  observed.distribution,
  simu_PFG,
  habitat.whole.area.df
)
}
\arguments{
\item{sim}{name of the single simulation to validate.}

\item{PFG.considered_PFG.compo}{a character vector of the list of PFG considered
in the validation.}

\item{strata.considered_PFG.compo}{a character vector of the list of precise 
strata considered in the validation.}

\item{habitat.considered_PFG.compo}{a character vector of the list of habitat(s)
considered in the validation.}

\item{observed.distribution}{PFG observed distribution table provides by \code{get.observed.distribution} function.}

\item{simu_PFG}{a \code{data frame} with simulated abundance for each PFG and strata 
(if option selected) and pixel ID, extracted from a \code{FATE} simulation (see \code{\link{POST_FATE.temporalEvolution}}).}

\item{habitat.whole.area.df}{a \code{data frame} which contain habitat names and code for each pixel that needs validation.}
}
\value{

}
\description{
This script is designed to compare the difference between the
PFG distribution in observed and simulated data. For a set of PFG, strata and
habitats chosen, the function computes distance between observed and simulated
distribution for a precise \code{FATE} simulation.
}
\details{
Firstly, this code merges \code{habitat.whole.area.df} data frame with the simulated PFG abundance 
\code{simu_PFG} data frame (with or without strata definition). 
After filtration of the required PFG, strata and habitats, the function transforms 
the data into relative metrics and, then, compute distribution per PFG, strata
and habitat (if necessary). Finally, the code computes proximity between observed 
and simulated data, per PFG, strata and habitat.
}
\author{
Matthieu Combaud, Maxime Delprat
}
