% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.adaptMaps.R
\name{.adaptMaps}
\alias{.adaptMaps}
\title{Adapt all raster maps of a \code{FATE} simulation folder (change NA 
to 0, and save as .tif)}
\usage{
.adaptMaps(name.simulation, extension.old, extension.new)
}
\arguments{
\item{name.simulation}{a \code{string} corresponding to the main directory 
or simulation name of the \code{FATE} simulation}

\item{extension.old}{a \code{string} corresponding to the extension of 
raster files to be found}

\item{extension.new}{a \code{string} (either \code{tif} or \code{img}) 
corresponding to the new extension to save all the maps}
}
\description{
This function scan all the raster files within a \code{FATE} 
simulation folder, change all NA values to 0 and save them with the 
specified extension.
}
\examples{

## Load example data

}
\author{
Maya Guéguen
}
