% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.getCutoff.R
\name{.getCutoff}
\alias{.getCutoff}
\title{Find cutoff to transform abundance values into binary values}
\usage{
.getCutoff(Obs, Fit)
}
\arguments{
\item{Obs}{a \code{vector} containing binary observed values (0/1)}

\item{Fit}{a \code{vector} containing abundance values}
}
\description{
This functions finds the best cutoff to transform abundance 
values into binary values while optimising sensitivity and specificity 
values based on observations
}
\examples{

}
