% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.getParam.R
\name{.getParam}
\alias{.getParam}
\title{Extract a parameter value(s) from a parameter file}
\usage{
.getParam(params.lines, flag, flag.split, is.num = TRUE)
}
\arguments{
\item{params.lines}{a \code{string} that corresponds to the 
name of the file from which to extract the parameter value}

\item{flag}{a \code{string} that corresponds to the parameter
name to be extracted and that must be present into the
\code{param.lines} file}

\item{flag.split}{either " " or "^--.*--$", depending on the
type of parameter file}

\item{is.num}{default \code{TRUE}. If \code{TRUE}, the 
extracted parameter is considered to be \code{numeric} and 
will be processed as such}
}
\description{
Extract a parameter value(s) from a parameter file
}
