% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.loadData.R
\name{.loadData}
\alias{.loadData}
\title{Load a dataset}
\usage{
.loadData(data.name, format.name)
}
\arguments{
\item{data.name}{a \code{string} corresponding to the name of the dataset 
that will be loaded}

\item{format.name}{a \code{string} corresponding to the downloading format way 
(either \code{RData} or \code{7z})}
}
\value{
\strong{\cr\cr#########################################################\cr}
\strong{>>> Champsaur_PFG}
\strong{\cr#########################################################\cr\cr}

A \code{list} object with 15 elements to help building the Plant Functional 
Group :
  
  \describe{
    \item{sp.observations}{a \code{data.frame} of dimension 
    \code{127257 x 3} \cr
    containing releves data about plant species in Champsaur \cr
    to be used with the \code{\link{PRE_FATE.selectDominant}} function
    \itemize{
      \item \strong{sites} : sites ID
      \item \strong{species} : species ID
      \item \strong{abund} : relative abundance obtained from the 
      \code{abund_BB} column with the 
      \code{\link{PRE_FATE.abundBraunBlanquet}} function
    }
    }
    \item{rules.selectDominant}{a \code{vector} containing parameters to be 
    given to the \code{\link{PRE_FATE.selectDominant}} function \cr
    (\code{doRuleA}, \code{rule.A1}, \code{rule.A2_quantile}, \code{doRuleB}, 
    \code{rule.B1_percentage}, \code{rule.B1_number}, \code{rule.B2}, \code{doRuleC})
    }
    \item{sp.SELECT}{a \code{list} obtained from the 
    \code{\link{PRE_FATE.selectDominant}} function run with 
    \code{sp.observations} and \code{rules.selectDominant}
    }
    \item{sp.traits}{a \code{data.frame} of dimension \code{250 x 23} \cr
    containing traits for dominant species \cr
    to be used with the \code{\link{PRE_FATE.speciesDistance}} function
    \itemize{
      \item \strong{species} : species ID
      \item \strong{GROUP} : rough generalization of Raunkier life-forms 
      (Phanerophyte, Chamaephyte, Herbaceous)
      \item \strong{MATURITY / LONGEVITY / HEIGHT / LDMC / LNC / SEEDM / 
      SLA} : numerical values
      \item \strong{LONGEVITY_log / HEIGHT_log / LDMC_log / LNC_log / 
      SEEDM_log / SLA_log} : log-transformed numerical values
      \item \strong{DISPERSAL / LIGHT / NITROGEN / NITROGEN_TOLERANCE / 
      MOISTURE / GRAZ_MOW_TOLERANCE / HABITAT / STRATEGY} : categorical 
      values
    }
    }
    \item{sp.traits.P}{a \code{data.frame} of dimension \code{19 x 9} \cr
    a subset of \code{sp.traits} (LONGEVITY_log / HEIGHT_log / SEEDM_log 
    / SLA_log / DISPERSAL / LIGHT / NITROGEN) for Phanerophyte dominant 
    species \cr
    to be used with the \code{\link{PRE_FATE.speciesDistance}} function
    }
    \item{sp.traits.C}{a \code{data.frame} of dimension \code{39 x 10} \cr
    a subset of \code{sp.traits} (LONGEVITY_log / HEIGHT_log / SEEDM_log 
    / SLA_log / LNC_log / DISPERSAL / LIGHT / NITROGEN) for Chamaephyte 
    dominant species \cr
    to be used with the \code{\link{PRE_FATE.speciesDistance}} function
    }
    \item{sp.traits.H}{a \code{data.frame} of dimension \code{192 x 7} \cr
    a subset of \code{sp.traits} (HEIGHT_log / SLA_log / LDMC_log / 
    LNC_log / LIGHT) for Herbaceous dominant species \cr
    to be used with the \code{\link{PRE_FATE.speciesDistance}} function
    }
    \item{mat.habitat}{\code{matrix} of dimension \code{240 x 240} \cr
    containing habitat dissimilarity distance (1 - Schoeners' D, obtained 
    with \code{\link[FD]{gowdis}}) for dominant species \cr
    to be used with the \code{\link{PRE_FATE.speciesDistance}} function}
    \item{mat.overlap}{\code{matrix} of dimension \code{243 x 243} \cr
    containing niche overlap distance (1 - Schoeners' D, obtained with 
    \code{\link[ecospat]{ecospat.niche.overlap}}) for dominant species \cr
    to be used with the \code{\link{PRE_FATE.speciesDistance}} function}
    \item{sp.DIST.P}{a \code{list} obtained from the 
    \code{\link{PRE_FATE.speciesDistance}} function run for Phanerophyte 
    dominant species with \code{mat.habitat}, \code{mat.overlap} and 
    \code{sp.traits.P} parameters}
    \item{sp.DIST.C}{a \code{list} obtained from the 
    \code{\link{PRE_FATE.speciesDistance}} function run for Chamaephyte 
    dominant species with \code{mat.habitat}, \code{mat.overlap} and 
    \code{sp.traits.C} parameters}
    \item{sp.DIST.H}{a \code{list} obtained from the 
    \code{\link{PRE_FATE.speciesDistance}} function run for Herbaceous 
    dominant species with \code{mat.habitat}, \code{mat.overlap} and 
    \code{sp.traits.H} parameters}
    
    \item{PFG.species}{\code{data.frame} of dimension \code{232 x 5} \cr
    containing dominant species information relative to PFG \cr
    obtained from the \code{\link{PRE_FATE.speciesClustering_step2}} 
    function
    \itemize{
      \item \strong{PFG} : name of assigned Plant Functional Group
      \item \strong{DETERMINANT} : is the species kept as determinant 
      species within the PFG ? \cr (see 
      \code{\link{PRE_FATE.speciesClustering_step2}} function for details)
      \item \strong{species} : species ID
      \item \strong{species_name} : species name (taxonomic)
      \item \strong{species_genus} : species genus (taxonomic)
    }
    }
    \item{PFG.traits}{\code{data.frame} of dimension \code{14 x 13} \cr
    containing traits for plant functional groups \cr
    obtained from the \code{\link{PRE_FATE.speciesClustering_step3}} 
    function
    \itemize{
      \item \strong{PFG} : Plant Functional Group short name
      \item \strong{no.species} : number of species within each group
      \item \strong{maturity} : MEAN age of maturity
      \item \strong{longevity} : MEAN age of lifespan
      \item \strong{height} : MEAN height (cm)
      \item \strong{light} : MEDIAN Landolt indicator value for light 
      preference (from 1 to 5)
      \item \strong{dispersal} : MEDIAN classes (from 1 to 7) based on 
      dispersal distances and types (Vittoz & Engler)
      \item \strong{soil_contrib} : MEAN Landolt indicator value for 
      nitrogen preference (from 1 to 5)
      \item \strong{soil_tol_min} : MIN tolerance value for nitrogen 
      preference (based on Landolt indicators, from 1 to 5)
      \item \strong{soil_tol_max} : MAX tolerance value for nitrogen 
      preference (based on Landolt indicators, from 1 to 5)
      \item \strong{LDMC} : MEAN LDMC
      \item \strong{LNC} : MEAN LNC
    }
    }
  }
  
\strong{\cr\cr#########################################################\cr}
\strong{>>> Champsaur_PARAM}
\strong{\cr#########################################################\cr\cr}
  
  

\strong{\cr\cr PNE_PARAM}

A \code{list} object with 13 elements to help building the simulation 
files and folders to run a \code{FATE} simulation :
  
  \describe{
    \item{masks}{a \code{\link[raster]{stack}} object of dimension 
    \code{782 x 619} with a resolution of \code{100m} and Lambers (lcc) 
    projection, containing 7 mask layers with binary values (0 or 1) to be 
    used in a \code{FATE} simulation :
    \itemize{
      \item \strong{maskEcrins} : simulation map, where occurs succession
      \item \strong{noDisturb} : perturbation map, when there is none
      \item \strong{mowing} : perturbation map, where occurs mowing
      \item \strong{grazing1} : perturbation map, where occurs light grazing
      \item \strong{grazing2} : perturbation map, where occurs extensive 
      grazing
      \item \strong{grazing3} : perturbation map, where occurs intensive 
      grazing
      \item \strong{grazingAll} :  perturbation map, where occurs grazing, 
      all types combined
    }
    }
    \item{HS_0}{a \code{\link[raster]{stack}} object of dimension 
    \code{782 x 619} with a resolution of \code{100m} and Lambers (lcc) 
    projection, containing 24 layers with probability values (between 0 and 
    1) representing Habitat Suitability for initialization phase and 
    \emph{equilibrium} or \emph{current} time (0) for each PFG 
    and to be used in a \code{FATE} simulation. \cr
    These maps are coming from Species Distribution Models and methods to 
    obtain them are described in Supplementary Materials of 
    \emph{\href{https://mayagueguen.github.io/FATE-WEBSITE/papers/Boulangeat_2014_GCB.pdf}{Boulangeat et al. 2014 GCB}}.}
    \item{HS_15}{same as HS_0 but 15 years after equilibrium}
    \item{HS_30}{same as HS_0 but 30 years after equilibrium}
    \item{HS_45}{same as HS_0 but 45 years after equilibrium}
    \item{HS_60}{same as HS_0 but 60 years after equilibrium}
    \item{HS_75}{same as HS_0 but 75 years after equilibrium}
    \item{HS_90}{same as HS_0 but 90 years after equilibrium}
    \item{strata_limits}{a \code{vector} of length \code{5} \cr
    containing height of stratum limits in centimeters \cr
    to be used with the \code{\link{PRE_FATE.params_PFGsuccession}} and 
    \code{\link{PRE_FATE.params_PFGlight}} functions \cr}
    \item{succ_light}{a \code{data.frame} of dimension \code{24 x 6} \cr
    containing traits for plant functional groups \cr
    obtained from the \code{\link{PRE_FATE.speciesClustering_step3}} 
    function \cr
    to be used with the \code{\link{PRE_FATE.params_PFGsuccession}} and 
    \code{\link{PRE_FATE.params_PFGlight}} functions \cr}
    \item{disp}{a \code{data.frame} of dimension \code{24 x 4} \cr
    containing dispersal values (in meters) for plant functional groups \cr
    to be used with the \code{\link{PRE_FATE.params_PFGdispersal}} function 
    \cr}
    \item{dist}{a \code{data.frame} of dimension \code{384 x 5} \cr
    containing response of plant functional groups to disturbances \cr
    to be used with the \code{\link{PRE_FATE.params_PFGdisturbance}} 
    function \cr}
    \item{global}{a \code{vector} of length \code{18} \cr
    containing global parameter values for a \code{FATE} simulation in 
    the PNE \cr
    to be used with the \code{\link{PRE_FATE.params_globalParameters}} 
    function \cr
    }
  }


\strong{\cr\cr PNE_RESULTS}

A \code{list} object with 5 elements :

\describe{
  \item{evaluation}{a \code{data.frame} of dimension \code{24 x 11} \cr
    containing statistical metrics to evaluate the predictive quality of 
    both Habitat Suitability models (HS) and \code{FATE} simulation 
    outputs \cr
    \itemize{
      \item \strong{PFG} : name of Plant Functional Group
      \item \strong{nb.obs.absences} : number of observed absences
      \item \strong{nb.obs.presences} : number of observed presences
      \item \strong{specificity.FATE} : proportion of actual negatives that 
      are correctly identified as such by the \code{FATE} model (true 
      negative rate)
      \item \strong{specificity.HS} : proportion of actual negatives that 
      are correctly identified as such by the Habitat Suitability model 
      (true negative rate)
      \item \strong{sensitivity.FATE} : proportion of actual positives that 
      are correctly identified as such by the \code{FATE} model (true 
      positive rate)
      \item \strong{sensitivity.HS} : proportion of actual positives that 
      are correctly identified as such by the Habitat Suitability model 
      (true positive rate)
      \item \strong{TSS.FATE} : True Skill Statistic for \code{FATE} 
      model
      \item \strong{TSS.HS} : True Skill Statistic for Habitat Suitability 
      model
      \item \strong{error.rate.FATE} : percentage of bad predictions for 
      \code{FATE} model
      \item \strong{error.rate.HS} : percentage of bad predictions for 
      Habitat Suitability model
    }
    }
  \item{abund_str.equilibrium}{a \code{\link[raster]{stack}} object of 
  dimension \code{782 x 619} with a resolution of \code{100m} and Lambers 
  (lcc) projection, containing 120 layers representing \code{FATE} 
  \strong{abundances for year 800 of initialization phase} (= 
  \emph{equilibrium} or \emph{current} time). \cr
  Maps are per PFG and per height stratum.}
  \item{forest_cover.init}{a \code{\link[raster]{stack}} object of 
  dimension \code{782 x 619} with a resolution of \code{100m} and Lambers 
  (lcc) projection, containing 16 layers representing \code{FATE} 
  \strong{relative forest cover through initialization phase (from year 50 
  to 800)}. \cr
  Maps are summing all PFGs through all height strata above 1.5 meters. \cr
  }
  \item{forest_cover \cr CC_BAU}{a \code{\link[raster]{stack}} object of 
  dimension \code{782 x 619} with a resolution of \code{100m} and Lambers 
  (lcc) projection, containing 14 layers representing \code{FATE} 
  \strong{relative forest cover through Climate Change (CC) + 
  Business-As-Usual (BAU) scenario (from year 850 to 1500)}. \cr
  Maps are summing all PFGs through all height strata above 1.5 meters. \cr
    }
  \item{forest_cover\cr CC_Abandon}{a \code{\link[raster]{stack}} object of 
  dimension \code{782 x 619} with a resolution of \code{100m} and Lambers 
  (lcc) projection, containing 14 layers representing \code{FATE} 
  \strong{relative forest cover through Climate Change (CC) + Abandonment 
  (BAU) scenario (from year 850 to 1500)}. \cr
  Maps are summing all PFGs through all height strata above 1.5 meters. \cr
  }
}
}
\description{

}
\author{
Maya Guéguen
}
