% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.loadData.R
\name{.loadData}
\alias{.loadData}
\title{Load a dataset}
\usage{
.loadData(data.name)
}
\arguments{
\item{data.name}{a \code{string} that corresponds to the 
name of the dataset that will be loaded}
}
\description{
This function loads one of the available data sets.
}
\author{
Maya Guéguen
}
