% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.loadPackage.R
\name{.loadPackage}
\alias{.loadPackage}
\title{Load a \code{R} package and install it if necessary}
\usage{
.loadPackage(package.name)
}
\arguments{
\item{package.name}{a \code{string} that corresponds to the 
name of the package that will be load or installed}
}
\description{
This function loads a \code{R} package and install it if 
necessary.
}
\author{
Maya Guéguen
}
