% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.setParam.R
\name{.setParam}
\alias{.setParam}
\title{Replace parameter value(s) from a parameter file}
\usage{
.setParam(params.lines, flag, flag.split, value)
}
\arguments{
\item{params.lines}{a \code{string} that corresponds to the 
name of the file from which to extract the parameter value}

\item{flag}{a \code{string} that corresponds to the parameter
name to be extracted and that must be present into the
\code{param.lines} file}

\item{flag.split}{either " " or "^--.*--$", depending on the
type of parameter file}

\item{value}{a \code{string} or a \code{numeric} value (it can also be 
a \code{vector}) containing the new value of the parameter to be changed}
}
\description{
This function finds in a text file the value(s) of a 
given parameter, and replace it with new value(s).
}
\author{
Maya Guéguen
}
