% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.upscaleMaps.R
\name{.upscaleMaps}
\alias{.upscaleMaps}
\alias{.cropMaps}
\title{Upscale / crop all raster maps of a \code{FATE} simulation folder}
\usage{
.upscaleMaps(name.simulation, resolution)
.cropMaps(name.simulation, extent)
}
\arguments{
\item{name.simulation}{a \code{string} corresponding to the main directory 
or simulation name of the \code{FATE} simulation}

\item{resolution}{an \code{integer} corresponding to the new resolution to 
upscale all the maps}

\item{extent}{a \code{vector} of 4 \code{numeric} values corresponding to 
the new extent to crop all the maps}
}
\description{
These functions scan all the raster files within a 
\code{FATE} simulation folder and upscale / crop them to the specified 
resolution / extent.
}
\examples{

## Load example data


}
\author{
Maya Guéguen
}
