% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.upscaleMaps.R
\name{.upscaleMaps}
\alias{.upscaleMaps}
\title{Upscale all raster maps of a \code{FATE-HD} simulation folder}
\usage{
.upscaleMaps(name.simulation, resolution)
}
\arguments{
\item{name.simulation}{a \code{string} that corresponds to the main
directory or simulation name of the \code{FATE-HD} simulation}

\item{resolution}{an \code{integer} that corresponds to the new 
resolution to upscale all the maps}
}
\description{
This function scans all the raster files within a 
\code{FATE-HD} simulation folder and upscale them to the specified 
resolution
}
\author{
Maya Guéguen
}
