% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.package_clValid.R
\name{dunn}
\alias{dunn}
\title{From clValid package O.7 : dunn function}
\usage{
dunn(distance = NULL, clusters)
}
\arguments{
\item{distance}{The distance matrix (as a matrix object) of the clustered 
observations. Required if Data is NULL.}

\item{clusters}{An integer vector indicating the cluster partitioning}
}
\description{
From clValid package O.7 : dunn function
}
\seealso{
\code{\link[clValid]{dunn}}
}
\keyword{clValid}
