% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.package_ecospat.R
\name{ecospat.niche.overlap}
\alias{ecospat.niche.overlap}
\alias{ecospat.kd}
\alias{ecospat.grid.clim.dyn}
\title{From ecospat package 3.2 : ecospat.kd, ecospat.grid.clim.dyn and 
ecospat.niche.overlap functions}
\usage{
ecospat.kd(x, ext, R = 100, th = 0, env.mask = c(), method = "adehabitat")
}
\description{
From ecospat package 3.2 : ecospat.kd, ecospat.grid.clim.dyn and 
ecospat.niche.overlap functions
}
\seealso{
\code{\link[ecospat]{ecospat.grid.clim.dyn}},
\code{\link[ecospat]{ecospat.niche.overlap}}
}
\keyword{ecospat,}
\keyword{niche}
\keyword{overlap}
