% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FATE}
\alias{FATE}
\title{FATE Wrapper}
\usage{
FATE(simulParam, no_CPU = 1L, verboseLevel = 0L)
}
\arguments{
\item{simulParam}{a \code{string} corresponding to the name of a 
parameter file that will be contained into the \code{PARAM_SIMUL} folder 
of the \code{FATE} simulation}

\item{no_CPU}{(\emph{optional}) default \code{1}. \cr The number of 
resources that can be used to parallelize the simulation}

\item{verboseLevel}{(\emph{optional}) default \code{0}. \cr The logger 
verbose level}
}
\value{
None
}
\description{
This function runs a FATE dynamical landscape vegetation simulation.
}
\examples{
\dontrun{FATE()}

}
\author{
Isabelle Boulangeat, \email{isabelle.boulangeat@inrae.fr}

Maya Guéguen, \email{maya.gueguen@univ-grenoble-alpes.fr}
}
