% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.get_observed_distribution.R
\name{get_observed_distribution}
\alias{get_observed_distribution}
\title{Compute distribution of relative abundance over observed releves.}
\usage{
get_observed_distribution(
  releves.PFG,
  hab.obs,
  studied.habitat = NULL,
  PFG.considered_PFG.compo,
  strata.considered_PFG.compo,
  habitat.considered_PFG.compo,
  perStrata,
  output.path
)
}
\arguments{
\item{releves.PFG}{a data frame with PFG abundances (column named abund) at each site, 
for each PFG and strata & coordinates of each site.}

\item{hab.obs}{a raster map of the extended studied map in the simulation, with same projection 
& resolution than simulation mask.}

\item{studied.habitat}{default \code{NULL}. a 2 columns data frame which contains 
the habitats (2nd column) and the ID (1st column) for each of them which 
will be taken into account for the validation.}

\item{PFG.considered_PFG.compo}{a character vector which contains the list of PFG considered
in the validation.}

\item{strata.considered_PFG.compo}{a character vector which contains the list of precise 
strata considered in the validation.}

\item{habitat.considered_PFG.compo}{a character vector which contains the list of habitat(s)
considered in the validation.}

\item{perStrata}{\code{Logical}. All strata together (FALSE) or per strata (TRUE).}

\item{name.simulation}{simulation folder name.}
}
\value{

}
\description{
This script is designed to compute distribution, per PFG/strata/habitat,
of relative abundance, from observed data.
}
\details{
The function takes the \code{releves.PFG} file and aggregate coverage per PFG. 
Then, the code gets habitat information from the \code{hab.obs} map & the \code{studied.habitat}
data frame, keep only interesting habitat(s), strata and PFG, and transforms 
the data into relative metrics. Finally, the script computes distribution per PFG 
and if required per strata/habitat (else all strata/habitat will be considered together).
}
\author{
Matthieu Combaud, Maxime Delprat
}
