% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.package_phyloclim.R
\name{niche.overlap}
\alias{niche.overlap}
\alias{di.pno}
\alias{di.enm}
\title{From phyloclim package 0.9.5 : niche.overlap and di.enm functions}
\usage{
di.pno(x, y)
di.enm(rn, x, y)
niche.overlap(x)
}
\arguments{
\item{x}{Might take one of the follwing forms: (1) a data frame containing 
predicted niche occupancy (PNO) profiles, e.g., as derived by 
\code{di.pno}; (2) a vector of filenames referencing to 
probability surfaces saved in ASCII-format with an ArcGIS-compatible header; 
(3) a list containing probability surfaces stored as objects of class 
\code{\link[sp]{SpatialGrid}}}
}
\description{
This function quantifies the degree of niche overlap using 
the statistics D and I (as proposed by Warren et al., 2008) based on 
Schoeners D (Schoener, 1968) and Hellinger Distances (van der Vaart, 1998). 
Niche overlap can be calculated either from the probability surfaces of 
entire ecological niche models (Warren et al., 2008) or from single 
dimensions of the climatic niches (Evans et al., 2009).
}
\seealso{
\code{\href{niche.overlap}{https://www.rdocumentation.org/packages/phyloclim/versions/0.9-4/topics/niche.overlap}}
}
\keyword{internal}
\keyword{phyloclim,}
