% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.plot_predicted_habitat.R
\name{plot_predicted_habitat}
\alias{plot_predicted_habitat}
\title{Create a raster map of habitat prediction for a specific \code{FATE} 
simulation at the last simulation year.}
\usage{
plot_predicted_habitat(
  predicted.habitat,
  col.df,
  simulation.map,
  output.path,
  sim.version
)
}
\arguments{
\item{predicted.habitat}{a csv file created by the do.habitat.validation function
which contain, for each pixel of the studied map, the simulated and observed habitat.}

\item{col.df}{a data frame with all the colors associated with the failure or 
success of each studied habitat prediction.}

\item{simulation.map}{a raster map of the whole studied area.}

\item{output.path}{access path to the for the folder where output files
will be created.}

\item{sim.version}{name of the simulation we want to validate.}
}
\value{
a synthetic.prediction.png file which contain the final prediction plot.
}
\description{
This script is designed to create a raster map of habitat prediction
based on a habitat prediction file. For each pixel, the habitat failure or success value
is associated to a color and then, the map is built.
}
\details{
The function determines true/false prediction ('failure' if false, 'success' if true)
and prepare a dataframe containing color and habitat code. Then, the script merge
the prediction dataframe with the color and code habitat dataframe. Finally,
the function draw a raster map and a plot of prediction habitat over it thanks
to the data prepared before.
}
\author{
Matthieu Combaud, Maxime Delprat
}
