% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PNE_RESULTS}
\alias{PNE_RESULTS}
\title{Ecrins National Park (PNE) data : outputs from \code{FATE-HD} 
simulation}
\format{A \code{list} object with 5 elements :

\describe{
  \item{evaluation}{a \code{data.frame} of dimension \code{24 x 11} \cr
    containing statistical metrics to evaluate the predictive quality of 
    both Habitat Suitability models (HS) and \code{FATE-HD} simulation 
    outputs \cr
    \itemize{
      \item \strong{PFG} : name of Plant Functional Group
      \item \strong{nb.obs.absences} : number of observed absences
      \item \strong{nb.obs.presences} : number of observed presences
      \item \strong{specificity.FATE} : proportion of actual negatives that 
      are correctly identified as such by the \code{FATE-HD} model (true 
      negative rate)
      \item \strong{specificity.HS} : proportion of actual negatives that 
      are correctly identified as such by the Habitat Suitability model 
      (true negative rate)
      \item \strong{sensitivity.FATE} : proportion of actual positives that 
      are correctly identified as such by the \code{FATE-HD} model (true 
      positive rate)
      \item \strong{sensitivity.HS} : proportion of actual positives that 
      are correctly identified as such by the Habitat Suitability model 
      (true positive rate)
      \item \strong{TSS.FATE} : True Skill Statistic for \code{FATE-HD} 
      model
      \item \strong{TSS.HS} : True Skill Statistic for Habitat Suitability 
      model
      \item \strong{error.rate.FATE} : percentage of bad predictions for 
      \code{FATE-HD} model
      \item \strong{error.rate.HS} : percentage of bad predictions for 
      Habitat Suitability model
    }
    }
  \item{abund_str.equilibrium}{a \code{\link[raster]{stack}} object of dimension 
    \code{782 x 619} with a resolution of \code{100m} and Lambers (lcc) 
    projection, containing 120 layers representing \code{FATE-HD} 
    \strong{abundances for year 800 of initialization phase} (= 
    \emph{equilibrium} or \emph{current} time). \cr
    Maps are per PFG and per height stratum.}
  \item{forest_cover.init}{a \code{\link[raster]{stack}} object of dimension 
    \code{782 x 619} with a resolution of \code{100m} and Lambers (lcc) 
    projection, containing 16 layers representing \code{FATE-HD} 
    \strong{relative forest cover through initialization phase (from year 50 
    to 800)}. \cr
    Maps are summing all PFGs through all height strata above 1.5 meters. \cr
    }
  \item{forest_cover \cr CC_BAU}{a \code{\link[raster]{stack}} object of dimension 
    \code{782 x 619} with a resolution of \code{100m} and Lambers (lcc) 
    projection, containing 14 layers representing \code{FATE-HD} 
    \strong{relative forest cover through Climate Change (CC) + 
    Business-As-Usual (BAU) scenario (from year 850 to 1500)}. \cr
    Maps are summing all PFGs through all height strata above 1.5 meters. \cr
    }
  \item{forest_cover\cr CC_Abandon}{a \code{\link[raster]{stack}} object of dimension 
    \code{782 x 619} with a resolution of \code{100m} and Lambers (lcc) 
    projection, containing 14 layers representing \code{FATE-HD} 
    \strong{relative forest cover through Climate Change (CC) + Abandonment (BAU) 
    scenario (from year 850 to 1500)}. \cr
    Maps are summing all PFGs through all height strata above 1.5 meters. \cr
    }
}}
\source{

}
\usage{
PNE_RESULTS
}
\description{
This dataset contains results obtained from outputs of 
\code{FATE-HD} simulation.
}
\author{
Maya Guéguen
}
\keyword{datasets}
