% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POST_FATE.graphic_evolutionCoverage.R
\name{POST_FATE.graphic_evolutionCoverage}
\alias{POST_FATE.graphic_evolutionCoverage}
\title{Create a graphical representation of the evolution of PFG coverage
and abundance through time for a \code{FATE} simulation}
\usage{
POST_FATE.graphic_evolutionCoverage(
  name.simulation,
  file.simulParam = NULL,
  opt.fixedScale = TRUE,
  opt.doPlot = TRUE
)
}
\arguments{
\item{name.simulation}{a \code{string} corresponding to the main directory 
or simulation name of the \code{FATE} simulation}

\item{file.simulParam}{default \code{NULL}. \cr A \code{string} 
corresponding to the name of a parameter file that will be contained into 
the \code{PARAM_SIMUL} folder of the \code{FATE} simulation}

\item{opt.fixedScale}{(\emph{optional}) default \code{TRUE}. \cr If 
\code{FALSE}, the ordinate scale will be adapted for each PFG for the 
graphical representation of the evolution of abundances through time}

\item{opt.doPlot}{(\emph{optional}) default \code{TRUE}. \cr If \code{TRUE}, 
plot(s) will be processed, otherwise only the calculation and reorganization 
of outputs will occur, be saved and returned}
}
\value{
A \code{list} containing two \code{data.frame} objects with the 
following columns, and two \code{ggplot2} objects :

\describe{
  \item{tab.spaceOccupancy}{
    \describe{
      \item{\code{PFG}}{concerned plant functional group (for abundance)}
      \item{\code{HAB}}{concerned habitat}
      \item{\emph{year}}{concerned simulation year}
      \item{\code{spaceOccupancy}}{number of occupied pixels divided by the 
      total number of pixels within the studied area}
    }
  }
  \item{tab.totalAbundance}{
    \describe{
      \item{\code{PFG}}{concerned plant functional group (for abundance)}
      \item{\code{HAB}}{concerned habitat}
      \item{\emph{year}}{concerned simulation year}
      \item{\code{totalAbundance}}{total abundance over all the pixels 
      within the studied area}
    }
  }
  \item{plot.spaceOccupancy}{\code{ggplot2} object, representing the 
  evolution of each PFG space occupancy}
  \item{plot.totalAbundance}{\code{ggplot2} object, representing the 
  evolution of each PFG total abundance \cr \cr}
}


Two \code{POST_FATE_TABLE_ZONE_evolution_[...].csv} files are created : 
\describe{
  \item{\file{spaceOccupancy}}{always, containing \code{tab.spaceOccupancy}}
  \item{\file{totalAbundance}}{always, containing \code{tab.totalAbundance}}
}


One \code{POST_FATE_GRAPHIC_A_evolution_coverage_[...].pdf} file is created 
containing two types of graphics : 
\describe{
  \item{spaceOccupancy}{to visualize for each PFG the evolution of its 
  occupation of the studied area through simulation time}
  \item{totalAbundance}{to visualize for each PFG the evolution of its 
  abundance within the whole studied area through simulation time}
}
}
\description{
This script is designed to produce two graphical 
representations for a \code{FATE} simulation : 1) the evolution through
time of the space occupation of each PFG ; 2) the evolution through time
of the abundance of each PFG. These graphics represent both the evolution 
over the whole area.
}
\details{
This function allows to obtain, for a specific \code{FATE} simulation and 
a specific parameter file within this simulation, two preanalytical 
graphics : 

\itemize{
  \item{the evolution of \strong{space occupancy} of each plant functional 
  group through simulation time, \cr with \emph{space occupancy} 
  representing the percentage of pixels within the mask of studied area 
  where the PFG is present
  }
  \item{the evolution of \strong{total abundance} of each plant functional 
  group through simulation time, \cr with \emph{total abundance} being the 
  sum over the whole studied area of the PFG abundances (\code{FATE} 
  \emph{arbitrary unit})
  }
}

If the information has been provided (see 
\code{\link{POST_FATE.temporalEvolution}}), the graphics will be also done 
per habitat. \cr \cr

\strong{It requires} that the \code{\link{POST_FATE.temporalEvolution}} 
function has been run and that the file 
\code{POST_FATE_TABLE_PIXEL_evolution_abundance.csv} exists.
}
\examples{

\dontrun{                      
POST_FATE.graphic_evolutionCoverage(name.simulation = "FATE_simulation"
                                    , file.simulParam = "Simul_parameters_V1.txt")
                                    
POST_FATE.graphic_evolutionCoverage(name.simulation = "FATE_simulation"
                                    , file.simulParam = "Simul_parameters_V1.txt"
                                    , opt.fixedScale = FALSE)
}



## ----------------------------------------------------------------------------------------- ##
## Load example data


}
\seealso{
\code{\link{POST_FATE.temporalEvolution}}
}
\author{
Maya Guéguen
}
\keyword{FATE,}
\keyword{abundance}
\keyword{outputs,}
\keyword{through}
\keyword{time}
