% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POST_FATE.graphic_evolutionCoverage.R
\name{POST_FATE.graphic_evolutionCoverage}
\alias{POST_FATE.graphic_evolutionCoverage}
\title{Create a graphical representation of the evolution of PFG coverage
and abundance through time for a \code{FATE-HD} simulation}
\usage{
POST_FATE.graphic_evolutionCoverage(name.simulation,
  file.simulParam = NULL, opt.abund_fixedScale = TRUE,
  opt.doPlot = TRUE)
}
\arguments{
\item{name.simulation}{a \code{string} that corresponds to the main 
directory or simulation name of the \code{FATE-HD} simulation}

\item{file.simulParam}{a \code{string} that corresponds to the name of a 
parameter file that will be contained into the \code{PARAM_SIMUL} folder 
of the \code{FATE-HD} simulation}

\item{opt.abund_fixedScale}{default \code{TRUE}. If \code{FALSE}, the 
ordinate scale will be adapted for each PFG for the graphical 
representation of the  evolution of abundances through time}

\item{opt.doPlot}{default \code{TRUE} (\emph{optional}). If TRUE, plot(s) 
will be processed, otherwise only the calculation and reorganization of 
outputs will occur, be saved and returned.}
}
\value{
Two \code{POST_FATE_[...].pdf} files are created : 
\describe{
  \item{\file{GRAPHIC_A \cr spaceOccupancy}}{to visualize for each PFG the
  evolution of its occupation of the studied area through simulation time}
  \item{\file{GRAPHIC_A \cr abundance}}{to visualize for each PFG the
  evolution of its abundance within the whole studied area through
  simulation time}
}
}
\description{
This script is designed to produce two graphical 
representations for a \code{FATE-HD} simulation : 1) the evolution through
time of the space occupation of each PFG ; 2) the evolution through time
of the abundance of each PFG. These graphics represent both the evolution 
over the whole area.
}
\details{
This function allows one to obtain, for a specific \code{FATE-HD} 
simulation and a specific parameter file within this simulation, two 
preanalytical graphics : 

\itemize{
  \item{the evolution of \strong{space occupancy} of each Plant Functional 
  Group through simulation time, with \emph{space occupancy} representing 
  the percentage of pixels within the mask of studied area where the PFG 
  is present
  }
  \item{the evolution of \strong{abundance} of each Plant Functional Group 
  through simulation time, with \emph{abundance} being the sum over the 
  whole studied area of the PFG abundances (\code{FATE-HD} \emph{arbitrary 
  unit})
  }
}

It requires that the \code{\link{POST_FATE.temporalEvolution}} has been run 
and that the \code{POST_FATE.evolution_abundance_PIXEL_[...].csv} exists.

If the information has been provided, the graphics will be also
done per habitat.
}
\examples{

\dontrun{                      
POST_FATE.graphic_evolutionCoverage(name.simulation = "FATE_simulation"
                                    , file.simulParam = "Simul_parameters_V1.txt"
                                    , opt.no_CPU = 1)
                                    
POST_FATE.graphic_evolutionCoverage(name.simulation = "FATE_simulation"
                                    , file.simulParam = "Simul_parameters_V1.txt"
                                    , opt.no_CPU = 4)
                                    
POST_FATE.graphic_evolutionCoverage(name.simulation = "FATE_simulation"
                                    , file.simulParam = "Simul_parameters_V1.txt"
                                    , opt.abund_fixedScale = FALSE
                                    , opt.no_CPU = 1)
}
                                    
                                    

}
\seealso{
\code{\link{POST_FATE.temporalEvolution}}
}
\author{
Maya Guéguen
}
\keyword{FATE,}
\keyword{abundance}
\keyword{outputs,}
\keyword{through}
\keyword{time}
