% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POST_FATE.graphic_evolutionCoverage.R
\name{POST_FATE.graphic_evolutionCoverage}
\alias{POST_FATE.graphic_evolutionCoverage}
\title{Create a graphical representation of the evolution of PFG coverage
and abundance through time for a \code{FATE-HD} simulation}
\usage{
POST_FATE.graphic_evolutionCoverage(name.simulation,
  file.simulParam = NULL, no.years = 10, opt.abund_fixedScale = TRUE,
  opt.ras_habitat = NULL, opt.no_CPU = 1)
}
\arguments{
\item{name.simulation}{a \code{string} that corresponds to the main directory
or simulation name of the \code{FATE-HD} simulation}

\item{file.simulParam}{a \code{string} that corresponds to the name of a
parameter file that will be contained into the \code{PARAM_SIMUL} folder
of the \code{FATE-HD} simulation}

\item{no.years}{an \code{integer} corresponding to the number of simulation
years that will be used to extract PFG abundance maps}

\item{opt.abund_fixedScale}{default \code{TRUE}. If \code{FALSE}, the
ordinate scale will be adapted for each PFG for the graphical representation
of the  evolution of abundances through time}

\item{opt.ras_habitat}{default NULL (\emph{optional}). A \code{string} that
corresponds to the file name of a raster mask, with an \code{integer} value
within each pixel, corresponding to a specific habitat}

\item{opt.no_CPU}{default 1 (\emph{optional}). The number of resources that 
can be used to parallelize the \code{unzip/zip} of raster files}
}
\value{
Two \code{POST_FATE_[...].pdf} files are created : 
\describe{
  \item{\file{GRAPHIC_A \cr spaceOccupancy}}{to visualize for each PFG the
  evolution of its occupation of the studied area through simulation time}
  \item{\file{GRAPHIC_A \cr abundance}}{to visualize for each PFG the
  evolution of its abundance within the whole studied area through
  simulation time}
}
}
\description{
This script is designed to produce two graphical 
representations for a \code{FATE-HD} simulation : 1) the evolution through
time of the space occupation of each PFG ; 2) the evolution through time
of the abundance of each PFG.
}
\details{
This function allows one to obtain, for a specific \code{FATE-HD} simulation
and a specific parameter file within this simulation, two preanalytical
graphics. \cr

For each PFG and each selected simulation year, raster maps are retrieved
from the results folder \code{ABUND_perPFG_allStrata} and unzipped.
Informations extracted lead to the production of two graphics before the
maps are compressed again :

\itemize{
  \item{the evolution of \strong{space occupancy} of each Plant Functional
  Group through simulation time, with \emph{space occupancy} representing
  the percentage of pixels within the mask of studied area where the PFG
  is present
  }
  \item{the evolution of \strong{abundance} of each Plant Functional
  Group through simulation time, with \emph{abundance} being the sum over
  the whole studied area of the PFG abundances (\code{FATE-HD} 
  \emph{arbitrary unit})
  }
}

If a raster mask for habitat has been provided, the graphics will be also
done per habitat.
}
\examples{

\dontrun{                      
POST_FATE.graphic_evolutionCoverage(name.simulation = "FATE_simulation"
                                    , file.simulParam = "Simul_parameters_V1.txt"
                                    , opt.no_CPU = 1)
                                    
POST_FATE.graphic_evolutionCoverage(name.simulation = "FATE_simulation"
                                    , file.simulParam = "Simul_parameters_V1.txt"
                                    , no.years = 50
                                    , opt.no_CPU = 4)
                                    
POST_FATE.graphic_evolutionCoverage(name.simulation = "FATE_simulation"
                                    , file.simulParam = "Simul_parameters_V1.txt"
                                    , no.years = 50
                                    , opt.abund_fixedScale = FALSE
                                    , opt.no_CPU = 1)
}
                                    
                                    

}
\seealso{
\code{\link{POST_FATE.relativeAbund_presenceAbsence}}
}
\author{
Maya Guéguen
}
\keyword{FATE,}
\keyword{abundance}
\keyword{outputs,}
\keyword{through}
\keyword{time}
