% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POST_FATE.graphic_evolutionPixels.R
\name{POST_FATE.graphic_evolutionPixels}
\alias{POST_FATE.graphic_evolutionPixels}
\title{Create a graphical representation of the evolution of PFG abundance
through time for 5 (or more) pixels of a \code{FATE-HD} simulation}
\usage{
POST_FATE.graphic_evolutionPixels(name.simulation,
  file.simulParam = NULL, opt.abund_fixedScale = TRUE,
  opt.cells_ID = NULL, opt.doPlot = TRUE)
}
\arguments{
\item{name.simulation}{a \code{string} that corresponds to the main directory
or simulation name of the \code{FATE-HD} simulation}

\item{file.simulParam}{a \code{string} that corresponds to the name of a
parameter file that will be contained into the \code{PARAM_SIMUL} folder
of the \code{FATE-HD} simulation}

\item{opt.abund_fixedScale}{default \code{TRUE}. If \code{FALSE}, the ordinate
scale will be adapted for each PFG for the graphical representation of the 
evolution of abundances through time}

\item{opt.cells_ID}{default NULL (\emph{optional}). The cells ID of the 
studied area for which PFG abundances will be extracted.}

\item{opt.doPlot}{default TRUE (\emph{optional}). If TRUE, plot(s) will be
processed, otherwise only the calculation and reorganization of outputs
will occur, be saved and returned.}
}
\value{
One \code{POST_FATE_[...].pdf} file is created : 
\describe{
  \item{\file{GRAPHIC_A \cr abundance}}{to visualize for each PFG the
  evolution of its abundance within each selected pixel through
  simulation time}
}
}
\description{
This script is designed to produce one graphical representation
for a \code{FATE-HD} simulation : the evolution through time of the 
abundance of each PFG for 5 (or more) randomly selected cells of the
studied area.
}
\details{
This function allows one to obtain, for a specific \code{FATE-HD} simulation
and a specific parameter file within this simulation, one preanalytical
graphics. \cr

For each PFG and each selected simulation year, raster maps are retrieved
from the results folder \code{ABUND_perPFG_allStrata} and unzipped.
Informations extracted lead to the production of one graphic before the
maps are compressed again :

\itemize{
  \item{the evolution of \strong{abundance} of each Plant Functional
  Group through simulation time, within 5 (or more) randomly selected pixels
  of the studied area (\code{FATE-HD} \emph{arbitrary unit})
  }
}
}
\examples{

\dontrun{                      
POST_FATE.graphic_evolutionPixels(name.simulation = "FATE_simulation"
                                    , file.simulParam = "Simul_parameters_V1.txt")
                                    
POST_FATE.graphic_evolutionPixels(name.simulation = "FATE_simulation"
                                    , file.simulParam = "Simul_parameters_V1.txt"
                                    , opt.abund_fixedScale = FALSE)
}
                                    
                                    

}
\seealso{
\code{\link{POST_FATE.relativeAbund}}, 
\code{\link{POST_FATE.graphic_validationStatistics}}
}
\author{
Maya Guéguen
}
\keyword{FATE,}
\keyword{abundance}
\keyword{outputs,}
\keyword{through}
\keyword{time}
