% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POST_FATE.graphic_mapPFG.R
\name{POST_FATE.graphic_mapPFG}
\alias{POST_FATE.graphic_mapPFG}
\title{Create a map related to plant functional group results (richness, 
relative cover, light or soil CWM) for one (or several) specific year of a 
\code{FATE} simulation}
\usage{
POST_FATE.graphic_mapPFG(
  name.simulation,
  file.simulParam = NULL,
  years,
  opt.stratum_min = 1,
  opt.stratum_max = 10,
  opt.doBinary = TRUE,
  opt.no_CPU = 1,
  opt.doPlot = TRUE
)
}
\arguments{
\item{name.simulation}{a \code{string} corresponding to the main directory 
or simulation name of the \code{FATE} simulation}

\item{file.simulParam}{default \code{NULL}. \cr A \code{string} 
corresponding to the name of a parameter file that will be contained into 
the \code{PARAM_SIMUL} folder of the \code{FATE} simulation}

\item{years}{an \code{integer}, or a \code{vector} of \code{integer}, 
corresponding to the simulation year(s) that will be used to extract PFG 
abundance and binary maps}

\item{opt.stratum_min}{(\emph{optional}) default \code{1}. \cr An 
\code{integer} corresponding to the lowest stratum from which PFG 
abundances will be summed up}

\item{opt.stratum_max}{(\emph{optional}) default \code{10}. \cr An 
\code{integer} corresponding to the highest stratum from which PFG 
abundances will be summed up}

\item{opt.doBinary}{(\emph{optional}) default \code{TRUE}. \cr If 
\code{TRUE}, abundance maps (absolute or relative) are systematically 
multiplied by binary maps (see 
\href{POST_FATE.graphic_mapPFG#details}{\code{Details}})}

\item{opt.no_CPU}{(\emph{optional}) default \code{1}. \cr The number of 
resources that can be used to parallelize the \code{unzip/zip} of raster 
files}

\item{opt.doPlot}{(\emph{optional}) default \code{TRUE}. \cr If \code{TRUE}, 
plot(s) will be processed, otherwise only the calculation and reorganization 
of outputs will occur, be saved and returned}
}
\value{
A \code{list} containing one or several (one for each simulation 
year) \code{list} of \code{raster} and \code{ggplot2} objects :

\describe{
  \item{tab}{ 
    \describe{
      \item{\code{cover}}{\code{raster} of relative coverage}
      \item{\code{DIV.0}}{\code{raster} of species richness}
      \item{\code{DIV.1}}{\code{raster} of Shannon entropy}
      \item{\code{DIV.2}}{\code{raster} of Simpson concentration}
      \item{\code{CWM.light}}{\code{raster} of light community weighted mean}
      \item{\code{CWM.soil}}{\code{raster} of soil community weighted mean}
    }
  }
  \item{plot}{
    \describe{
      \item{\code{cover}}{\code{ggplot2} object, representing \code{cover} 
      raster}
      \item{\code{richness}}{\code{ggplot2} object, representing 
      \code{DIV.0} raster}
      \item{\code{CWM.light}}{\code{ggplot2} object, representing 
      \code{CWM.light} raster}
      \item{\code{CWM.soil}}{\code{ggplot2} object, representing 
      \code{CWM.soil} raster \cr \cr}
    }
  }
}


\file{POST_FATE_GRAPHIC_C_map_PFG_[...].pdf} file is created containing up 
to four graphics : 
\describe{
  \item{\file{map_PFGcover}}{to visualize the PFG cover within the studied 
  area}
  \item{\file{map_PFGrichness}}{to visualize the PFG richness within the 
  studied area}
  \item{\file{PFGlight}}{to visualize the light CWM within the studied area}
  \item{\file{PFGsoil}}{to visualize the soil CWM within the studied area}
}

Three \file{PFGrichness_YEAR_[...]_STRATA_all_q[...].tif} files are created 
into the simulation results folder :
\describe{
  \item{\file{q0}}{PFG richness}
  \item{\file{q1}}{PFG Shannon entropy}
  \item{\file{q2}}{PFG Simpson concentration}
}

Raster files are also created for cover, and light and soil CWM if those 
modules were selected (see \code{\link{PRE_FATE.params_globalParameters}}).
}
\description{
This script is designed to produce one (or several) raster map 
related to plant functional group results (richness, relative cover, light 
or soil CWM) for one (or several) specific \code{FATE} simulation year.
}
\details{
This function allows to obtain, for a specific \code{FATE} simulation and 
a specific parameter file within this simulation, \strong{up to six raster 
maps} and preanalytical graphics. \cr \cr

For each PFG and each selected simulation year, raster maps are retrieved 
from the results folders \code{ABUND_perPFG_perStrata} and 
\code{BIN_perPFG_perStrata} and unzipped. 
Informations extracted lead to the production of up to six graphics before 
the maps are compressed again :

\itemize{
  \item{map of \strong{PFG richness} within each pixel, representing the sum 
  of binary maps \cr
  \strong{Richness} is calculated with the 
  \strong{\href{http://www.jstor.org/stable/23143936}{Leinster & Cobbold 
  2012 Ecology} framework} which allows to give more or less importance to 
  the commun species through the \code{q} parameter :
  \describe{
    \item{\code{q = 0}}{species richness}
    \item{\code{q = 1}}{Shannon entropy}
    \item{\code{q = 2}}{Simpson concentration \cr \cr}
  }
  }
  \item{map of \strong{PFG relative cover}, representing the sum of relative 
  abundance maps of all PFG \cr (potentially above a height threshold 
  defined by \code{opt.stratum_min})}
  \item{\strong{if light was activated} (see 
  \code{\link{PRE_FATE.params_globalParameters}}), \strong{community 
  weighted mean of PFG light preferences} (extracted from \code{LIGHT} 
  parameter within \code{LIGHT} files, see 
  \code{\link{PRE_FATE.params_PFGlight}})}
  \item{\strong{if soil was activated} (see 
  \code{\link{PRE_FATE.params_globalParameters}}), \strong{community 
  weighted mean of PFG soil preferences} (extracted from \code{SOIL_CONTRIB} 
  parameter within \code{SOIL} files, see 
  \code{\link{PRE_FATE.params_PFGsoil}}) \cr \cr}
}


\strong{It requires} that the \code{\link{POST_FATE.relativeAbund}}, 
(\code{\link{POST_FATE.graphic_validationStatistics}}) and 
\code{\link{POST_FATE.binaryMaps}} functions have been run and that the 
folders \code{BIN_perPFG_allStrata} and \code{BIN_perPFG_perStrata} exist. 
\cr \cr

If \code{opt.doBinary = TRUE}, abundance maps (absolute or relative) are 
systematically multiplied by binary maps extracted from 
\code{BIN_perPFG_allStrata} and \code{BIN_perPFG_perStrata} folders and 
produced by \code{\link{POST_FATE.binaryMaps}} function. 
This way, produced raster maps reflect the \emph{validated/refined} predictions. 
\code{opt.doBinary} can be set to \code{FALSE} to reflect \emph{pure} 
simulation results.
}
\examples{

\dontrun{                      
POST_FATE.graphic_mapPFG(name.simulation = "FATE_simulation"
                         , file.simulParam = "Simul_parameters_V1.txt"
                         , years = 850
                         , opt.stratum_min = 3
                         , opt.no_CPU = 1)
                                    
POST_FATE.graphic_mapPFG(name.simulation = "FATE_simulation"
                         , file.simulParam = "Simul_parameters_V1.txt"
                         , year = c(850, 950)
                         , opt.doBinary = FALSE)
}
                                    
                                    

}
\seealso{
\code{\link{POST_FATE.relativeAbund}}, 
\code{\link{POST_FATE.graphic_validationStatistics}},
\code{\link{POST_FATE.binaryMaps}}
}
\author{
Maya Guéguen
}
\keyword{FATE,}
\keyword{abundance,}
\keyword{community}
\keyword{cover,}
\keyword{forest}
\keyword{light,}
\keyword{mean,}
\keyword{outputs,}
\keyword{relative}
\keyword{richness,}
\keyword{soil}
\keyword{weighted}
