% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POST_FATE.graphic_mapPFGcover.R
\name{POST_FATE.graphic_mapPFGcover}
\alias{POST_FATE.graphic_mapPFGcover}
\title{Create a map of the Plant Functional Group cover 
 \cr for one (or several) specific year of a \code{FATE-HD} simulation}
\usage{
POST_FATE.graphic_mapPFGcover(name.simulation, file.simulParam = NULL,
  year, strata_min, opt.no_CPU = 1)
}
\arguments{
\item{name.simulation}{a \code{string} that corresponds to the main directory
or simulation name of the \code{FATE-HD} simulation}

\item{file.simulParam}{a \code{string} that corresponds to the name of a
parameter file that will be contained into the \code{PARAM_SIMUL} folder
of the \code{FATE-HD} simulation}

\item{year}{an \code{integer} corresponding to the simulation year(s) that 
will be used to extract PFG abundance and binary maps}

\item{strata_min}{an \code{integer} corresponding to the lowest stratum from
which PFG abundances are summed up to the highest stratum}

\item{opt.no_CPU}{default 1 (\emph{optional}). The number of resources that 
can be used to parallelize the \code{unzip/zip} of raster files}
}
\value{
One \code{POST_FATE_[...].pdf} file is created : 
\describe{
  \item{\file{GRAPHIC_D \cr PFGcover}}{to visualize the PFG cover
  within the studied area}
}
}
\description{
This script is designed to produce a raster map of PFG cover
for one (or several) specific \code{FATE-HD} simulation year.
}
\details{
This function allows one to obtain, for a specific \code{FATE-HD} simulation
and a specific parameter file within this simulation, one preanalytical
graphic. \cr

For each PFG and each selected simulation year, raster maps are retrieved
from the results folders \code{ABUND_perPFG_perStrata} and 
\code{BIN_perPFG_perStrata} and unzipped.
Informations extracted lead to the production of one graphic before the
maps are compressed again :

\itemize{
  \item{the map of \strong{Plant Functional Group cover} for each selected
  simulation year(s), representing the relative cumulated abundance of PFG 
  present in each pixel above a height threshold
  }
}
}
\examples{

\dontrun{                      
POST_FATE.graphic_mapPFGcover(name.simulation = "FATE_simulation"
                              , file.simulParam = "Simul_parameters_V1.txt"
                              , year = 850
                              , strata_min = 3
                              , opt.no_CPU = 1)
                                    
POST_FATE.graphic_mapPFGcover(name.simulation = "FATE_simulation"
                              , file.simulParam = "Simul_parameters_V1.txt"
                              , year = c(850, 950)
                              , strata_min = 3
                              , opt.no_CPU = 1)
}
                                    
                                    

}
\author{
Maya Guéguen
}
