% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POST_FATE.graphic_mapPFGrichness.R
\name{POST_FATE.graphic_mapPFGrichness}
\alias{POST_FATE.graphic_mapPFGrichness}
\title{Create a map of the Plant Functional Group richness 
 \cr for one (or several) specific year of a \code{FATE-HD} simulation}
\usage{
POST_FATE.graphic_mapPFGrichness(name.simulation, file.simulParam = NULL,
  year, opt.no_CPU = 1, opt.doPlot = TRUE)
}
\arguments{
\item{name.simulation}{a \code{string} that corresponds to the main 
directory or simulation name of the \code{FATE-HD} simulation}

\item{file.simulParam}{a \code{string} that corresponds to the name of a 
parameter file that will be contained into the \code{PARAM_SIMUL} folder 
of the \code{FATE-HD} simulation}

\item{year}{an \code{integer} corresponding to the simulation year(s) that 
will be used to extract PFG binary maps}

\item{opt.no_CPU}{default \code{1} (\emph{optional}). The number of 
resources that can be used to parallelize the \code{unzip/zip} of raster 
files}

\item{opt.doPlot}{default \code{TRUE} (\emph{optional}). If \code{TRUE}, 
plot(s) will be processed, otherwise only the calculation and 
reorganization of outputs will occur, be saved and returned.}
}
\value{
A \code{list} containing one or several (one for each simulation 
year) \code{list} of \code{ggplot2} objects, representing the Plant 
Functional Group richness for each value of \code{q} (0, 1 or 2). \cr \cr

Three \code{PFGrichness_YEAR_[...]_STRATA_all_q[...].tif} files are 
created : 
\describe{
  \item{\file{q0}}{PFG richness}
  \item{\file{q1}}{PFG Shannon entropy}
  \item{\file{q2}}{PFG Simpson concentration}
}

One \code{POST_FATE_[...].pdf} file is created : 
\describe{
  \item{\file{GRAPHIC_C \cr map_PFGrichness}}{to visualize the PFG 
  richness within the studied area}
}
}
\description{
This script is designed to produce a raster map of PFG 
richness for one (or several) specific \code{FATE-HD} simulation year.
}
\details{
This function allows one to obtain, for a specific \code{FATE-HD} 
simulation and a specific parameter file within this simulation, three 
raster maps of PFG richness and one preanalytical graphic. \cr \cr

For each PFG and each selected simulation year, raster maps are retrieved 
from the results folders \code{ABUND_REL_perPFG_allStrata} and 
\code{BIN_perPFG_allStrata} and unzipped.
Informations extracted lead to the production of richness maps and one 
graphic before the maps are compressed again :

\itemize{
  \item{for each selected simulation year(s), the map of \strong{Plant 
  Functional Group richness} representing the number of PFG present in 
  each pixel}
  \item{\strong{richness} is calculated with the 
  \strong{\href{http://www.jstor.org/stable/23143936}{Leinster & Cobbold 
  2012 Ecology} framework} which allows to give more or less importance to 
  the commun species through the \code{q} parameter :
  \describe{
    \item{\code{q = 0}}{species richness}
    \item{\code{q = 1}}{Shannon entropy}
    \item{\code{q = 2}}{Simpson concentration \cr \cr}
  }
  }
}

It requires that the \code{\link{POST_FATE.relativeAbund}} and 
\code{\link{POST_FATE.graphic_validationStatistics}} functions have 
been run and that the folder \code{BIN_perPFG_allStrata} exists.
}
\examples{

\dontrun{                      
POST_FATE.graphic_mapPFGrichness(name.simulation = "FATE_simulation"
                                 , file.simulParam = "Simul_parameters_V1.txt"
                                 , year = 850
                                 , opt.no_CPU = 1)
                                    
POST_FATE.graphic_mapPFGrichness(name.simulation = "FATE_simulation"
                                 , file.simulParam = "Simul_parameters_V1.txt"
                                 , year = c(850, 950)
                                 , opt.no_CPU = 1)
}
                                    
                                    

}
\seealso{
\code{\link{POST_FATE.relativeAbund}}, 
\code{\link{POST_FATE.graphic_validationStatistics}}
}
\author{
Maya Guéguen
}
\keyword{FATE,}
\keyword{PFG}
\keyword{outputs,}
\keyword{richness}
