% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POST_FATE.graphic_validationStatistics.R
\name{POST_FATE.graphic_validationStatistics}
\alias{POST_FATE.graphic_validationStatistics}
\title{Create a graphical representation of several statistics for each PFG 
to asses the quality of the model for one (or several) specific year of 
a \code{FATE-HD} simulation}
\usage{
POST_FATE.graphic_validationStatistics(name.simulation,
  file.simulParam = NULL, year, mat.PFG.obs, opt.no_CPU = 1)
}
\arguments{
\item{name.simulation}{a \code{string} that corresponds to the main directory
or simulation name of the \code{FATE-HD} simulation}

\item{file.simulParam}{a \code{string} that corresponds to the name of a
parameter file that will be contained into the \code{PARAM_SIMUL} folder
of the \code{FATE-HD} simulation}

\item{year}{an \code{integer} corresponding to the simulation year(s) that 
will be used to extract PFG abundance maps}

\item{mat.PFG.obs}{a}

\item{opt.no_CPU}{default 1 (\emph{optional}). The number of resources that 
can be used to parallelize the \code{unzip/zip} of raster files}
}
\value{
One \code{POST_FATE_[...].pdf} file is created : 
\describe{
  \item{\file{GRAPHIC_E \cr validationStatistics}}{to assess the modeling 
  quality of each PFG based on given observations within the studied area}
}
}
\description{
This script is designed to produce a graphical representation
of several statistics (sensitivity, specificity, TSS, AUC) for quality
assessment for one (or several) specific \code{FATE-HD} simulation year.
}
\details{
This function allows one to obtain, for a specific \code{FATE-HD} simulation
and a specific parameter file within this simulation, one preanalytical
graphic. \cr

For each PFG and each selected simulation year, raster maps are retrieved
from the results folder \code{ABUND_perPFG_allStrata} and unzipped.
Informations extracted lead to the production of one graphic before the
maps are compressed again :

\itemize{
  \item{the value of \strong{several statistics for the predictive quality
  of the model for each Plant Functional Group} and for each selected
  simulation year(s)
  }
}
}
\examples{

\dontrun{                      
POST_FATE.graphic_validationStatistics(name.simulation = "FATE_simulation"
                                 , file.simulParam = "Simul_parameters_V1.txt"
                                 , year = 850
                                 , opt.no_CPU = 1)
                                    
POST_FATE.graphic_validationStatistics(name.simulation = "FATE_simulation"
                                 , file.simulParam = "Simul_parameters_V1.txt"
                                 , year = c(850, 950)
                                 , opt.no_CPU = 1)
                                    
POST_FATE.graphic_validationStatistics(name.simulation = "FATE_simulation"
                                 , file.simulParam = "Simul_parameters_V1.txt"
                                 , year = 850
                                 , opt.no_CPU = 1
                                 , opt.strata = 2)
}
                                    
                                    

}
\author{
Maya Guéguen
}
