% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POST_FATE.graphic_validationStatistics.R
\name{POST_FATE.graphic_validationStatistics}
\alias{POST_FATE.graphic_validationStatistics}
\title{Create a graphical representation of several statistics for each PFG 
to asses the quality of the model 
 \cr for one (or several) specific year of a \code{FATE-HD} simulation}
\usage{
POST_FATE.graphic_validationStatistics(name.simulation,
  file.simulParam = NULL, year, mat.PFG.obs, opt.no_CPU = 1)
}
\arguments{
\item{name.simulation}{a \code{string} that corresponds to the main directory
or simulation name of the \code{FATE-HD} simulation}

\item{file.simulParam}{a \code{string} that corresponds to the name of a
parameter file that will be contained into the \code{PARAM_SIMUL} folder
of the \code{FATE-HD} simulation}

\item{year}{an \code{integer} corresponding to the simulation year(s) that 
will be used to extract PFG binary maps}

\item{mat.PFG.obs}{a \code{data.frame} with 4 columns : PFG, X, Y, obs}

\item{opt.no_CPU}{default 1 (\emph{optional}). The number of resources that 
can be used to parallelize the \code{unzip/zip} of raster files}
}
\value{
A \code{data.frame} with the following columns :
\describe{
  \item{\code{PFG}}{the concerned Plant Functional Group}
  \item{\code{AUC.sd}}{standard deviation of the AUC values}
  \item{\code{sensitivity.sd}}{standard deviation of the sensitivity values}
  \item{\code{specificity.sd}}{standard deviation of the specificity values}
  \item{\code{variable}}{name of the calculated statistic among 'sensitivity',
  'specificity', 'TSS' and 'AUC'}
  \item{\code{value}}{value of the corresponding statistic}
}

One folder is created :
\describe{
  \item{\file{BIN_perPFG \cr_allStrata}}{containing presence / absence  
  raster maps for each PFG across all strata}
}

One \code{POST_FATE_[...].pdf} file is created : 
\describe{
  \item{\file{GRAPHIC_C \cr validationStatistics}}{to assess the modeling 
  quality of each PFG based on given observations within the studied area}
}
}
\description{
This script is designed to produce a graphical representation
of several statistics (sensitivity, specificity, TSS, AUC) for quality
assessment for one (or several) specific \code{FATE-HD} simulation year.
}
\details{
This function allows one to obtain, for a specific \code{FATE-HD} simulation
and a specific parameter file within this simulation, one preanalytical
graphic. \cr

For each PFG and each selected simulation year, raster maps are retrieved
from the results folder \code{ABUND_REL_perPFG_allStrata} and unzipped.
Informations extracted lead to the production of presence/absence maps and 
one graphic before the maps are compressed again :

\itemize{
  \item{the value of \strong{several statistics for the predictive quality
  of the model for each Plant Functional Group} and for each selected
  simulation year(s)
  }
}

Observation records (presences and absences) are required for each PFG 
within the \code{mat.PFG.obs} object :

\describe{
  \item{\code{PFG}}{the concerned Plant Functional Group}
  \item{\code{X} and \code{Y}}{the coordinates of each observations,
  matching with the projection of the mask of \code{name.simulation}}
  \item{\code{obs}}{either 0 or 1 to indicate presences or absences}
}
}
\examples{

\dontrun{                      
POST_FATE.graphic_validationStatistics(name.simulation = "FATE_simulation"
                                       , file.simulParam = "Simul_parameters_V1.txt"
                                       , year = 850
                                       , opt.no_CPU = 1)
                                    
POST_FATE.graphic_validationStatistics(name.simulation = "FATE_simulation"
                                       , file.simulParam = "Simul_parameters_V1.txt"
                                       , year = c(850, 950)
                                       , opt.no_CPU = 1)
}
                                    
                                    

}
\seealso{
\code{\link{POST_FATE.relativeAbund_presenceAbsence}}
}
\author{
Maya Guéguen
}
\keyword{FATE,}
\keyword{area}
\keyword{binary,}
\keyword{curve,}
\keyword{outputs,}
\keyword{sensitivity,}
\keyword{skill}
\keyword{specificity,}
\keyword{statistic}
\keyword{true}
\keyword{under}
