% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POST_FATE.graphics.R
\name{POST_FATE.graphics}
\alias{POST_FATE.graphics}
\title{Create all possible graphical representations for a \code{FATE-HD} 
simulation}
\usage{
POST_FATE.graphics(name.simulation, file.simulParam, year, no.years,
  opt.strata_min = 1, opt.strata = "ALL", opt.mat.PFG.succ = NULL,
  opt.mat.PFG.obs = NULL, opt.mat.CWM.obs = NULL,
  opt.mat.cover.obs = NULL, opt.mat.light.obs = NULL,
  opt.mat.soil.obs = NULL, opt.ras.cover.obs = NULL,
  opt.ras.light.obs = NULL, opt.ras.soil.obs = NULL,
  opt.ras_habitat = NULL, opt.cells_ID = NULL,
  opt.abund_fixedScale = TRUE, opt.no_CPU = 1, opt.doPlot = TRUE,
  opt.doFunc.evolutionCoverage = TRUE,
  opt.doFunc.evolutionPixels = TRUE, opt.doFunc.validation = TRUE,
  opt.doFunc.mapPFGvsHS = TRUE, opt.doFunc.mapPFGrichness = TRUE,
  opt.doFunc.mapPFGcover = TRUE, opt.doFunc.mapPFGlight = TRUE,
  opt.doFunc.mapPFGsoil = TRUE)
}
\arguments{
\item{name.simulation}{a \code{string} that corresponds to the main 
directory or simulation name of the \code{FATE-HD} simulation}

\item{file.simulParam}{a \code{string} that corresponds to the name of a
parameter file that will be contained into the \code{PARAM_SIMUL} folder
of the \code{FATE-HD} simulation}

\item{year}{an \code{integer} corresponding to the simulation year(s) that 
will be used to extract PFG abundance and/or binary maps}

\item{no.years}{an \code{integer} corresponding to the number of simulation 
years that will be used to extract PFG abundance maps}

\item{opt.strata_min}{default 1 (\emph{optional}). An \code{integer} 
corresponding to the lowest stratum from which PFG abundances are summed up 
to the highest stratum}

\item{opt.strata}{default ALL (\emph{optional}). The stratum number from 
which to extract PFG binary maps}

\item{opt.mat.PFG.succ}{default NULL (\emph{optional}). A \code{data.frame} 
with 2 or 3 columns : PFG, light, soil_contrib}

\item{opt.mat.PFG.obs}{default NULL (\emph{optional}). A \code{data.frame} 
with 4 columns : PFG, X, Y, obs (0/1)}

\item{opt.mat.CWM.obs}{default NULL (\emph{optional}). A \code{data.frame} 
with 3 or 4 columns : X, Y, light (CWM), soil (CWM)}

\item{opt.mat.cover.obs}{default NULL (\emph{optional}). A \code{data.frame}
with 3 columns : X, Y, obs (0/1)}

\item{opt.mat.light.obs}{default NULL (\emph{optional}). A \code{data.frame}
with 3 columns : X, Y, obs (0/1)}

\item{opt.mat.soil.obs}{default NULL (\emph{optional}). A \code{data.frame}
with 3 columns : X, Y, obs (0/1)}

\item{opt.ras.cover.obs}{default NULL (\emph{optional}). A \code{string} 
that corresponds to the file name of a raster containing observed values for 
vegetation cover}

\item{opt.ras.light.obs}{default NULL (\emph{optional}). A \code{string} 
that corresponds to the file name of a raster containing observed values for 
community weighted mean of light}

\item{opt.ras.soil.obs}{default NULL (\emph{optional}). A \code{string} 
that corresponds to the file name of a raster containing observed values for 
community weighted mean of soil}

\item{opt.ras_habitat}{default NULL (\emph{optional}). A \code{string} that 
corresponds to the file name of a raster mask, with an \code{integer} value 
within each pixel, corresponding to a specific habitat}

\item{opt.cells_ID}{default NULL (\emph{optional}). The cells ID of the 
studied area for which light resources will be extracted.}

\item{opt.abund_fixedScale}{default \code{TRUE}. If \code{FALSE}, the 
ordinate scale will be adapted for each PFG for the graphical representation 
of the  evolution of abundances through time}

\item{opt.no_CPU}{default 1 (\emph{optional}). The number of resources that 
can be used to parallelize the \code{unzip/zip} of raster files}

\item{opt.doPlot}{default TRUE (\emph{optional}). If TRUE, plot(s) will be 
processed, otherwise only the calculation and reorganization of outputs 
will occur, be saved and returned.}

\item{opt.doFunc.evolutionCoverage}{default TRUE (\emph{optional}). If TRUE, 
corresponding \code{POST_FATE.graphic_...} function will be run.}

\item{opt.doFunc.evolutionPixels}{default TRUE (\emph{optional}).If TRUE, 
corresponding \code{POST_FATE.graphic_...} function will be run.
 @param opt.doFunc.validation default TRUE (\emph{optional}). If TRUE, 
corresponding \code{POST_FATE.graphic_...} function will be run.}

\item{opt.doFunc.mapPFGvsHS}{default TRUE (\emph{optional}). If TRUE, 
corresponding \code{POST_FATE.graphic_...} function will be run.}

\item{opt.doFunc.mapPFGrichness}{default TRUE (\emph{optional}).If TRUE, 
corresponding \code{POST_FATE.graphic_...} function will be run.}

\item{opt.doFunc.mapPFGcover}{default TRUE (\emph{optional}). If TRUE, 
corresponding \code{POST_FATE.graphic_...} function will be run.}

\item{opt.doFunc.mapPFGlight}{default TRUE (\emph{optional}). If TRUE, 
corresponding \code{POST_FATE.graphic_...} function will be run.}

\item{opt.doFunc.mapPFGsoil}{default TRUE (\emph{optional}). If TRUE, 
corresponding \code{POST_FATE.graphic_...} function will be run.}
}
\description{
This script is designed to produce a set of graphical 
representations for a \code{FATE-HD} simulation. Graphics can be of three 
types : 1) representing an evolution through time (of abundance, light, 
soil) ; 2) representing a spatial distribution for a specific year 
(richness, abundance, light, soil) ; 3) or vizualising the goodness of the 
modelisation (presence/absence, validation statistics).
}
\details{
This function allows one to obtain, for a specific \code{FATE-HD} simulation 
and a specific parameter file within this simulation, up to ten 
preanalytical graphics. \cr

For each PFG and each selected simulation year, raster maps are retrieved
from the results folders (\code{ABUND_perPFG_allStrata} or 
\code{ABUND_perPFG_allStrata} or \code{ABUND_REL_perPFG_allStrata} or 
\code{BIN_perPFG_perStrata} or \code{BIN_perPFG_allStrata} or \code{LIGHT} 
or \code{SOIL}) and unzipped.
Informations extracted lead to the production of the following graphics 
before the maps are compressed again :

\itemize{
  \item{the evolution of \strong{space occupancy} of each Plant Functional
  Group through simulation time, with \emph{space occupancy} representing
  the percentage of pixels within the mask of studied area where the PFG
  is present (see \code{\link{POST_FATE.graphic_evolutionCoverage}})
  }
  \item{the evolution of \strong{abundance} of each Plant Functional
  Group through simulation time, with \emph{abundance} being the sum over
  the whole studied area of the PFG abundances (\code{FATE-HD} 
  \emph{arbitrary unit}) (see 
  \code{\link{POST_FATE.graphic_evolutionCoverage}})
  }
  \item{the evolution of \strong{abundance} of each Plant Functional
  Group through simulation time, within 5 (or more) randomly selected pixels
  of the studied area (\code{FATE-HD} \emph{arbitrary unit}), as well as 
  \strong{light resources} of each stratum (\emph{1: Low, 2: Medium, 3: High}) 
  and \strong{soil resources} if these modules are selected 
  (see \code{\link{POST_FATE.graphic_evolutionPixels}})
  }
  \item{the value of \strong{several statistics for the predictive quality
  of the model for each Plant Functional Group} and for each selected
  simulation year(s) (see 
  \code{\link{POST_FATE.graphic_validationStatistics}})
  }
  \item{the maps of \strong{Plant Functional Group habitat suitability and
  occurrences} for each selected simulation year(s), representing the 
  probability of presence of each PFG in each pixel compared to its
  simulated presence (see \code{\link{POST_FATE.graphic_mapPFGvsHS}})
  }
  \item{the map of \strong{Plant Functional Group richness} for each selected
  simulation year(s), representing the number of PFG present in each pixel 
  (see \code{\link{POST_FATE.graphic_mapPFGrichness}})
  }
  \item{the map of \strong{Plant Functional Group cover} for each selected
  simulation year(s), representing the relative cumulated abundance of PFG 
  present in each pixel above a height threshold (see 
  \code{\link{POST_FATE.graphic_mapPFGcover}})
  }
  \item{the map of \strong{light Community Weighted Mean} for each selected
  simulation year(s), representing the simulated value of light (Landolt)
  within each pixel above a height threshold (see 
  \code{\link{POST_FATE.graphic_mapPFGlight}})
  }
  \item{the map of \strong{soil Community Weighted Mean} for each selected
  simulation year(s), representing the simulated value of soil (Landolt)
  within each pixel (see \code{\link{POST_FATE.graphic_mapPFGsoil}})
  }
}
}
\examples{

\dontrun{                      
POST_FATE.graphics(name.simulation = "FATE_simulation"
                                    , file.simulParam = "Simul_parameters_V1.txt"
                                    , year = 100
                                    , no.years = 10)
                                    
POST_FATE.graphics(name.simulation = "FATE_simulation"
                                    , file.simulParam = "Simul_parameters_V1.txt"
                                    , year = 100
                                    , no.years = 10)
                                    , opt.abund_fixedScale = FALSE
                                    , opt.no_CPU = 4)
}
                                    
                                    

}
\seealso{
\code{\link{POST_FATE.relativeAbund}}, 
\code{\link{POST_FATE.graphic_validationStatistics}}, 
\code{\link{POST_FATE.graphic_mapPFGvsHS}}, 
\code{\link{POST_FATE.graphic_mapPFGcover}}, 
\code{\link{POST_FATE.graphic_mapPFGrichness}}, 
\code{\link{POST_FATE.graphic_mapPFGlight}}, 
\code{\link{POST_FATE.graphic_mapPFGsoil}}, 
\code{\link{POST_FATE.graphic_evolutionCoverage}}, 
\code{\link{POST_FATE.graphic_evolutionAbund_pixels}}, 
\code{\link{POST_FATE.graphic_evolutionLight_pixels}}, 
\code{\link{POST_FATE.graphic_evolutionSoil_pixels}}
}
\author{
Maya Guéguen
}
\keyword{FATE,}
\keyword{abundance}
\keyword{outputs,}
\keyword{through}
\keyword{time}
