% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POST_FATE.validation_habitat.R
\name{POST_FATE.validation.habitat}
\alias{POST_FATE.validation.habitat}
\alias{POST_FATE.validation_habitat}
\title{Compute habitat performance and create a prediction plot of habitat
for a whole map of a \code{FATE} simulation.}
\usage{
POST_FATE.validation_habitat(
  name.simulation,
  sim.version,
  obs.path,
  releves.PFG,
  releves.sites,
  hab.obs,
  validation.mask
)
}
\arguments{
\item{name.simulation}{simulation folder name.}

\item{sim.version}{name of the simulation we want to validate (it works with
only one sim.version).}

\item{obs.path}{the function needs observed data, please create a folder for them in your 
simulation folder and then indicate in this parameter the access path to this new folder.}

\item{releves.PFG}{name of file which contain the observed Braund-Blanquet abundance at each site
and each PFG and strata (with extension).}

\item{hab.obs}{name of the file which contain the extended studied map in the simulation (with extension).}

\item{validation.mask}{name of the file which contain a raster mask that specified which pixels need validation (with extension).}

\item{releves.sites}{name of the file which contain coordinates and a description of
the habitat associated with the dominant species of each site in the studied map (with extension).}
}
\value{
Two folders are created in name.simulation folder :
\describe{
  \item{\file{VALIDATION/HABITAT/sim.version}}{containing the prepared CBNA data,
  RF model, the performance analyzes (confusion matrix and TSS) for the training and 
testing parts of the RF model, the habitat performance file, the habitat prediction file with 
observed and simulated habitat for each pixel of the whole map and the final prediction plot.}
  \item{\file{DATA_OBS}}{maps of observed habitat and csv files of PFG and sites releves.}
}
}
\description{
This script compare habitat simulations and observations and
create a map to visualize this comparison with all the the \code{FATE} and
observed data.
}
\details{
The observed habitat is derived from the cesbio map, the simulated habitat 
is derived from FATE simulated relative abundance, based on a random forest 
algorithm trained on CBNA data. To compare observations and simulations, the function
compute confusion matrix between observation and prediction and then compute the TSS 
for each habitat h (number of prediction of habitat h/number of observation 
of habitat h + number of non-prediction of habitat h/number of non-observation 
of habitat h). The final metrics this script use is the mean of TSS per habitat over all 
habitats, weighted by the share of each habitat in the observed habitat distribution.
}
\author{
Matthieu Combaud, Maxime Delprat
}
