% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POST_FATE.validation.R
\name{POST_FATE.validation}
\alias{POST_FATE.validation}
\title{Compute validation data for habitat, PFG richness and composition for a \code{FATE} simulation.}
\usage{
POST_FATE.validation(
  name.simulation,
  file.simulParam,
  year,
  mat.obs,
  mat.hab,
  ras.habitat,
  doRichness = TRUE,
  doComposition = TRUE,
  doHabitat = TRUE,
  RF.seed = 123,
  RF.training = 0.7,
  doHabitat.allMap = FALSE,
  opt.ras_validation = NULL,
  opt.keep_PFG = NULL,
  opt.keep_strata = NULL
)
}
\arguments{
\item{name.simulation}{a \code{string} corresponding to the main directory 
or simulation name of the \code{FATE} simulation}

\item{file.simulParam}{default \code{NULL}. \cr A \code{string} 
corresponding to the name of a parameter file that will be contained into 
the \code{PARAM_SIMUL} folder of the \code{FATE} simulation}

\item{year}{an \code{integer} corresponding to the simulation year that will 
be used to extract PFG abundance table}

\item{mat.obs}{a \code{data.frame} with at least 5 columns : \cr 
\code{site}, \code{x}, \code{y}, \code{PFG}, \code{abund} 
\cr (\emph{and optionally, \code{strata}, \code{code.habitat}}) 
(see \href{POST_FATE.validation#details}{\code{Details}})}

\item{mat.hab}{a \code{data.frame} with 2 columns : \code{code.habitat}, 
\code{habitat}}

\item{ras.habitat}{a \code{string} corresponding to the file name of a 
raster mask, with an \code{integer} value within each pixel, corresponding 
to a specific habitat}

\item{doRichness}{(\code{logical}) default \code{TRUE}. \cr If \code{TRUE}, 
difference in observed and simulated PFG richness will be run.}

\item{doComposition}{(\code{logical}) default \code{TRUE}. \cr If \code{TRUE}, 
difference in observed and simulated PFG distribution over habitat will be run.}

\item{doHabitat}{(\code{logical}) default \code{TRUE}. \cr If \code{TRUE}, 
a random forest model to predict habitat from PFG abundances will be run.}

\item{RF.seed}{default \code{123}. \cr An \code{integer} to be given to 
\code{\link[base]{set.seed}} function, in order to fix the produced results 
if needed, as dataset will be divided randomly into training and testing 
datasets for the random forest model}

\item{RF.training}{default \code{0.7}. \cr A \code{numeric} between \code{0} 
and \code{1} corresponding to the percentage of data that will be used as 
training dataset in the random forest model}

\item{doHabitat.allMap}{(\code{logical}) default \code{TRUE}. \cr If 
\code{TRUE}, habitat prediction over the whole simulation map will be run.}

\item{opt.ras_validation}{(\emph{optional}) default \code{NULL}. \cr 
A \code{string} corresponding to the file name of a raster mask, with either 
\code{0} or \code{1} within each pixel, \code{1} corresponding to the cells 
of the studied area in which the validation will take place}

\item{opt.keep_PFG}{(\emph{optional}) default \code{NULL}. \cr 
A \code{vector} of \code{character} corresponding to the names of the PFG to 
keep for the validation}

\item{opt.keep_strata}{(\emph{optional}) default \code{NULL}. \cr 
A \code{list} with names corresponding to the strata to keep for the 
validation, and each element containing correspondence with \code{FATE} 
strata definition}
}
\value{
\describe{
  \item{Habitat}{into the \code{name.simulation/VALIDATION/HABITAT/ directory :} \cr
  1 .csv file containing the prepared observed data (relative abundances & habitat information). \cr
  1 .rds file containing the RF model. \cr
  5 .csv files containing the performance analysis (confusion matrix and TSS) for the training & 
  testing parts of the RF model and the final habitat performance. \cr
  If option selected, 1 .csv file containing the habitat prediction within each pixel, 1 .csv file with 
  sucess or failure of the prediction and 1 .png prediction map.}
}
\describe{
  \item{PFG Composition}{into the \code{name.simulation/VALIDATION/PFG COMPOSITION/ directory :} \cr
  1 .csv file containing the proximity between observed and simulated data computed for each PFG/strata/habitat. \cr 
  1 .csv file containing the observed releves transformed into relative metrics. \cr 
  1 .csv file containing the final output with the distribution per PFG, strata and habitat.}
}
\describe{
  \item{PFG richness}{into the \code{name.simulation/VALIDATION/PFG RICHNESS/ directory :} \cr
  1 .csv file of PFG richness in a \code{FATE} simulation. \cr 
  1 .csv fie of the PFG extinction frequency in a \code{FATE} simulation. \cr 
  1 .rds file which is the abundance per PFG file.}
}
}
\description{
This script is designed to compare observed and simulated 
PFG distribution for one specific \code{FATE} simulation year by computing 
a) difference in PFG richness ; b) similarity between distribution quantiles ; 
c) a random forest model to predict habitat classes.
}
\details{
This function allows to obtain, for a specific \code{FATE} simulation and 
a specific parameter file within this simulation, \strong{PFG validation 
habitat model and predictions} and one preanalytical graphic. \cr \cr





\describe{
  \item{Habitat validation}{ \cr
  \describe{
    \item{Observed habitat}{is provided by the \strong{ras.habitat} map. The final set of habitats taken into account
    in the validation is provided by \strong{mat.hab} table which contains 
    all the \strong{habitats}, with their corresponding \strong{ID}.}
    \item{Simulated habitat}{is determined from \code{FATE} simulated relative abundances 
    (see \code{\link{POST_FATE.temporalEvolution}}), thanks to a \code{Random Forest} algorithm. 
    The model is trained on \strong{mat.obs} data. Information about \strong{PFG abundances} at each
    \strong{sites} with \strong{xy} coordinates are necessary. PFG abundances can be given in absolute abundance
    (only in sites where at least one PFG is present) or in presence-absence data.\cr
    Eventually, \strong{habitat ID} information can be provided, as well as \strong{strata} names.
    If not, habitat information will be taken from \strong{ras.habitat} map, and PFG abundance will be considered for all strata.}
  }
  To compare observations and simulations, the function computes confusion matrix between 
  observations and predictions and then computes the TSS for each habitat h.
  \deqn{TSS_{\text{ h}} = (\frac{\text{True Positives}_{\text{ h}}}{\text{True Positives}_{\text{ h}}{\text{ + False Negatives}_{\text{ h}}}} + \frac{\text{True Negatives}_{\text{ h}}}{\text{True Negatives}_{\text{ h}}{\text{ + False Positives}_{\text{ h}}}}) - \text{1}}
  The final metric used is the mean of TSS per habitat over all habitats, weighted by the share of each habitat in the observed 
  habitat distribution. The habitat validation also provides a visual comparison of observed and 
  simulated habitat on the whole studied area, if option selected.}
  
  \item{PFG composition validation}{ \cr 
  \describe{
    \item{Observed composition}{is computed for a set of PFG, habitat & strata
  (for all strata if strata definition is not activated) by computing 4 quartiles of the composition, based
  on \strong{mat.obs} data provided. Habitats are chosen in \code{mat.hab},
  PFG list is defined in the file \code{file.simulParam} and, if defined, \code{mat.obs} strata
  are taken into account. Optionally, PFGs contained in \code{opt.keep_PFG} can be excluded from the analysis.}
    \item{Simulated composition}{is computed in the same way than the observed composition, with 
    \code{FATE} abundances \cr (see \code{\link{POST_FATE.temporalEvolution}}).}
  }
  Then, a composition similarity between each habitat/strata combinations is provided by computing a 
  pseudo-distance between observed and simulated quartiles for each PFG y.
  \deqn{S_{\text{ habitat, strata}} = \sum S_{\text{ y}{\text{, }habitat}{\text{, }strata}}}
  with
  \deqn{S_{\text{ y}{\text{, }habitat}{\text{, }strata}} = 1 - \frac{\text{1}}{4} * \sum abs(Q_{\text{ i}{\text{, }sim}} - Q_{\text{ i}{\text{, }obs}})}
  with i varying from 1 to 4.
  }
  \item{PFG richness validation}{the observed PFG richness corresponds to the PFG list contained in the file \code{file.simulParam}, 
  the simulated PFG richness is the number of PFG for which abundance over the simulation area 
  is strictly superior to zero for the simulation year under scrutiny. \cr
  Then, observed and simulated richness are compared for all the PFGs in the simulation in order to quantify the PFG mortality.
  Optionally, PFGs contained in \code{opt.keep_PFG} can be excluded from the analysis.}
}
}
\examples{

library(raster)

## Create a simulation folder
PRE_FATE.skeletonDirectory(name.simulation = "FATE_Champsaur")

## Load example data
Champsaur_params = .loadData("Champsaur_params", "RData")
.loadData("Champsaur_results_V1", "7z")

## Please extract results files in the 'FATE_Champsaur/RESULTS' folder

## Define a vector to choose habitats taken into account
mat.hab = data.frame(ID = c(6, 5, 7, 8)
                             , habitat = c("coniferous.forest"
                                           , "deciduous.forest"
                                           , "natural.grassland"
                                           , "woody.heatland"))

## Habitat & validation maps
ras.habitaterved = Champsaur_params$stk.mask$habitat
ras_simulation = Champsaur_params$stk.mask$Champsaur
ras.habitat = projectRaster(from = ras.habitaterved, to = ras_simulation, method = "ngb")
writeRaster(ras_simulation, filename = "FATE_Champsaur/DATA/MASK/MASK_Champsaur.tif")

## Observed data
mat.obs = Champsaur_params$tab.releves

## Transform observed PFG abundances into relative abundances
mat.obs$abund = PRE_FATE.abundBraunBlanquet(mat.obs$abund)/100
mat.obs = aggregate(abund ~ site + PFG + strata + x + y
                        , data = mat.obs, FUN = "sum")
                        
## Create Global and Simulation parameters
PRE_FATE.params_globalParameters(name.simulation = "FATE_Champsaur"
                                , opt.saving_abund_PFG_stratum = TRUE
                                , opt.saving_abund_PFG = TRUE
                                , opt.saving_abund_stratum = FALSE
                                , required.no_PFG = 15
                                , required.no_strata = 7
                                , required.simul_duration = 2000
                                , required.seeding_duration = 1000
                                , required.seeding_timestep = 1
                                , required.seeding_input = 100
                                , required.potential_fecundity = 1
                                , required.max_abund_low = 1000
                                , required.max_abund_medium = 2000
                                , required.max_abund_high = 3000
                                , doDispersal = TRUE
                                , DISPERSAL.mode = 1
                                , DISPERSAL.saving = FALSE
                                , doHabSuitability = TRUE
                                , HABSUIT.mode = 1)
                                
PRE_FATE.params_simulParameters(name.simulation = "FATE_Champsaur"
                                , name.MASK = "MASK_Champsaur.tif")

param = "Simul_parameters_V1.txt"
simul.param = paste0("FATE_Champsaur/PARAM_SIMUL/", param)

POST_FATE.validation(name.simulation = "FATE_Champsaur"
                     , file.simulParam = simul.param
                     , year = 2000
                     , doHabitat = TRUE
                     , mat.obs = mat.obs
                     , ras.habitat = ras.habitat
                     , mat.hab = mat.hab
                     , opt.keep_strata = opt.keep_strata
                     , doHabitat.allMap = TRUE
                     , doComposition = TRUE
                     , doRichness = TRUE)


}
\author{
Matthieu Combaud, Maxime Delprat, Maya Guéguen
}
