% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POST_FATE.validation.R
\name{POST_FATE.validation}
\alias{POST_FATE.validation}
\title{Computes validation data for habitat, PFG richness and composition for a \code{FATE} simulation.}
\usage{
POST_FATE.validation(
  name.simulation,
  sim.version,
  year,
  perStrata = FALSE,
  opt.no_CPU = 1,
  doHabitat = TRUE,
  releves.PFG,
  hab.obs,
  studied.habitat,
  predict.all.map = FALSE,
  seed,
  validation.mask = NULL,
  list.strata.simulations = NULL,
  doComposition = TRUE,
  PFG.considered_PFG.compo,
  habitat.considered_PFG.compo,
  strata.considered_PFG.compo = "A",
  doRichness = TRUE,
  list.PFG,
  exclude.PFG = NULL
)
}
\arguments{
\item{name.simulation}{Simulation folder name.}

\item{sim.version}{A character vector with the name(s) of the simulation(s) to validate.}

\item{year}{Year of simulation for validation.}

\item{perStrata}{\code{Logical}. \cr Default \code{FALSE}. If \code{TRUE}, PFG abundance is defined by strata. 
If \code{FALSE}, PFG abundance defined for all strata (habitat & PFG composition & PFG richness validation).}

\item{opt.no_CPU}{Default \code{1}. \cr The number of resources that can be used to 
parallelize the computation of prediction performance for habitat & richness validation.}

\item{doHabitat}{\code{Logical}. Default \code{TRUE}. \cr If \code{TRUE}, habitat validation module is activated,
if \code{FALSE}, habitat validation module is disabled.}

\item{releves.PFG}{A \code{data.frame} with at least 5 columns : \cr
\code{site}, \code{x}, \code{y}, which contain respectively ID, x coordinate & y coordinate of each site of the study area. \cr
\code{abund} & \code{PFG} which contain respectively abundance (can be absolute abundance, Braun-Blanquet abundance or presence-absence) 
& name of PFG.
\cr (\emph{and optionally, \code{strata}}) which contains the number of strata at each the abundance is noted. 
(habitat & PFG composition validation).}

\item{hab.obs}{A raster map of the extended studied map in the simulation, with same projection 
& resolution than simulation mask (habitat & PFG composition validation).}

\item{studied.habitat}{A \code{data.frame} with 2 columns : 
\cr \code{ID} which contains the habitat ID, & \code{habitat} which contains the habitat names which will be taken into account 
for the validation (habitat & PFG composition validation).}

\item{predict.all.map}{Default \code{FALSE}. \cr If \code{TRUE}, the function will compute habitat prediction
over the whole map and will provide a prediction map.}

\item{seed}{\code{Numerical}. Number of seeds to set in order to generate a Random Forest model.}

\item{validation.mask}{(\code{Optional}). Default \code{NULL}. \cr A raster mask (with 0 or 1 in each pixel) that specified on 
which pixels the performance of the prediction will be compute, with same projection & resolution than simulation mask 
(habitat & PFG composition validation). \cr
If \code{NULL}, the function will take the simulation mask (which means that the performance will be compute over the whole map)}

\item{list.strata.simulations}{(\code{Optional}). Default \code{NULL}. \cr If \code{perStrata} = \code{TRUE}, 
a character vector which contain \code{FATE} strata definition and correspondence with observed strata definition. 
If \code{perStrata} = \code{FALSE}, please specify \code{NULL} value.}

\item{doComposition}{\code{Logical}. Default \code{TRUE}. \cr If \code{TRUE}, PFG composition validation module is activated,
if \code{FALSE}, PFG composition validation module is disabled.}

\item{PFG.considered_PFG.compo}{A character vector of the list of PFG considered
in the validation (PFG composition validation).}

\item{habitat.considered_PFG.compo}{A character vector of the list of habitat(s)
considered in the validation (PFG composition validation).}

\item{strata.considered_PFG.compo}{Default \code{"A"}. \cr If \code{perStrata} = \code{FALSE}, a character vector with value "A" 
(selection of one or several specific strata disabled). If \code{perStrata} = \code{TRUE}, a character 
vector with at least one of the observed strata (PFG composition validation).}

\item{doRichness}{\code{Logical}. Default \code{TRUE}. \cr If \code{TRUE}, PFG richness validation module is activated,
if \code{FALSE}, PFG richness validation module is disabled.}

\item{list.PFG}{A character vector which contain all the PFGs taken account in
the simulation and observed in the simulation area (PFG richness validation).}

\item{exclude.PFG}{(\code{Optional}). Default \code{NULL}. \cr A character vector containing the names 
of the PFG you want to exclude from the analysis (PFG richness validation).}
}
\value{

}
\description{
This script is designed to compute validation data for : 
\cr \code{Habitat} : Compares habitat simulations and observations and
create a map to visualize this comparison with all the \code{FATE} and
observed data (if option selected).
\cr \code{PFG Composition} : Produced a computation of observed distribution 
of relative abundance in the simulation area and a computation of distance between
observed and simulated distribution.
\cr \code{PFG Richness} : Computes the PFG richness over the whole simulation area 
for a \code{FATE} simulation and computes the difference between observed and simulated PFG richness.
}
\details{
\describe{
  \item{Habitat validation}{The observed habitat is derived from a map of the area or, if defined, 
from \code{studied.habitat}, the simulated habitat is derived from \code{FATE} simulated relative 
abundance, based on a random forest algorithm trained on observed releves data (see \code{\link{train_RF_habitat}}). \cr
To compare observations and simulations, the function computes confusion matrix between 
observations and predictions and then compute the TSS for each habitat h 
(number of prediction of habitat h/number of observation of habitat h + number of non-prediction 
of habitat h/number of non-observation of habitat h). \cr The final metrics this script use is the 
mean of TSS per habitat over all habitats, weighted by the share of each habitat in the observed 
habitat distribution. The habitat validation also provides a visual comparison of observed and 
simulated habitat on the whole studied area, if option selected (see \code{\link{do_habitat_validation}} &
\code{\link{plot_predicted_habitat}}).} \cr
  \item{PFG composition validation}{This code firstly run the \code{get_observed_distribution} 
function in order to have a \code{obs.distri} file which contain the observed distribution 
per PFG, strata and habitat. This file is also an argument for the \code{do_PFG_composition_validation} 
function run next. This second sub function provides the computation of distance between observed 
and simulated distribution.}
  \item{PFG richness validation}{Firstly, the function updates the \code{list.PFG} with \code{exclude.PFG} vector.
Then, the script takes the abundance per PFG (and per strata if option selected) file from the 
results of the \code{FATE} simulation and computes the difference between the \code{list.PFG} 
and all the PFG which are presents in the abundance file, in order to obtain the PFG richness 
for a simulation. The function also determine if an observed PFG is missing in the results of the 
simulation at a specific year.}
}
}
\examples{

## Load example data
Champsaur_params = .loadData('Champsaur_params', 'RData')

## Please run a FATE simulation and the function POST_FATE.temporalEvolution before

## Define a vector to choose habitats taken into account
studied.habitat = data.frame(ID = c(6, 5, 7, 8), habitat = c("coniferous.forest", "deciduous.forest", "natural.grassland", "woody.heatland"))
## Habitat & validation maps
hab.observed = Champsaur_params$stk.mask$habitat
simulation.map = Champsaur_params$stk.mask$Champsaur
hab.obs = projectRaster(from = hab.observed, res = res(simulation.map)[1], crs = crs(projection(simulation.map)), method = "ngb")
## Observed data
releves.PFG = Champsaur_params$tab.releves
## List of PFG taken into account in a FATE simulation
list.PFG = as.factor(c("C1", "C2", "C3", "C4", "H1", "H2", "H3", "H4", "H5", "H6", "P1", "P2", "P3", "P4", "P5"))
## Habitat, strata and PFG considered in PFG compo validation
habitat.considered = c("coniferous.forest", "deciduous.forest", "natural.grassland", "woody.heatland")
PFG.considered_PFG.compo = as.factor(c("H1", "H2", "H3", "H4", "H5", "H6", "P1", "P2", "P3", "P4", "P5"))

POST_FATE.validation(name.simulation = "FATE_Champsaur"
                     , sim.version = "SIMUL_V1"
                     , year = 2000
                     , doHabitat = TRUE
                     , releves.PFG = releves.PFG
                     , hab.obs = hab.obs
                     , studied.habitat = studied.habitat
                     , predict.all.map = TRUE
                     , seed = 123
                     , doComposition = TRUE
                     , PFG.considered_PFG.compo = PFG.considered_PFG.compo
                     , habitat.considered_PFG.compo = habitat.considered
                     , doRichness = TRUE
                     , list.PFG = list.PFG)

}
\author{
Matthieu Combaud, Maxime Delprat
}
