% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POST_FATE.validation.R
\name{POST_FATE.validation}
\alias{POST_FATE.validation}
\title{Computes validation data for habitat, PFG richness and composition for a \code{FATE} simulation.}
\usage{
POST_FATE.validation(
  name.simulation,
  sim.version,
  year,
  doHabitat = TRUE,
  obs.path,
  releves.PFG,
  releves.sites,
  hab.obs,
  validation.mask,
  studied.habitat = NULL,
  doComposition = TRUE,
  PFG.considered_PFG.compo,
  habitat.considered_PFG.compo,
  doRichness = TRUE,
  list.PFG,
  exclude.PFG = NULL
)
}
\arguments{
\item{name.simulation}{simulation folder name.}

\item{sim.version}{name of the simulation to validate (it works with only one \code{sim.version}).}

\item{year}{year of simulation for validation.}

\item{doHabitat}{logical. Default \code{TRUE}. If \code{TRUE}, habitat validation module is activated,
if \code{FALSE}, habitat validation module is disabled.}

\item{obs.path}{the function needs observed data, please create a folder for them in your 
simulation folder and then indicate in this parameter the access path to this new folder (habitat & PFG composition validation).}

\item{releves.PFG}{name of file which contain the observed Braund-Blanquet abundance at each site
and each PFG and strata (habitat & PFG composition validation).}

\item{hab.obs}{name of the file which contain the extended studied map in the simulation (habitat & PFG composition validation).}

\item{validation.mask}{name of the file which contain a raster mask that specified which pixels need validation 
(habitat & PFG composition validation).}

\item{studied.habitat}{default \code{NULL}. If \code{NULL}, the function will
take into account of all habitats in the \code{hab.obs} map. Otherwise, please specify 
in a vector habitats that will be take into account for the validation (habitat validation).}

\item{doComposition}{logical. Default \code{TRUE}. If \code{TRUE}, PFG composition validation module is activated,
if \code{FALSE}, PFG composition validation module is disabled.}

\item{PFG.considered_PFG.compo}{a character vector of the list of PFG considered
in the validation (PFG composition validation).}

\item{habitat.considered_PFG.compo}{a character vector of the list of habitat(s)
considered in the validation (PFG composition validation).}

\item{doRichness}{logical. Default \code{TRUE}. If \code{TRUE}, PFG richness validation module is activated,
if \code{FALSE}, PFG richness validation module is disabled.}

\item{list.PFG}{a character vector which contain all the PFGs taken account in
the simulation and observed in the simulation area (PFG richness validation).}

\item{exclude.PFG}{default \code{NULL}. A character vector containing the names 
of the PFG you want to exclude from the analysis (PFG richness validation).}

\item{releves.site}{name of the file which contain coordinates and a description of
the habitat associated with the dominant species of each site in the studied map (habitat & PFG composition validation).}
}
\value{

}
\description{
This script is designed to compute validation data for : 
\code{Habitat} : compares habitat simulations and observations and
create a map to visualize this comparison with all the \code{FATE} and
observed data.
\code{PFG Composition} : produced a computation of observed distribution 
of relative abundance in the simulation area and a computation of distance between
observed and simulated distribution.
\code{PFG Richness} : computes the PFG richness over the whole simulation area 
for a \code{FATE} simulation and computes the difference between observed and simulated PFG richness.
}
\details{
\describe{
  \item{Habitat validation}{The observed habitat is derived from the cesbio map, the simulated habitat 
is derived from FATE simulated relative abundance, based on a random forest 
algorithm trained on CBNA data. To compare observations and simulations, the function
compute confusion matrix between observation and prediction and then compute the TSS 
for each habitat h (number of prediction of habitat h/number of observation 
of habitat h + number of non-prediction of habitat h/number of non-observation 
of habitat h). The final metrics this script use is the mean of TSS per habitat over all 
habitats, weighted by the share of each habitat in the observed habitat distribution.}
  \item{PFG composition validation}{This code firstly run the \code{get.observed.distribution} function in order to have
a \code{obs.distri} file which contain the observed distribution per PFG, strata and habitat.
This file is also an argument for the \code{do.PFG.composition.validation} function run next.
This second sub function provide the computation of distance between observed and simulated distribution. \cr 
NB : The argument \code{strata.considered_PFG.compo} is by default "A" in the 2 sub functions because 
it's easier for a \code{FATE} simulation to provide PFG abundances for all strata. \cr The argument 
\code{perStrata.compo} is by default \code{NULL} for the same reasons.}
  \item{PFG richness validation}{Firstly, the function updates the \code{list.PFG} with \code{exclude.PFG} vector.
Then, the script takes the abundance per PFG file from the results of the \code{FATE} 
simulation and computes the difference between the \code{list.PFG} and all the PFG 
which are presents in the abundance file, in order to obtain the PFG richness for a simulation.
The function also determine if an observed PFG is missing in the results of the simulation at 
a specific year.}
}
}
\examples{

## Habitat validation ---------------------------------------------------------------------------------
POST_FATE.validation(name.simulation = "FATE_Champsaur"
                      , sim.version = "SIMUL_V4.1"
                      , year = 2000
                      , doHabitat = TRUE
                      , obs.path = "FATE_Champsaur/DATA_OBS/"
                      , releves.PFG = "releves.PFG.abundance.csv"
                      , releves.sites = "releves.sites.shp"
                      , hab.obs = "simplified.cesbio.map.grd"
                      , validation.mask = "certain.habitat.100m.restricted.grd"
                      , studied.habitat = NULL
                      , doComposition = FALSE
                      , doRichness = FALSE)
                      
## PFG composition validation --------------------------------------------------------------------------
list.PFG<-as.factor(c("C1","C2","C3","C4","H1","H2","H3","H4","H5","H6","P1","P2","P3","P4","P5"))
habitat.considered = c("coniferous.forest", "deciduous.forest", "natural.grassland", "woody.heatland")
POST_FATE.validation(name.simulation = "FATE_Champsaur"
                      , sim.version = "SIMUL_V4.1"
                      , year = 2000
                      , doHabitat = FALSE
                      , obs.path = "FATE_Champsaur/DATA_OBS/"
                      , releves.PFG = "releves.PFG.abundance.csv"
                      , releves.sites = "releves.sites.shp"
                      , hab.obs = "simplified.cesbio.map.grd"
                      , validation.mask = "certain.habitat.100m.restricted.grd"
                      , studied.habitat = NULL
                      , doComposition = TRUE
                      , PFG.considered_PFG.compo = list.PFG
                      , habitat.considered_PFG.compo = habitat.considered
                      , doRichness = FALSE)
                      
## PFG richness validation -----------------------------------------------------------------------------
list.PFG<-as.factor(c("C1","C2","C3","C4","H1","H2","H3","H4","H5","H6","P1","P2","P3","P4","P5"))
POST_FATE.validation(name.simulation = "FATE_CHampsaur"
                      , sim.version = "SIMUL_V4.1"
                      , year = 2000
                      , doHabitat = FALSE
                      , doComposition = FALSE
                      , doRichness = TRUE
                      , list.PFG = list.PFG
                      , exclude.PFG = NULL)

}
\author{
Matthieu Combaud, Maxime Delprat
}
