% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POST_FATE.validation.R
\name{POST_FATE.validation}
\alias{POST_FATE.validation}
\title{Compute validation data for habitat, PFG richness and composition for a \code{FATE} simulation.}
\usage{
POST_FATE.validation(
  name.simulation,
  file.simulParam,
  year,
  perStrata = FALSE,
  opt.no_CPU = 1,
  doHabitat = TRUE,
  releves.PFG,
  hab.obs,
  studied.habitat,
  predict.all.map = FALSE,
  RF.seed = 123,
  RF.training = 0.7,
  validation.mask = NULL,
  list.strata.simulations = NULL,
  doComposition = TRUE,
  doRichness = TRUE,
  exclude.PFG = NULL
)
}
\arguments{
\item{name.simulation}{a \code{string} corresponding to the simulation folder name.}

\item{file.simulParam}{default \code{NULL}. \cr A \code{string} 
corresponding to the name of a parameter file that will be contained into 
the \code{PARAM_SIMUL} folder of the \code{FATE} simulation.}

\item{year}{an \code{integer} corresponding to the year of simulation for validation.}

\item{perStrata}{(\code{logical}) default \code{FALSE}. \cr 
If \code{TRUE}, habitat & PFG composition are computed with PFG abundance defined by strata. \cr
If \code{FALSE}, habitat & PFG composition are computed with PFG abundance defined for all strata.}

\item{opt.no_CPU}{default \code{1}. \cr An \code{integer} corresponding to the number of resources that can be used to 
parallelize the computation of prediction performance for habitat & richness validation.}

\item{doHabitat}{(\code{logical}) default \code{TRUE}. \cr If \code{TRUE}, habitat validation module is activated,
if \code{FALSE}, habitat validation module is disabled.}

\item{releves.PFG}{a \code{data.frame} with at least 5 columns : \cr
\code{site}, \code{x}, \code{y}, \code{abund}, \code{PFG}
\cr (\emph{and optionally, \code{strata}, \code{code.habitat}})
\cr (see \href{POST_FATE.validation#details}{\code{Details}}).}

\item{hab.obs}{a \code{raster} map of the habitats within the studied area, with same projection 
& resolution than simulation mask.}

\item{studied.habitat}{a \code{data.frame} with 2 columns : \cr
\code{ID} ,\code{habitat}
\cr (see \href{POST_FATE.validation#details}{\code{Details}}).}

\item{predict.all.map}{(\code{logical}) default \code{FALSE}. \cr If \code{TRUE}, the function will compute habitat prediction 
& performance over the whole map and will provide a prediction map.}

\item{RF.seed}{default \code{123}. \cr An \code{integer} corresponding to the number of seeds to set 
in order to generate a \code{Random Forest} model.}

\item{RF.training}{default \code{0.7}. \cr A \code{double} between 0 & 1 corresponding to the part of the data used for 
training a \code{Random Forest} model on \code{releves.PFG} data.}

\item{validation.mask}{(\code{optional}) default \code{NULL}. \cr A \code{raster} map (with 0 or 1 in each pixel) that specified on 
which pixels the performance of the habitat prediction will be compute, with same projection & resolution than simulation mask 
\cr If \code{NULL}, the function will take the simulation mask (which means that the performance will be compute over the whole map).}

\item{list.strata.simulations}{(\code{optional}) default \code{NULL}. \cr If \code{perStrata} = \code{TRUE}, 
a \code{list} which contains, for each stratum defined in \code{releves.PFG}, the correspondence with \code{FATE} strata definition. \cr
If \code{perStrata} = \code{FALSE}, please specify \code{NULL} value.}

\item{doComposition}{(\code{logical}) default \code{TRUE}. \cr If \code{TRUE}, PFG composition validation module is activated.}

\item{doRichness}{(\code{logical}) default \code{TRUE}. \cr If \code{TRUE}, PFG richness validation module is activated.}

\item{exclude.PFG}{(\code{optional}) default \code{NULL}. \cr A \code{character} vector containing the name(s) 
of the PFG(s) that will be excluded from the composition and richness analysis.}
}
\value{
\describe{
  \item{Habitat}{into the \code{name.simulation/VALIDATION/HABITAT/ directory :} \cr
  1 .csv file containing the prepared observed data (relative abundances & habitat information). \cr
  1 .rds file containing the RF model. \cr
  5 .csv files containing the performance analysis (confusion matrix and TSS) for the training & 
  testing parts of the RF model and the final habitat performance. \cr
  If option selected, 1 .csv file containing the habitat prediction within each pixel, 1 .csv file with 
  sucess or failure of the prediction and 1 .png prediction map.}
}
\describe{
  \item{PFG Composition}{into the \code{name.simulation/VALIDATION/PFG COMPOSITION/ directory :} \cr
  1 .csv file containing the proximity between observed and simulated data computed for each PFG/strata/habitat. \cr 
  1 .csv file containing the observed releves transformed into relative metrics. \cr 
  1 .csv file containing the final output with the distribution per PFG, strata and habitat.}
}
\describe{
  \item{PFG richness}{into the \code{name.simulation/VALIDATION/PFG RICHNESS/ directory :} \cr
  1 .csv file of PFG richness in a \code{FATE} simulation. \cr 
  1 .csv fie of the PFG extinction frequency in a \code{FATE} simulation. \cr 
  1 .rds file which is the abundance per PFG file.}
}
}
\description{
This script is designed to compute validation data for : \cr
\describe{
  \item{Habitat}{Compares simulated and observed habitat and
create a map to visualize this comparison with all the \code{FATE} and
observed data \cr (if option selected).}
  \item{PFG Composition}{Computes the similarity between observed and simulated 
  PFG composition for a chosen set of PFG/habitat/strata combinations.}
  \item{PFG Richness}{Computes the PFG richness over the whole simulation area 
for a \code{FATE} simulation and computes the difference between observed and simulated PFG richness.}
}
}
\details{
\describe{
  \item{Habitat validation}{ \cr
  \describe{
    \item{Observed habitat}{is provided by the \strong{hab.obs} map. The final set of habitats taken into account
    in the validation is provided by \strong{studied.habitat} table which contains 
    all the \strong{habitats}, with their corresponding \strong{ID}.}
    \item{Simulated habitat}{is determined from \code{FATE} simulated relative abundances 
    (see \code{\link{POST_FATE.temporalEvolution}}), thanks to a \code{Random Forest} algorithm. 
    The model is trained on \strong{releves.PFG} data. Information about \strong{PFG abundances} at each
    \strong{sites} with \strong{xy} coordinates are necessary. \cr
    Eventually, \strong{habitat ID} information can be provided, as well as \strong{strata} name.
    If not, habitat information will be taken from \strong{hab.obs} map, and PFG abundance will be considered for all strata.}
  }
  To compare observations and simulations, the function computes confusion matrix between 
  observations and predictions and then computes the TSS for each habitat h.
  \deqn{TSS_{\text{ h}} = (\frac{\text{number of prediction}_{\text{ h}}}{\text{number of observation}_{\text{ h}}} + \frac{\text{number of non-prediction}_{\text{ h}}}{\text{number of non-observation}_{\text{ h}}}) - \text{1}}
  The final metric used is the mean of TSS per habitat over all habitats, weighted by the share of each habitat in the observed 
  habitat distribution. The habitat validation also provides a visual comparison of observed and 
  simulated habitat on the whole studied area, if option selected.}
  
  \item{PFG composition validation}{ \cr 
  \describe{
    \item{Observed composition}{is computed for a set of PFG, habitat & strata
  (for all strata if strata definition is not activated) by computing 4 quartiles of the composition, based
  on \strong{releves.PFG} data provided. Habitats are chosen in \code{studied.habitat},
  PFG list is defined in the file \code{file.simulParam} and, if defined, \code{releves.PFG} strata
  are taken into account. Optionally, PFGs contained in \code{exclude.PFG} can be excluded from the analysis.}
    \item{Simulated composition}{is computed in the same way than the observed composition, with 
    \code{FATE} abundances \cr (see \code{\link{POST_FATE.temporalEvolution}}).}
  }
  Then, a composition similarity between each habitat/strata combinations is provided by computing a 
  pseudo-distance between observed and simulated quartiles for each PFG y.
  \deqn{S_{\text{ habitat, strata}} = \sum S_{\text{ y}{\text{, }habitat}{\text{, }strata}}}
  with
  \deqn{S_{\text{ y}{\text{, }habitat}{\text{, }strata}} = 1 - \frac{\text{1}}{4} * \sum abs(Q_{\text{ i}{\text{, }sim}} - Q_{\text{ i}{\text{, }obs}})}
  with i varying from 1 to 4.
  }
  \item{PFG richness validation}{the observed PFG richness corresponds to the PFG list contained in the file \code{file.simulParam}, 
  the simulated PFG richness is the number of PFG for which abundance over the simulation area 
  is strictly superior to zero for the simulation year under scrutiny. \cr
  Then, observed and simulated richness are compared for all the PFGs in the simulation in order to quantify the PFG mortality.
  Optionally, PFGs contained in \code{exclude.PFG} can be excluded from the analysis.}
}
}
\examples{

library(raster)

## Load example data
Champsaur_params = .loadData("Champsaur_params", "RData")

## Create a skeleton folder
PRE_FATE.skeletonDirectory(name.simulation = "FATE_Champsaur")

## Load results from a simulation
.loadData("Champsaur_results_V1", "7z")

## Please extract files from the 7z folder in 'FATE_CHAMPSAUR/RESULTS'

## Define a vector to choose habitats taken into account
studied.habitat = data.frame(ID = c(6, 5, 7, 8), habitat = c("coniferous.forest", "deciduous.forest", "natural.grassland", "woody.heatland"))

## Habitat & validation maps
hab.observed = Champsaur_params$stk.mask$habitat
simulation.map = Champsaur_params$stk.mask$Champsaur
hab.obs = projectRaster(from = hab.observed, res = res(simulation.map)[1], crs = crs(projection(simulation.map)), method = "ngb")
writeRaster(simulation.map, filename = "FATE_Champsaur/DATA/MASK/MASK_Champsaur.tif")

## Observed data
releves.PFG = Champsaur_params$tab.releves

## List of PFG taken into account in a FATE simulation
list.PFG = c("C1", "C2", "C3", "C4", "H1", "H2", "H3", "H4", "H5", "H6", "P1", "P2", "P3", "P4", "P5")

POST_FATE.validation(name.simulation = "FATE_Champsaur"
                     , sim.version = "SIMUL_V1.1"
                     , year = 2000
                     , doHabitat = TRUE
                     , releves.PFG = releves.PFG
                     , hab.obs = hab.obs
                     , studied.habitat = studied.habitat
                     , predict.all.map = TRUE
                     , doComposition = TRUE
                     , doRichness = TRUE)


}
\author{
Matthieu Combaud, Maxime Delprat, Maya Guéguen
}
