% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRE_FATE.params_globalParameters.R
\name{PRE_FATE.params_globalParameters}
\alias{PRE_FATE.params_globalParameters}
\title{Create \emph{Global_parameters} parameter file for a \code{FATE-HD}
simulation}
\usage{
PRE_FATE.params_globalParameters(name.simulation, opt.no_CPU = 1,
  opt.replacePrevious = FALSE, required.no_PFG, required.no_STRATA,
  required.simul_duration, required.seeding_duration,
  required.seeding_timestep, doLight = FALSE, doDispersal = FALSE,
  doHabSuitability = FALSE, HABSUIT.ref_option, doDisturbances = FALSE,
  DIST.no, DIST.no_sub, DIST.freq, doSoil = FALSE, SOIL.no_categories,
  doFire = FALSE, doDrought = FALSE, doHabStability = FALSE,
  doAliens = FALSE)
}
\arguments{
\item{name.simulation}{a \code{string} that corresponds to the main directory
or simulation name of the \code{FATE-HD} simulation}

\item{opt.no_CPU}{default 1 (\emph{optional}). The number of resources that 
can be used to parallelize the \code{FATE-HD} simulation}

\item{opt.replacePrevious}{default \code{FALSE} (\emph{optional}). If \code{TRUE},
pre-existing files inside \code{name.simulation/DATA/GLOBAL_PARAMETERS} folder
will be replaced}

\item{required.no_PFG}{an \code{integer} corresponding to the number of PFG}

\item{required.no_STRATA}{an \code{integer} corresponding to the number of 
height strata}

\item{required.simul_duration}{an \code{integer} corresponding to the 
duration of simulation (\emph{in years})}

\item{required.seeding_duration}{an \code{integer} corresponding to the 
duration of seeding (\emph{in years})}

\item{required.seeding_timestep}{an \code{integer} corresponding to the 
time interval at which occurs the seeding, and until the seeding duration
 is not over (\emph{in years})}

\item{doLight}{default \code{FALSE}. If \code{TRUE}, light competition is 
activated in the \code{FATE-HD} simulation, and associated parameters are 
required}

\item{doDispersal}{default \code{FALSE}. If \code{TRUE}, seed dispersal is 
activated in the \code{FATE-HD} simulation, and associated parameters are 
required}

\item{doHabSuitability}{default \code{FALSE}. If \code{TRUE}, habitat
suitability is activated in the \code{FATE-HD} simulation, and associated 
parameters are required}

\item{HABSUIT.ref_option}{an \code{integer} corresponding to the way of 
simulating the habitat suitability variation between years for each PFG, 
either random (1) or PFG specific (2)}

\item{doDisturbances}{default \code{FALSE}. If \code{TRUE}, disturbances 
are applied in the \code{FATE-HD} simulation, and associated parameters 
are required}

\item{DIST.no}{the number of disturbances}

\item{DIST.no_sub}{the number of way a PFG could react to a disturbance}

\item{DIST.freq}{the frequency of each disturbance (\emph{in years})}

\item{doSoil}{default \code{FALSE}. If \code{TRUE}, soil competition is
activated in the \code{FATE-HD} simulation, and associated parameters 
are required}

\item{SOIL.no_categories}{the number of soil composition classes}

\item{doFire}{\emph{to be filled}}

\item{doDrought}{\emph{to be filled}}

\item{doHabStability}{\emph{to be filled}}

\item{doAliens}{\emph{to be filled}}
}
\value{
A \code{.txt} file into the \code{name.simulation/DATA/GLOBAL_PARAMETERS}
directory with the following parameters :

\itemize{
  \item NB_CPUS
  \item NB_FG
  \item NB_STRATUM
  \item SIMULATION_DURATION
  \item SEEDING_DURATION
  \item SEEDING_TIMESTEP \cr \cr
}

If the simulation includes \emph{light competition} :

\itemize{
  \item DO_LIGHT_COMPETITION
}

If the simulation includes \emph{dispersal} :

\itemize{
  \item DO_DISPERSAL
}

If the simulation includes \emph{habitat suitability} :

\itemize{
  \item DO_HAB_SUITABILITY
  \item HABSUIT_OPTION
}

If the simulation includes \emph{disturbances} :

\itemize{
  \item DO_DISTURBANCES
  \item NB_DISTURBANCES
  \item NB_SUBDISTURBANCES
  \item FREQ_DISTURBANCES
}

If the simulation includes \emph{soil competition} :

\itemize{
  \item DO_SOIL_COMPETITION
  \item NB_SOIL_CATEGORIES
}

If the simulation includes \emph{drought disturbances} :

\itemize{
  \item DO_DROUGHT_DISTURBANCES
  \item NB_SUBDROUGHT
  \item CHRONO_POST_DROUGHT
  \item CHRONO_CURR_DROUGHT
}

If the simulation includes \emph{habitat stability check} :

\itemize{
  \item DO_HAB_STABILITY
  \item NB_HABITATS
}

If the simulation includes \emph{aliens introduction} :

\itemize{
  \item DO_ALIENS_DISTURBANCE
  \item FREQ_ALIENS
}
}
\description{
This script is designed to create one parameter file
containing \code{GLOBAL PARAMETERS} used in \code{FATE-HD} model.
}
\details{
The core module of \code{FATE-HD} requires several parameters to define 
general characteristics of the simulation :

\describe{
  \item{Studied system}{
  \itemize{
    \item \emph{required.no_PFG} : the number of Plant Functional Groups 
    that will be included into the simulation.  
    This number should match with the number of files that will be given 
    with the characteristics of each group (SUCC, DISP, ...).
    \item \emph{required.no_STRATA} : the number of height strata that 
    will be used into the succession module.  
    This number should match with the maximum number of strata possible
    defined into the PFG SUCC files. \cr \cr
  }
  }
  \item{Simulation timing}{
  \itemize{
    \item \emph{required.simul_duration} : the duration of simulation 
    (\emph{in years})
    \item \emph{required.seeding_duration} : the duration of seeding 
    (\emph{in years})
    \item \emph{required.seeding_timestep} : the time interval at which 
    occurs the seeding, and until the seeding duration is not over 
    (\emph{in years}) \cr \cr
  }
  }
}


The other modules of \code{FATE-HD} can be activated within this file, and
if so, some additional parameters will be required :

\describe{
  \item{LIGHT}{= to influence plants fecundity and seed 
  recruitment according to PFG preferences for light conditions \cr
  = light resources are calculated as a proxy of PFG abundances
  within each height stratum \cr
  \cr \cr
  }
  \item{DISPERSAL}{= to allow plants to disperse seeds according
  to 3 user-defined distances \cr
  \cr \cr
  }
  \item{HABITAT SUITABILITY}{= to influence plants fecundity and seed 
  recruitment according to PFG preferences for habitat conditions \cr
  = filter based 
  on maps given for each PFG within the \emph{Simul_parameters} file 
  with the \code{PFG_HAB_MASK} flag. \cr \cr
  These maps must contain values 
  between 0 and 1 corresponding to the probability of presence of the 
  group in each pixel. Each year (timestep), this value will be compared 
  to a reference value, and if superior, the PFG will be able to grow and 
  survive.
  \itemize{
    \item \emph{HABSUIT.ref_option} : the habitat suitability reference
    value can be set in two possible ways :
    \enumerate{
      \item \emph{random} : for each pixel, the reference value is drawn from
      a uniform distribution, and the same value is used for each PFG 
      within this pixel.
      \item \emph{PFG specific} : for each PFG, a mean value and a standard 
      deviation value are drawn from a uniform distribution. For each 
      pixel and for each PFG, the reference value is drawn from a normal 
      distribution of parameters the mean and standard deviation of the PFG.
       \cr \cr
    }
  }
  }
  \item{DISTURBANCES}{= defined for events such as mowing, grazing, but also
  urbanization, crops, etc. \cr \cr
  The impact zone is defined with a mask (map with 0 or 1) and the user will
  have to define how each PFG will be impacted depending on age and life stage. 
  \itemize{
    \item \emph{DIST.no} : the number of different disturbances
    \item \emph{DIST.no_sub} : the number of way a PFG could react to a 
    perturbation
    \item \emph{DIST.freq} : the frequency of each disturbance (\emph{in years})
  }
  }
  \item{SOIL}{= to influence plants mortality according to PFG preferences 
  for soil conditions (fertility) \cr
  = soil composition is calculated as the weighted mean of each PFG's
  contribution \cr \cr
  \itemize{
    \item \emph{SOIL.no_categories} : the number of soil composition classes
    that will be used into the soil module.
    This number should match with the maximum value of contribution or
    tolerance possible defined into the PFG SOIL files. \cr \cr
  }
  }
  \item{DROUGHT}{= to experience extreme events with a direct and a delayed
  response on PFG. \cr \cr
  The PFG's past drought exposure is compared to current moisture values as 
  a proxy of drought index. Developed canopy closure helps to reduce these 
  values. Each PFG is impacted if submitted one year to drought, and also (and  
  more consequently) if submitted several years in a row.
  \itemize{
    \item \emph{DROUGHT.no_sub} : the number of way a PFG could react to a 
    perturbation
    \item \emph{DROUGHT.chrono_post} : if post-drought effects will occur 
    previously to succession (0) or after (1)
    \item \emph{DROUGHT.chrono_curr} : if current-drought effects will occur 
    previously to succession (0) or after (1)
  }
  }
  \item{HABITAT STABILITY}{= to evaluate through the evolution of the PFG 
  composition in different habitats. \cr \cr
  The habitat distribution is given with a categorical map with each value
  corresponding to a specific habitat. Every year, the abundance of each PFG
  within each habitat is recorded, as well as the evenness of the habitat :
  \deqn{
  \text{evenness} = - SUM(\text{PFG.proportion} * log(\text{PFG.proportion})) / log(\text{no.PFG} - 1)
  }
  \deqn{\text{with PFG.proportion} = \text{abund.PFG.in.Habitat} / \text{abund.allPFG.in.Habitat}}
  Every 5 years, a stability check is done to see if the abundance or the
  evenness have values out of the usual distribution.
  \itemize{
    \item \emph{HABSTAB.no_hab} : the number of different habitats
  }
  }
  \item{INVASIVE \cr INTRODUCTION}{= to add new PFG during the simulation. \cr \cr
  The introduction areas are defined with a mask (map with 0 or 1). If the 
  habitat suitability filter is on, suitability maps will also be needed for
  these new groups.
  \itemize{
    \item \emph{ALIEN.freq} : the frequency of each introduction (\emph{in years})
  }
  }
}
}
\examples{

## Create a skeleton folder with the default name ('FATE_simulation')
PRE_FATE.skeletonDirectory()

## Create a Global_parameters file
PRE_FATE.params_globalParameters(name.simulation = "FATE_simulation"
                                 , required.no_PFG = 3
                                 , required.no_STRATA = 5
                                 , required.simul_duration = 100
                                 , required.seeding_duration = 10
                                 , required.seeding_timestep = 1
                                 , doLightCompetition = TRUE
                                 , doDispersal = TRUE
                                 , doHabSuitability = TRUE
                                 , HABSUIT.ref_option = 1
                                 )
                                   
## Create SEVERAL Global_parameters files
PRE_FATE.params_globalParameters(name.simulation = "FATE_simulation"
                                 , required.no_PFG = 3
                                 , required.no_STRATA = 5
                                 , required.simul_duration = 100
                                 , required.seeding_duration = 10
                                 , required.seeding_timestep = 1
                                 , doLightCompetition = TRUE
                                 , doDispersal = TRUE
                                 , doHabSuitability = TRUE
                                 , HABSUIT.ref_option = c(1,2)
                                 )

}
\author{
Maya Guéguen
}
