% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRE_FATE.params_multipleSet.R
\name{PRE_FATE.params_multipleSet}
\alias{PRE_FATE.params_multipleSet}
\title{Create multiple set(s) of parameter files for a \code{FATE-HD} 
simulation}
\usage{
PRE_FATE.params_multipleSet(name.simulation.1, name.simulation.2 = NULL,
  file.simulParam.1, file.simulParam.2 = NULL, no_simulations,
  opt.percent_max = 0.5, opt.percent_seeding = 0.5,
  opt.percent_light = 0.5, opt.percent_soil = 0.5,
  do.max_abund_low = TRUE, do.max_abund_medium = TRUE,
  do.max_abund_high = TRUE, do.seeding_duration = TRUE,
  do.seeding_timestep = TRUE, do.seeding_input = TRUE,
  do.nb_stratum = TRUE, do.LIGHT.thresh_medium = TRUE,
  do.LIGHT.thresh_low = TRUE, do.SOIL.init = TRUE,
  do.SOIL.retention = TRUE, do.DISPERSAL.mode = TRUE,
  do.HABSUIT.ref_option = TRUE)
}
\arguments{
\item{name.simulation.1}{a \code{string} that corresponds to the main 
directory or simulation name of the \code{FATE-HD} simulation from which to 
retrieve the first parameter simulation file (\code{file.simulParam.1}), 
and the second if given (\code{file.simulParam.2}) and no other directory 
provided (\code{name.simulation.2 = NULL})}

\item{name.simulation.2}{default \code{NULL} (\emph{optional}). \cr 
A \code{string} that corresponds to the main directory or simulation name 
of the \code{FATE-HD} simulation from which to retrieve the second 
parameter simulation file (\code{file.simulParam.2})}

\item{file.simulParam.1}{a \code{string} that corresponds to the name of 
the simulation parameter file from which to retrieve parameter values that 
will be used to build the multiple set of new parameters}

\item{file.simulParam.2}{default \code{NULL} (\emph{optional}). \cr 
A \code{string} that corresponds to the name of the second simulation 
parameter file from which to retrieve parameter values that will be used 
to build parameter ranges in comparison with values from 
\code{file.simulParam.1}}

\item{no_simulations}{an \code{integer} corresponding to the number of set 
of parameters that will be produced according to Latin Hypercube Sampling 
(LHS)}

\item{opt.percent_max}{default \code{0.5}. Amount of variation (between 0 
and 1) around the original value of \code{MAX_ABUND_LOW}, 
\code{MAX_ABUND_MEDIUM}, \code{MAX_ABUND_HIGH} if selected}

\item{opt.percent_seeding}{default \code{0.5}. Amount of variation (between 
0 and 1) around the original value of \code{SEEDING_DURATION}, 
\code{SEEDING_TIMESTEP}, \code{SEEDING_INPUT} if selected}

\item{opt.percent_light}{default \code{0.5}. Amount of variation (between 0 
and 1) around the original value of \code{LIGHT_THRESH_MEDIUM}, 
\code{LIGHT_THRESH_LOW} if selected}

\item{opt.percent_soil}{default \code{0.5}. Amount of variation (between 0 
and 1) around the original value of \code{SOIL_INIT}, 
\code{SOIL_RETENTION} if selected}

\item{do.max_abund_low}{default \code{TRUE}. If \code{TRUE}, 
\code{MAX_ABUND_LOW} parameter within \code{Global_parameters} file will be 
declined into a range of values}

\item{do.max_abund_medium}{default \code{TRUE}. If \code{TRUE}, 
\code{MAX_ABUND_MEDIUM} parameter within \code{Global_parameters} file will 
be declined into a range of values}

\item{do.max_abund_high}{default \code{TRUE}. If \code{TRUE}, 
\code{MAX_ABUND_HIGH} parameter within \code{Global_parameters} file will 
be declined into a range of values}

\item{do.seeding_duration}{default \code{TRUE}. If \code{TRUE}, 
\code{SEEDING_DURATION} parameter within \code{Global_parameters} file will 
be declined into a range of values}

\item{do.seeding_timestep}{default \code{TRUE}. If \code{TRUE}, 
\code{SEEDING_TIMESTEP} parameter within \code{Global_parameters} file will 
be declined into a range of values}

\item{do.seeding_input}{default \code{TRUE}. If \code{TRUE}, 
\code{SEEDING_INPUT} parameter within \code{Global_parameters} file will 
be declined into a range of values}

\item{do.nb_stratum}{default \code{TRUE}. If \code{TRUE}, 
\code{NB_STRATUM} parameter within \code{Global_parameters} file will 
be declined into a range of values, with potential impact on some 
parameters within PFG succession (and light) files (parameters 
\code{STRATA}, \code{MAX_ABUNDANCE}, \code{IMM_SIZE}, \code{CHANG_STR_AGES} 
(and \code{SHADE_TOL})) \cr
(see \code{\link{PRE_FATE.params_PFGsuccession}} 
(and \code{\link{PRE_FATE.params_PFGlight}}))}

\item{do.LIGHT.thresh_medium}{default \code{TRUE}. If \code{TRUE}, 
\code{LIGHT_THRESH_MEDIUM} parameter within \code{Global_parameters} file 
will be declined into a range of values \cr (see 
\code{\link{PRE_FATE.params_globalParameters}})}

\item{do.LIGHT.thresh_low}{default \code{TRUE}. If \code{TRUE}, 
\code{LIGHT_THRESH_LOW} parameter within \code{Global_parameters} file 
will be declined into a range of values \cr (see 
\code{\link{PRE_FATE.params_globalParameters}})}

\item{do.SOIL.init}{default \code{TRUE}. If \code{TRUE}, 
\code{SOIL_INIT} parameter within \code{Global_parameters} file 
will be declined into a range of values \cr (see 
\code{\link{PRE_FATE.params_globalParameters}})}

\item{do.SOIL.retention}{default \code{TRUE}. If \code{TRUE}, 
\code{SOIL_RETENTION} parameter within \code{Global_parameters} file 
will be declined into a range of values \cr (see 
\code{\link{PRE_FATE.params_globalParameters}})}

\item{do.DISPERSAL.mode}{default \code{TRUE}. If \code{TRUE}, 
\code{DISPERSAL_MODE} parameter within \code{Global_parameters} file will 
be declined into its 3 possible values (either uniform kernel (\code{1}), 
exponential kernel (\code{2}) or exponential kernel with probability 
(\code{3})) \cr (see \code{\link{PRE_FATE.params_globalParameters}})}

\item{do.HABSUIT.ref_option}{default \code{TRUE}. If \code{TRUE}, 
\code{HABSUIT_OPTION} parameter within \code{Global_parameters} file will 
be declined into its 2 possible values (either random (\code{1}) or PFG 
specific (\code{2})) \cr (see 
\code{\link{PRE_FATE.params_globalParameters}})}
}
\value{
A new folder containing the different sets of parameters asked. 

Depending on what elements have been asked to be varied, three types of 
files can have been modified :
\enumerate{
  \item the global parameter file
  \item the PFG succession files
  \item the PFG light succession files \cr \cr
}

Below are listed the parameters that can change (if selected) within each 
file :

\itemize{
  \item Into the \code{name.simulation/DATA/GLOBAL_PARAMETERS} folder :
  \itemize{
    \item NB_STRATUM
    \item SEEDING_DURATION
    \item SEEDING_TIMESTEP
    \item SEEDING_INPUT
    \item MAX_ABUND_LOW
    \item MAX_ABUND_MEDIUM 
    \item MAX_ABUND_HIGH \cr \cr
  }
  If the simulation includes \emph{light competition} :
  \itemize{
    \item LIGHT_THRESH_MEDIUM
    \item LIGHT_THRESH_LOW
  }
  If the simulation includes \emph{soil competition} :
  \itemize{
    \item SOIL_INIT
    \item SOIL_RETENTION
  }
  If the simulation includes \emph{dispersal} :
  \itemize{
    \item DISPERSAL_MODE
  }
  If the simulation includes \emph{habitat suitability} :
  \itemize{
    \item HABSUIT_OPTION \cr \cr
  }
  \item Into the \code{name.simulation/DATA/PFGS/SUCC} folder :
  \itemize{
    \item STRATA
    \item MAX_ABUNDANCE
    \item IMM_SIZE
    \item CHANG_STR_AGES \cr \cr
  }
  \item Into the \code{name.simulation/DATA/PFGS/LIGHT} folder :
  \itemize{
    \item SHADE_TOL
  }
}
}
\description{
This script is designed to create multiple sets of parameters 
using Latin Hypercube Sampling to help find best combination of parameters 
(see \code{Details})
}
\details{
A \code{FATE-HD} simulation requires several parameters to define general 
characteristics of the simulation : they are saved within a 
\code{Global_parameters} file (see 
\code{\link{PRE_FATE.params_globalParameters}}). To fit the model to a 
particular area and set of Plant Functional Groups (PFG), these are the 
parameters that should be optimized, since they are not data-dependant, 
unlike, for example, parameters related to PFG (height, maturity, dispersal 
distances, soil tolerance, etc).

(\emph{Note : this is true, except when varying the number of strata, which 
will have an impact on some parameters within SUCC and LIGHT PFG parameter 
files.})

\strong{The main idea is to start from a complete simulation folder, to 
select the parameters that should vary, and to create new parameter files 
with new parameter values based on pre-existing values. \cr}

Three possible scenarios are available :
\describe{
  \item{1 folder - \cr 1 simulation file}{
  \itemize{
    \item requested parameter values are extracted from the given 
    simulation file
    \item ranges are assigned to each parameter according to the specified 
    value \cr e.g. : if \code{opt.percent_seeding = 0.5}, and 
    \code{do.seeding_duration} is asked, values will be generated for this 
    parameter between : 
    \deqn{\text{SEEDING_DURATION} \pm \text{SEEDING_DURATION} * 
    \frac{50}{100}}
    \item according to the required number of parameter sets to be produced 
    (\code{no_simulations}), Latin Hypercube Sampling is applied to select 
    each new parameter values
    \item parameter files are created for these new parameter values \cr \cr
  }
  }
  \item{1 folder - \cr 2 simulation files}{
  \itemize{
    \item same as 1st scenario
    \item ranges assigned to each parameter correspond to the extracted 
    values (e.g. : if do.seeding_duration is asked, values will be 
    generated for this parameter between : \code{SEEDING_DURATION 
    (file_simulation.1)} and \code{SEEDING_DURATION (file_simulation.2)} 
    \cr \cr
  }
  }
  \item{2 folders - \cr 2 simulation files}{
  \itemize{
    \item same as 2nd scenario, except that the two given simulation files 
    come from two different simulation folders
  }
  }
}

\strong{Latin Hypercube Sampling} is a statistical method to generate a 
sampling of parameter values from a multidimensional space, while ensuring 
a good representation of the real variability.
The range of each parameter is known, and depending on the number of set of 
parameters asked to be obtained at the end, each range is more or less 
finely cut and values are drawn in order to explore the whole space of 
combinations.
}
\examples{


## Load example data
PNE_PARAM = .loadData("PNE_PARAM")

## PNE_PARAM$succ_light : data.frame
## PNE_PARAM$strata_limits : vector
## PNE_PARAM$disp : data.frame
## PNE_PARAM$dist : data.frame
## PNE_PARAM$global : vector
## PNE_PARAM$masks : rasterStack


## Create a skeleton folder
PRE_FATE.skeletonDirectory(name.simulation = "FATE_PNE")

## Create PFG succession parameter files : predefined of strata limits
tab = PNE_PARAM$succ_light[, c("PFG", "type", "height", "maturity", "longevity")]
PRE_FATE.params_PFGsuccession(name.simulation = "FATE_PNE"
                              , mat.PFG.succ = tab
                              , strata.limits = PNE_PARAM$strata_limits
                              , strata.limits_reduce = FALSE)

## Create PFG light parameter files : predefined of strata limits
tab = PNE_PARAM$succ_light[, c("PFG", "type", "height", "maturity", "longevity", "light")]
PRE_FATE.params_PFGlight(name.simulation = "FATE_PNE"
                         , mat.PFG.succ = tab
                         , strata.limits = PNE_PARAM$strata_limits
                         , strata.limits_reduce = FALSE)

## Create PFG dispersal parameter files
PRE_FATE.params_PFGdispersal(name.simulation = "FATE_PNE"
                             , mat.PFG.disp = PNE_PARAM$disp)

## Create PFG disturbance parameter files
PRE_FATE.params_PFGdisturbance(name.simulation = "FATE_PNE"
                               , mat.PFG.dist = PNE_PARAM$dist)

## Create a Global_parameters file
PRE_FATE.params_globalParameters(name.simulation = "FATE_PNE"
                                 , required.no_PFG = PNE_PARAM$global["NB_FG"]
                                 , required.no_STRATA = PNE_PARAM$global["NB_STRATUM"]
                                 , required.simul_duration = PNE_PARAM$global["SIMULATION_DURATION"]
                                 , required.seeding_duration = PNE_PARAM$global["SEEDING_DURATION"]
                                 , required.seeding_timestep = PNE_PARAM$global["SEEDING_TIMESTEP"]
                                 , required.seeding_input = PNE_PARAM$global["SEEDING_INPUT"]
                                 , required.max_abund_low = PNE_PARAM$global["MAX_ABUND_LOW"]
                                 , required.max_abund_medium = PNE_PARAM$global["MAX_ABUND_MEDIUM"]
                                 , required.max_abund_high = PNE_PARAM$global["MAX_ABUND_HIGH"]
                                 , doLight = TRUE
                                 , LIGHT.thresh_medium = PNE_PARAM$global["LIGHT.thresh_medium"]
                                 , LIGHT.thresh_low = PNE_PARAM$global["LIGHT.thresh_low"]
                                 , doDispersal = TRUE
                                 , DISPERSAL.mode = PNE_PARAM$global["DISPERSAL.mode"]
                                 , doHabSuitability = TRUE
                                 , HABSUIT.ref_option = PNE_PARAM$global["HABSUIT.ref_option"]
                                 , doDisturbances = TRUE
                                 , DIST.no = PNE_PARAM$global["DIST.no"]
                                 , DIST.no_sub = PNE_PARAM$global["DIST.no_sub"]
                                 , DIST.freq = rep(PNE_PARAM$global["DIST.freq"]
                                                   , PNE_PARAM$global["DIST.no"])
)

## Create simulation masks
library(raster)
writeRaster(PNE_PARAM$masks$maskEcrins
            , file = "FATE_PNE/DATA/MASK/mask.tif"
            , overwrite = TRUE)
writeRaster(PNE_PARAM$masks$noDisturb
            , file = "FATE_PNE/DATA/MASK/noDisturb.tif"
            , overwrite = TRUE)

## Create simulation parameters file
PRE_FATE.params_simulParameters(name.simulation = "FATE_PNE"
                                , name.mask = "mask.tif"
                                , name.dist = "noDisturb.tif")

## Create multiple set of parameters
PRE_FATE.params_multipleSet(name.simulation.1 = "FATE_PNE"
                            , file.simulParam.1 = "Simul_parameters_V1.txt"
                            , no_simulations = 30
)

if (dir.exists("FATE_simulation_MULTIPLE_SET")){
  unlink("FATE_simulation_MULTIPLE_SET", recursive = TRUE)
}
PRE_FATE.params_multipleSet(name.simulation.1 = "FATE_PNE"
                            , file.simulParam.1 = "Simul_parameters_V1.txt"
                            , no_simulations = 5
                            , do.max_abund_low = FALSE
                            , do.max_abund_medium = FALSE
                            , do.max_abund_high = FALSE
                            , do.seeding_duration = FALSE
                            , do.seeding_timestep = FALSE
                            , do.seeding_input = FALSE
                            , do.nb_stratum = FALSE
                            , do.LIGHT.thresh_medium = FALSE
                            , do.LIGHT.thresh_low = FALSE
                            , do.SOIL.init = FALSE
                            , do.SOIL.retention = FALSE
                            , do.DISPERSAL.mode = FALSE
                            , do.HABSUIT.ref_option = FALSE
)


}
\seealso{
\code{\link[SPOT]{designLHD}},
\code{\link{PRE_FATE.skeletonDirectory}},
\code{\link{PRE_FATE.params_PFGsuccession}},
\code{\link{PRE_FATE.params_PFGlight}},
}
\author{
Maya Guéguen
}
\keyword{FATE,}
\keyword{Hypercube}
\keyword{Latin}
\keyword{Sampling}
\keyword{simulation,}
