% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRE_FATE.params_PFGdispersal.R
\name{PRE_FATE.params_PFGdispersal}
\alias{PRE_FATE.params_PFGdispersal}
\title{Create \emph{DISPERSAL} parameter files for a \code{FATE} 
simulation}
\usage{
PRE_FATE.params_PFGdispersal(
  name.simulation,
  mat.PFG.disp,
  opt.folder.name = NULL
)
}
\arguments{
\item{name.simulation}{a \code{string} corresponding to the main directory 
or simulation name of the \code{FATE} simulation}

\item{mat.PFG.disp}{a \code{data.frame} with 4 columns : \code{PFG}, 
\code{d50}, \code{d99}, \code{ldd}}

\item{opt.folder.name}{(\emph{optional}) \cr a \code{string} corresponding 
to the name of the folder that will be created into the 
\code{name.simulation/DATA/PFGS/DISP/} directory to store the results}
}
\value{
A \code{.txt} file per PFG into the 
\code{name.simulation/DATA/PFGS/DISP/} directory with the following 
parameters :

\describe{
  \item{DISPERS_DIST}{dispersal distances (3 values) \emph{(in meters)} 
  \cr \cr}
}


A \file{DISP_COMPLETE_TABLE.csv} file summarizing information for all 
groups into the \file{name.simulation/DATA/PFGS/} directory.

If the \code{opt.folder.name} has been used, the files will be into the 
folder \file{name.simulation/DATA/PFGS/DISP/opt.folder.name/}.
}
\description{
This script is designed to create parameter files containing 
dispersal distances for each PFG (one file for each of them) used in the 
dispersal module of \code{FATE}.
}
\details{
A \strong{dispersal module} is available to make the \code{FATE} model 
spatially explicit by dispersing seeds of each PFG with a kernel (see 
\code{\link{PRE_FATE.params_globalParameters}}). \cr \cr

Dispersal distances are needed for each PFG to quantify the amount of seeds 
dispersed into 3 different concentric circles :

\describe{
  \item{d50}{the distance at which \code{50\%} of seeds are dispersed}
  \item{d99}{the distance at which \code{99\%} of seeds are dispersed}
  \item{ldd}{the long dispersal distance at which \code{100\%} of seeds are dispersed}
}
}
\examples{

## Create a skeleton folder with the default name ('FATE_simulation')
PRE_FATE.skeletonDirectory()

## Create PFG dispersal parameter files
PRE_FATE.params_PFGdispersal(name.simulation = "FATE_simulation"
                            , mat.PFG.disp = data.frame(PFG = c("PFG1", "PFG2", "PFG3")
                                                        , d50 = c(50, 50, 10)
                                                        , d99 = c(1000, 1500, 2000)
                                                        , ldd = c(10000, 5000, 10000)))



## ----------------------------------------------------------------------------------------- ##
  
## Load example data

}
\seealso{
\code{\link{PRE_FATE.skeletonDirectory}}, 
\code{\link{PRE_FATE.params_globalParameters}}
}
\author{
Maya Guéguen
}
\keyword{FATE,}
\keyword{dispersal}
\keyword{distance}
\keyword{simulation,}
