% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRE_FATE.params_PFGdispersal.R
\name{PRE_FATE.params_PFGdispersal}
\alias{PRE_FATE.params_PFGdispersal}
\title{Create \emph{DISPERSAL} parameter files for a \code{FATE-HD}
simulation}
\usage{
PRE_FATE.params_PFGdispersal(name.simulation, mat.PFG.disp)
}
\arguments{
\item{name.simulation}{a \code{string} that corresponds to the main directory
or simulation name of the \code{FATE-HD} simulation}

\item{mat.PFG.disp}{a \code{data.frame} with 4 columns : PFG, d50, d99, ldd}
}
\value{
A \code{.txt} file per PFG into the \code{name.simulation/DATA/PFGS/DISP/}
directory with the following parameters :

\itemize{
  \item DISPERS_DIST
}
}
\description{
This script is designed to create parameter files containing 
dispersal distances for each PFG (one file for each of them) used in the
dispersal module of \code{FATE-HD}.
}
\details{
For each PFG, a dispersal module is available to disperse seeds
with a kernel. \cr
Dispersal distances are needed to quantify the amount of
seeds dispersed into 3 different concentric circles :

\describe{
  \item{d50}{the distance at which 50\% of seeds are dispersed}
  \item{d99}{the distance at which 49\% of seeds are dispersed}
  \item{ldd}{the long dispersal distance at which 1\% of seeds are dispersed}
}
}
\examples{

## Create a skeleton folder with the default name ('FATE_simulation')
PRE_FATE.skeletonDirectory()

## Create a Namespace_constants parameter file
PRE_FATE.params_PFGdispersal(name.simulation = "FATE_simulation"
                            , mat.PFG.disp = data.frame(PFG = c("PFG1", "PFG2", "PFG3")
                                                        , d50 = c(500, 500, 100)
                                                        , d99 = c(10000, 15000, 20000)
                                                        , ldd = c(100000, 50000, 100000)))

}
\author{
Maya Guéguen
}
