% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRE_FATE.params_PFGdisturbance.R
\name{PRE_FATE.params_PFGdisturbance}
\alias{PRE_FATE.params_PFGdisturbance}
\title{Create \emph{DISTURBANCE} parameter files for a \code{FATE}
simulation}
\usage{
PRE_FATE.params_PFGdisturbance(
  name.simulation,
  mat.PFG.dist = NULL,
  mat.PFG.tol,
  opt.folder.name = NULL
)
}
\arguments{
\item{name.simulation}{a \code{string} corresponding to the main directory 
or simulation name of the \code{FATE} simulation}

\item{mat.PFG.dist}{(\emph{optional}) \cr 
a \code{data.frame} with 5 columns : \cr 
\code{PFG}, \code{type}, \code{maturity}, \code{longevity}, 
\code{age_above_150cm} (see 
\href{PRE_FATE.params_PFGdisturbance.html#details}{\code{Details}})}

\item{mat.PFG.tol}{a \code{data.frame} with 3 to 7 columns : \cr 
\itemize{
  \item \code{nameDist},
  \item \code{PFG},
  \item (\emph{\code{responseStage}, \code{breakAge}, \code{resproutAge}}), 
  \item \code{responseStage}, \code{killedIndiv}, \code{resproutIndiv}  
  (\emph{or \code{strategy_tol}})
}
(see \href{PRE_FATE.params_PFGdisturbance.html#details}{\code{Details}})}

\item{opt.folder.name}{(\emph{optional}) \cr a \code{string} corresponding 
to the name of the folder that will be created into the 
\code{name.simulation/DATA/PFGS/DIST/} directory to store the results}
}
\value{
A \code{.txt} file per PFG into the 
\code{name.simulation/DATA/PFGS/DIST/} directory with the following 
parameters :

\describe{
  \item{BREAK_AGE}{ages at which the PFG changes of response stage 
  \emph{(in years)}}
  \item{RESPR_AGE}{resprouting age table (in a single row) \cr
  This is a vector of \code{no.DIST * no.responseStages} numbers 
  corresponding \cr to the age at which the PFG can be rejuvenated 
  (younger than the actual one) :
  \itemize{
    \item at different response stages \emph{(\code{RS})}
    \item for each disturbance \emph{(\code{DI})}.
  }
  These parameters should be given in this order (e.g. with 3 response 
  stages) : \cr \code{DI1_RS1, DI1_RS2, DI1_RS3, DI2_RS1...} \emph{(in 
  years)}. 
  }
  \item{FATES}{disturbance response table (in a single row) \cr
  This is a vector of \code{no.DIST * no.responseStages * 2} numbers 
  corresponding \cr to the proportion of individuals :
  \itemize{
    \item that will be killed \emph{(\code{Ki})} or resprout 
    \emph{(\code{Re})}
    \item at different response stages \emph{(\code{RS})}
    \item for each disturbance \emph{(\code{DI})}.
  }
  These parameters should be given in this order (e.g. with 3 response 
  stages) : \cr \code{DI1_RS1_Ki, DI1_RS1_Re, DI1_RS2_Ki, DI1_RS2_Re, 
  DI1_RS3_Ki, DI1_RS3_Re, DI2_RS1_Ki...}
  \cr \emph{(from \code{0} to \code{10}, corresponding to 0 to 100\%)}. 
  }
  \item{PROP_KILLED}{proportion of propagules killed by each disturbance \cr
  \emph{(from \code{0} to \code{10}, corresponding to 0 to 100\%)}}
  \item{ACTIVATED_SEED}{proportion of seeds activated by each disturbance \cr
  \emph{(from \code{0} to \code{10}, corresponding to 0 to 100\%)} \cr \cr}
}

A \code{DIST_COMPLETE_TABLE.csv} file summarizing information for all 
groups into the \code{name.simulation/DATA/PFGS/} directory.  

If the \code{opt.folder.name} has been used, the files will be into the 
folder \code{name.simulation/DATA/PFGS/DIST/opt.folder.name/}.
}
\description{
This script is designed to create parameter files containing 
response to disturbance parameters for each PFG (one file for each of them) 
used in the disturbance module of \code{FATE}.
}
\details{
The \strong{disturbance module} allows the user to simulate spatial 
perturbation(s) that will impact each PFG in terms of \emph{resprouting} and 
\emph{mortality} at different response stages. \cr \cr


Several parameters, given within \code{mat.PFG.dist} or \code{mat.PFG.tol}, 
are required for each PFG in order to set up these responses :

\describe{
  \item{PFG}{the concerned plant functional group \cr \cr}
  
  \item{(\emph{type})}{or life-form, based on Raunkier. \cr It should be 
  either \code{H} (herbaceous), \code{C} (chamaephyte) or \code{P} 
  (phanerophyte) for now}
  \item{(\emph{maturity})}{the age from which the PFG can reproduce}
  \item{(\emph{longevity})}{the maximum or average lifespan of the PFG}
  \item{(\emph{age_above_150cm})}{the maximum height stratum that the PFG 
  can reach \cr \cr}
  
  \item{nameDist}{the name of each perturbation (several can be defined at 
  the same time) \cr \cr}
  
  \item{(\emph{responseStage})}{an \code{integer} corresponding to the 
  concerned response class}
  \item{(\emph{breakAge})}{the age from which the PFG is associated with 
  this response class}
  \item{(\emph{resproutAge})}{the age at which the plants will grow back, 
  if they grow back \cr \cr}
  
  \item{responseStage}{an \code{integer} corresponding to the concerned 
  response class}
  \item{killedIndiv}{an \code{integer} between \code{0} and \code{10} 
  corresponding to the proportion of killed individuals}
  \item{resproutIndiv}{an \code{integer} between \code{0} and \code{10} 
  corresponding to the proportion of resprouting individuals \cr \cr}
  
  \item{(\emph{strategy_tol})}{a \code{string} to choose the response to 
  disturbance strategy : \cr \code{indifferent}, \code{mowing_herbs}, 
  \code{mowing_trees}, \code{grazing_herbs_1}, \code{grazing_herbs_2}, 
  \code{grazing_herbs_3}, \code{grazing_trees_1}, \code{grazing_trees_2}, 
  \code{grazing_trees_3} \cr \cr}
}


These values will allow to calculate or define a set of characteristics for 
each PFG :

\describe{
  \item{BREAK_AGE}{ = each PFG can respond to a disturbance in several 
  different ways that depend on the PFG age \cr
   = ages at which each PFG changes of response stage \cr \cr
  Two methods to define these ages are available :
  \itemize{
    \item from \strong{predefined rules} (using \code{type}, 
    \code{maturity}, \code{longevity}, \code{age_above_150cm}) : \cr \cr
    4 classes are defined that can be labelled as : \cr \strong{JustBorn 
    (\code{1})}, \strong{Juveniles (\code{2})}, \strong{Matures (\code{3})}, 
    \strong{Senescents (\code{4})} \cr \cr
    \tabular{rcc}{
       \tab \strong{\code{H} (herbaceous)} \tab \strong{\code{C} 
       (chamaephyte) or \code{P} (phanerophyte)} \cr
      \strong{from class \code{1} to \code{2}} \tab \code{maturity - 2} 
      \tab \code{1} \cr
      \strong{from class \code{2} to \code{3}} \tab \code{maturity} \tab 
      \code{min}(\code{maturity - 2 , age_above_150cm}) \cr
      \strong{from class \code{3} to \code{4}} \tab \code{longevity - 2} 
      \tab \code{min}(\code{longevity - 2 , age_above_150cm})
    }
    
    Some corrections are made for short-living plants (annuals and 
    biennials) :
    \itemize{
      \item as they die after 1 or 2 years, they are not affected 
      differently according to life stages
      \item break ages from class \code{1} to \code{3} are set to \code{1}, 
      and break age from \code{3} to \code{4} is set to their longevity 
      (\code{1} or \code{2}) \cr \cr
    }
    \item from \strong{user data} : \cr
      \emph{with the values contained within the \code{breakAge} column, 
      if provided \cr \cr}
  }
  }
  \item{RESPR_AGE}{ = when subject to a perturbation, each PFG can either 
  stay undisturbed, be killed, or resprout at a particular age 
  \emph{(in years)} \cr
   = ages at which each PFG will be rejuvenated by a disturbance \cr \cr
  Two methods to define these ages are available :
  \itemize{
    \item from \strong{predefined rules} (using \code{maturity}, 
    \code{longevity}, \code{age_above_150cm}) :
    \itemize{
      \item classes \code{1} and \code{2} : too young to resprout
      \item class \code{3} : 
      \code{min}(\code{maturity - 2 , age_above_150cm})
      \item class \code{4} : \code{longevity - 2}
      \item short-living plants (annuals and biennials) always start back 
      at \code{0} \cr \cr
    }
    \item from \strong{user data} : \cr
      \emph{with the values contained within the \code{resproutAge} column, 
      if provided \cr \cr}
  }
  }
  \item{FATES}{ = proportion of killed and resprouting individuals \cr
   = for each disturbance and for each response stage \cr \cr
  Two methods to define these tolerances are available :
  \itemize{
    \item from \strong{predefined scenarios} (using 
    \code{strategy_tol}) : \cr
      \itemize{
        \item the values give the percentage of killed or resprouting 
        individuals
        \item with \code{1, 2, 3, 4}: response classes
        \item with \code{K}: killed individuals, \code{R}: resprouting 
        individuals \cr \cr
      }
      \describe{
        \item{}{\strong{\code{| ___1___ | ___2___ | ___3___ | ___4___ |}}}
        \item{}{\strong{\code{| _K_ _R_ | _K_ _R_ | _K_ _R_ | _K_ _R_ |}}}
        \item{}{\code{_________________________________________}}
        \item{indifferent}{\code{| _0_ _0_ | _0_ _0_ | _0_ _0_ | _0_ _0_ |}}
        \item{}{\code{_________________________________________}}
        \item{mowing_herbs}{\code{| _0_ _0_ | _0_ _0_ | 50\% 50\% | 100\% 
        0_ |}}
        \item{mowing_trees}{\code{| _0_ _0_ | 100\% 0_ | 100\% 0_ | 100\% 
        0_ |}}
        \item{}{\code{_________________________________________}}
        \item{grazing_herbs_1}{\code{| _0_ _0_ | 10\% _0_ | _0_ 50\% | _0_ 
        10\% |}}
        \item{grazing_herbs_2}{\code{| _0_ _0_ | 50\% _0_ | _0_ 80\% | 10\% 
        50\% |}}
        \item{grazing_herbs_3}{\code{| _0_ _0_ | 90\% _0_ | 10\% 90\% | 50\% 
        50\% |}}
        \item{}{\code{_________________________________________}}
        \item{grazing_trees_1}{\code{| 40\% _0_ | _0_ _0_ | _0_ _0_ | _0_ 
        _0_ |}}
        \item{grazing_trees_2}{\code{| 80\% _0_ | _0_ _0_ | _0_ _0_ | _0_ 
        _0_ |}}
        \item{grazing_trees_3}{\code{| 100\% 0_ | 40\% _0_ | _0_ _0_ | _0_ 
        _0_ |}}
      }
    \item from \strong{user data} : \cr
      \emph{with the values contained within the \code{responseStage}, 
      \code{killedIndiv} and \code{resproutIndiv} columns, if provided \cr
      The \code{PFG} column can contain either the life form (\code{H}, 
      \code{C} or \code{P}) or the PFG name. Both methods can be combined 
      (but are applied in the order given by the \code{PFG} column). \cr \cr
      }
  }
  }
  \item{PROP_KILLED}{ = the proportion of propagules killed by each 
  disturbance \cr
  (\emph{currently set to \code{0} for all PFG and disturbances})
  }
  \item{ACTIVATED_SEED}{ = the proportion of seeds activated by each 
  disturbance \cr
  (\emph{currently set to \code{0} for all PFG and disturbances})
  }
}
}
\examples{


## ----------------------------------------------------------------------------------------- ##
                                    
## Load example data
                                                           

}
\seealso{
\code{\link{PRE_FATE.skeletonDirectory}}, 
\code{\link{PRE_FATE.params_globalParameters}}
}
\author{
Maya Guéguen
}
\keyword{FATE,}
\keyword{disturbance,}
\keyword{killing,}
\keyword{resprouting}
\keyword{simulation,}
