% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRE_FATE.params_PFGdisturbance.R
\name{PRE_FATE.params_PFGdisturbance}
\alias{PRE_FATE.params_PFGdisturbance}
\title{Create \emph{DISTURBANCE} parameter files for a \code{FATE-HD}
simulation}
\usage{
PRE_FATE.params_PFGdisturbance(name.simulation, mat.PFG.dist,
  mat.PFG.succ = paste0(name.simulation,
  "/DATA/PFGS/SUCC_COMPLETE_TABLE.csv"), opt.folder.name = NULL)
}
\arguments{
\item{name.simulation}{a \code{string} that corresponds to the main 
directory or simulation name of the \code{FATE-HD} simulation}

\item{mat.PFG.dist}{a \code{data.frame} with 5 columns : \cr \code{name}, 
\code{responseStage}, \code{PFG}, \code{KilledIndiv}, \code{ResproutIndiv} 
(see \code{Details})}

\item{mat.PFG.succ}{a \code{data.frame} with at least 6 columns : \cr
\code{PFG}, \code{type}, \code{MATURITY}, \code{LONGEVITY}, \code{STRATA} 
and \code{CHANG_STR_AGES_to_str_3}. \cr
Such an object can be obtained with the 
\code{\link{PRE_FATE.params_PFGsuccession}} function.}

\item{opt.folder.name}{(\emph{optional}) \cr a \code{string} that 
corresponds to the name of the folder that will be created into the 
\code{name.simulation/DATA/PFGS/DIST/} directory to store the results}
}
\value{
A \code{.txt} file per PFG into the \code{name.simulation/DATA/PFGS/DIST/}
directory with the following parameters :

\describe{
  \item{BREAK_AGE}{the age when each PFG changes of response stage \emph{(in years)}}
  \item{RESPR_AGE}{the PFG resprouting age table (in a single row) \cr
  This is a vector of \code{no.DIST * no.responseStages} numbers \emph{(in years)}
  corresponding to the age at which each living PFG can be rejuvenated (younger 
  than the actual one) :
  \itemize{
    \item at different response stages \emph{(!currently 4, not yet editable!)}
    \item for each disturbance.
  }
  }
  \item{FATES}{the PFG disturbance response table (in a single row) \cr
  This is a vector of \code{no.DIST * no.responseStages * 2} numbers \cr
  \emph{(0: 0\% 1: 10\% 2: 20\% 3: 30\% 4: 40\% 5: 50\% 6: 60\% 7: 70\% 8: 80\% 9: 
  90\% 10: 100\%)} \cr
  corresponding to the proportion of individuals :
  \itemize{
    \item that will be killed or resprout
    \item at different response stages \emph{(!currently 4, not yet editable!)}
    \item for each disturbance.
  }
  }
  \item{PROP_KILLED}{the proportion of propagules killed by each disturbance \cr
  \emph{(0: 0\% 1: 10\% 2: 20\% 3: 30\% 4: 40\% 5: 50\% 6: 60\% 7: 70\% 8: 80\% 9: 
  90\% 10: 100\%)}}
  \item{ACTIVATED_SEED}{the proportion of seeds activated by each disturbance \cr
  \emph{(0: 0\% 1: 10\% 2: 20\% 3: 30\% 4: 40\% 5: 50\% 6: 60\% 7: 70\% 8: 80\% 9: 
  90\% 10: 100\%)} \cr \cr}
}

A \code{DIST_COMPLETE_TABLE.csv} file summarizing information for all groups into the
\code{name.simulation/DATA/PFGS/} directory.

If the \code{opt.folder.name} has been used, the files will be into the folder
\code{name.simulation/DATA/PFGS/DIST/opt.folder.name/}
}
\description{
This script is designed to create parameter files containing 
response to disturbance parameters for each PFG (one file for each of them) 
used in the disturbance module of \code{FATE-HD}.
}
\details{
The disturbance module of \code{FATE-HD} allows the user to simulate 
spatial perturbation(s) that will impact each PFG in terms of 
\emph{resprouting} and \emph{mortality} on the different response stages 
\emph{(!currently 4, not yet editable!)}. \cr

Several parameters, given within \code{mat.PFG.succ}, are required for each 
PFG in order to set up these responses :

\describe{
  \item{type}{or life-form, based on Raunkier. It should be either \code{H} 
  (herbaceous), \code{C} (chamaephyte) or \code{P} (phanerophyte) for now}
  \item{MATURITY}{the age from which the PFG can reproduce}
  \item{LONGEVITY}{the maximum or average lifespan of the PFG}
  \item{STRATA}{the maximum height stratum that the PFG can reach}
  \item{CHANG_STR_AGES}{at what age each PFG goes into the upper stratum}
}


These values will allow to calculate or define a set of characteristics for 
each PFG :

\describe{
  \item{BREAK_AGE}{ = each PFG can respond to a disturbance in several 
  different ways \emph{(!currently 4, not yet editable!)} that depend 
  on the PFG age \cr
   = ages at which each PFG changes of response stage \cr \cr
   
   These response classes are the same for all disturbances for each PFG. 
   \cr These ages are defined according to the maturity and longevity of 
   the PFG, its life form and its ability to grow above 1.5 meter :
   \itemize{
     \item age from class \code{1} to \code{2} :
     \itemize{
       \item \code{maturity - 2} for herbaceous
       \item \code{1} for chamaephytes and phanerophytes
     }
     \item age from class \code{2} to \code{3} :
     \itemize{
       \item \code{maturity} for herbaceous
       \item \code{min(maturity - 2 , age_above_150cm)} for chamaephytes 
       and phanerophytes
     }
     \item age from class \code{3} to \code{4} :
     \itemize{
       \item \code{longevity - 2} for herbaceous
       \item \code{min(longevity - 2 , age_above_150cm)} for chamaephytes 
       and phanerophytes
     }
   }
   Some corrections are made for short-living plants (annuals and 
   biennials) : as they die after 1 or 2 years, they are not affected 
   differently according to life stages. Break ages from class \code{1} to 
   \code{3} are set to 1, and break age from \code{3} to \code{4} is set 
   to their longevity (1 or 2).
  }
  \item{RESPR_AGE}{ = when subject to a perturbation, each PFG can either 
  stay undisturbed, be killed, or resprout \emph{(in years)} \cr
   = ages at which each PFG will be rejuvenated by a disturbance \cr \cr
  It does not impact dead inviduals, only living ones. \cr
  It is defined according to the maturity and longevity of the PFG, and 
  its ability to grow above 1.5 meter :
  \itemize{
    \item individuals within response stages \code{1} and \code{2} are too 
    young to resprout
    \item individuals within response stage \code{3}: \cr
    \code{min(maturity - 2, age_above_150cm)}
    \item individuals within response stage \code{4} : \code{longevity - 2}
    \item short-living plants (annuals and biennials) always start back 
    at 0
  }
  }
}

A second file, \code{mat.PFG.dist}, is required to define the importance of 
the response of each PFG to each disturbance :

\describe{
  \item{name}{the name of each perturbation (several can be defined at the 
  same time)}
  \item{responseStage}{the concerned response class \emph{(!currently 4, 
  not yet editable!)}}
  \item{PFG}{the concerned plant functional group \emph{(!should match with 
  the ones given within \code{mat.PFG.succ}!)}}
  \item{KilledIndiv}{the proportion of killed individuals}
  \item{ResproutIndiv}{the proportion of resprouting individuals}
}

These values will allow to define a third parameter for each PFG :

\describe{
  \item{FATES}{ = proportion of killed or resprouting individuals \cr
   = for each disturbance and for each response stage \cr
   Two methods are available to give the proportion of killed and 
   resprouting individuals :
   \itemize{
     \item for each life form : \code{H} (herbaceous), \code{C}
     (chamaephyte) or \code{P} (phanerophyte)
     \item for each PFG
   }
   Both methods can be combined (but are applied in the order given by the 
   PFG column).
  }
}

Two parameters are also defined, but currently set to 0 :

\describe{
  \item{PROP_KILLED}{ = the proportion of propagules killed by each 
  disturbance \cr \cr
  It is currently set to 0 for all PFG and disturbances.
  }
  \item{ACTIVATED_SEED}{ = the proportion of seeds activated by each 
  disturbance \cr \cr
  It is currently set to 0 for all PFG and disturbances.}
}
}
\examples{

## Create a skeleton folder with the default name ('FATE_simulation')
PRE_FATE.skeletonDirectory()

## Create PFG succession parameter files
PRE_FATE.params_PFGsuccession(name.simulation = "FATE_simulation"
                            , mat.PFG.succ = data.frame(PFG = paste0("PFG",1:6)
                                                        , type = c("C", "C", "H", "H", "P", "P")  
                                                        , height = c(10, 250, 36, 68, 1250, 550)
                                                        , maturity = c(5, 5, 3, 3, 8, 9)
                                                        , longevity = c(12, 200, 25, 4, 110, 70)))
                                                        
## Create PFG disturbance parameter files
tab.dist = data.frame(name = rep(c("DIST1","DIST2"), each = 4 * 3)
                      , responseStage = rep(1:4, 2 * 3)
                      , PFG = rep(c("C", "H", "P"), each = 2 * 4)
                      , KilledIndiv = c(c(0,10,10,10,1,1,0,0)
                                        , c(0,0,0,0,1,1,0,0)
                                        , c(10,10,10,10,10,0,0,0))
                      , ResproutIndiv = c(c(0,0,0,0,0,0,5,1)
                                          , c(0,0,9,10,0,0,5,1)
                                          , c(0,0,0,0,0,0,0,0)))

PRE_FATE.params_PFGdisturbance(name.simulation = "FATE_simulation"
                               , mat.PFG.dist = tab.dist)



## ----------------------------------------------------------------------------------------- ##
                                    
## Load example data
PNE_PARAM = .loadData("PNE_PARAM")

## PNE_PARAM$succ_light : data.frame
## PNE_PARAM$strata_limits : vector
## PNE_PARAM$dist : data.frame

tab = PNE_PARAM$succ_light[, c("PFG", "type", "height", "maturity", "longevity")]

## Create PFG succession parameter files : predefined of strata limits
PRE_FATE.params_PFGsuccession(name.simulation = "FATE_simulation"
                            , mat.PFG.succ = tab
                            , strata.limits = PNE_PARAM$strata_limits
                            , strata.limits_reduce = FALSE)
                            
## Create PFG disturbance parameter files
PRE_FATE.params_PFGdisturbance(name.simulation = "FATE_simulation"
                               , mat.PFG.dist = PNE_PARAM$dist)
                                                           

}
\seealso{
\code{\link{PRE_FATE.skeletonDirectory}}
}
\author{
Maya Guéguen
}
\keyword{FATE,}
\keyword{disturbance,}
\keyword{killing,}
\keyword{resprouting}
\keyword{simulation,}
