% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRE_FATE.params_PFGlight.R
\name{PRE_FATE.params_PFGlight}
\alias{PRE_FATE.params_PFGlight}
\title{Create \emph{LIGHT} parameter files for a \code{FATE} simulation}
\usage{
PRE_FATE.params_PFGlight(
  name.simulation,
  mat.PFG.light,
  mat.PFG.tol = NULL,
  opt.folder.name = NULL
)
}
\arguments{
\item{name.simulation}{a \code{string} corresponding to the main directory 
or simulation name of the \code{FATE} simulation}

\item{mat.PFG.light}{a \code{data.frame} with 2 to 6 columns : \cr 
\itemize{
  \item \code{PFG},
  \item \code{type}, (\emph{or \code{active_germ_low}, 
  \code{active_germ_medium}, \code{active_germ_high}}) (\emph{or
  \code{strategy_ag}})
  \item \emph{\code{type}, \code{light_need}}
}
(see \href{PRE_FATE.params_PFGlight.html#details}{\code{Details}})}

\item{mat.PFG.tol}{(\emph{optional}) \cr 
a \code{data.frame} with 2 to 4 columns : \cr 
\itemize{
  \item \code{PFG},
  \item \code{lifeStage}, \code{resources}, \code{tolerance} 
  (\emph{or \code{strategy_tol}})
}
(see \href{PRE_FATE.params_PFGlight.html#details}{\code{Details}})}

\item{opt.folder.name}{(\emph{optional}) \cr a \code{string} corresponding 
to the name of the folder that will be created into the 
\code{name.simulation/DATA/PFGS/LIGHT/} directory to store the results}
}
\value{
A \code{.txt} file per PFG into the 
\code{name.simulation/DATA/PFGS/LIGHT/} directory with the following 
parameters :

\describe{
  \item{NAME}{name of the PFG}
  \item{ACTIVE_GERM}{germination rates depending on light conditions 
  \cr \emph{(from \code{0} to \code{10}, corresponding to 0 to 100\%)}}
  \item{LIGHT}{light value or strategy of the PFG}
  \item{LIGHT_TOL}{light tolerance table (in a single row). \cr 
  This is a vector of 9 numbers corresponding to the ability of the PFG to 
  survive or not :
  \itemize{
    \item at different life stages \emph{(Germinant (\code{Ge}), Immature 
    (\code{Im}), Mature (\code{Ma}))}
    \item under different light conditions \emph{(Low (\code{L}), Medium 
    (\code{M}) or High (\code{H}))}.
  }
  These parameters should be given in this order : \code{GeL, GeM, GeH, ImL, 
  ImM, ImH, MaL, MaM, MaH}
  \cr \emph{(from \code{0} to \code{10}, corresponding to 0 to 100\%)}. 
  \cr \cr}
}

A \code{LIGHT_COMPLETE_TABLE.csv} file summarizing information for all 
groups into the \code{name.simulation/DATA/PFGS/} directory.  

If the \code{opt.folder.name} has been used, the files will be into the 
folder \code{name.simulation/DATA/PFGS/LIGHT/opt.folder.name/}.
}
\description{
This script is designed to create parameter files containing 
light-related parameters for each PFG (one file for each of them) used in 
the light module of \code{FATE}.
}
\details{
The \strong{light module} allows the user to add the effect of 
\strong{light competition} within a primary vegetation succession. \cr \cr

Several parameters, given within \code{mat.PFG.light} or \code{mat.PFG.tol}, 
are required for each PFG in order to set up this light competition :

\describe{
  \item{PFG}{the concerned plant functional group \cr \cr}
  
  \item{type}{or life-form, based on Raunkier. \cr It should be either 
  \code{H} (herbaceous), \code{C} (chamaephyte) or \code{P} (phanerophyte) 
  for now}
  \item{(\emph{active_germ_low})}{an \code{integer} between \code{0} and 
  \code{10} corresponding to the proportion of seeds that will germinate for 
  \code{Low} light condition}
  \item{(\emph{active_germ_medium})}{an \code{integer} between \code{0} and 
  \code{10} corresponding to the proportion of seeds that will germinate for 
  \code{Medium} light condition}
  \item{(\emph{active_germ_high})}{an \code{integer} between \code{0} and 
  \code{10} corresponding to the proportion of seeds that will germinate for 
  \code{High} light condition}
  \item{(\emph{strategy_ag})}{a \code{string} to choose the germination 
  strategy : \cr \code{light_lover}, \code{indifferent}, \code{shade_lover} 
  \cr \cr}
  
  \item{(\emph{light_need})}{an \code{integer} between \code{0} and \code{5} 
  corresponding to the light preference of the PFG (e.g. from Flora 
  Indicativa)\cr \cr}
  
  \item{lifeStage}{the concerned life stage (\code{Germinant}, 
  \code{Immature}, \code{Mature})}
  \item{resources}{the concerned light condition (\code{Low}, 
  \code{Medium}, \code{High})}
  \item{tolerance}{an \code{integer} between \code{0} and \code{10} 
  corresponding to the proportion of surviving individuals}
  \item{(\emph{strategy_tol})}{a \code{string} to choose the tolerance 
  strategy : \cr \code{full_light}, \code{pioneer}, \code{ubiquist}, 
  \code{semi_shade}, \code{undergrowth} \cr \cr}
}


These values will allow to calculate or define a set of characteristics for 
each PFG :

\describe{
  \item{ACTIVE_GERM}{proportion of seeds that will germinate for each light 
  condition (\code{Low}, \code{Medium}, \code{High}) \cr \cr
  Three methods to define these proportions are available :
  \itemize{
    \item from \strong{predefined scenarios} (using \code{strategy_ag}) :
    \describe{
      \item{}{\strong{\code{| _L_ _M_ _H_ |}}}
      \item{}{\code{_______________}}
      \item{light_lover}{\code{| 50\% 80\% 90\% |}}
      \item{indifferent}{\code{| 90\% 90\% 90\% |}}
      \item{shade_lover}{\code{| 90\% 80\% 50\% |}}
    }
    \item from \strong{predefined rules} (using \code{type}) :
    \itemize{
      \item for \code{H} (herbaceous) : \code{50\%, 80\%, 90\%}
      \item for \code{C} (chamaephyte) or \code{P} (phanerophyte): 
      \code{90\%, 90\%, 90\%}
    }
    \item from \strong{user data} : \cr
    \emph{with the values contained within the \code{active_germ_low}, 
    \code{active_germ_medium} and \code{active_germ_high} columns, if 
    provided \cr \cr}
  }
  }

  \item{LIGHT_TOL}{ defined for each life stage (\code{Germinant}, 
  \code{Immature}, \code{Mature}) \cr and each soil condition (\code{Low}, 
  \code{Medium}, \code{High}) \cr \cr
  Three methods to define these tolerances are available :
  \itemize{
    \item from \strong{predefined scenarios} (using 
    \code{strategy_tol}) : \cr
      \itemize{
        \item \code{.} means \emph{Not tolerant}, \code{1} means 
        \emph{Tolerant} (\code{100\%})
        \item with \code{g}: Germinant, \code{i}: Immature, \code{m}: Mature
        \item with \code{L}: low light, \code{M}: medium light, \code{H}: 
        high light \cr \cr
      }
    \describe{
      \item{}{\strong{\code{| _ g _ | _ i _ | _ m _ |}}}
      \item{}{\strong{\code{| L M H | L M H | L M H |}}}
      \item{}{\code{_________________________}}
      \item{full_light}{\code{| 1 1 1 | . . 1 | . . 1 |}}
      \item{pioneer}{\code{| 1 1 1 | . 1 1 | . 1 1 |}}
      \item{ubiquist}{\code{| 1 1 1 | 1 1 1 | 1 1 1 |}}
      \item{semi_shade}{\code{| 1 1 . | 1 1 . | 1 1 1 |}}
      \item{undergrowth}{\code{| 1 1 . | 1 1 . | 1 1 . |}}
    }
    \item from \strong{predefined rules} (using \code{type} and 
    \code{light_need}):
      \describe{
        \item{(A)}{PFG are tolerant to \code{Low} light if \code{light <= 2}}
        \item{(A)}{PFG are tolerant to \code{Medium} light if 
        \code{2 <= light <= 4}}
        \item{(A)}{PFG are tolerant to \code{High} light if 
        \code{light >= 3}}
        \item{(B)}{all germinants are assumed to be tolerant to \code{Low} 
        light}
        \item{(C)}{all mature trees or chamaephytes are assumed to be 
        tolerant to \code{Medium} and \code{High} light conditions}
        \item{\emph{(D)}}{\emph{all immature trees that grow in the 
        penultimate stratum are assumed to be tolerant to \code{High} light} 
        \strong{!! desactivated !!}}
      }
      \itemize{
        \item \code{.} means \emph{Not tolerant}
        \item \code{A, B, C, D} mean \emph{Tolerant} according to one of 
        the rule defined above
        \item with \code{g}: Germinant, \code{i}: Immature, \code{m}: Mature
        \item with \code{L}: low light, \code{M}: medium light, \code{H}: 
        high light \cr \cr
      }
      \describe{
        \item{}{\strong{\code{| _ g _ | _ i _ | _ m _ |}}}
        \item{}{\strong{\code{| L M H | L M H | L M H |}}}
        \item{}{\code{_________________________}}
        \item{1}{\code{| A . . | A . D | A C C |}}
        \item{2}{\code{| A A . | A A D | A A C |}}
        \item{3}{\code{| B A . | . A D | . A C |}}
        \item{4}{\code{| B A A | . A A | . A A |}}
        \item{5}{\code{| B . A | . . A | . C A |}}
      }
    \item from \strong{user data} : \cr
      \emph{with the values contained within the \code{lifeStage}, 
      \code{resources} and \code{tolerance} columns, if provided}
  }
  }
}
}
\examples{


## ----------------------------------------------------------------------------------------- ##

## Load example data
                         

}
\seealso{
\code{\link{PRE_FATE.skeletonDirectory}},
\code{\link{PRE_FATE.params_globalParameters}}, 
\code{\link{PRE_FATE.params_PFGsuccession}}
}
\author{
Maya Guéguen
}
\keyword{FATE,}
\keyword{light}
\keyword{simulation,}
\keyword{tolerance}
