% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRE_FATE.params_PFGlight.R
\name{PRE_FATE.params_PFGlight}
\alias{PRE_FATE.params_PFGlight}
\title{Create \emph{LIGHT} parameter files for a \code{FATE-HD}
simulation}
\usage{
PRE_FATE.params_PFGlight(name.simulation, mat.PFG.succ)
}
\arguments{
\item{name.simulation}{a \code{string} that corresponds to the main 
directory or simulation name of the \code{FATE-HD} simulation}

\item{mat.PFG.succ}{a \code{data.frame} with 6 columns : PFG, type, height,
maturity, longevity, light}
}
\value{
A \code{.txt} file per PFG into the 
\code{name.simulation/DATA/PFGS/LIGHT/} directory with the following 
parameters :

\itemize{
  \item NAME : name of the PFG
  \item ACTIVE_GERM : the germination rates depending on light conditions
  \cr \emph{(0: 0\% 1: 10\% 2: 20\% 3: 30\% 4: 40\% 5: 50\% 6: 60\% 7: 70\% 
  8: 80\% 9: 90\% 10: 100\%)}
  \item SHADE_TOL : the PFG shade tolerance table (in a single row). \cr 
  This is a vector of 9 numbers \emph{(0: Die 1: Survive)} corresponding
  to the ability of the PFG to survive or not :
  \itemize{
    \item at different life stages \emph{(Germinant (Ge), Immature (Im), 
    Mature (Ma))}
    \item under different light conditions \emph{(Low (L), Medium (M) or 
    High (H))}.
  }
  These parameters should be given in this order : GeL, GeM, GeH, ImL, ImM,
  ImH, MaL, MaM, MaH.
}

A \code{LIGHT_COMPLETE_TABLE.csv} file summarizing information for all groups into the
\code{name.simulation/DATA/PFGS/} directory.
}
\description{
This script is designed to create parameter files containing 
light tolerance parameters for each PFG (one file for each of them) used 
in the light module of \code{FATE-HD}.
}
\details{
The light module allows the user to simulate a primary vegetation succession
based on light competition. \cr
Several parameters are required for each PFG in order to set up this life 
cycle :

\describe{
  \item{type}{or life-form, based on Raunkier. It should be either \code{H} 
  (herbaceous), \code{C} (chamaephyte) or \code{P} (phanerophyte) for now}
  \item{height}{the maximum or average height that reach the PFG}
  \item{maturity}{the age from which the PFG can reproduce}
  \item{longevity}{the maximum or average lifespan of the PFG \cr \cr}
  \item{light}{a value between 0 and 10 corresponding to the Ellenberg light 
  value of the PFG (or between 0 and 5 corresponding to Flora Indicativa light value)\cr \cr}
}




These values will allow to calculate or define a set of characteristics for 
each PFG :

\describe{
  \item{STRATA_LIMITS}{(see \code{\link{PRE_FATE.params_PFGsuccession}})}
  \item{STRATA}{(see \code{\link{PRE_FATE.params_PFGsuccession}})}
  \item{MAX_ABUNDANCE}{(see \code{\link{PRE_FATE.params_PFGsuccession}})}
  \item{IMM_SIZE}{(see \code{\link{PRE_FATE.params_PFGsuccession}})}
  \item{CHANG_STR_AGES}{(see \code{\link{PRE_FATE.params_PFGsuccession}})}
  \item{ACTIVE_GERM}{proportion of seeds that will germinate for each light
  condition (Low, Medium, High)}
  \item{SHADE_TOL}{ defined for each life stage (Germinant, Immature, 
  Mature) and each light condition (Low, Medium, High) :
  \itemize{
    \item PFG are tolerant to low light if \code{light <= 3}
    \item PFG are tolerant to medium light if \code{light <= 4}
    \item PFG are tolerant to high light if \code{light >= 3}
    \item all germinants are assumed to be tolerant to Low and Medium light
    \item all mature trees are assumed to be tolerant to High light in the 
    upper strata
    \item all immature trees that grow in the penultimate stratum are 
    assumed to be tolerant to High light
  }
  }
}
}
\examples{

## Create a skeleton folder with the default name ('FATE_simulation')
PRE_FATE.skeletonDirectory()

## Create PFG succession parameter files
PRE_FATE.params_PFGlight(name.simulation = "FATE_simulation"
                            , mat.PFG.succ = data.frame(PFG = paste0("PFG",1:6)
                                                        , type = c("C", "C", "H", "H", "P", "P")  
                                                        , height = c(10, 250, 36, 68, 1250, 550)
                                                        , maturity = c(5, 5, 3, 3, 8, 9)
                                                        , longevity = c(12, 200, 25, 4, 110, 70)
                                                        , light = c(4, 6, 3, 6, 5, 5)))

}
\seealso{
\code{\link{PRE_FATE.skeletonDirectory}}
}
\author{
Maya Guéguen
}
\keyword{FATE,}
\keyword{light,}
\keyword{shade}
\keyword{simulation,}
\keyword{tolerance}
