% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRE_FATE.params_PFGsoil.R
\name{PRE_FATE.params_PFGsoil}
\alias{PRE_FATE.params_PFGsoil}
\title{Create \emph{SOIL} parameter files for a \code{FATE} simulation}
\usage{
PRE_FATE.params_PFGsoil(
  name.simulation,
  mat.PFG.soil,
  mat.PFG.tol = NULL,
  opt.folder.name = NULL
)
}
\arguments{
\item{name.simulation}{a \code{string} corresponding to the main directory 
or simulation name of the \code{FATE} simulation}

\item{mat.PFG.soil}{a \code{data.frame} with 3 to 7 columns : \cr 
\itemize{
  \item \code{PFG},
  \item \code{type}, (\emph{or \code{active_germ_low}, 
  \code{active_germ_medium}, \code{active_germ_high}}) (\emph{or
  \code{strategy_ag}})
  \item \code{soil_contrib}, \code{soil_tol_min}, \code{soil_tol_max} 
  (\emph{or \code{strategy_contrib}})
}
(see \href{PRE_FATE.params_PFGsoil.html#details}{\code{Details}})}

\item{mat.PFG.tol}{(\emph{optional}) \cr 
a \code{data.frame} with 2 to 4 columns : \cr 
\itemize{
  \item \code{PFG},
  \item \code{lifeStage}, \code{resources}, \code{tolerance} 
  (\emph{or \code{strategy_tol}})
}
(see \href{PRE_FATE.params_PFGsoil.html#details}{\code{Details}})}

\item{opt.folder.name}{(\emph{optional}) \cr a \code{string} corresponding 
to the name of the folder that will be created into the 
\code{name.simulation/DATA/PFGS/SOIL/} directory to store the results}
}
\value{
A \code{.txt} file per PFG into the 
\code{name.simulation/DATA/PFGS/SOIL/} directory with the following 
parameters :

\describe{
  \item{NAME}{name of the PFG}
  \item{ACTIVE_GERM}{germination rates depending on soil conditions
  \cr \emph{(from \code{0} to \code{10}, corresponding to 0 to 100\%)}}
  \item{SOIL_CONTRIB}{contribution to the soil value of the pixel}
  \item{SOIL_LOW}{lower value of soil supported by the PFG, \cr 
  defining the limit between \code{Low} and \code{Medium} soil resources 
  for this PFG}
  \item{SOIL_HIGH}{upper value of soil supported by the PFG, \cr 
  defining the limit between \code{Medium} and \code{High} soil resources 
  for this PFG}
  \item{SOIL_TOL}{soil tolerance table (in a single row). \cr 
  This is a vector of 9 numbers corresponding to the ability of the PFG to 
  survive or not :
  \itemize{
    \item at different life stages \emph{(Germinant (\code{Ge}), Immature 
    (\code{Im}), Mature (\code{Ma}))}
    \item under different soil conditions \emph{(Low (\code{L}), Medium 
    (\code{M}) or High (\code{H}))}.
  }
  These parameters should be given in this order : \code{GeL, GeM, GeH, ImL, 
  ImM, ImH, MaL, MaM, MaH}
  \cr \emph{(from \code{0} to \code{10}, corresponding to 0 to 100\%)}. 
  \cr \cr}
}

A \code{SOIL_COMPLETE_TABLE.csv} file summarizing information for all 
groups into the \code{name.simulation/DATA/PFGS/} directory.  

If the \code{opt.folder.name} has been used, the files will be into the 
folder \code{name.simulation/DATA/PFGS/SOIL/opt.folder.name/}.
}
\description{
This script is designed to create parameter files containing 
soil contribution and tolerance for each PFG (one file for each of them) 
used in the soil module of \code{FATE}.
}
\details{
The \strong{soil module} allows the user to add the effect of 
\strong{soil competition} within a primary vegetation succession. \cr \cr

Several parameters, given within \code{mat.PFG.soil} or \code{mat.PFG.tol}, 
are required for each PFG in order to set up the soil competition :

\describe{
  \item{PFG}{the concerned plant functional group \cr \cr}
  
  \item{type}{or life-form, based on Raunkier. \cr It should be either 
  \code{H} (herbaceous), \code{C} (chamaephyte) or \code{P} (phanerophyte) 
  for now}
  \item{(\emph{active_germ_low})}{an \code{integer} between \code{0} and 
  \code{10} corresponding to the proportion of seeds that will germinate for 
  \code{Low} soil condition}
  \item{(\emph{active_germ_medium})}{an \code{integer} between \code{0} and 
  \code{10} corresponding to the proportion of seeds that will germinate for 
  \code{Medium} soil condition}
  \item{(\emph{active_germ_high})}{an \code{integer} between \code{0} and 
  \code{10} corresponding to the proportion of seeds that will germinate for 
  \code{High} soil condition}
  \item{(\emph{strategy_ag})}{a \code{string} to choose the germination 
  strategy : \cr \code{poor_lover}, \code{indifferent}, \code{rich_lover} 
  \cr \cr}
  
  \item{soil_contrib}{a value corresponding to the PFG preference for soil 
  fertility \cr (e.g. nitrogen value from Ellenberg or Flora Indicativa)}
  \item{soil_tol_min}{the minimum soil value tolerated by the PFG (on the 
  same scale than \code{soil_contrib})}
  \item{soil_tol_max}{the maximum soil value tolerated by the PFG (on the 
  same scale than \code{soil_contrib})}
  \item{(\emph{strategy_contrib})}{a \code{string} to choose the 
  contribution strategy : \cr \code{ubiquist}, \code{tobedefined} \cr \cr}
  
  \item{lifeStage}{the concerned life stage (\code{Germinant}, 
  \code{Immature}, \code{Mature})}
  \item{resources}{the concerned soil condition (\code{Low}, 
  \code{Medium}, \code{High})}
  \item{tolerance}{an \code{integer} between \code{0} and \code{10} 
  corresponding to the proportion of surviving individuals}
  \item{(\emph{strategy_tol})}{a \code{string} to choose the tolerance 
  strategy : \cr \code{ubiquist}, \code{tobedefined} \cr \cr}
}

These values will allow to calculate or define a set of characteristics for 
each PFG :

\describe{
  \item{ACTIVE_GERM}{proportion of seeds that will germinate for each soil 
  condition (\code{Low}, \code{Medium}, \code{High}) \cr \cr
  Three methods to define these proportions are available :
  \itemize{
    \item from \strong{predefined scenarios} (using \code{strategy_ag}) :
    \describe{
      \item{}{\strong{\code{| _L_ _M_ _H_ |}}}
      \item{}{\code{_______________}}
      \item{poor_lover}{\code{| 80\% 90\% 50\% |}}
      \item{indifferent}{\code{| 90\% 90\% 90\% |}}
      \item{rich_lover}{\code{| 50\% 90\% 80\% |}}
    }
    \item from \strong{predefined rules} (using \code{type}) :
    \itemize{
      \item for \code{H} (herbaceous) : \code{80\%, 100\%, 50\%}
      \item for \code{C} (chamaephyte) or \code{P} (phanerophyte) : 
      \code{90\%, 100\%, 90\%}
    }
    \item from \strong{user data} : \cr
    \emph{with the values contained within the \code{active_germ_low}, 
    \code{active_germ_medium} and \code{active_germ_high} columns, if 
    provided \cr \cr}
  }
  }
  
  \item{SOIL_CONTRIB \cr SOIL_LOW \cr SOIL_HIGH}{
  Two methods to define these values are available :
  \itemize{
    \item from \strong{predefined scenarios} (using 
    \code{strategy_contrib}) : \cr
      \itemize{
        \item the values give the \code{soil_tol_min}, \code{soil_contrib} 
        and \code{soil_tol_max}
        \item with \code{L}: low soil, \code{M}: medium soil, \code{H}: 
        high soil \cr \cr
      }
      \describe{
        \item{}{\strong{\code{| ___ L ___ | ___ M ___ | ___ H ___ |}}}
        \item{}{\code{_____________________________________}}
        \item{ubiquist}{\code{__________ 1.5 _ 2.5 _ 4.5 __________}}
        \item{to be filled}{}
      }
    \item from \strong{user data} : \cr
      \emph{with the values contained within the \code{soil_contrib}, 
      \code{soil_tol_min} and \code{soil_tol_max} columns, if provided \cr \cr}
  }
  }
  
  \item{SOIL_TOL}{ defined for each life stage (\code{Germinant}, 
  \code{Immature}, \code{Mature}) \cr and each soil condition (\code{Low}, 
  \code{Medium}, \code{High}) \cr \cr
  Three methods to define these tolerances are available :
  \itemize{
    \item from \strong{predefined scenarios} (using 
    \code{strategy_tol}) : \cr
      \itemize{
        \item the values give the percentage of surviving individuals to the 
        concerned conditions
        \item with \code{g}: Germinant, \code{i}: Immature, \code{m}: Mature
        \item with \code{L}: low soil, \code{M}: medium soil, \code{H}: 
        high soil \cr \cr
      }
      \describe{
        \item{}{\strong{\code{| ___ g ___ | ___ i ___ | ___ m ___ |}}}
        \item{}{\strong{\code{| _L _M_ H_ | _L _M_ H_ | _L _M_ H_ |}}}
        \item{}{\code{_____________________________________}}
        \item{ubiquist}{\code{| 90 100 90 | 90 100 90 | 90 100 90 |}}
        \item{to be filled}{}
      }
    \item from \strong{predefined rules} :
      \describe{
        \item{(A)}{germinants are severely impacted by wrong soil conditions}
        \item{(B)}{immatures are half impacted by wrong soil conditions}
        \item{(C)}{matures are little impacted by wrong soil conditions}
      }
      \itemize{
        \item the values give the percentage of surviving individuals to the 
        concerned conditions
        \item with \code{g}: Germinant, \code{i}: Immature, \code{m}: Mature
        \item with \code{L}: low soil, \code{M}: medium soil, \code{H}: 
        high soil \cr \cr
      }
      \describe{
        \item{}{\strong{\code{| ___ g ___ | ___ i ___ | ___ m ___ |}}}
        \item{}{\strong{\code{| _L _M_ H_ | _L _M_ H_ | _L _M_ H_ |}}}
        \item{}{\code{_____________________________________}}
        \item{}{\code{| 10 100 30 | 40 100 60 | 70 100 90 |}}
      }
    \item from \strong{user data} : \cr
      \emph{with the values contained within the \code{lifeStage}, 
      \code{resources} and \code{tolerance} columns, if provided}
  }
  }
}
}
\examples{

## Create a skeleton folder with the default name ('FATE_simulation')
PRE_FATE.skeletonDirectory()

## Create PFG soil parameter files
PRE_FATE.params_PFGsoil(name.simulation = "FATE_simulation"
                        , mat.PFG.soil = data.frame(PFG = c("PFG1", "PFG2", "PFG3")
                                                    , type = c("H", "H", "C")
                                                    , soil_contrib = c(2.5, 3, 4.8)
                                                    , soil_tol_min = c(2, 3, 3)
                                                    , soil_tol_max = c(3, 3, 6)))

}
\seealso{
\code{\link{PRE_FATE.skeletonDirectory}},
\code{\link{PRE_FATE.params_PFGsuccession}}
}
\author{
Maya Guéguen
}
\keyword{FATE,}
\keyword{nitrogen}
\keyword{simulation,}
\keyword{soil}
\keyword{tolerance,}
