% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRE_FATE.params_PFGsoil.R
\name{PRE_FATE.params_PFGsoil}
\alias{PRE_FATE.params_PFGsoil}
\title{Create \emph{SOIL} parameter files for a \code{FATE-HD}
simulation}
\usage{
PRE_FATE.params_PFGsoil(name.simulation, mat.PFG.soil,
  no.class = max(mat.PFG.soil$soil_tol_max))
}
\arguments{
\item{name.simulation}{a \code{string} that corresponds to the main directory
or simulation name of the \code{FATE-HD} simulation}

\item{mat.PFG.soil}{a \code{data.frame} with 5 columns : PFG, soil_contrib, 
soil_tol_min, soil_tol_max}

\item{no.class}{an \code{integer} that corresponds to the number of soil
classes. Default value is the maximum value of soil_tol_max in \code{mat.PFG.soil}}
}
\value{
A \code{.txt} file per PFG into the \code{name.simulation/DATA/PFGS/SOIL/}
directory with the following parameters :

\itemize{
  \item SOIL_CONTRIB
  \item SOIL_TOL
}
}
\description{
This script is designed to create parameter files containing 
soil contribution and tolerance for each PFG (one file for each of them) 
used in the soil module of \code{FATE-HD}.
}
\details{
The soil module allows the user to simulate a primary vegetation succession 
based on soil competition. \cr
Several parameters are required for each PFG in order to set up the soil competition :

\describe{
  \item{soil_contrib}{a value between 0 and 10 corresponding to the Ellenberg 
  nitrogen value of the PFG}
  \item{soil_tol_min}{the minimum Ellenberg nitrogen value tolerated by the PFG}
  \item{soil_tol_max}{the maximum Ellenberg nitrogen value tolerated by the PFG}
}
}
\examples{

## Create a skeleton folder with the default name ('FATE_simulation')
PRE_FATE.skeletonDirectory()

## Create PFG soil parameter files
PRE_FATE.params_PFGsoil(name.simulation = "FATE_simulation"
                        , mat.PFG.soil = data.frame(PFG = c("PFG1", "PFG2", "PFG3")
                                                    , soil_contrib = c(2.5, 3, 4.8)
                                                    , soil_tol_min = c(2, 3, 3)
                                                    , soil_tol_max = c(3, 3, 6)))

}
\seealso{
\code{\link{PRE_FATE.skeletonDirectory}}
}
\author{
Maya Guéguen
}
\keyword{FATE,}
\keyword{nitrogen}
\keyword{simulation,}
\keyword{soil}
\keyword{tolerance,}
