% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRE_FATE.params_PFGsoil.R
\name{PRE_FATE.params_PFGsoil}
\alias{PRE_FATE.params_PFGsoil}
\title{Create \emph{SOIL} parameter files for a \code{FATE-HD} simulation}
\usage{
PRE_FATE.params_PFGsoil(name.simulation, mat.PFG.soil,
  mat.PFG.tol = NULL, opt.folder.name = NULL)
}
\arguments{
\item{name.simulation}{a \code{string} that corresponds to the main 
directory or simulation name of the \code{FATE-HD} simulation}

\item{mat.PFG.soil}{a \code{data.frame} with 5 columns : \code{PFG}, 
\code{type}, \code{soil_contrib}, \code{soil_tol_min}, \code{soil_tol_max}}

\item{mat.PFG.tol}{a \code{data.frame} with 4 columns : \code{PFG}, 
\code{lifeStage}, \code{soilResources}, \code{soil_tol}}

\item{opt.folder.name}{(\emph{optional}) \cr a \code{string} that
corresponds to the name of the folder that will be created into the 
\code{name.simulation/DATA/PFGS/SOIL/} directory to store the results}
}
\value{
A \code{.txt} file per PFG into the 
\code{name.simulation/DATA/PFGS/SOIL/} directory with the following 
parameters :

\describe{
  \item{NAME}{name of the PFG}
  \item{SOIL_CONTRIB}{the contribution (influence) of the PFG on the
  soil value of the pixel}
  \item{SOIL_LOW}{the lower value of soil supported by the PFG}
  \item{SOIL_HIGH}{the upper value of soil supported by the PFG}
  \item{ACTIVE_GERM}{the germination rates depending on soil conditions
  \cr \emph{(0: 0\% 1: 10\% 2: 20\% 3: 30\% 4: 40\% 5: 50\% 6: 60\% 7: 70\% 
  8: 80\% 9: 90\% 10: 100\%)}}
  \item{SOIL_TOL}{the PFG soil tolerance table (in a single row). \cr 
  This is a vector of 9 numbers corresponding to the ability of the PFG to 
  survive or not :
  \itemize{
    \item at different life stages \emph{(Germinant (Ge), Immature (Im), 
    Mature (Ma))}
    \item under different soil conditions \emph{(Low (L), Medium (M) or 
    High (H))}.
  }
  These parameters should be given in this order : GeL, GeM, GeH, ImL, ImM, 
  ImH, MaL, MaM, MaH.
  \cr \emph{(0: 0\% 1: 10\% 2: 20\% 3: 30\% 4: 40\% 5: 50\% 6: 60\% 7: 70\% 
  8: 80\% 9: 90\% 10: 100\%)}}
}

A \code{SOIL_COMPLETE_TABLE.csv} file summarizing information for all groups
into the \code{name.simulation/DATA/PFGS/} directory.

If the \code{opt.folder.name} has been used, the files will be into the folder
\code{name.simulation/DATA/PFGS/SOIL/opt.folder.name/}
}
\description{
This script is designed to create parameter files containing 
soil contribution and tolerance for each PFG (one file for each of them) 
used in the soil module of \code{FATE-HD}.
}
\details{
The soil module allows the user to simulate a primary vegetation succession 
based on soil competition. \cr

Several parameters, given within \code{mat.PFG.soil}, are required for each 
PFG in order to set up the soil competition :

\describe{
  \item{type}{or life-form, based on Raunkier. It should be either \code{H} 
  (herbaceous), \code{C} (chamaephyte) or \code{P} (phanerophyte) for now}
  \item{soil_contrib}{a value corresponding to the PFG preference for soil 
  fertility \cr (e.g. nitrogen value from Ellenberg or Flora Indicativa)}
  \item{soil_tol_min}{the minimum soil value tolerated by the PFG (on the 
  same scale than \emph{soil_contrib})}
  \item{soil_tol_max}{the maximum soil value tolerated by the PFG (on the 
  same scale than \emph{soil_contrib})}
}


These values will allow to calculate or define a set of characteristics for 
each PFG :

\describe{
  \item{SOIL_CONTRIB}{PFG contribution to the soil value}
  \item{SOIL_LOW}{minimum soil value tolerated by the PFG}
  \item{SOIL_HIGH}{maximum soil value tolerated by the PFG}
  \item{ACTIVE_GERM}{proportion of seeds that will germinate for each soil 
  condition (\code{Low}, \code{Medium}, \code{High}) :
  \itemize{
    \item woody species have little variation in germination rate : 
    \code{90\%} for \code{Low} and \code{High} conditions, \code{100\%} for 
    \code{Medium} condition
    \item herbaceous germinate less in richer soil : \code{80\%} for 
    \code{Low}, \code{100\%} for \code{Medium} and \code{50\%} for 
    \code{High} conditions
  }
  }
}

A second file, \code{mat.PFG.tol}, can be given to define the importance of 
the response of each PFG to each soil condition :

\describe{
  \item{PFG}{the name of the PFG concerned}
  \item{lifeStage}{the concerned life stage (Germinant, Immature, Mature)}
  \item{soilResources}{the concerned soil condition (\code{Low}, 
  \code{Medium}, \code{High})}
  \item{soil_tol}{the proportion of surviving individuals}
}

If these values are not given by the user, default rules are defined below :

\describe{
  \item{SOIL_TOL}{ defined for each life stage (Germinant, Immature, 
  Mature) and each soil condition (\code{Low}, \code{Medium}, \code{High}) :
  \itemize{
    \item (A) germinants are severely impacted by wrong soil conditions :
    \itemize{
      \item only \code{10\%} of the cohorts survive in \code{Low} soil 
      condition
      \item all cohorts die in \code{High} soil condition
    }
    \item (B) immatures are half impacted by wrong soil conditions :
    \itemize{
      \item \code{50\%} of the cohorts survive in \code{Low} soil condition
      \item \code{40\%} of the cohorts survive in \code{High} soil condition
    }
    \item (C) matures are little affected by wrong soil conditions :
    \itemize{
      \item \code{90\%} of the cohorts survive in \code{Low} soil condition
      \item \code{80\%} of the cohorts survive in \code{High} soil condition
    }
  }
  }
}
}
\examples{

## Create a skeleton folder with the default name ('FATE_simulation')
PRE_FATE.skeletonDirectory()

## Create PFG soil parameter files
PRE_FATE.params_PFGsoil(name.simulation = "FATE_simulation"
                        , mat.PFG.soil = data.frame(PFG = c("PFG1", "PFG2", "PFG3")
                                                    , type = c("H", "H", "C")
                                                    , soil_contrib = c(2.5, 3, 4.8)
                                                    , soil_tol_min = c(2, 3, 3)
                                                    , soil_tol_max = c(3, 3, 6)))

}
\seealso{
\code{\link{PRE_FATE.skeletonDirectory}}
}
\author{
Maya Guéguen
}
\keyword{FATE,}
\keyword{nitrogen}
\keyword{simulation,}
\keyword{soil}
\keyword{tolerance,}
