% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRE_FATE.params_PFGsuccession.R
\name{PRE_FATE.params_PFGsuccession}
\alias{PRE_FATE.params_PFGsuccession}
\title{Create \emph{SUCCESSION} parameter files for a \code{FATE-HD}
simulation}
\usage{
PRE_FATE.params_PFGsuccession(name.simulation, mat.PFG.succ)
}
\arguments{
\item{name.simulation}{a \code{string} that corresponds to the main directory
or simulation name of the \code{FATE-HD} simulation}

\item{mat.PFG.succ}{a \code{data.frame} with 6 columns : PFG, type, height,
maturity, longevity, dispersal}
}
\value{
A \code{.txt} file per PFG into the \code{name.simulation/DATA/PFGS/SUCC/}
directory with the following parameters :

\itemize{
  \item MATURITY : the maturity age of the PFG (in years)
  \item LONGEVITY : the PFG life span (in years)
  \item MAX_ABUNDANCE : the maximal (qualitative) shade that the PFG is able to produce (1: Low 2: Medium 3: High)
  \item IMM_SIZE : the relative size of the immature PFG (0: 0% 1: 10% 2: 50% 3: 90% 4: 100% 5: 40% 6: 80%)
  \item CHANG_STR_AGES : the ages at which the PFG goes in the upper stratum (in years, put a value higher than the PFG life span if it is not supposed to rise a stratum)
  \item WIDE_DISPERS : is the PFG able to disperse everywhere (i.e. no dispersal limits) (0: No 1: Yes)
  \item ACTIVE_GERM : the germination rates depending on light conditions (0: 0% 1: 10% 2: 50% 3: 90% 4: 100% 5: 40% 6: 80%)
  \item SHADE_TOL : the PFG shade tolerance table (in a single row). This is a vector of 9 numbers (0: Die 1: Survive) corresponding to the ability of the PFG to survive or not at different life stages (Germinate (Ge), Immature (Im), Mature (Ma)) under different light conditions (Low (L), Medium (M) or High (H) light resources). These parameters should be given in this order : GeL, GeM, GeH, ImL, ImM, ImH, MaL, MaM, MaH.
  \item SEED_POOL_LIFE : the maximal number of years seeds are able to survive
  \item SEED_DORMANCY
}
}
\description{
This script is designed to create parameter files containing 
succession parameters for each PFG (one file for each of them) used in the
core module of \code{FATE-HD}.
}
\details{
For each PFG,
}
\examples{

## Create a skeleton folder with the default name ('FATE_simulation')
PRE_FATE.skeletonDirectory()

## Create a Namespace_constants parameter file
PRE_FATE.params_PFGsuccession(name.simulation = "FATE_simulation"
                            , mat.PFG.succ = data.frame(PFG = c("PFG1", "PFG2", "PFG3")
                                                        , type = c("H", "C", "P")  
                                                        , height = c(50, 20, 100)
                                                        , maturity = c(5, 5, 12)
                                                        , longevity = c(25, 9, 54)
                                                        , dispersal = c(1, 1, 1)))

}
\author{
Maya Guéguen
}
