% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRE_FATE.params_PFGsuccession.R
\name{PRE_FATE.params_PFGsuccession}
\alias{PRE_FATE.params_PFGsuccession}
\title{Create \emph{SUCCESSION} parameter files for a \code{FATE-HD}
simulation}
\usage{
PRE_FATE.params_PFGsuccession(name.simulation, mat.PFG.succ,
  strata.limits = c(0, 20, 50, 150, 400, 1000, 2000, 5000, 10000),
  strata.limits_reduce = TRUE, opt.folder.name = NULL)
}
\arguments{
\item{name.simulation}{a \code{string} that corresponds to the main 
directory or simulation name of the \code{FATE-HD} simulation}

\item{mat.PFG.succ}{a \code{data.frame} with 5 columns : PFG, type, height,
maturity, longevity}

\item{strata.limits}{a \code{vector} of \code{integer} containing values 
among which height strata limits will be chosen}

\item{strata.limits_reduce}{default \code{TRUE}. If \code{TRUE}, stratum 
height limits are checked to try and bring several PFGs together in a same
stratum}

\item{opt.folder.name}{(\emph{optional}) \cr a \code{string} that
corresponds to the name of the folder that will be created into the 
\code{name.simulation/DATA/PFGS/SUCC/} directory to store the results}
}
\value{
A \code{.txt} file per PFG into the 
\code{name.simulation/DATA/PFGS/SUCC/} directory with the following 
parameters :

\itemize{
  \item NAME : name of the PFG
  \item MATURITY : the maturity age of the PFG \emph{(in years)}
  \item LONGEVITY : the PFG life span \emph{(in years)}
  \item MAX_ABUNDANCE : the maximal (qualitative) abundance / space that 
  the PFG is able to produce / occupy \cr (qualitative) 
  \emph{(1: Low 2: Medium 3: High)}
  \item IMM_SIZE : the relative size of the immature PFG \cr
  \emph{(0: 0\% 1: 10\% 2: 20\% 3: 30\% 4: 40\% 5: 50\% 6: 60\% 7: 70\% 
  8: 80\% 9: 90\% 10: 100\%)}
  \item CHANG_STR_AGES : the ages at which the PFG goes in the upper stratum
  \cr \emph{(in years, put a value higher than the PFG life span if it is 
  not supposed to rise a stratum)}
  \item SEED_POOL_LIFE : the maximal number of years seeds are able to
  survive (for active and dormant pool)
  \item SEED_DORMANCY : are the seeds dormant or not \emph{(0: No 1: Yes)}
  \item POTENTIAL_FECUNDITY : the maximum number of seeds produced by the
  PFG \cr \cr
}

A \code{SUCC_COMPLETE_TABLE.csv} file summarizing information for all groups
into the \code{name.simulation/DATA/PFGS/} directory.  
This file can be used to parameterize the disturbance files.

If the \code{opt.folder.name} has been used, the files will be into the folder
\code{name.simulation/DATA/PFGS/SUCC/opt.folder.name/}
}
\description{
This script is designed to create parameter files containing 
succession parameters for each PFG (one file for each of them) used in the
core module of \code{FATE-HD}.
}
\details{
The core module of \code{FATE-HD} allows the user to simulate a primary 
vegetation succession based on life history traits.
Several parameters are required for each PFG in order to set up this life 
cycle :

\describe{
  \item{type}{or life-form, based on Raunkier. It should be either \code{H} 
  (herbaceous), \code{C} (chamaephyte) or \code{P} (phanerophyte) for now}
  \item{height}{the maximum or average height that reach the PFG}
  \item{maturity}{the age from which the PFG can reproduce}
  \item{longevity}{the maximum or average lifespan of the PFG \cr \cr}
}




These values will allow to calculate or define a set of characteristics for
each PFG :

\describe{
  \item{STRATA_LIMITS}{= the height values that define each stratum.\cr \cr
  The steps are detailed below and try to homogenize the number of PFG
  within each stratum :
  \itemize{
    \item the average number of PFG per stratum should be close to 
    the square root of the total number of PFG (\code{no.PFG.perStrata})
    \item \code{strata limits} should go exponentially and will be selected
    among \cr \code{0, 20, 50, 150, 400, 1000, 2000, 5000, 10000}
    \item PFG are divided according to their \code{height} and these 
    \code{strata limits} and then grouped in order to have per stratum a 
    number of PFG \code{>= (no.PFG.perStrata - 2)}
  }
  }
  \item{STRATA}{the maximum stratum that each PFG can reach}
  \item{MAX_ABUNDANCE}{= maximum abundance of mature PFG in favorable 
  conditions \cr \cr
  It can be seen as a proxy of maximum carrying capacity (\emph{and 
  therefore as a proxy a quantity of shade produced as well, if the light
  module is activated}), and it is defined according to the number of
  strata potentially occupied by a PFG and its life form :
  \itemize{
    \item herbaceous take little place (\emph{and make little shade}) (1),
    \cr chamaephytes take medium place (\emph{and make medium shade}) (2)
    \cr and phanerophytes take a lot of place (\emph{and make lot of shade})
    (3)
    \item all plants in first stratum take little place (\emph{and make 
    little shade}) (1)
    \item plants other than herbaceous in stratum 2 take medium place 
    (\emph{and make medium shade}) (2)
    \item herbaceous in stratum > 2 take medium place (\emph{and make 
    medium shade}) (2)
    \item chamaephytes in stratum > 3 take lot of place (\emph{and make 
    lot of shade}) (3)
  }
  }
  \item{IMM_SIZE}{= relative size of immature versus mature plants \cr
  \itemize{
    \item immature herbaceous take as much space as mature herbaceous \cr
    (\emph{and contribute to shade in the same way}) (100 \%)
    \item immature chamaephytes take two times less space than mature
    chamaephytes \cr (\emph{and contribute to shade half less}) (50 \%)
    \item immature phanerophytes take only 10 \% of their full space
    (\emph{and shade}) capacity
    
    \item intermediate percentages for herbaceous in stratum 2 (80 \%) and 
    in stratum > 2 (50 \%)
    \item immature chamaephytes in 1st stratum take as much space as mature
    chamaephytes \cr (\emph{and contribute to shade in the same way})
    (100 \%)
    \item immature phanerophytes with height < 10m take two times less space
    than mature phanerophytes \cr (\emph{and contribute to shade half less})
    (50 \%)
  }
  }
  \item{CHANG_STR_AGES}{= at what age each PFG goes into the upper stratum.
  \cr \cr
  It is defined using a logistic growth curve with 2 points to parameterize
  it :
  \enumerate{
    \item at \code{age = maturity/2}, \code{height = IMM_SIZE * height}
    \item at \code{age = longevity}, \code{height = height}
  }
  }
  \item{POTENTIAL_FECUNDITY}{= maximum number of seeds produced by the PFG.
  \cr \cr
  By default, defined to the same value for all PFG (100), but can be
  changed by hand directly into the produced files.
  }
}
}
\examples{

## Create a skeleton folder with the default name ('FATE_simulation')
PRE_FATE.skeletonDirectory()

## Create PFG succession parameter files
PRE_FATE.params_PFGsuccession(name.simulation = "FATE_simulation"
                            , mat.PFG.succ = data.frame(PFG = paste0("PFG",1:6)
                                                        , type = c("C", "C", "H", "H", "P", "P")  
                                                        , height = c(10, 250, 36, 68, 1250, 550)
                                                        , maturity = c(5, 5, 3, 3, 8, 9)
                                                        , longevity = c(12, 200, 25, 4, 110, 70)))

}
\seealso{
\code{\link{PRE_FATE.skeletonDirectory}}
}
\author{
Maya Guéguen
}
\keyword{FATE,}
\keyword{height,}
\keyword{longevity,}
\keyword{maturity}
\keyword{simulation,}
