% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRE_FATE.params_simulParameters.R
\name{PRE_FATE.params_simulParameters}
\alias{PRE_FATE.params_simulParameters}
\title{Create \emph{Simul_parameters} parameter file for a \code{FATE} 
simulation}
\usage{
PRE_FATE.params_simulParameters(
  name.simulation,
  name.MASK,
  name.SAVED_STATE = NULL,
  name.DIST = NULL,
  name.DROUGHT = NULL,
  name.FIRE = NULL,
  name.ELEVATION = NULL,
  name.SLOPE = NULL,
  opt.global.name = NULL,
  opt.folder.name = NULL
)
}
\arguments{
\item{name.simulation}{a \code{string} corresponding to the main directory 
or simulation name of the \code{FATE} simulation}

\item{name.MASK}{a \code{string} corresponding to the file name of a raster 
mask, with either \code{0} or \code{1} within each pixel, \code{1} 
corresponding to the cells of the studied area in which the succession 
(core) module of the \code{FATE} simulation will take place (see 
\href{PRE_FATE.params_globalParameters.html#details}{\code{PRE_FATE.params_globalParameters}})}

\item{name.SAVED_STATE}{(\emph{optional}) \cr a \code{string} corresponding 
to the file name of a \code{FATE} object, obtained from a previous 
simulation and from which to restart this new simulation}

\item{name.DIST}{(\emph{optional}) \cr a \code{string} corresponding to the 
file name of a raster mask, with either \code{0} or \code{1} within each 
pixel, \code{1} corresponding to the cells of the studied area in which the 
disturbance module of the \code{FATE} simulation will take place (see 
\href{PRE_FATE.params_globalParameters.html#details}{\code{PRE_FATE.params_globalParameters}})}

\item{name.DROUGHT}{(\emph{optional}) \cr a \code{string} corresponding to 
the name of a raster file, with a \code{numeric} value within each pixel 
corresponding to the drought intensity experienced by this pixel through 
the drought (or fire) disturbance module of the \code{FATE} simulation (see 
\href{PRE_FATE.params_globalParameters.html#details}{\code{PRE_FATE.params_globalParameters}})}

\item{name.FIRE}{(\emph{optional}) \cr a \code{string} corresponding to the 
file name of a raster mask, with either \code{0} or \code{1} within each 
pixel, \code{1} corresponding to the cells of the studied area in which the 
fire disturbance module of the \code{FATE} simulation will take place (see 
\href{PRE_FATE.params_globalParameters.html#details}{\code{PRE_FATE.params_globalParameters}})}

\item{name.ELEVATION}{(\emph{optional}) \cr a \code{string} corresponding to 
the name of a raster file, with a \code{numeric} value within each pixel 
corresponding to the elevation of this pixel and used by the fire 
disturbance module of the \code{FATE} simulation (see 
\href{PRE_FATE.params_globalParameters.html#details}{\code{PRE_FATE.params_globalParameters}})}

\item{name.SLOPE}{(\emph{optional}) \cr a \code{string} corresponding to 
the name of a raster file, with a \code{numeric} value within each pixel 
corresponding to the slope of this pixel and used by the fire 
disturbance module of the \code{FATE} simulation (see 
\href{PRE_FATE.params_globalParameters.html#details}{\code{PRE_FATE.params_globalParameters}})}

\item{opt.global.name}{(\emph{optional}) \cr a \code{string} corresponding 
to the name of the global parameter file in the folder 
\code{name.simulation/DATA/GLOBAL_PARAMETERS/} that will be used to build 
the simulation parameter file}

\item{opt.folder.name}{(\emph{optional}) \cr a \code{string} corresponding 
to the name of the folder in each \code{name.simulation/DATA/PFGS/module/} 
from which PFG file names will be extracted to build the simulation 
parameter file}
}
\value{
A \code{.txt} file into the \code{name.simulation/PARAM_SIMUL/}
directory with the following parameters :

\itemize{
  \item \strong{--GLOBAL_PARAMS--}
  \item \strong{--SAVING_DIR--}
  \itemize{
  \item --SAVED_STATE-- (\emph{optional})
  \item --SAVING_YEARS_MAPS-- (\emph{optional})
  \item --SAVING_YEARS_OBJECTS-- (\emph{optional})
  }
  \item \strong{--MASK--}
  \itemize{
  \item --MASK_CHANGEMASK_YEARS-- (\emph{optional})
  \item --MASK_CHANGEMASK_FILES-- (\emph{optional})
  }
  \item \strong{--PFG_PARAMS_LIFE_HISTORY--}
  \itemize{
  \item --PFG_PARAMS_LIGHT-- (\emph{optional})
  \item --PFG_PARAMS_SOIL-- (\emph{optional})
  \item --PFG_PARAMS_DISPERSAL-- (\emph{optional})
  \item --PFG_MASK_HABSUIT-- (\emph{optional})
  \item --HABSUIT_CHANGEMASK_YEARS-- (\emph{optional})
  \item --HABSUIT_CHANGEMASK_FILES-- (\emph{optional})
  \item --PFG_PARAMS_DISTURBANCES-- (\emph{optional})
  \item --DIST_MASK-- (\emph{optional})
  \item --DIST_CHANGEMASK_YEARS-- (\emph{optional})
  \item --DIST_CHANGEMASK_FILES-- (\emph{optional})
  \item --PFG_PARAMS_DROUGHT-- (\emph{optional})
  \item --DROUGHT_MASK-- (\emph{optional})
  \item --DROUGHT_CHANGEMASK_YEARS-- (\emph{optional})
  \item --DROUGHT_CHANGEMASK_FILES-- (\emph{optional})
  \item --PFG_MASK_ALIENS-- (\emph{optional})
  \item --ALIENS_CHANGEMASK_YEARS-- (\emph{optional})
  \item --ALIENS_CHANGEMASK_FILES-- (\emph{optional})
  \item --ALIENS_CHANGEFREQ_YEARS-- (\emph{optional})
  \item --ALIENS_CHANGEFREQ_FILES-- (\emph{optional})
  \item --PFG_PARAMS_FIRE-- (\emph{optional})
  \item --FIRE_MASK-- (\emph{optional})
  \item --FIRE_CHANGEMASK_YEARS-- (\emph{optional})
  \item --FIRE_CHANGEMASK_FILES-- (\emph{optional})
  \item --FIRE_CHANGEFREQ_YEARS-- (\emph{optional})
  \item --FIRE_CHANGEFREQ_FILES-- (\emph{optional})
  \item --ELEVATION_MASK-- (\emph{optional})
  \item --SLOPE_MASK-- (\emph{optional})
}
  \item \strong{--END_OF_FILE--}
}
}
\description{
This script is designed to create one (or several) parameter 
file containing \code{PARAMETER FILENAMES} used in \code{FATE} model.
}
\details{
The \code{FATE} software takes only one input parameter : a file 
containing links to other files containing all the parameters and data 
needed by the program to run.


\describe{
  \item{GLOBAL_PARAMS}{file where parameters related to the simulation 
  definition are referred (e.g. number of PFG involved, number of height 
  strata, simulation duration, computer resources, modules loaded, etc) \cr
  (see \code{\link{PRE_FATE.params_globalParameters}}) \cr \cr
  }
  \item{SAVING_DIR}{directory where simulation outputs will be stored}
  \item{SAVED_STATE (\emph{optional})}{file containing the results of a 
  previous \code{FATE} simulation from which to restart this new simulation 
  \cr }
  \item{SAVING_YEARS_ \cr ARRAYS (\emph{optional})}{file containing the 
  years for which simulation maps will be saved \cr
  (see \code{\link{PRE_FATE.params_savingYears}})
  }
  \item{SAVING_YEARS_ \cr OBJECTS (\emph{optional})}{file containing the 
  years for which simulation outputs will be saved \cr
  (see \code{\link{PRE_FATE.params_savingYears}}) \cr \cr
  }
  \item{MASK}{raster mask that will define the study area}
  \item{MASK_CHANGEMASK_YEARS \cr (\emph{optional})}{file containing the years 
  to change rasters for the succession module \cr
  (see \code{\link{PRE_FATE.params_changingYears}})
  }
  \item{MASK_CHANGEMASK_FILES \cr (\emph{optional})}{file containing the files 
  to change rasters for the succession module \cr
  (see \code{\link{PRE_FATE.params_changingYears}}) \cr \cr
  }
  \item{PFG_PARAMS_ \cr LIFE_HISTORY}{PFG life history related parameters 
  (one by PFG) \cr
  (see \code{\link{PRE_FATE.params_PFGsuccession}}) \cr \cr
  }
  \item{PFG_PARAMS_ \cr LIGHT (\emph{optional})}{PFG light preferences and 
  tolerance related parameters (one by PFG) \cr
  (see \code{\link{PRE_FATE.params_PFGlight}}) \cr \cr
  }
  \item{PFG_PARAMS_ \cr SOIL (\emph{optional})}{PFG soil contribution and 
  tolerance related parameters (one by PFG) \cr
  (see \code{\link{PRE_FATE.params_PFGsoil}}) \cr \cr
  }
  \item{PFG_PARAMS_ \cr DISPERSAL (\emph{optional})}{PFG dispersal 
  capacity related parameters (one by PFG) \cr
  (see \code{\link{PRE_FATE.params_PFGdispersal}}) \cr \cr
  }
  \item{PFG_MASK_HABSUIT \cr (\emph{optional})}{raster masks (one by PFG) 
  containing PFG habitat suitability for the study area}
  \item{HABSUIT_CHANGEMASK_YEARS \cr (\emph{optional})}{file containing the years 
  to change rasters for the habitat suitability module \cr
  (see \code{\link{PRE_FATE.params_changingYears}})
  }
  \item{HABSUIT_CHANGEMASK_FILES \cr (\emph{optional})}{file containing the files 
  to change rasters for the habitat suitability module \cr
  (see \code{\link{PRE_FATE.params_changingYears}}) \cr \cr
  }
  \item{PFG_PARAMS_ \cr DISTURBANCES (\emph{optional})}{PFG disturbance 
  related parameters in terms of resprouting and mortality (one by PFG) \cr
  (see \code{\link{PRE_FATE.params_PFGdisturbance}})
  }
  \item{DIST_MASK \cr (\emph{optional})}{raster masks that will define the 
  disturbance areas}
  \item{DIST_CHANGEMASK_YEARS \cr (\emph{optional})}{file containing the years 
  to change rasters for the disturbance module \cr
  (see \code{\link{PRE_FATE.params_changingYears}})
  }
  \item{DIST_CHANGEMASK_FILES \cr (\emph{optional})}{file containing the files 
  to change rasters for the disturbance module \cr
  (see \code{\link{PRE_FATE.params_changingYears}}) \cr \cr
  }
  \item{PFG_PARAMS_DROUGHT \cr (\emph{optional})}{PFG drought disturbance 
  related parameters in terms of resprouting and mortality (one by PFG) \cr
  (see \code{\link{PRE_FATE.params_PFGdrought}})}
  \item{DROUGHT_MASK \cr (\emph{optional})}{raster mask that will define the 
  drought intensity of the area}
  \item{DROUGHT_CHANGEMASK_YEARS \cr (\emph{optional})}{file containing the 
  years to change rasters for the drought disturbances module \cr
  (see \code{\link{PRE_FATE.params_changingYears}})}
  \item{DROUGHT_CHANGEMASK_FILES \cr (\emph{optional})}{file containing the 
  files to change rasters for the drought disturbances module \cr
  (see \code{\link{PRE_FATE.params_changingYears}}) \cr \cr}
  \item{PFG_MASK_ALIENS \cr (\emph{optional})}{raster masks (one by alien) 
  containing alien introduction zones for the study area}
  \item{ALIENS_CHANGEMASK_YEARS \cr (\emph{optional})}{file containing the 
  years to change rasters for the aliens introduction module \cr
  (see \code{\link{PRE_FATE.params_changingYears}})}
  \item{ALIENS_CHANGEMASK_FILES \cr (\emph{optional})}{file containing the 
  files to change rasters for the aliens introduction module \cr
  (see \code{\link{PRE_FATE.params_changingYears}})}
  \item{ALIENS_CHANGEFREQ_YEARS \cr (\emph{optional})}{file containing the 
  years to change frequencies for the aliens introduction module \cr
  (see \code{\link{PRE_FATE.params_changingYears}})}
  \item{ALIENS_CHANGEFREQ_FILES \cr (\emph{optional})}{file containing the 
  files to change frequencies for the aliens introduction module \cr
  (see \code{\link{PRE_FATE.params_changingYears}}) \cr \cr}
  \item{PFG_PARAMS_FIRE \cr (\emph{optional})}{PFG fire disturbance 
  related parameters in terms of resprouting and mortality (one by PFG) \cr
  (see \code{\link{PRE_FATE.params_PFGdisturbance}})}
  \item{FIRE_MASK \cr (\emph{optional})}{raster mask that will define the 
  fire disturbance areas}
  \item{FIRE_CHANGEMASK_YEARS \cr (\emph{optional})}{file containing the 
  years to change rasters for the fire disturbances module \cr
  (see \code{\link{PRE_FATE.params_changingYears}})}
  \item{FIRE_CHANGEMASK_FILES \cr (\emph{optional})}{file containing the 
  files to change rasters for the fire disturbances module \cr
  (see \code{\link{PRE_FATE.params_changingYears}})}
  \item{FIRE_CHANGEFREQ_YEARS \cr (\emph{optional})}{file containing the 
  years to change frequencies for the fire disturbances module \cr
  (see \code{\link{PRE_FATE.params_changingYears}})}
  \item{FIRE_CHANGEFREQ_FILES \cr (\emph{optional})}{file containing the 
  files to change frequencies for the fire disturbances module \cr
  (see \code{\link{PRE_FATE.params_changingYears}})}
  \item{ELEVATION_MASK \cr (\emph{optional})}{raster mask that will define 
  the elevation of the area}
  \item{SLOPE_MASK \cr (\emph{optional})}{raster mask that will define the 
  slope of the area}
}
}
\note{
\itemize{
  \item The function produces links to files that are \strong{NOT absolute 
  paths BUT relative ones}. \cr
  \emph{When relative paths are used}, the user should be careful of the 
  folder from which the simulation is launched. \cr
  \emph{A function allows to transform these relative paths into absolute 
  ones. (see examples of \code{\link{.setPattern}})}.
  \item \strong{The order of files matters!} \cr
  For instance the first link below \code{--PFG_PARAMS_LIFE_HISTORY--} flag 
  (e.g. \emph{PFG Albert}) has to match with the first item below the 
  \code{--PFG_PARAMS_DISPERSAL--} flag (must be \emph{PFG Albert} too).
}
}
\examples{

           
## ----------------------------------------------------------------------------------------- ##
                                
## Load example data

}
\seealso{
\code{\link{PRE_FATE.skeletonDirectory}},
\code{\link{PRE_FATE.params_globalParameters}},
\code{\link{PRE_FATE.params_PFGsuccession}},
\code{\link{PRE_FATE.params_PFGlight}},
\code{\link{PRE_FATE.params_PFGsoil}},
\code{\link{PRE_FATE.params_PFGdispersal}},
\code{\link{PRE_FATE.params_PFGdisturbance}},
\code{\link{PRE_FATE.params_PFGdrought}},
\code{\link{PRE_FATE.params_savingYears}},
\code{\link{PRE_FATE.params_changingYears}},
\code{\link{.setPattern}}
}
\author{
Maya Guéguen
}
\keyword{FATE,}
\keyword{simulation}
