% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRE_FATE.params_simulParameters.R
\name{PRE_FATE.params_simulParameters}
\alias{PRE_FATE.params_simulParameters}
\title{Create \emph{Simul_parameters} parameter file for a \code{FATE-HD}
simulation}
\usage{
PRE_FATE.params_simulParameters(name.simulation)
}
\arguments{
\item{name.simulation}{a \code{string} that corresponds to the main directory
or simulation name of the \code{FATE-HD} simulation}
}
\value{
A \code{.txt} file into the \code{name.simulation/PARAM_SIMUL/}
directory with the following parameters :

\itemize{
  \item --NAMESPACE_CONSTANTS--
  \item --GLOBAL_PARAMS--
  \item --ARRAYS_SAVING_YEARS-- (\emph{optional})
  \item --OBJECTS_SAVING_YEARS-- (\emph{optional})
  \item --PFG_LIFE_HISTORY_PARAMS--
  \item --PFG_DISPERSAL_PARAMS--
  \item --PFG_DISTURBANCES_PARAMS-- (\emph{optional})
  \item --PFG_ENVSUIT-- (\emph{optional})
  \item --END_OF_FILE--
}
}
\description{
This script is designed to create one (or several) parameter file
containing \code{PARAMETER FILENAMES} used in \code{FATE-HD} model.
}
\details{
\describe{
  \item{To get abundances \cr per PFG per pixel}{
  }
}
}
\examples{

## Create a skeleton folder with the default name ('FATE_simulation')
if (dir.exists("FATE_simulation")) system("rm -r FATE_simulation/")
PRE_FATE.skeletonDirectory()

## Create a Namespace_constants parameter file
PRE_FATE.params_namespaceConstants(name.simulation = "FATE_simulation"
                                   , global.abund.low = 1000000
                                   , global.abund.med = 5000000
                                   , global.abund.high = 8000000
                                   , global.max.by.cohort = 5000000
                                   , global.resource.thresh.med = 13000000
                                   , global.resource.thresh.low = 19000000)

## Create a Global_parameters file
PRE_FATE.params_globalParameters(name.simulation = "FATE_simulation"
                                 , required.succ_option = "fateh"
                                 , required.no_PFG = 6
                                 , required.no_STRATA = 5
                                 , required.hs_option = 1
                                 , required.seeding_timestep = 1
                                 , required.seeding_duration = c(10,50)
                                 , required.simul_duration = 100)

## Create a SAVE_year_maps or/and SAVE_year_objects parameter file
PRE_FATE.params_saveYears(name.simulation = "FATE_simulation"
                          , years.maps = c(100, 150, 200)
                          , years.objects = 200)

## Create a Changing_times parameter file
PRE_FATE.params_changingYears(name.simulation = "FATE_simulation"
                              , type.changing = "DIST"
                              , mat.changing = data.frame(year = c(50,50,80,80)
                                                          , order = c(1,2,1,2)
                                                          , file.name = c("MASK_DIST1_50.tif"
                                                                          , "MASK_DIST2_50.tif"
                                                                          , "MASK_DIST1_80.tif"
                                                                          , "MASK_DIST2_80.tif")))

## Create PFG succession parameter files
PRE_FATE.params_PFGsuccession(name.simulation = "FATE_simulation"
                              , mat.PFG.succ = data.frame(PFG = paste0("PFG",1:6)
                                                          , type = c("C", "C", "H", "H", "P", "P")
                                                          , height = c(10, 250, 36, 68, 1250, 550)
                                                          , maturity = c(5, 5, 3, 3, 8, 9)
                                                          , longevity = c(12, 200, 25, 4, 110, 70)
                                                          , dispersal = 1
                                                          , light = c(4, 6, 3, 6, 5, 5)))

## Create PFG dispersal parameter files
PRE_FATE.params_PFGdispersal(name.simulation = "FATE_simulation"
                             , mat.PFG.disp = data.frame(PFG = paste0("PFG", 1:6)
                                                         , d50 = rep(c(500, 500, 100),2)
                                                         , d99 = rep(c(10000, 15000, 20000),2)
                                                         , ldd = rep(c(100000, 50000, 100000),2)))

## Create PFG disturbance parameter files
PRE_FATE.params_PFGdisturbance(name.simulation = "FATE_simulation"
                               , mat.PFG.dist = data.frame(name = rep(c("DIST1","DIST2"), each = 4)
                                                           , responseStage = rep(1:4, 2)
                                                           , KilledIndiv_H = c(0,0,0,0,1,1,0,0)
                                                           , KilledIndiv_C = c(0,10,10,10,1,1,0,0)
                                                           , KilledIndiv_P = c(10,10,10,10,10,0,0,0)
                                                           , ResproutIndiv_H = c(0,0,9,10,0,0,5,1)
                                                           , ResproutIndiv_C = c(0,0,0,0,0,0,5,1)
                                                           , ResproutIndiv_P = c(0,0,0,0,0,0,0,0)))

## Create a Simulation parameter file
PRE_FATE.params_simulParameters(name.simulation = "FATE_simulation")
                                   

}
\author{
Maya Guéguen
}
