% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRE_FATE.params_simulParameters.R
\name{PRE_FATE.params_simulParameters}
\alias{PRE_FATE.params_simulParameters}
\title{Create \emph{Simul_parameters} parameter file for a \code{FATE-HD}
simulation}
\usage{
PRE_FATE.params_simulParameters(name.simulation, name.mask,
  name.dist = NULL)
}
\arguments{
\item{name.simulation}{a \code{string} that corresponds to the main 
directory or simulation name of the \code{FATE-HD} simulation}

\item{name.mask}{a \code{string} that corresponds to the file name of a 
raster mask, with either 0 or 1 within each pixel, 1 corresponding to the 
cells of the studied area in which the succession module of the 
\code{FATE-HD} simulation will take place}

\item{name.dist}{(\emph{optional}) \cr a \code{string} that corresponds 
to the file name of a raster mask, with either 0 or 1 within each pixel, 
1 corresponding to the cells of the studied area in which the succession 
module of the \code{FATE-HD} simulation will take place}
}
\value{
A \code{.txt} file into the \code{name.simulation/PARAM_SIMUL/}
directory with the following parameters :

\itemize{
  \item \strong{--GLOBAL_PARAMS--}
  \item \strong{--SAVE_DIR--}
  \item --ARRAYS_SAVING_YEARS-- (\emph{optional})
  \item --OBJECTS_SAVING_YEARS-- (\emph{optional})
  \item \strong{--MASK--}
  \item --MASK_CHANGE_TIME-- (\emph{optional})
  \item --MASK_CHANGE_MASK-- (\emph{optional})
  \item \strong{--PFG_LIFE_HISTORY_PARAMS--}
  \item --PFG_DISPERSAL_PARAMS-- (\emph{optional})
  \item --PFG_HAB_MASK-- (\emph{optional})
  \item --HAB_CHANGE_TIME-- (\emph{optional})
  \item --HAB_CHANGE_MASK-- (\emph{optional})
  \item --PFG_DISTURBANCES_PARAMS-- (\emph{optional})
  \item --DIST_MASK-- (\emph{optional})
  \item --DIST_CHANGE_TIME-- (\emph{optional})
  \item --DIST_CHANGE_MASK-- (\emph{optional})
  \item --PFG_SOIL_PARAMS-- (\emph{optional})
  \item --END_OF_FILE--
}
}
\description{
This script is designed to create one (or several) parameter 
file containing \code{PARAMETER FILENAMES} used in \code{FATE-HD} model.
}
\details{
The \code{FATE-HD} software takes only one input parameter : a file containing
links to other files containing all the parameters and data needed by the 
program to run.


\describe{
  \item{GLOBAL_PARAMS}{file where parameters related to the simulation 
  definition are referred (e.g. number of PFG involved, number of height 
  strata, simulation duration, computer resources, modules loaded, ...) \cr
  (see \code{\link{PRE_FATE.params_globalParameters}}) \cr \cr
  }
  \item{SAVE_DIR}{directory where simulation outputs will be stored}
  \item{ARRAYS_SAVING_ \cr YEARS (\emph{optional})}{file containing the 
  years for which simulation maps will be saved \cr
  (see \code{\link{PRE_FATE.params_saveYears}})
  }
  \item{OBJECTS_SAVING_ \cr YEARS (\emph{optional})}{file containing the 
  years for which simulation outputs will be saved \cr
  (see \code{\link{PRE_FATE.params_saveYears}}) \cr \cr
  }
  \item{MASK}{raster mask that will define the study area}
  \item{MASK_CHANGE_TIME \cr (\emph{optional})}{file containing the years to 
  change rasters for the succession module \cr
  (see \code{\link{PRE_FATE.params_changingYears}})
  }
  \item{MASK_CHANGE_MASK \cr (\emph{optional})}{file containing the files to 
  change rasters for the succession module \cr
  (see \code{\link{PRE_FATE.params_changingYears}}) \cr \cr
  }
  \item{PFG_LIFE_HISTORY_ \cr PARAMS}{PFG life history related parameters 
  (one by PFG) \cr
  (see \code{\link{PRE_FATE.params_PFGsuccession}}) \cr \cr
  }
  \item{PFG_DISPERSAL_ \cr PARAMS (\emph{optional})}{PFG dispersal capacity related parameters 
  (one by PFG) \cr
  (see \code{\link{PRE_FATE.params_PFGdispersal}}) \cr \cr
  }
  \item{PFG_HAB_MASK \cr (\emph{optional})}{raster masks (one by PFG) containing 
  PFG habitat suitability for the study area}
  \item{HAB_CHANGE_TIME \cr (\emph{optional})}{file containing the years to 
  change rasters for the habitat suitability module \cr
  (see \code{\link{PRE_FATE.params_changingYears}})
  }
  \item{HAB_CHANGE_MASK \cr (\emph{optional})}{file containing the files to 
  change rasters for the habitat suitability module \cr
  (see \code{\link{PRE_FATE.params_changingYears}}) \cr \cr
  }
  \item{PFG_DISTURBANCES_ \cr PARAMS (\emph{optional})}{PFG disturbance 
  related parameters in terms of resprouting and mortality (one by PFG) \cr
  (see \code{\link{PRE_FATE.params_PFGdisturbance}})
  }
  \item{DIST_MASK \cr (\emph{optional})}{raster masks that will define the 
  disturbance areas}
  \item{DIST_CHANGE_TIME \cr (\emph{optional})}{file containing the years to 
  change rasters for the disturbance module \cr
  (see \code{\link{PRE_FATE.params_changingYears}})
  }
  \item{DIST_CHANGE_MASK \cr (\emph{optional})}{file containing the files to 
  change rasters for the disturbance module \cr
  (see \code{\link{PRE_FATE.params_changingYears}})
  }
  \item{PFG_SOIL_ \cr PARAMS (\emph{optional})}{PFG soil contribution and tolerance
   related parameters (one by PFG) \cr
  (see \code{\link{PRE_FATE.params_PFGsoil}}) \cr \cr
  }
}
}
\note{
\itemize{
  \item The links to files are \strong{NOT absolute paths}. \cr
  \emph{When relative paths are used}, the user should be careful of the
  folder from which the simulation is launched. \cr
  \emph{A function allows to transform these relative paths into absolute
  ones. (!not available yet!)}
  \item \strong{The order of files matters!} \cr
  For instance the first link below \code{--PFG_LIFE_HISTORY_PARAMS--} flag 
  has to match (the same PFG) with the first item below the 
  \code{--PFG_DISPERSAL_PARAMS--} flag.
}
}
\examples{

## Create a skeleton folder with the default name ('FATE_simulation')
if (dir.exists("FATE_simulation")) unlink("FATE_simulation", recursive = TRUE)
PRE_FATE.skeletonDirectory()

## Create a Global_parameters file
PRE_FATE.params_globalParameters(name.simulation = "FATE_simulation"
                                 , required.no_PFG = 6
                                 , required.no_STRATA = 5
                                 , required.simul_duration = 100
                                 , required.seeding_duration = c(10,50)
                                 , required.seeding_timestep = 1
                                 , required.seeding_input = 100
                                 , required.max_by_cohort = 5000000
                                 , required.max_abund_low = 3000000
                                 , required.max_abund_medium = 5000000
                                 , required.max_abund_high = 9000000)

## Create a SAVE_year_maps or/and SAVE_year_objects parameter file
PRE_FATE.params_saveYears(name.simulation = "FATE_simulation"
                          , years.maps = c(100, 150, 200)
                          , years.objects = 200)

## Create a Changing_times parameter file
PRE_FATE.params_changingYears(name.simulation = "FATE_simulation"
                              , type.changing = "DIST"
                              , mat.changing = data.frame(year = c(50,50,80,80)
                                                          , order = c(1,2,1,2)
                                                          , file.name = c("MASK_DIST1_50.tif"
                                                                          , "MASK_DIST2_50.tif"
                                                                          , "MASK_DIST1_80.tif"
                                                                          , "MASK_DIST2_80.tif")))

## Create PFG succession parameter files
PRE_FATE.params_PFGsuccession(name.simulation = "FATE_simulation"
                              , mat.PFG.succ = data.frame(PFG = paste0("PFG",1:6)
                                                          , type = c("C", "C", "H", "H", "P", "P")
                                                          , height = c(10, 250, 36, 68, 1250, 550)
                                                          , maturity = c(5, 5, 3, 3, 8, 9)
                                                          , longevity = c(12, 200, 25, 4, 110, 70)))

## Create PFG dispersal parameter files
PRE_FATE.params_PFGdispersal(name.simulation = "FATE_simulation"
                            , mat.PFG.disp = data.frame(PFG = paste0("PFG", 1:6)
                                                        , MODE = 1
                                                        , d50 = rep(c(50, 50, 10),2)
                                                        , d99 = rep(c(1000, 1500, 2000),2)
                                                        , ldd = rep(c(10000, 5000, 10000), 2)))

## Create PFG disturbance parameter files
PRE_FATE.params_PFGdisturbance(name.simulation = "FATE_simulation"
                               , mat.PFG.dist = data.frame(name = rep(c("DIST1","DIST2"), each = 4)
                                                           , responseStage = rep(1:4, 2)
                                                           , KilledIndiv_H = c(0,0,0,0,1,1,0,0)
                                                           , KilledIndiv_C = c(0,10,10,10,1,1,0,0)
                                                           , KilledIndiv_P = c(10,10,10,10,10,0,0,0)
                                                           , ResproutIndiv_H = c(0,0,9,10,0,0,5,1)
                                                           , ResproutIndiv_C = c(0,0,0,0,0,0,5,1)
                                                           , ResproutIndiv_P = c(0,0,0,0,0,0,0,0)))
                                                           
## Create PFG soil parameter files
PRE_FATE.params_PFGsoil(name.simulation = "FATE_simulation"
                        , mat.PFG.soil = data.frame(PFG = paste0("PFG", 1:6)
                                                    , type = c("C", "C", "H", "H", "P", "P")
                                                    , soil_contrib = c(2.5, 3, 4.8, 2.5, 3, 4.8)
                                                    , soil_tol_min = c(2, 3, 3, 2, 3, 3)
                                                    , soil_tol_max = c(3, 3, 6, 3, 3, 6)))

## Create a Simulation parameter file
file.create("FATE_simulation/DATA/MASK/mask.tif")
PRE_FATE.params_simulParameters(name.simulation = "FATE_simulation"
                                , name.mask = "mask.tif")

for (pfg in 1:6){
  file.create(paste0("FATE_simulation/DATA/PFGS/HABSUIT/mask_PFG", pfg, ".tif"))
}
PRE_FATE.params_simulParameters(name.simulation = "FATE_simulation"
                                , name.mask = "mask.tif")
                                   

}
\seealso{
\code{\link{PRE_FATE.skeletonDirectory}},
\code{\link{PRE_FATE.params_globalParameters}},
\code{\link{PRE_FATE.params_PFGsuccession}},
\code{\link{PRE_FATE.params_PFGlight}},
\code{\link{PRE_FATE.params_PFGdispersal}},
\code{\link{PRE_FATE.params_PFGdisturbance}},
\code{\link{PRE_FATE.params_PFGsoil}},
\code{\link{PRE_FATE.params_saveYears}},
\code{\link{PRE_FATE.params_changingYears}}
}
\author{
Maya Guéguen
}
\keyword{FATE,}
\keyword{simulation}
