% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRE_FATE.selectDominant.R
\name{PRE_FATE.selectDominant}
\alias{PRE_FATE.selectDominant}
\title{Selection of dominant species from abundance releves}
\usage{
PRE_FATE.selectDominant(mat.site.species.abund = NULL, sites = NULL,
  species = NULL, abund = NULL, landclass = NULL,
  selectionRule.quanti = 0, selectionRule.min_mean_abund = 0,
  selectionRule.min_no_abund_over25 = 0, doHabitatSelection = FALSE,
  selectionRule.min_percent_habitat = 0.05,
  selectionRule.min_no_habitat = 5)
}
\arguments{
\item{mat.site.species.abund}{a \code{data.frame} with at least 3 columns :
sites ID, species ID, abundance values as frequence values(e.g. Braun Blanquet)
with \code{NA} when no information, habitat ID (optional)}

\item{sites}{a \code{vector} with sites ID}

\item{species}{a \code{vector} with species ID}

\item{abund}{a \code{vector} with abundance values (e.g. Braun Blanquet) with \code{NA}
when no information}

\item{selectionRule.quanti}{minimum quantile of total number of sites that
must occupy the species (between 0 and 1)}

\item{selectionRule.min_mean_abund}{minimal average abundance of the species}

\item{selectionRule.min_no_abund_over25}{minimum number of sites where species
is dominant (>= 25 \% of coverage)}

\item{doHabitatSelection}{default \code{FALSE}. If \code{TRUE}, selection is also done including
constraints on habitat class}

\item{selectionRule.min_percent_habitat}{minimum percentage of habitat class
occupied by the species (between 0 and 1)}

\item{selectionRule.min_no_habitat}{minimum number of sites of specific habitat
occupied by the species}

\item{habitat}{a \code{vector} with habitat ID}
}
\value{
A \code{data.frame} with all the species selected and the values of
parameters used to make the selection :

\describe{
  \item{stat.no_sites_recorded}{number of sites with information
  (presence-absence or abundance)}
  \item{stat.no_sites_abund}{number of sites with abundance information}
  \item{stat.abund_median}{median abundance (coverage percentage)}
  \item{stat.abund_mean}{mean abundance (coverage percentage)}
  \item{stat.abund_max}{maximal abundance (coverage percentage)}
  \item{stat.no_sites \cr abund_max}{number of sites with maximal abundance}
  \item{stat.no_sites \cr abund_over25}{number of sites with abundance >= 25}
  \item{SELECTION}{dataset in which the species has been selected as dominant
  \cr (global = all data, or habitat, or global + habitat...)}
}


The information is written in \code{PRE_FATE_DOMINANT_species_selected_[...].csv} files :
\describe{
  \item{\file{COMPLETE_TABLE}}{the complete
  table of selected species with all the statistics described above}
  \item{\file{SPECIES_ONLY}}{only the names /
  ID of the species selected}
}

Two \code{PRE_FATE_DOMINANT_[...].pdf} files are also created : 
\describe{
  \item{\file{STEP_1 \cr selectionCriteria}}{to visualize the values of
  species metrics, and the criteria used to make the selection}
  \item{\file{STEP_2 \cr selectedSpecies}}{to visualize the values of
  species metrics, for the selected dominant species}
}
}
\description{
This script is designed to select dominant species from
             abundance records, and habitat if the information is available.
}
\details{
This function provides a way to \emph{select dominant species based on
abundance sampling information}. \cr
This information can be given either directly in the form of a \code{data.frame},
or indirectly with a \code{vector} for each required information (sites,
species, abundance, habitat class - optional).


2 rules are applied to make the species selection :

\describe{
  \item{\strong{1. Presence releves}}{
  \describe{
    \item{\strong{on number of sites}}{the species should be found in a certain
    number of sites, which corresponds to the quantile \code{selectionRule.quanti}
    of all the number of records per species \cr \emph{This should ensure that
    the selected species are covering all the studied area (or at least a
    determining part of it, assuming that the releves are well distributed over
    all the zone).}}
  }
  }
  \item{\strong{2. Abundance releves : }}{both conditions must be fullfilled
  \describe{
    \item{\strong{on average abundance}}{the species should have a mean
    abundance superior or equal to \code{selectionRule.min_mean_abund}}
    \item{\strong{on dominancy}}{the species should be dominant (i.e. represent
    at least 25 \% of the coverage of the site) in at least
    \code{selectionRule.min_no_abund_over25} sites}
    \item{}{\emph{This should ensure that the selected species, besides being
    abundant, are} frequently \emph{abundant, meaning not ultra specific to a
    particular area.}}
  }
  }
}

If habitat information is available (e.g. type of environment : urban,
desert, grassland... ; type of vegetation : shrubs, forest, alpine grasslands... ;
etc), 1 rule with 2 conditions can be added to the selection steps,
\emph{in order to keep species that are not dominant at the large scale
but could be representative of a specific habitat} :

\describe{
  \item{\strong{3. Releves per habitat : }}{both conditions must be fullfilled
  \describe{
    \item{\strong{on occupancy}}{the species should occupy at least
    \code{selectionRule.min_percent_habitat} \% of a landclass}
    \item{\strong{on representation}}{the species should be found in at least
    \code{selectionRule.min_no_habitat} sites of a landclass}
    \item{}{\emph{In the same way than for rule 2., this should allow to select
    species that are not globally dominant but that can be representative of a
    specific habitat, both in terms of abundance and dominancy.}}
  }
  }
}
}
\examples{

## Load example data
data(MontBlanc)
str(MontBlanc)

## MontBlanc$mat.releves : data.frame

## Transformation of Braun-Blanquet abundances
MontBlanc$mat.releves$abund = PRE_FATE.abundBraunBlanquet(abund = MontBlanc$mat.releves$abund)

## Selection of dominant species

#########################################################################################
## EXAMPLE 1 : With default parametrization
#########################################################################################

sp.DOM = PRE_FATE.selectDominant(mat.site.species.abund = MontBlanc$mat.releves)

#########################################################################################
## EXAMPLE 2 : Redefinition of global selection rules
#########################################################################################

sp.DOM = PRE_FATE.selectDominant(mat.site.species.abund = MontBlanc$mat.releves
                                 , selectionRule.quanti = 0.9
                                 , selectionRule.min_mean_abund = 25
                                 , selectionRule.min_no_abund_over25 = 10)

#########################################################################################
## EXAMPLE 3 : Add landclass selection with default parametrization
#########################################################################################

sp.DOM = PRE_FATE.selectDominant(mat.site.species.abund = MontBlanc$mat.releves
                                 , selectionRule.quanti = 0.9
                                 , selectionRule.min_mean_abund = 25
                                 , selectionRule.min_no_abund_over25 = 10
                                 , doHabitatSelection = TRUE)
                          
#########################################################################################       
## EXAMPLE 4 : Redefinition of landclass selection rules
#########################################################################################

sp.DOM = PRE_FATE.selectDominant(mat.site.species.abund = MontBlanc$mat.releves
                                 , selectionRule.quanti = 0.9
                                 , selectionRule.min_mean_abund = 25
                                 , selectionRule.min_no_abund_over25 = 10
                                 , doHabitatSelection = TRUE
                                 , selectionRule.min_percent_habitat = 0.05
                                 , selectionRule.min_no_habitat = 10)
                                 
#########################################################################################

## 2 .csv and 2 .pdf files have been produced
str(sp.DOM)

}
\seealso{
\code{\link{PRE_FATE.abundBraunBlanquet}}
}
\author{
Maya Guéguen
}
\keyword{abundance,}
\keyword{class}
\keyword{dominant}
\keyword{habitat}
\keyword{quantile,}
\keyword{species,}
