% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRE_FATE.skeletonDirectory.R
\name{PRE_FATE.skeletonDirectory}
\alias{PRE_FATE.skeletonDirectory}
\title{Create the skeleton folder for a \code{FATE-HD} simulation}
\usage{
PRE_FATE.skeletonDirectory(name.simulation = "FATE_simulation")
}
\arguments{
\item{name.simulation}{a \code{string} that will be used as the main directory
and simulation name}
}
\value{
A directory tree with folder to contain the parameter files, the
simulation files and the results.
}
\description{
This script is designed to create a user-friendly directory
tree to run a \code{FATE-HD} simulation.
}
\details{
\code{FATE-HD} requires only one input parameter, which is a file containing
the names of parameter files, which may themselves contain parameters or other
file names. The point is : the user could give names of files stored everywhere
on a machine, and does not have to put them all in one same place.

But as this is more practical, this function proposes a way to organize all
those files or parameter files that will or could be used by a \code{FATE-HD}
simulation.

The tree structure is detailed below :

\describe{
  \item{DATA}{this folder will contain all the data or parameters that are
  needed by the model
  \itemize{
    \item \strong{GLOBAL_PARAMETERS} : files containing global parameters for the simulation
    \item \strong{NAMESPACE_CONSTANTS} : files containing constants used in the model
    \item \strong{MASK} : all maps used in the model
    \item \strong{SCENARIO} : files containing information about changes in
    input data ( e.g. habitat suitability maps, disturbances maps, etc) 
    \item \strong{SAVE} : files containing information about times to save outputs
    \item \strong{PFGS} :
    \itemize{
      \item \strong{SUCC} : all the PFG life history parameter files
      \item \strong{DISP} : all the PFG dispersal parameter files
      \item \strong{DIST} : all the PFG-response to disturbances parameter files
      \item \strong{SOIL} : all the PFG soil parameter files
      \item \strong{HABSUIT} : all the PFG habitat suitability maps
    }
  }
  }
  \item{PARAM_SIMUL}{this folder will contain simulation files that can be
  given as input to the software}
  \item{RESULTS}{this folder will collect all the results produced by the
  software with a folder for each simulation}
}
}
\examples{

## Create a skeleton folder with the default name ('FATE_simulation')
PRE_FATE.skeletonDirectory()

}
\author{
Maya Guéguen
}
