% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRE_FATE.speciesClustering_step2.R
\name{PRE_FATE.speciesClustering_step2}
\alias{PRE_FATE.speciesClustering_step2}
\title{Choose clusters and select determinant species}
\usage{
PRE_FATE.speciesClustering_step2(
  clust.dendrograms,
  no.clusters,
  mat.species.DIST
)
}
\arguments{
\item{clust.dendrograms}{a dendrogram, or a \code{list} of dendrograms (one 
for each \code{GROUP} value). \cr Such an object can be obtained 
with the \code{\link{PRE_FATE.speciesClustering_step1}} function.}

\item{no.clusters}{an \code{integer}, or a \code{vector} of \code{integer} 
(one for each \code{GROUP} value), with the number of clusters to be kept}

\item{mat.species.DIST}{a \code{dist} object, or a \code{list} of 
\code{dist} objects (one for each \code{GROUP} value), corresponding to the 
distance between each pair of species. \cr Such an object can be obtained 
with the \code{\link{PRE_FATE.speciesDistance}} function.}
}
\value{
A \code{list} containing one \code{vector}, one \code{data.frame} 
with the following columns, and two \code{ggplot2} objects :

\describe{
  \item{determ.sp}{the names of all determinant species}
  \item{determ.all}{(\emph{determinant and non-determinant species})
  \describe{
    \item{\code{PFG}}{ID of the plant functional group 
    (\code{GROUP} + \code{ID.cluster})}
    \item{\code{GROUP}}{name of data subset}
    \item{\code{ID.cluster}}{cluster number}
    \item{\code{species}}{name of species}
    \item{\code{ID.species}}{species number in each PFG}
    \item{\code{sp.mean.dist}}{species mean distance to other species of 
    the same PFG}
    \item{\code{allSp.mean}}{\eqn{mean(\text{sp.mean.dist})} within the PFG}
    \item{\code{allSp.min}}{\eqn{mean(\text{sp.mean.dist}) - 1.64 * 
    sd(\text{sp.mean.dist})} within the PFG}
    \item{\code{allSp.max}}{\eqn{mean(\text{sp.mean.dist}) + 1.64 * 
    sd(\text{sp.mean.dist})} within the PFG}
    \item{\code{DETERMINANT}}{\code{TRUE} if determinant species, \code{FALSE} 
    otherwise}
  }
  }
  \item{plot.distance}{\code{ggplot2} object, representing the distribution 
  of mean distances between species for each functional group} 
  \item{plot.PCO}{\code{list} of \code{ggplot2} objects, representing the 
  PFG within the functional space \cr \cr}
}

One \file{PRE_FATE_CLUSTERING_STEP_2_distantSpecies_PCO.pdf} file is created 
containing two types of graphics : 
\describe{
  \item{distantSpecies}{to visualize in each PFG the distribution of mean 
  distance of each species to other species, and non-determinant species 
  which are outside the distribution}
  \item{PCO}{to visualize in each PFG the distribution of species, with and 
  without non-determinant species}
}
}
\description{
This script is designed to obtain functional groups by : 1) 
selecting the number of clusters to be kept from an object obtained with 
the \code{\link{PRE_FATE.speciesClustering_step1}} function ; 2) refining 
these groups by identifying determinant species in each of them.
}
\details{
This function allows to obtain a classification of \strong{dominant} 
species into Plant Functional Groups (PFG), and the \strong{determinant} 
species based on these PFG. \cr \cr

\strong{What is the difference between \code{dominant} and
\code{determinant} species ?}

\itemize{
  \item{\strong{Dominant} species are species representative of an 
  environment or a studied area, in terms of number of releves or 
  abundance values. They can be found with the 
  \code{\link{PRE_FATE.selectDominant}} function of this package. These 
  dominant species are used to build PFG with the 
  \code{\link{PRE_FATE.speciesClustering_step1}} function.
  }
  \item{Once PFG are built, \strong{determinant} species are defined as 
  refined subsets of dominant species within each PFG. \cr The process is 
  detailed below :
  \itemize{
    \item each dominant species is assigned to a PFG
    \item within each PFG :
    \itemize{
      \item for each species, compute its mean distance to the other 
      species within the PFG (\code{sp.mean.dist})
      \item calculate the mean value of all these mean distances
      (\code{allSp.mean})
      \item calculate the deviation values around this mean value
      (\code{allSp.min} and \code{allSp.max})
      \item determinant species are the ones that are included between 
      those deviation values
    }
  }
  }
}
}
\examples{

## Load example data
data(DATASET_Bauges_PFG)

## Species dissimilarity distance (niche overlap + traits distance)
tab.dist = DATASET_Bauges_PFG$dom.dist_total
str(tab.dist)
as.matrix(tab.dist[[1]])[1:5, 1:5]


## Build dendrograms -------------------------------------------------------------------------
sp.CLUST = PRE_FATE.speciesClustering_step1(mat.species.DIST = tab.dist)
names(sp.CLUST)


## Number of clusters per group
plot(sp.CLUST$plot.clustNo)
no.clusters = c(5, 5, 9)

## Find determinant species ------------------------------------------------------------------
sp.DETERM = PRE_FATE.speciesClustering_step2(clust.dendrograms = sp.CLUST$clust.dendrograms
                                             , no.clusters = no.clusters
                                             , mat.species.DIST = tab.dist)
names(sp.DETERM)

str(sp.DETERM$determ.sp)
str(sp.DETERM$determ.all)

## Species names
sp.NAMES = DATASET_Bauges_PFG$sp.names
sp.NAMES$species = paste0("X", sp.NAMES$species)
determ = merge(sp.DETERM$determ.all, sp.NAMES, by = "species", all.x = TRUE)
str(determ)

plot(sp.DETERM$plot.distance)
plot(sp.DETERM$plot.PCO$Chamaephyte)
plot(sp.DETERM$plot.PCO$Herbaceous)
plot(sp.DETERM$plot.PCO$Phanerophyte)


}
\seealso{
\code{\link[stats]{cutree}},
\code{\link[ade4]{quasieuclid}},
\code{\link[ade4]{dudi.pco}},
\code{\link{PRE_FATE.speciesDistance}},
\code{\link{PRE_FATE.speciesClustering_step1}},
\code{\link{PRE_FATE.speciesClustering_step3}}
}
\author{
Maya Guéguen
}
\keyword{Component}
\keyword{Ordination}
\keyword{Principal}
\keyword{clustering,}
\keyword{hierarchical}
