% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRE_FATE.speciesClustering_step2.R
\name{PRE_FATE.speciesClustering_step2}
\alias{PRE_FATE.speciesClustering_step2}
\title{Choose}
\usage{
PRE_FATE.speciesClustering_step2(clust.dendograms, no.clusters,
  mat.species.DIST)
}
\arguments{
\item{clust.dendograms}{A dendogram or a \code{list} of dendograms with one
for each \code{GROUP} value, as can be obtained with the 
\code{PRE_FATE.speciesClustering_step1} function.}

\item{no.clusters}{A \code{vector} with the number of clusters to be kept
in each subset of data (if the data is split).}

\item{mat.species.DIST}{A \code{dist} object corresponding to the distance
between each pair of species, or a \code{list} of \code{dist} objects,
one for each \code{GROUP} value. Such an object can be obtained with the
\code{PRE_FATE.speciesDistance} function.}
}
\value{
A \code{list} object with 2 elements :

\item{determ.sp}{a \code{vector} with the names of all determinant species}
\item{determ.all}{a \code{data.frame} containing all species (determinant
and non-determinant) with 10 columns :
\itemize{
  \item \code{pfg} : the ID of the PFG (group + no.cluster)
  \item \code{group} : name of sub-dataset
  \item \code{no.cluster} : cluster number
  \item \code{sp} : name of species
  \item \code{ID} : species number in each PFG
  \item \code{sp.mean.dist} : species mean distance to other species of
  the same PFG
  \item \code{allSp.mean} : mean(sp.mean.dist) within the PFG
  \item \code{allSp.min} : mean(sp.mean.dist) - 1.64 * sd(sp.mean.dist)
   within the PFG
  \item \code{allSp.max} : mean(sp.mean.dist) + 1.64 * sd(sp.mean.dist)
   within the PFG
  \item \code{toSuppr} : 0 if determinant species, 1 otherwise
}
}

Two .pdf files are also created : 

\describe{
  \item{\file{PRE_FATE_CLUSTERING_STEP_2C_distantSpecies}}{to account
  for the chosen clustering method}
  \item{\file{PRE_FATE_CLUSTERING_STEP_2C_PCO}}{for decision
  support, to help the user to choose the adequate number of clusters
  to be used into the \code{hclust} method}
}
}
\description{
This script is designed to obtain functional groups by : 1)
select the number of clusters to be kept from an object obtained with the
\code{PRE_FATE.speciesClustering_step1} function ; 2) refine these groups
by identifying determinant species in each one.
}
\details{
This function allows one to obtain
}
\examples{

## Load example data
data(MontBlanc)

## MontBlanc$mat.traits : data.frame
## MontBlanc$mat.nicheOverlap : niolap object

## Calculate distance between species
sp.DIST = PRE_FATE.speciesDistance(mat.species.traits = MontBlanc$mat.traits,
                                   mat.species.overlap = MontBlanc$mat.nicheOverlap)
                                   
str(sp.DIST)
             
## Run hierarchical clustering and number of clusters' selection                      
sp.CLUST = PRE_FATE.speciesClustering_step1(mat.species.DIST = sp.DIST)

## 2 .pdf files have been produced
str(sp.CLUST)

## Select number of clusters and find determinant species
sp.DETERM = PRE_FATE.speciesClustering_step2(clust.dendograms = sp.CLUST$clust.dendograms
                                             , no.clusters = c(11, 7, 8)
                                             , mat.species.DIST = sp.DIST)

## 2 .pdf files have been produced
str(sp.DETERM)

}
\seealso{
\code{\link{PRE_FATE.speciesDistance}},
\code{\link{PRE_FATE.speciesClustering_step1}}
}
\author{
Maya Guéguen
}
\keyword{Component}
\keyword{Ordination}
\keyword{Principal}
\keyword{clustering,}
\keyword{hierarchical}
