% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRE_FATE.speciesClustering_step2.R
\name{PRE_FATE.speciesClustering_step2}
\alias{PRE_FATE.speciesClustering_step2}
\title{Choose clusters and select determinant species}
\usage{
PRE_FATE.speciesClustering_step2(clust.dendograms, no.clusters,
  mat.species.DIST)
}
\arguments{
\item{clust.dendograms}{a dendogram or a \code{list} of dendograms with one
for each \code{GROUP} value, as can be obtained with the 
\code{PRE_FATE.speciesClustering_step1} function.}

\item{no.clusters}{a \code{vector} with the number of clusters to be kept
in each subset of data (if the data is split).}

\item{mat.species.DIST}{a \code{dist} object corresponding to the distance
between each pair of species, or a \code{list} of \code{dist} objects,
one for each \code{GROUP} value. Such an object can be obtained with the
\code{PRE_FATE.speciesDistance} function.}
}
\value{
A \code{list} object with 2 elements :

\describe{
  \item{determ.sp}{a \code{vector} with the names of all determinant species}
  \item{determ.all}{a \code{data.frame} containing all species (determinant
  and non-determinant) with 10 columns :
  \itemize{
    \item \code{pfg} : the ID of the PFG (group + no.cluster)
    \item \code{group} : name of sub-dataset
    \item \code{no.cluster} : cluster number
    \item \code{sp} : name of species
    \item \code{ID} : species number in each PFG
    \item \code{sp.mean.dist} : species mean distance to other species of
    the same PFG
    \item \code{allSp.mean} : mean(sp.mean.dist) within the PFG
    \item \code{allSp.min} : mean(sp.mean.dist) - 1.64 * sd(sp.mean.dist)
    within the PFG
    \item \code{allSp.max} : mean(sp.mean.dist) + 1.64 * sd(sp.mean.dist)
    within the PFG
    \item \code{toSuppr} : 0 if determinant species, 1 otherwise
  }
  }
}

Two \code{PRE_FATE_CLUSTERING_[...].pdf} files are created : 
\describe{
  \item{\file{STEP_2C \cr distantSpecies}}{to visualize in each PFG the
  distribution of mean distance of each species to other species, and
  non-determinant species which are outside the distribution}
  \item{\file{STEP_2C \cr PCO}}{to visualize in each PFG the distribution
  of species, with and without non-determinant species}
}
}
\description{
This script is designed to obtain functional groups by : 1)
select the number of clusters to be kept from an object obtained with the
\code{\link{PRE_FATE.speciesClustering_step1}} function ; 2) refine these 
groups by identifying determinant species in each of them.
}
\details{
This function allows one to obtain a classification of \emph{dominant}
species into Plant Functional Groups (PFG), and the \emph{determinant}
species based on these PFGs.

\strong{What is the difference between \code{dominant} and
\code{determinant} species ?}

\itemize{
  \item{\strong{Dominant} species are species representative of an
  environment or a studied area, in terms of number of releves or
  abundance values. They can be found with the \code{PRE_FATE.selectDominant}
  function of this package. These dominant species are used to build PFG
  with the \code{PRE_FATE.speciesClustering_step1} function.
  }
  \item{Once PFG are built, \strong{determinant} species are defined as
  refined subsets of dominant species within each PFG. The process is
  detailed below :
  \itemize{
    \item each dominant species is assigned to a PFG
    \item within each PFG :
    \itemize{
      \item for each species, compute its mean distance to the other
      species within the PFG (\code{sp.mean.dist})
      \item calculate the mean value of all these mean distances
      (\code{allSp.mean})
      \item calculate the deviation values around this mean value
      (\code{allSp.min} and \code{allSp.max})
      \item determinant species are the ones that are included between
      these deviation values
    }
  }
  }
}
}
\examples{

## Load example data
data("PNE_PFG")

## PNE_PFG$dom.traits : data.frame
## PNE_PFG$dom.dist_overlap : niolap object

## Calculate distance between species
sp.DIST = PRE_FATE.speciesDistance(mat.species.traits = PNE_PFG$dom.traits
                                   , mat.species.overlap = PNE_PFG$dom.dist_overlap
                                   , opt.max.percent.NA = 0.9
                                   , opt.max.percent.similarSpecies = 0.25
                                   , opt.min.sd = 0.3)
                                   
str(sp.DIST)
             
## Run hierarchical clustering and number of clusters' selection                      
sp.CLUST = PRE_FATE.speciesClustering_step1(mat.species.DIST = sp.DIST)

## 2 .pdf files have been produced
str(sp.CLUST)

## Select number of clusters and find determinant species
sp.DETERM = PRE_FATE.speciesClustering_step2(clust.dendograms = sp.CLUST$clust.dendograms
                                             , no.clusters = PNE_PFG$nb.clusters
                                             , mat.species.DIST = sp.DIST)

## 2 .pdf files have been produced
str(sp.DETERM)

}
\seealso{
\code{\link[stats]{cutree}},
\code{\link[ade4]{dudi.pco}},
\code{\link{PRE_FATE.speciesDistance}},
\code{\link{PRE_FATE.speciesClustering_step1}},
\code{\link{PRE_FATE.speciesClustering_step3}}
}
\author{
Maya Guéguen
}
\keyword{Component}
\keyword{Ordination}
\keyword{Principal}
\keyword{clustering,}
\keyword{hierarchical}
