% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRE_FATE.speciesClustering_step3.R
\name{PRE_FATE.speciesClustering_step3}
\alias{PRE_FATE.speciesClustering_step3}
\title{Calculate PFG traits values based on determinant species traits 
values}
\usage{
PRE_FATE.speciesClustering_step3(mat.traits)
}
\arguments{
\item{mat.traits}{a \code{data.frame} with at least 3 columns :
\describe{
  \item{\code{species}}{the ID of each determinant species (see 
  \code{\link{PRE_FATE.speciesClustering_step2}})}
  \item{\code{PFG}}{the corresponding Plant Functional Group (see 
  \code{\link{PRE_FATE.speciesClustering_step2}})}
  \item{\code{...}}{one column for each functional trait (see 
  \href{PRE_FATE.speciesClustering_step3#details}{\code{Details}})}
}}
}
\value{
A \code{list} containing one \code{data.frame} with the following 
columns, and one \code{list} with as many \code{ggplot2} objects as 
functional traits given in \code{mat.traits} :

\describe{
  \item{tab}{ \cr
  \describe{
    \item{\code{PFG}}{the concerned plant functional group}
    \item{\code{no.species}}{the number of species contained in this PFG}
    \item{\code{...}}{one column for each functional trait, computed as the 
    \code{mean} (for numeric traits) or the \code{median} (for categorical 
    traits) of the values of the determinant species of this PFG}
  }
  }
  \item{plot}{\cr
  \describe{
    \item{\code{...}}{one for each functional trait, 'specific' cases 
    excepted (see 
    \href{PRE_FATE.speciesClustering_step3#details}{\code{Details}})}
  }
  }
}
    
The information is written in \file{PRE_FATE_PFG_TABLE_traits.csv} and 
\file{PRE_FATE_CLUSTERING_STEP_3_PFGtraitsValues.pdf} files. \cr
This \code{.csv} file can be used to build parameter files to run a 
\code{FATE} simulation (e.g. \code{\link{PRE_FATE.params_PFGsuccession}}).
}
\description{
This script is designed to calculate PFG traits values based 
on determinant species traits values. Either the mean or the 
median is used depending on the trait class (i.e. numeric or 
categorical).
}
\details{
This function allows to \strong{obtain '\emph{average}' functional trait 
values for each Plant Functional Group}, based on values at the determinant 
species level. \cr \cr


\emph{A graphic is automatically produced for each functional 
trait given, with boxplot representing the values of determinant species, 
and colored points the values calculated for each PFG. \cr
However, some traits can have 'specific' representation, as long as their 
names within \code{mat.traits} match one of the configuration detailed 
below :}
\describe{
  \item{\code{maturity}, \code{longevity}}{to visualize the difference 
  between these two values, for the maturity time has an impact on the 
  fecundity of the PFG within \code{FATE} (see 
  \href{../articles/fate_tutorial_3_MODULES.html#core-module-succession}{CORE module})
  \cr \emph{If there is NO values for longevity within one PFG, and some 
  maturity values are available, some values might be inferred as 
  \eqn{\text{maturity} * 2}. If there is NO values for maturity within 
  one PFG, and some longevity values are available, some values might be 
  inferred as \eqn{\text{longevity} / 2}.}
  }
  \item{\code{height}, \code{light}}{to visualize the PFG light 
  preference, and help decide and understand the choice of the height 
  limits of strata in \code{FATE} (see 
  \href{../articles/fate_tutorial_3_MODULES.html#light-module-competition}{LIGHT 
  competition module})}
  \item{\code{soil_contrib}, \cr \code{soil_tol_min}, \cr 
  \code{soil_tol_max}}{to visualize the PFG soil preference, and help 
  parameterize the global parameters of the soil competition module 
  within \code{FATE} (see 
  \href{../articles/fate_tutorial_3_MODULES.html#soil-module-competition}{SOIL 
  competition module})}
  \item{\cr \code{soil_contrib}, \cr \code{soil_tolerance}}{same as the 
  previous one, but \code{soil_tol_min} and \code{soil_tol_max} values 
  are obtained by adding or removing \code{soil_tolerance} to 
  \code{soil_contrib}}
}
}
\examples{

## Load example data
data(DATASET_Bauges_PFG)

## Species traits
tab.traits = DATASET_Bauges_PFG$dom.traits
str(tab.traits)

## Determinant species
sp.DETERM = DATASET_Bauges_PFG$dom.determ
str(sp.DETERM)

## Merge traits and PFG informations
mat.traits = merge(sp.DETERM[, c("species", "PFG")]
                   , tab.traits
                   , by = "species", all.x = TRUE)


## Compute traits per PFG : no specific graphic ----------------------------------------------
sp.PFG = PRE_FATE.speciesClustering_step3(mat.traits = mat.traits)
names(sp.PFG)
str(sp.PFG$tab)
names(sp.PFG$plot)
plot(sp.PFG$plot$DISPERSAL)
plot(sp.PFG$plot$LIGHT)
plot(sp.PFG$plot$MOISTURE)


## Compute traits per PFG : with one specific graphic ----------------------------------------
colnames(mat.traits) = c("species", "PFG", "GROUP", "DISPERSAL"
                         , "light", "NITROGEN", "MOISTURE", "height")
sp.PFG = PRE_FATE.speciesClustering_step3(mat.traits = mat.traits)
names(sp.PFG)
str(sp.PFG$tab)
names(sp.PFG$plot)
plot(sp.PFG$plot$height_light)


}
\seealso{
\code{\link{PRE_FATE.speciesClustering_step1}},
\code{\link{PRE_FATE.speciesClustering_step2}}
}
\author{
Maya Guéguen
}
\keyword{functional}
\keyword{group,}
\keyword{traits}
