% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRE_FATE.speciesClustering_step3.R
\name{PRE_FATE.speciesClustering_step3}
\alias{PRE_FATE.speciesClustering_step3}
\title{Calculate PFG traits values based on determinant species traits 
values}
\usage{
PRE_FATE.speciesClustering_step3(mat.species.traits)
}
\arguments{
\item{mat.species.traits}{a \code{data.frame} with at least 3 columns :
\describe{
  \item{\code{species}}{the ID of each determinant species (see 
  \code{\link{PRE_FATE.speciesClustering_step2}})}
  \item{\code{PFG}}{a factor variable containing grouping information to
  divide the species into Plant Functional Groups (see 
  \code{\link{PRE_FATE.speciesClustering_step2}})}
  \item{\code{...}}{one column for each functional trait (see 
  \code{Details})}
}}
}
\value{
A \code{data.frame} containing new trait values for each Plant 
Functional Group (PFG) :

\describe{
  \item{type}{either \code{H} (herbaceous), \code{C} (chamaephyte) or 
  \code{P} (phanerophyte) for now}
  \item{height}{mean}
  \item{maturity}{mean}
  \item{longevity}{mean}
  \item{dispersal}{median}
  \item{light}{median}
  \item{soil_contrib}{mean + round to one digit}
  \item{soil_tolerance}{mean + round to one digit}
  \item{palatability}{mean + round to one digit
   \cr \cr}
}

The information is written in \code{PRE_FATE_PFG_TRAITS_TABLE.csv} file. \cr
This file can be used to build parameter files to run a \code{FATE-HD} 
simulation.
}
\description{
This script is designed to calculate PFG traits values based 
on determinant species traits values. Either the \code{mean} or the 
\code{median} is used depending on the trait class (i.e. numeric or 
categorical).
}
\details{
This function allows one to obtain '\emph{average}' functional trait values 
for each Plant Functional Group, based on values at the determinant species
level.

Those functional traits can be :

\describe{
  \item{PFG}{name of the PFG}
  \item{type}{or life-form, based on Raunkier. \cr It should be either 
  \code{H} (herbaceous), \code{C} (chamaephyte) or \code{P} (phanerophyte) 
  for now}
  \item{height}{the maximum or average height that reach the species}
  \item{maturity}{the age from which the species can reproduce}
  \item{longevity}{the maximum or average lifespan of the species}
  \item{dispersal}{the age from which the species can reproduce}
  \item{light}{a value corresponding to the light preference of the species 
  (from shade-lover to preference for full light)}
  \item{soil_contrib}{a value corresponding to the preference of the 
  species for soil fertility (e.g. Ellenberg, Flora Indicativa)}
  \item{soil_tolerance}{the range of soil tolerance : 1 (small) or 2 
  (large)}
  \item{palatability}{the appetence of each species from null to high
   \cr \cr}
}
}
\examples{

## Load example data
PNE_PFG = .loadData("PNE_PFG")

## PNE_PFG$dom.traits : data.frame
## PNE_PFG$dom.determ : data.frame

tab1 = PNE_PFG$dom.determ[which(PNE_PFG$dom.determ$determinant == TRUE), c("species", "PFG")]
tab2 = PNE_PFG$dom.traits[, c("species", "height", "dispersal", "palatability", "light")]
determ.traits = merge(tab1, tab2, by = "species")

PFG.traits = PRE_FATE.speciesClustering_step3(mat.species.traits = determ.traits)
summary(PFG.traits)

}
\seealso{
\code{\link{PRE_FATE.speciesClustering_step1}},
\code{\link{PRE_FATE.speciesClustering_step2}}
}
\author{
Maya Guéguen
}
\keyword{functional}
\keyword{group,}
\keyword{traits}
