% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRE_FATE.speciesClustering_step3.R
\name{PRE_FATE.speciesClustering_step3}
\alias{PRE_FATE.speciesClustering_step3}
\title{Calculate PFG traits values based on determinant species traits 
values}
\usage{
PRE_FATE.speciesClustering_step3(mat.species.traits)
}
\arguments{
\item{mat.species.traits}{A \code{data.frame} with at least 3 columns :
\describe{
  \item{\code{species}}{the ID of each determinant species (see 
  \code{\link{PRE_FATE.speciesClustering_step2}})}
  \item{\code{PFG}}{a factor variable containing grouping information to
  divide the species into Plant Functional Groups (see 
  \code{\link{PRE_FATE.speciesClustering_step2}})}
  \item{\code{...}}{one column for each functional trait (see 
  \code{Details})}
}}
}
\value{
A \code{.txt} file per PFG into the 
\code{name.simulation/DATA/PFGS/SUCC/} directory with the following 
parameters :

\itemize{
  \item NAME : name of the PFG
  \item MATURITY : the maturity age of the PFG \emph{(in years)}
  \item LONGEVITY : the PFG life span \emph{(in years)}
  \item MAX_ABUNDANCE : the maximal (qualitative) shade that the PFG is able
  to produce \cr \emph{(1: Low 2: Medium 3: High)}
  \item IMM_SIZE : the relative size of the immature PFG \cr
  \emph{(0: 0\% 1: 10\% 2: 20\% 3: 30\% 4: 40\% 5: 50\% 6: 60\% 7: 70\% 
  8: 80\% 9: 90\% 10: 100\%)}
  \item CHANG_STR_AGES : the ages at which the PFG goes in the upper stratum
  \cr \emph{(in years, put a value higher than the PFG life span if it is 
  not supposed to rise a stratum)}
  \item SEED_POOL_LIFE : the maximal number of years seeds are able to
  survive (for active and dormant pool)
  \item SEED_DORMANCY : are the seeds dormant or not \emph{(0: No 1: Yes)
  \cr \cr}
}

A \code{SUCC_COMPLETE_TABLE.csv} file summarizing information for all groups
into the \code{name.simulation/DATA/PFGS/} directory.  
This file can be used to parameterize the disturbance files.
}
\description{
This script is designed to create
}
\details{
This function allows one to obtain 'average' functional trait values for
each Plant Functional Group, based on values at the determinant species
level.

Those functional traits can be :

\describe{
  \item{type}{or life-form, based on Raunkier. It should be either \code{H} 
  (herbaceous), \code{C} (chamaephyte) or \code{P} (phanerophyte) for now}
  \item{height}{the maximum or average height that reach the species}
  \item{maturity}{the age from which the species can reproduce}
  \item{longevity}{the maximum or average lifespan of the species}
  \item{dispersal}{the age from which the species can reproduce}
  \item{light}{a value corresponding to the light preference of the species
  (from preference for shade to full light)}
  \item{soil_contrib}{a value between 0 and 10 corresponding to the 
  Ellenberg nitrogen value of the PFG}
  \item{soil_tolerance}{the range of soil tolerance : small(1) or large(2)}
  \item{palatability}{the appetence of each species from null to high
   \cr \cr}
}
}
\examples{

## Load example data
.loadData("PNE_PFG")

## PNE_PFG$dom.traits : data.frame
## PNE_PFG$dom.determ : data.frame

tab1 = PNE_PFG$dom.determ[which(PNE_PFG$dom.determ$determinant == TRUE), c("species", "PFG")]
tab2 = PNE_PFG$dom.traits[, c("species", "height", "dispersal", "palatability", "light")]
determ.traits = merge(tab1, tab2, by = "species")

PFG.traits = PRE_FATE.speciesClustering_step3(mat.species.traits = determ.traits)
summary(PFG.traits)

}
\seealso{
\code{\link{PRE_FATE.speciesClustering_step1}},
\code{\link{PRE_FATE.speciesClustering_step2}}
}
\author{
Maya Guéguen
}
\keyword{functional}
\keyword{group,}
\keyword{traits}
