% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRE_FATE.speciesDistance.R
\name{PRE_FATE.speciesDistance}
\alias{PRE_FATE.speciesDistance}
\title{Computation of distances between species
based on traits and niche overlap}
\usage{
PRE_FATE.speciesDistance(mat.species.traits, mat.species.overlap)
}
\arguments{
\item{mat.species.traits}{A \code{data.frame} with at least 3 columns :
\describe{
\item{\code{species}}{the ID of each studied species}
\item{\code{GROUP}}{a factor variable containing grouping information to
  divide the species into sub data sets (see \code{Details})}
\item{\code{...}}{one column for each functional trait}
}}

\item{mat.species.overlap}{Three options :
\itemize{
  \item a \code{data.frame} with 2 columns :
  \describe{
    \item{\code{species}}{the ID of each studied species}
    \item{\code{raster}}{path to raster file with species distribution}
  }
  \item a dissimilarity structure representing the niche overlap between
  each pair of species. It can be a \code{dist} object, a \code{niolap}
  object, or simply a \code{matrix}.
}}
}
\value{
A \code{dist} object corresponding to the distance between each pair
of species, or a \code{list} of \code{dist} objects, one for each
\code{GROUP} value.
}
\description{
This script is designed to create a distance matrix between
species, combining functional distances (based on functional trait values)
and niche overlap (based on co-occurrence of species).
}
\details{
This function allows one to obtain a distance matrix between species, based
on two types of distance information :

\enumerate{
  \item{\strong{Functional traits : }}{
  \itemize{
    \item The \code{GROUP} column is required if species must be separated
    to have one final distance matrix per \code{GROUP} value. If the column
    is missing, all species will be considered as part of a unique dataset.
    \item The traits can be qualitative or quantitative, but previously
    identified as such (i.e. with the use of functions such as
    \code{as.numeric}, \code{as.factor} and \code{ordered}).
    \item Functional distance matrix is calculated with Gower dissimilarity,
    using the \code{gowdis} function from \pkg{FD} package.
  }
  }
  \item{\strong{Niche overlap : }}{
  \itemize{
    \item If a \code{data.frame} is given, the degree of niche overlap will
    be computed using the \code{niche.overlap} function from \pkg{phyloclim}
    package.
  }
  }
}

Functional distances and niche overlap informations are then \strong{combined}
according to the following formula :

\deqn{mat.DIST_{sub-group} = [ mat.OVERLAP_{sub-group} +
mat.FUNCTIONAL_{sub-group} * n_{traits} ] / [ n_{traits} + 1 ]}

meaning that distance matrix obtained from functional information is
weighted by the number of traits used.
}
\note{
Informations must be complete to include a species into the distance
calculation. Hence, \strong{if a species is missing a trait value, or information
about distribution or niche overlap, it will be removed from the computation}.
}
\examples{
## Load example data
data(MontBlanc)
str(MontBlanc)

## MontBlanc$mat.traits : data.frame
## MontBlanc$mat.nicheOverlap : niolap object
sp.DIST = PRE_FATE.speciesDistance(mat.species.traits = MontBlanc$mat.traits,
                                   mat.species.overlap = MontBlanc$mat.nicheOverlap)

str(sp.DIST)

}
\author{
Maya Guéguen
}
\keyword{Gower}
\keyword{distance}
