% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRE_FATE.speciesDistanceOverlap.R
\name{PRE_FATE.speciesDistanceOverlap}
\alias{PRE_FATE.speciesDistanceOverlap}
\title{Computation of niche overlap distances between species}
\usage{
PRE_FATE.speciesDistanceOverlap(mat.overlap.option, mat.overlap.object)
}
\arguments{
\item{mat.overlap.option}{a \code{string} corresponding to the way to 
calculate the distance between species based on niche overlap (either 
\code{PCA} or \code{raster}, see 
\href{PRE_FATE.speciesDistance#details}{\code{Details}})}

\item{mat.overlap.object}{two options, depending on the value of 
\code{mat.overlap.option} :
\itemize{
  \item (\code{PCA} option) a \code{list} with 2 elements :
  \describe{
    \item{\code{tab.dom.PA}}{a \code{matrix} or \code{data.frame} with 
    sites in rows and species in columns, containing either \code{NA}, 
    \code{0} or \code{1} (see \code{\link{PRE_FATE.selectDominant}})}
    \item{\code{tab.env}}{a \code{matrix} or \code{data.frame} with 
    sites in rows and environmental variables in columns}
  }
  \item (\code{raster} option) a \code{data.frame} with 2 columns :
  \describe{
    \item{\code{species}}{the ID of each studied species}
    \item{\code{raster}}{path to raster file with species distribution}
  }
}}
}
\value{
A \code{matrix} containing overlap distances between each pair 
of species, calculated as \code{1 - Schoeners D}.
}
\description{
This script is designed to create a distance matrix between 
species, based on co-occurrence of species.
}
\details{
This function allows to obtain a \strong{distance matrix between species} 
(1 - Schoeners D), based on niche overlap information :

\itemize{
  \item If \code{PCA} option is selected, the degree of niche overlap will 
  be computed using the \code{\link[ecospat]{ecospat.niche.overlap}}. 
  \item If \code{raster} option is selected, the degree of niche overlap will 
  be computed using the 
  \href{https://www.rdocumentation.org/packages/phyloclim/versions/0.9-4/topics/niche.overlap}{\code{niche.overlap}}
  . \cr \cr \cr
}
}
\examples{

## Load example data
Champsaur_PFG = .loadData('Champsaur_PFG', 'RData')

## Prepare sites x species table
## Add absences in community sites
sites = Champsaur_PFG$sp.observations
tab.dom.PA = Champsaur_PFG$tab.dom.PA
for (si in sites$sites[which(sites$TYPE == "COMMUNITY")])
{
  ind = which(rownames(tab.dom.PA) == si)
  tab.dom.PA[ind, which(is.na(tab.dom.PA[ind, ]))] = 0
}

## Prepare environmental table
tab.env = Champsaur_PFG$tab.env

## Calculate niche overlap distances -----------------------------------------
list.over = list(tab.dom.PA[, 1:10], tab.env)
DIST.overlap = PRE_FATE.speciesDistanceOverlap(mat.overlap.option = "PCA"
                                               , mat.overlap.object = list.over)
(DIST.overlap[1:5, 1:5])

}
\seealso{
\code{\link[ecospat]{ecospat.niche.overlap}}, 
\href{https://www.rdocumentation.org/packages/phyloclim/versions/0.9-4/topics/niche.overlap}{\code{niche.overlap}}
}
\author{
Maya Guéguen
}
\keyword{D}
\keyword{Schoeners}
\keyword{niche}
\keyword{overlap,}
