% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFate-package.R
\docType{package}
\name{RFate}
\alias{RFate}
\title{\packageTitle{RFate}}
\description{
\dQuote{FATE-HD is a dynamic landscape vegetation model that simulates
 interactions between plant modelling entities (e.g. species or plant
 functional groups), their population dynamics and dispersal, whilst taking
 into account external drivers such as disturbance regimes, and environmental
 variations. The model is built on past conceptual and technical developments
 (\href{https://besjournals.onlinelibrary.wiley.com/doi/full/10.1111/j.1365-2664.2007.01416.x}{Albert et al., 2008};
 \href{https://onlinelibrary.wiley.com/doi/full/10.1111/j.1600-0587.2009.06000.x}{Midgley et al., 2010})
 but has been entirely recoded in C++ and revisited.}
(\href{https://onlinelibrary.wiley.com/doi/full/10.1111/gcb.12466}{Boulangeat, 2014})
}
\section{RFate functions}{

\code{\link{runFate}}
}

\author{
\packageAuthor{RFate}

Maintainer: \packageMaintainer{RFate}
}
\keyword{internal}
\keyword{package}
