% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.train_RF_habitat.R
\name{train_RF_habitat}
\alias{train_RF_habitat}
\title{Create a random forest algorithm trained on CBNA data.}
\usage{
train_RF_habitat(
  releves.PFG,
  hab.obs,
  external.training.mask = NULL,
  studied.habitat = NULL,
  RF.param,
  output.path,
  perStrata,
  sim.version
)
}
\arguments{
\item{releves.PFG}{a data frame with abundance (column named abund) at each site
and for each PFG and strata.}

\item{hab.obs}{a raster map of the observed habitat in the
extended studied area.}

\item{external.training.mask}{default \code{NULL}. (optional) Keep only
releves data in a specific area.}

\item{studied.habitat}{default \code{NULL}. If \code{NULL}, the function will
take into account of habitats define in the \code{hab.obs} map. Otherwise, please specify 
in a 2 columns data frame the habitats (2nd column) and the ID (1st column) for each of them which will be taken 
into account for the validation.}

\item{RF.param}{a list of 2 parameters for random forest model :
share.training defines the size of the trainig part of the data base.
ntree is the number of trees build by the algorithm, it allows to reduce
the prediction error.}

\item{output.path}{access path to the for the folder where output files
will be created.}

\item{perStrata}{\code{Logical}. If TRUE, the PFG abundance is defined
by strata in each site. If FALSE, PFG abundance is defined for all strata.}

\item{sim.version}{name of the simulation we want to validate.}
}
\value{
2 prepared observed releves files are created before the building of the random
forest model in a habitat validation folder.
5 more files are created at the end of the script to save the RF model and
the performance analyzes (confusion matrix and TSS) for the training and 
testing parts.
}
\description{
This script is designed to produce a random forest model
trained on observed PFG abundance, sites releves and a map of observed
habitat.
}
\details{
This function transform PFG abundance in relative abundance,
get habitat information from the releves map of from a vector previously defined, 
keep releves on interesting habitat and then builds a random forest model. Finally, 
the function analyzes the model performance with computation of confusion matrix and TSS between
the training and testing sample.
}
\author{
Matthieu Combaud, Maxime Delprat
}
