% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.train_RF_habitat.R
\name{train_RF_habitat}
\alias{train_RF_habitat}
\title{Create a random forest algorithm trained on observed vegetation data}
\usage{
train_RF_habitat(
  releves.PFG,
  hab.obs.RF = NULL,
  external.training.mask = NULL,
  studied.habitat,
  RF.param,
  output.path,
  perStrata = FALSE,
  seed
)
}
\arguments{
\item{releves.PFG}{a \code{data.frame} with at least 5 columns : \cr
\code{site}, \code{x}, \code{y}, \code{abund}, \code{PFG}
\cr (\emph{and optionally, \code{strata}, \code{code.habitat}})
\cr (see \href{train_RF_habitat#details}{\code{Details}})}

\item{external.training.mask}{(optional) default \code{NULL}. 
\cr a \code{raster} map for keeping releves data only in a specific area.}

\item{studied.habitat}{a \code{data.frame} with 2 columns : \cr
\code{ID} ,\code{habitat}
\cr (see \href{train_RF_habitat#details}{\code{Details}})}

\item{RF.param}{a \code{list} of 2 parameters to fit a random forest model : \cr
\code{share.training} defines the size of the training part of the data base. \cr
\code{ntree} is the number of trees build by the algorithm, it allows to reduce the prediction error.}

\item{output.path}{access path to the folder where output files will be created.}

\item{perStrata}{(\code{Logical}) default \code{FALSE}. 
\cr If \code{TRUE}, the PFG abundance must be defined
by strata in each site. If \code{FALSE}, PFG abundance must be defined for all strata.}
}
\value{
2 prepared observed releves files are created before the building of the random
forest model in a folder previously defined. \cr
5 more files are created at the end of the script to save the RF model and
the performance analyzes (confusion matrix and TSS) for the training and 
testing parts.
}
\description{
This script is designed to produce a random forest model
trained on observed PFG abundance and a map of observed habitat.
}
\details{
This function transforms PFG abundance in relative abundance,
gets habitat information from an habitat data frame previously defined and a habitat map, 
keep releves on interesting habitat(s) and then builds a random forest model. Finally, 
the function analyzes the model performance with computation of confusion matrix and TSS between
the training and testing sample.
}
\author{
Matthieu Combaud, Maxime Delprat
}
