% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.train_RF_habitat.R
\name{train_RF_habitat}
\alias{train_RF_habitat}
\title{Create a random forest algorithm trained on observed vegetation data}
\usage{
train_RF_habitat(
  releves.PFG,
  hab.obs.RF = NULL,
  external.training.mask = NULL,
  studied.habitat,
  RF.param,
  output.path,
  perStrata = FALSE,
  seed
)
}
\arguments{
\item{releves.PFG}{a \code{data.frame} with at least 5 columns : \cr
\code{site}, \code{x}, \code{y}, \code{abund}, \code{PFG}
\cr (\emph{and optionally, \code{strata}, \code{code.habitat}})
\cr (see \href{train_RF_habitat#details}{\code{Details}})}

\item{external.training.mask}{(optional) default \code{NULL}. 
\cr a \code{raster} map for keeping releves data only in a specific area.}

\item{studied.habitat}{a \code{data.frame} with 2 columns : \cr
\code{ID} ,\code{habitat}
\cr (see \href{train_RF_habitat#details}{\code{Details}})}

\item{RF.param}{a \code{list} of 2 parameters to fit a random forest model : \cr
\code{share.training} defines the proportion of the data base used for training the model. \cr
\code{ntree} is the number of trees built by the algorithm, it allows to reduce the prediction error.}

\item{output.path}{access path to the folder where output files will be created.}

\item{perStrata}{(\code{logical}) default \code{FALSE}. 
\cr If \code{TRUE}, the PFG abundance must be defined
by strata in each site. If \code{FALSE}, PFG abundance must be defined for all strata.}
}
\value{
into the \code{name.simulation/VALIDATION/HABITAT/ directory :} \cr
1 .csv file containing the prepared observed data (relative abundances & habitat information). \cr
1 .rds file containing the RF model. \cr
4 .csv files containing the performance analysis (2 TSS per habitat files, and 2 aggregate TSS files) for the training & 
testing samples.
}
\description{
This script is designed to produce a \code{Random Forest} model
trained on observed PFG abundances & habitats.
}
\details{
The \code{Random Forest} algorithm is trained on \strong{releves.PFG} data. \cr
Information about \strong{PFG abundances} at each \strong{sites} with \strong{xy} coordinates are necessary. \cr
Eventually, \strong{habitat ID} information can be provided. If not, observed habitat is provided by the 
\strong{hab.obs.RF} map. The final set of habitats taken into account in the validation is provided by 
\strong{studied.habitat} table containing all the \strong{habitats}, with their corresponding \strong{ID}.
Finally, a performance analysis is provided by computing TSS between training and testing samples of the data base.
}
\author{
Matthieu Combaud, Maxime Delprat
}
